Struct glib::auto::main_context::MainContext

pub struct MainContext(Shared<GMainContext, MemoryManager>);

Tuple Fields
0: Shared<GMainContext, MemoryManager>
Implementations

impl MainContext

pub fn new() -> MainContext

pub fn acquire(&self) -> bool

pub fn dispatch(&self)

pub fn is_owner(&self) -> bool

pub fn iteration(&self, may_block: bool) -> bool

pub fn pending(&self) -> bool

pub fn pop_thread_default(&self)

pub fn push_thread_default(&self)

pub fn release(&self)

pub fn wakeup(&self)

pub fn default() -> MainContext

pub fn get_thread_default() -> Option<MainContext>

pub fn ref_thread_default() -> MainContext

impl MainContext

pub fn prepare(&self) -> (bool, i32)

pub fn find_source_by_id(&self, source_id: &SourceId) -> Option<Source>

pub fn invoke<F>(&self, func: F) where
F: FnOnce() + Send + 'static,

Invokes func on the main context.

pub fn invoke_with_priority<F>(&self, priority: Priority, func: F) where
F: FnOnce() + Send + 'static,

Invokes func on the main context with the given priority.

pub fn invoke_local<F>(&self, func: F) where
F: FnOnce() + 'static,

Invokes func on the main context.

Different to invoke(), this does not require func to be Send but can only be called from the thread that owns the main context.

This function panics if called from a different thread than the one that owns the main context.

pub fn invoke_local_with_priority<F>(&self, priority: Priority, func: F) where
F: FnOnce() + 'static,

Invokes func on the main context with the given priority.

Different to invoke_with_priority(), this does not require func to be Send but can only be called from the thread that owns the main context.

This function panics if called from a different thread than the one that owns the main context.

unsafe fn invoke_unsafe<F>(&self, priority: Priority, func: F) where
F: FnOnce() + 'static,

pub fn with_thread_default<R, F: Sized>(&self, func: F) -> R where
F: FnOnce() -> R,

Calls closure with context configured as the thread default one.

Thread default context is changed in panic-safe manner by calling push_thread_default before calling closure and pop_thread_default afterwards regardless of whether closure panicked or not.

impl MainContext

pub fn channel<T>(priority: Priority) -> (Sender<T>, Receiver<T>)

Creates a channel for a main context.

The Receiver has to be attached to a main context at a later time, together with a closure that will be called for every item sent to a Sender.

The Sender can be cloned and both the Sender and Receiver can be sent to different threads as long as the item type implements the Send trait.

When the last Sender is dropped the channel is removed from the main context. If the Receiver is dropped and not attached to a main context all sending to the Sender will fail.

The returned Sender behaves the same as std::sync::mpsc::Sender.

pub fn sync_channel<T>(
priority: Priority,
bound: usize
) -> (SyncSender<T>, Receiver<T>)

Creates a synchronous channel for a main context with a given bound on the capacity of the channel.

The Receiver has to be attached to a main context at a later time, together with a closure that will be called for every item sent to a SyncSender.

The SyncSender can be cloned and both the SyncSender and Receiver can be sent to different threads as long as the item type implements the Send trait.

When the last SyncSender is dropped the channel is removed from the main context. If the Receiver is dropped and not attached to a main context all sending to the SyncSender will fail.

The returned SyncSender behaves the same as std::sync::mpsc::SyncSender.

impl MainContext

pub fn spawn<F: Future<Output = ()> + Send + 'static>(&self, f: F)

Spawn a new infallible Future on the main context.

This can be called from any thread and will execute the future from the thread where main context is running, e.g. via a MainLoop.

pub fn spawn_local<F: Future<Output = ()> + 'static>(&self, f: F)

Spawn a new infallible Future on the main context.

The given Future does not have to be Send.

This can be called only from the thread where the main context is running, e.g. from any other Future that is executed on this main context, or after calling push_thread_default or acquire on the main context.

pub fn spawn_with_priority<F: Future<Output = ()> + Send + 'static>(
&self,
priority: Priority,
f: F
)

Spawn a new infallible Future on the main context, with a non-default priority.

This can be called from any thread and will execute the future from the thread where main context is running, e.g. via a MainLoop.

pub fn spawn_local_with_priority<F: Future<Output = ()> + 'static>(
&self,
priority: Priority,
f: F
)

Spawn a new infallible Future on the main context, with a non-default priority.

The given Future does not have to be Send.

This can be called only from the thread where the main context is running, e.g. from any other Future that is executed on this main context, or after calling push_thread_default or acquire on the main context.

pub fn block_on<F: Future>(&self, f: F) -> F::Output

Runs a new, infallible Future on the main context and block until it finished, returning the result of the Future.

The given Future does not have to be Send or 'static.

This must only be called if no MainLoop or anything else is running on this specific main context.