Struct glib::char::Char

pub struct Char(pub c_char);

Wrapper for values where C functions expect a plain C char

Consider the following C function prototype from glib:

void g_key_file_set_list_separator (GKeyFile *key_file, gchar separator);

This function plainly expects a byte as the separator argument. However, having this function exposed to Rust as the following would be inconvenient:

impl KeyFile {
    pub fn set_list_separator(&self, separator: libc:c_char) { }
}

This would be inconvenient because users would have to do the conversion from a Rust char to an libc::c_char by hand, which is just a type alias for i8 on most system.

This Char type is a wrapper over an libc::c_char, so that we can pass it to Glib or C functions. The check for whether a Rust char (a Unicode scalar value) actually fits in a libc::c_char is done in the new function; see its documentation for details.

The inner libc::c_char (which is equivalent to i8 can be extracted with .0, or by calling my_char.to_glib().

Methods

impl Char

pub fn new(c: char) -> Option<Char>

Creates a Some(Char) if the given char is representable as an libc::c_char

Example

extern "C" fn have_a_byte(b: libc::c_char);

let a = Char::new('a').unwrap();
assert!(a.0 == 65);
have_a_byte(a.to_glib());

let not_representable = Char::new('☔');
assert!(not_representable.is_none());