Trait glib::clone::Upgrade

pub trait Upgradewhere
    Self: Sized,{
    type Strong;

    // Required method
    fn upgrade(&self) -> Option<Self::Strong>;
}

Trait for generalizing upgrading a weak reference to a strong reference.

Required Associated Types

type Strong

Strong reference type.

Required Methods

fn upgrade(&self) -> Option<Self::Strong>

Try upgrading a weak reference to a strong reference.