Struct glib::collections::slist::SList

pub struct SList<T: TransparentPtrType> { /* private fields */ }

A list of items of type T.

Behaves like an Iterator<Item = T> but allows modifications.

Implementations

impl<T: TransparentPtrType> SList<T>

pub unsafe fn from_glib_none(list: *const GSList) -> SList<T>

Create a new SList around a list.

pub unsafe fn from_glib_container(list: *mut GSList) -> SList<T>

Create a new SList around a list.

pub unsafe fn from_glib_full(list: *mut GSList) -> SList<T>

Create a new SList around a list.

pub fn new() -> Self

Creates a new empty list.

pub fn iter(&self) -> Iter<'_, T>

Create a non-destructive iterator over the SList.

pub fn iter_mut(&mut self) -> IterMut<'_, T>

Create a non-destructive mutable iterator over the SList.

pub fn is_empty(&self) -> bool

Check if the list is empty.

This operation is O(1).

pub fn len(&self) -> usize

Returns the length of the list.

This operation is O(n).

pub fn front(&self) -> Option<&T>

Returns a reference to the first item of the list, if any.

This operation is O(1).

pub fn front_mut(&mut self) -> Option<&mut T>

Returns a mutable reference to the first item of the list, if any.

This operation is O(1).

pub fn pop_front(&mut self) -> Option<T>

Removes the front item from the list, if any.

This operation is O(1).

pub fn push_front(&mut self, item: T)

Prepends the new item to the front of the list.

This operation is O(1).

pub fn back(&self) -> Option<&T>

Returns a reference to the last item of the list, if any.

This operation is O(n).

pub fn back_mut(&mut self) -> Option<&mut T>

Returns a mutable reference to the last item of the list, if any.

This operation is O(n).

pub fn pop_back(&mut self) -> Option<T>

Removes the back item from the list, if any.

This operation is O(n).

pub fn push_back(&mut self, item: T)

Appends the new item to the back of the list.

this operation is O(n).

pub fn reverse(&mut self)

Reverse the list.

This operation is O(n).

pub fn sort(&mut self)
where T: Ord,

Sorts the list.

This operation is O(n * log n).

pub fn sort_by<F: FnMut(&T, &T) -> Ordering>(&mut self, f: F)

Sorts the list.

This operation is O(n * log n).

pub fn clear(&mut self)

Removes all items from the list.

pub fn retain(&mut self, f: impl FnMut(&T) -> bool)

Only keeps the item in the list for which f returns true.

pub fn as_ptr(&self) -> *const GSList

Returns the underlying pointer.

pub fn as_mut_ptr(&mut self) -> *mut GSList

Returns the underlying pointer.

pub fn into_raw(self) -> *mut GSList

Consumes the list and returns the underlying pointer.