glib::collections::strv

Struct StrV

pub struct StrV { /* private fields */ }

NULL-terminated array of NULL-terminated strings.

The underlying memory is always NULL-terminated.

This can be used like a &[&str], &mut [&str] and Vec<&str>.

Implementations

impl StrV

pub unsafe fn from_glib_borrow<'a>( ptr: *const *const c_char, ) -> &'a [GStringPtr]

Borrows a C array.

pub unsafe fn from_glib_borrow_num<'a>( ptr: *const *const c_char, len: usize, ) -> &'a [GStringPtr]

Borrows a C array.

pub unsafe fn from_glib_none_num( ptr: *const *const c_char, len: usize, _null_terminated: bool, ) -> Self

Create a new StrV around a C array.

pub unsafe fn from_glib_container_num( ptr: *mut *const c_char, len: usize, null_terminated: bool, ) -> Self

Create a new StrV around a C array.

pub unsafe fn from_glib_full_num( ptr: *mut *mut c_char, len: usize, null_terminated: bool, ) -> Self

Create a new StrV around a C array.

pub unsafe fn from_glib_none(ptr: *const *const c_char) -> Self

Create a new StrV around a NULL-terminated C array.

pub unsafe fn from_glib_container(ptr: *mut *const c_char) -> Self

Create a new StrV around a NULL-terminated C array.

pub unsafe fn from_glib_full(ptr: *mut *mut c_char) -> Self

Create a new StrV around a NULL-terminated C array.

pub fn new() -> Self

Creates a new empty slice.

pub fn with_capacity(capacity: usize) -> Self

Creates a new empty slice with the given capacity.

pub fn as_ptr(&self) -> *const *mut c_char

Returns the underlying pointer.

This is guaranteed to be NULL-terminated.

pub fn into_raw(self) -> *mut *mut c_char

Consumes the slice and returns the underlying pointer.

This is guaranteed to be NULL-terminated.

pub fn len(&self) -> usize

Gets the length of the slice.

pub fn is_empty(&self) -> bool

Returns true if the slice is empty.

pub fn capacity(&self) -> usize

Returns the capacity of the slice.

This includes the space that is reserved for the NULL-terminator.

pub unsafe fn set_len(&mut self, len: usize)

Sets the length of the slice to len.

SAFETY

There must be at least len valid items and a NULL-terminator after the last item.

pub fn reserve(&mut self, additional: usize)

Reserves at least this much additional capacity.

pub const fn as_slice(&self) -> &[GStringPtr]

Borrows this slice as a &[GStringPtr].

pub fn clear(&mut self)

Removes all items from the slice.

pub fn extend_from_slice<S: AsRef<str>>(&mut self, other: &[S])

Clones and appends all elements in slice to the slice.

pub fn insert(&mut self, index: usize, item: GString)

Inserts item at position index of the slice, shifting all elements after it to the right.

pub fn push(&mut self, item: GString)

Pushes item to the end of the slice.

pub fn remove(&mut self, index: usize) -> GString

Removes item from position index of the slice, shifting all elements after it to the left.

pub fn swap(&mut self, index: usize, new_item: GString) -> GString

Swaps item from position index of the slice and returns it.

pub fn pop(&mut self) -> Option<GString>

Removes the last item of the slice and returns it.

pub fn truncate(&mut self, len: usize)

Shortens the slice by keeping the last len items.

If there are fewer than len items then this has no effect.

pub fn join(&self, separator: Option<impl IntoGStr>) -> GString

Joins the strings into a longer string, with an optional separator

pub fn contains(&self, s: impl IntoGStr) -> bool

Checks whether the StrV contains the specified string