glib

Module convert

Structs

IConv

Enums

CvtError

A wrapper for ConvertError that can hold an offset into the input string.

IConvError

A wrapper for std::io::Error that can hold an offset into an input string.

Functions

casefold

convert

convert_with_fallback

filename_charsets

filename_from_utf8

filename_to_utf8

locale_from_utf8

locale_to_utf8

normalize

utf8_to_utf32

utf32_to_utf8