glib

Enum FileError

#[non_exhaustive]
pub enum FileError {

    Exist,
    Isdir,
    Acces,
    Nametoolong,
    Noent,
    Notdir,
    Nxio,
    Nodev,
    Rofs,
    Txtbsy,
    Fault,
    Loop,
    Nospc,
    Nomem,
    Mfile,
    Nfile,
    Badf,
    Inval,
    Pipe,
    Again,
    Intr,
    Io,
    Perm,
    Nosys,
    Failed,

}

Values corresponding to @errno codes returned from file operations on UNIX. Unlike @errno codes, GFileError values are available on all systems, even Windows. The exact meaning of each code depends on what sort of file operation you were performing; the UNIX documentation gives more details. The following error code descriptions come from the GNU C Library manual, and are under the copyright of that manual.

Itโs not very portable to make detailed assumptions about exactly which errors will be returned from a given operation. Some errors donโt occur on some systems, etc., sometimes there are subtle differences in when a system will report a given error, etc.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Exist

Operation not permitted; only the owner of the file (or other resource) or processes with special privileges can perform the operation.

Isdir

File is a directory; you cannot open a directory for writing, or create or remove hard links to it.

Acces

Permission denied; the file permissions do not allow the attempted operation.

Nametoolong

Filename too long.

Noent

No such file or directory. This is a โfile doesnโt existโ error for ordinary files that are referenced in contexts where they are expected to already exist.

Notdir

A file that isnโt a directory was specified when a directory is required.

Nxio

No such device or address. The system tried to use the device represented by a file you specified, and it couldnโt find the device. This can mean that the device file was installed incorrectly, or that the physical device is missing or not correctly attached to the computer.

Nodev

The underlying file system of the specified file does not support memory mapping.

Rofs

The directory containing the new link canโt be modified because itโs on a read-only file system.

Txtbsy

Text file busy.

Fault

You passed in a pointer to bad memory. (GLib wonโt reliably return this, donโt pass in pointers to bad memory.)

Loop

Too many levels of symbolic links were encountered in looking up a file name. This often indicates a cycle of symbolic links.

Nospc

No space left on device; write operation on a file failed because the disk is full.

Nomem

No memory available. The system cannot allocate more virtual memory because its capacity is full.

Mfile

The current process has too many files open and canโt open any more. Duplicate descriptors do count toward this limit.

Nfile

There are too many distinct file openings in the entire system.

Badf

Bad file descriptor; for example, I/O on a descriptor that has been closed or reading from a descriptor open only for writing (or vice versa).

Inval

Invalid argument. This is used to indicate various kinds of problems with passing the wrong argument to a library function.

Pipe

Broken pipe; there is no process reading from the other end of a pipe. Every library function that returns this error code also generates a โSIGPIPEโ signal; this signal terminates the program if not handled or blocked. Thus, your program will never actually see this code unless it has handled or blocked โSIGPIPEโ.

Again

Resource temporarily unavailable; the call might work if you try again later.

Intr

Interrupted function call; an asynchronous signal occurred and prevented completion of the call. When this happens, you should try the call again.

Io

Input/output error; usually used for physical read or write errors. i.e. the disk or other physical device hardware is returning errors.

Perm

Operation not permitted; only the owner of the file (or other resource) or processes with special privileges can perform the operation.

Nosys

Function not implemented; this indicates that the system is missing some functionality.

Failed

Does not correspond to a UNIX error code; this is the standard โfailed for unspecified reasonโ error code present in all #GError error code enumerations. Returned if no specific code applies.