Struct glib::enums::FlagsBuilder

pub struct FlagsBuilder<'a>(/* private fields */);

Builder for conveniently setting/unsetting flags and returning a Value.

Example for getting a flags property, unsetting some flags and setting the updated flags on the object again:

let flags = obj.property("flags").unwrap();
let flags_class = FlagsClass::new(flags.type_()).unwrap();
let flags = flags_class.builder_with_value(flags).unwrap()
    .unset_by_nick("some-flag")
    .unset_by_nick("some-other-flag")
    .build()
    .unwrap();
obj.set_property("flags", &flags).unwrap();

If setting/unsetting any value fails, build() returns None.

Implementations

impl<'a> FlagsBuilder<'a>

pub fn set(self, f: u32) -> Self

Set flags corresponding to integer value f.

pub fn set_by_name(self, name: &str) -> Self

Set flags corresponding to string name name.

pub fn set_by_nick(self, nick: &str) -> Self

Set flags corresponding to string nick nick.

pub fn unset(self, f: u32) -> Self

Unsets flags corresponding to integer value f.

pub fn unset_by_name(self, name: &str) -> Self

Unset flags corresponding to string name name.

pub fn unset_by_nick(self, nick: &str) -> Self

Unset flags corresponding to string nick nick.

pub fn build(self) -> Option<Value>

Converts to the final Value, unless any previous setting/unsetting of flags failed.