glib
Module error

Error binding and helper trait.

Structs

BoolError

Error

A generic error capable of representing various error domains (types). The GError structure contains information about an error that has occurred. The GError structure contains information about an error that has occurred.

Traits

ErrorDomain

GLib error domain.