Struct glib::error::Error

pub struct Error(_);

A generic error capable of representing various error domains (types).

Methods

impl Error

pub fn new<T: ErrorDomain>(error: T, message: &str) -> Error

Creates an error with supplied error enum variant and message.

pub fn is<T: ErrorDomain>(&self) -> bool

Checks if the error domain matches T.

pub fn kind<T: ErrorDomain>(&self) -> Option<T>

Tries to convert to a specific error enum.

Returns Some if the error belongs to the enum's error domain and None otherwise.

Examples

if let Some(file_error) = error.kind::<FileError>() {
    match file_error {
        FileError::Exist => ...
        FileError::Isdir => ...
        ...
    }
}

match error {
    Some(FileError::Exist) => ...
    Some(FileError::Isdir) => ...
    ...
}

pub fn wrap(ptr: *mut GError) -> Error