Trait glib::error::ErrorDomain

pub trait ErrorDomain: Copy {
    fn domain() -> Quark;
    fn code(self) -> i32;
    fn from(code: i32) -> Option<Self>
    where
        Self: Sized;
}

GLib error domain.

This trait is implemented by error enums that represent error domains (types).

Required Methods

fn domain() -> Quark

Returns the quark identifying the error domain.

As returned from g_some_error_quark.

fn code(self) -> i32

Gets the integer representation of the variant.

fn from(code: i32) -> Option<Self> where
Self: Sized,

Tries to convert an integer code to an enum variant.

By convention, the Failed variant, if present, is a catch-all, i.e. any unrecognized codes map to it.