Struct glib::ffi::GDate

pub struct GDate;

Represents a day between January 1, Year 1 and a few thousand years in the future. None of its members should be accessed directly.

If the GDate is obtained from g_date_new(), it will be safe to mutate but invalid and thus not safe for calendrical computations.

If it's declared on the stack, it will contain garbage so must be initialized with g_date_clear(). g_date_clear() makes the date invalid but sane. An invalid date doesn't represent a day, it's "empty." A date becomes valid after you set it to a Julian day or you set a day, month, and year.