glib

Function access

pub fn access(filename: impl AsRef<Path>, mode: i32) -> i32

A wrapper for the POSIX access() function. This function is used to test a pathname for one or several of read, write or execute permissions, or just existence.

On Windows, the file protection mechanism is not at all POSIX-like, and the underlying function in the C library only checks the FAT-style READONLY attribute, and does not look at the ACL of a file at all. This function is this in practise almost useless on Windows. Software that needs to handle file permissions on Windows more exactly should use the Win32 API.

See your C library manual for more details about access().

filename

a pathname in the GLib file name encoding (UTF-8 on Windows)

mode

as in access()

Returns

zero if the pathname refers to an existing file system object that has all the tested permissions, or -1 otherwise or on error. A wrapper for the POSIX access() function. This function is used to test a pathname for one or several of read, write or execute permissions, or just existence.

On Windows, the file protection mechanism is not at all POSIX-like, and the underlying function in the C library only checks the FAT-style READONLY attribute, and does not look at the ACL of a file at all. This function is this in practise almost useless on Windows. Software that needs to handle file permissions on Windows more exactly should use the Win32 API.

See your C library manual for more details about access().

filename

a pathname in the GLib file name encoding (UTF-8 on Windows)

mode

as in access()

Returns

zero if the pathname refers to an existing file system object that has all the tested permissions, or -1 otherwise or on error.