glib

Function base64_decode

pub fn base64_decode(text: &str) -> Vec<u8> 

Decode a sequence of Base-64 encoded text into binary data. Note that the returned binary data is not necessarily zero-terminated, so it should not be used as a character string.

text

zero-terminated string with base64 text to decode

Returns

          newly allocated buffer containing the binary data
          that @text represents. The returned buffer must
          be freed with g_free().