glib

Function dcgettext

pub fn dcgettext(domain: Option<&str>, msgid: &str, category: i32) -> GString

This is a variant of g_dgettext() that allows specifying a locale category instead of always using LC_MESSAGES. See g_dgettext() for more information about how this functions differs from calling dcgettext() directly.

domain

the translation domain to use, or None to use the domain set with textdomain()

msgid

message to translate

category

a locale category

Returns

the translated string for the given locale category