Function glib::dngettext

pub fn dngettext(
    domain: Option<&str>, 
    msgid: &str, 
    msgid_plural: &str, 
    n: c_ulong
) -> GString

This function is a wrapper of dngettext() which does not translate the message if the default domain as set with textdomain() has no translations for the current locale.

See dgettext() for details of how this differs from dngettext() proper.

domain

the translation domain to use, or None to use the domain set with textdomain()

msgid

message to translate

msgid_plural

plural form of the message

n

the quantity for which translation is needed

Returns

The translated string