glib
Function dpgettext2

pub fn dpgettext2(domain: Option<&str>, context: &str, msgid: &str) -> GString

This function is a variant of g_dgettext() which supports a disambiguating message context. GNU gettext uses the ‘\004’ character to separate the message context and message id in @msgctxtid.

This uses g_dgettext() internally. See that functions for differences with dgettext() proper.

This function differs from C_() in that it is not a macro and thus you may use non-string-literals as context and msgid arguments.

domain

the translation domain to use, or None to use the domain set with textdomain()

context

the message context

msgid

the message

Returns

The translated string