glib
Function environ

pub fn environ() -> Vec<OsString>

Gets the list of environment variables for the current process.

The list is None terminated and each item in the list is of the form ‘NAME=VALUE’.

This is equivalent to direct access to the ‘environ’ global variable, except portable.

The return value is freshly allocated and it should be freed with g_strfreev() when it is no longer needed.

Returns

the list of environment variables