Function glib::file_set_contents

pub fn file_set_contents<P: AsRef<Path>>(
    filename: P, 
    contents: &[u8]
) -> Result<(), Error>

Writes all of contents to a file named filename. This is a convenience wrapper around calling file_set_contents_full() with flags set to G_FILE_SET_CONTENTS_CONSISTENT | G_FILE_SET_CONTENTS_ONLY_EXISTING and mode set to 0666.

filename

name of a file to write contents to, in the GLib file name encoding

contents

string to write to the file

Returns

true on success, false if an error occurred