Function glib::filename_from_utf8

pub fn filename_from_utf8(
    utf8string: impl IntoGStr
) -> Result<(PathBuf, usize), CvtError>

Converts a string from UTF-8 to the encoding GLib uses for filenames. Note that on Windows GLib uses UTF-8 for filenames; on other platforms, this function indirectly depends on the [current locale][setlocale].

The input string shall not contain nul characters even if the len argument is positive. A nul character found inside the string will result in error ConvertError::IllegalSequence. If the filename encoding is not UTF-8 and the conversion output contains a nul character, the error ConvertError::EmbeddedNul is set and the function returns None.

utf8string

a UTF-8 encoded string.

len

the length of the string, or -1 if the string is nul-terminated.

Returns

The converted string, or None on an error.

bytes_read

location to store the number of bytes in the input string that were successfully converted, or None. Even if the conversion was successful, this may be less than len if there were partial characters at the end of the input. If the error ConvertError::IllegalSequence occurs, the value stored will be the byte offset after the last valid input sequence.

bytes_written

the number of bytes stored in the output buffer (not including the terminating nul).