glib

Function filename_to_utf8

pub fn filename_to_utf8(
    opsysstring: impl AsRef<Path>,
) -> Result<(GString, usize), CvtError>

Converts a string which is in the encoding used by GLib for filenames into a UTF-8 string. Note that on Windows GLib uses UTF-8 for filenames; on other platforms, this function indirectly depends on the current locale.

The input string shall not contain nul characters even if the @len argument is positive. A nul character found inside the string will result in error ConvertError::IllegalSequence. If the source encoding is not UTF-8 and the conversion output contains a nul character, the error ConvertError::EmbeddedNul is set and the functGets properties of an object.

In general, a copy is made of the property contents and the caller is responsible for freeing the memory in the appropriate manner for the type, for instance by calling g_free() or g_object_unref().

Here is an example of using [get()][Self::get()] to get the contents of three properties: an integer, a string and an object:

⚠️ The following code is in C ⚠️

 gint intval;
 guint64 uint64val;
 gchar *strval;
 GObject *objval;

 g_object_get (my_object,
               "int-property", &intval,
               "uint64-property", &uint64val,
               "str-property", &strval,
               "obj-property", &objval,
               NULL);

 // Do something with intval, uint64val, strval, objval

 g_free (strval);
 g_object_unref (objval);

first_property_name

name of the first property to get Converts a string which is in the encoding used by GLib for filenames into a UTF-8 string. Note that on Windows GLib uses UTF-8 for filenames; on other platforms, this function indirectly depends on the current locale.

The input string shall not contain nul characters even if the @len argument is positive. A nul character found inside the string will result in error ConvertError::IllegalSequence. If the source encoding is not UTF-8 and the conversion output contains a nul character, the error ConvertError::EmbeddedNul is set and the functGets properties of an object.

In general, a copy is made of the property contents and the caller is responsible for freeing the memory in the appropriate manner for the type, for instance by calling g_free() or g_object_unref().

Here is an example of using [get()][Self::get()] to get the contents of three properties: an integer, a string and an object:

⚠️ The following code is in C ⚠️

 gint intval;
 guint64 uint64val;
 gchar *strval;
 GObject *objval;

 g_object_get (my_object,
               "int-property", &intval,
               "uint64-property", &uint64val,
               "str-property", &strval,
               "obj-property", &objval,
               NULL);

 // Do something with intval, uint64val, strval, objval

 g_free (strval);
 g_object_unref (objval);

first_property_name

name of the first property to get