Function glib::find_program_in_path

pub fn find_program_in_path(program: impl AsRef<Path>) -> Option<PathBuf>

Locates the first executable named program in the user’s path, in the same way that execvp() would locate it. Returns an allocated string with the absolute path name, or None if the program is not found in the path. If program is already an absolute path, returns a copy of program Increase the reference count of self, and possibly remove the [floating][floating-ref] reference, if self has a floating reference.

In other words, if the object is floating, then this call “assumes ownership” of the floating reference, converting it to a normal reference by clearing the floating flag while leaving the reference count unchanged. If the object is not floating, then this call adds a new normal reference increasing the reference count by one.

Since GLib 2.56, the type of self will be propagated to the return type under the same conditions as for g_object_ref().

Returns

self