glib
Function hostname_to_unicode

pub fn hostname_to_unicode(hostname: &str) -> Option<GString>

Converts @hostname to its canonical presentation form; a UTF-8 string in Unicode normalization form C, containing no uppercase letters, no forbidden characters, and no ASCII-encoded segments, and not ending with a trailing dot.

Of course if @hostname is not an internationalized hostname, then the canonical presentation form will be entirely ASCII.

hostname

a valid UTF-8 or ASCII hostname

Returns

a UTF-8 hostname, which must be freed, or None if @hostname is in some way invalid.