glib

Function interval_stream_with_priority

pub fn interval_stream_with_priority(
    priority: Priority,
    value: Duration,
) -> Pin<Box<dyn Stream<Item = ()> + Send + 'static>>

Create a Stream that will provide a value every given number of milliseconds.

The Future must be spawned on an Executor backed by a glib::MainContext.