Function glib::language_names_with_category

pub fn language_names_with_category(category_name: &str) -> Vec<GString>

This is supported on crate feature v2_58 only.

Computes a list of applicable locale names with a locale category name, which can be used to construct the fallback locale-dependent filenames or search paths. The returned list is sorted from most desirable to least desirable and always contains the default locale “C”.

This function consults the environment variables LANGUAGE, LC_ALL, category_name, and LANG to find the list of locales specified by the user.

language_names() returns g_get_language_names_with_category(“LC_MESSAGES”).

category_name

a locale category name

Returns

                         0, 100,
                                         50,
                                         G_PARAM_READWRITE);
g_object_class_install_property (gobject_class,
                                 PROP_FOO,
                                 properties[PROP_FOO]);

}

and then notify a change on the "foo" property with:

**⚠️ The following code is in C ⚠️**

```C
  g_object_notify_by_pspec (self, properties[PROP_FOO]);

pspec

the ParamSpec of a property installed on the class of self.