Function glib::log_default_handler

pub fn log_default_handler(
    log_domain: Option<&str>, 
    log_level: LogLevel, 
    message: Option<&str>
)

The default log handler set up by GLib; g_log_set_default_handler() allows to install an alternate default log handler. This is used if no log handler has been set for the particular log domain and log level combination. It outputs the message to stderr or stdout and if the log level is fatal it calls G_BREAKPOINT(). It automatically prints a new-line character after the message, so one does not need to be manually included in message.

The behavior of this log handler can be influenced by a number of environment variables:

G_MESSAGES_PREFIXED: A :-separated list of log levels for which messages should be prefixed by the program name and PID of the application.

G_MESSAGES_DEBUG: A space-separated list of log domains for which debug and informational messages are printed. By default these messages are not printed.

stderr is used for levels LogLevelFlags::LEVEL_ERROR, LogLevelFlags::LEVEL_CRITICAL, LogLevelFlags::LEVEL_WARNING and LogLevelFlags::LEVEL_MESSAGE. stdout is used for the rest, unless stderr was requested by g_log_writer_default_set_use_stderr().

This has no effect if structured logging is enabled; see [Using Structured Logging][using-structured-logging].

log_domain

the log domain of the message, or None for the default “” application domain

log_level

the level of the message

message

the message

unused_data

data passed from g_log() which is unused