Function glib::log_set_always_fatal

pub fn log_set_always_fatal(fatal_levels: LogLevels) -> LogLevels

Sets the message levels which are always fatal, in any log domain. When a message with any of these levels is logged the program terminates. You can only set the levels defined by GLib to be fatal. LogLevelFlags::LEVEL_ERROR is always fatal.

You can also make some message levels fatal at runtime by setting the G_DEBUG environment variable (see Running GLib Applications).

Libraries should not call this function, as it affects all messages logged by a process, including those from other libraries.

Structured log messages (using g_log_structured() and [log_structured_array()][crate::log_structured_array()]) are fatal only if the default log writer is used; otherwise it is up to the writer function to determine which log messages are fatal. See [Using Structured Logging][using-structured-logging].

fatal_mask

the mask containing bits set for each level of error which is to be fatal

Returns

the old fatal mask