Function glib::log_set_fatal_mask

pub fn log_set_fatal_mask(
    log_domain: Option<&str>, 
    fatal_levels: LogLevels
) -> LogLevels

Sets the log levels which are fatal in the given domain. LogLevelFlags::LEVEL_ERROR is always fatal.

This has no effect on structured log messages (using g_log_structured() or [log_structured_array()][crate::log_structured_array()]). To change the fatal behaviour for specific log messages, programs must install a custom log writer function using g_log_set_writer_func(). See [Using Structured Logging][using-structured-logging].

This function is mostly intended to be used with LogLevelFlags::LEVEL_CRITICAL. You should typically not set LogLevelFlags::LEVEL_WARNING, LogLevelFlags::LEVEL_MESSAGE, LogLevelFlags::LEVEL_INFO or LogLevelFlags::LEVEL_DEBUG as fatal except inside of test programs.

log_domain

the log domain

fatal_mask

the new fatal mask

Returns

the old fatal mask for the log domain