glib

Function log_writer_default_set_use_stderr

pub unsafe fn log_writer_default_set_use_stderr(use_stderr: bool)

Sets whether GLib log functions output to stderr or stdout.

By default, log messages of level LogLevel::Info and LogLevel::Debug are sent to stdout, and other log messages are sent to stderr. Passing true will send all messages to stderr.

Safety

This function sets global state and is not thread-aware, as such it should be called before any threads may try to use GLib logging.