Function glib::log_writer_default_would_drop

pub fn log_writer_default_would_drop(
    log_level: LogLevelFlags, 
    log_domain: Option<&str>
) -> bool

This is supported on crate feature v2_68 only.

Check whether g_log_writer_default() and g_log_default_handler() would ignore a message with the given domain and level.

As with g_log_default_handler(), this function drops debug and informational messages unless their log domain (or all) is listed in the space-separated G_MESSAGES_DEBUG environment variable.

This can be used when implementing log writers with the same filtering behaviour as the default, but a different destination or output format:

⚠️ The following code is in C ⚠️

  if (g_log_writer_default_would_drop (log_level, log_domain))
    return G_LOG_WRITER_HANDLED;

or to skip an expensive computation if it is only needed for a debugging message, and G_MESSAGES_DEBUG is not set:

⚠️ The following code is in C ⚠️

  if (!g_log_writer_default_would_drop (G_LOG_LEVEL_DEBUG, G_LOG_DOMAIN))
    {
      gchar *result = expensive_computation (my_object);

      g_debug ("my_object result: %s", result);
      g_free (result);
    }

log_level

log level, either from LogLevelFlags, or a user-defined level

log_domain

log domain

Returns

true if the log message would be dropped by GLib’s default log handlers