Function glib::mkdtemp_full

pub fn mkdtemp_full(tmpl: impl AsRef<Path>, mode: i32) -> Option<PathBuf>

Creates a temporary directory. See the mkdtemp() documentation on most UNIX-like systems.

The parameter is a string that should follow the rules for mkdtemp() templates, i.e. contain the string “XXXXXX”. mkdtemp_full() is slightly more flexible than mkdtemp() in that the sequence does not have to occur at the very end of the template and you can pass a mode. The X string will be modified to form the name of a directory that didn’t exist. The string should be in the GLib file name encoding. Most importantly, on Windows it should be in UTF-8.

If you are going to be creating a temporary directory inside the directory returned by tmp_dir(), you might want to use g_dir_make_tmp() instead.

tmpl

template directory name

mode

permissions to create the temporary directory with

Returns

A pointer to tmpl, which has been modified to hold the directory name. In case of errors, None is returned, and errno will be set.