Function glib::os_info

pub fn os_info(key_name: &str) -> Option<GString>

Available on crate feature v2_64 only.

Get information about the operating system.

On Linux this comes from the /etc/os-release file. On other systems, it may come from a variety of sources. You can either use the standard key names like G_OS_INFO_KEY_NAME or pass any UTF-8 string key name. For example, /etc/os-release provides a number of other less commonly used values that may be useful. No key is guaranteed to be provided, so the caller should always check if the result is None.

key_name

a key for the OS info being requested, for example G_OS_INFO_KEY_NAME.

Returns

The associated value for the requested key or None if this information is not provided.