Function glib::random_set_seed

pub fn random_set_seed(seed: u32)

Sets the seed for the global random number generator, which is used by the g_random_* functions, to @seed.

seed

a value to reinitialize the global random number generator