glib
Function spaced_primes_closest

pub fn spaced_primes_closest(num: u32) -> u32

Gets the smallest prime number from a built-in array of primes which is larger than @num. This is used within GLib to calculate the optimum size of a #GHashTable.

The built-in array of primes ranges from 11 to 13845163 such that each prime is approximately 1.5-2 times the previous prime.

num

a #guint

Returns

the smallest prime number from a built-in array of primes which is larger than @num