Function glib::unicode_script_to_iso15924

pub fn unicode_script_to_iso15924(script: UnicodeScript) -> u32

Looks up the ISO 15924 code for script. ISO 15924 assigns four-letter codes to scripts. For example, the code for Arabic is ‘Arab’. The four letter codes are encoded as a guint32 by this function in a big-endian fashion. That is, the code returned for Arabic is 0x41726162 (0x41 is ASCII code for ‘A’, 0x72 is ASCII code for ‘r’, etc).

See Codes for the representation of names of scripts for details.

script

a Unicode script

Returns

the ISO 15924 code for script, encoded as an integer, of zero if script is UnicodeScript::InvalidCode or ISO 15924 code ‘Zzzz’ (script code for UNKNOWN) if script is not understood.