glib

Function unlink

pub fn unlink(filename: impl AsRef<Path>) -> i32

A wrapper for the POSIX unlink() function. The unlink() function deletes a name from the filesystem. If this was the last link to the file and no processes have it opened, the diskspace occupied by the file is freed.

See your C library manual for more details about unlink(). Note that on Windows, it is in general not possible to delete files that are open to some process, or mapped into memory.

filename

a pathname in the GLib file name encoding (UTF-8 on Windows)

Returns

0 if the name was successfully deleted, -1 if an error occurred