glib

Function user_special_dir

pub fn user_special_dir(directory: UserDirectory) -> Option<PathBuf>

Returns the full path of a special directory using its logical id.

On UNIX this is done using the XDG special user directories. For compatibility with existing practise, [UserDirectory::DirectoryDesktop][crate::UserDirectory::DirectoryDesktop] falls back to $HOME/Desktop when XDG special user directories have not been set up.

Depending on the platform, the user might be able to change the path of the special directory without requiring the session to restart; GLib will not reflect any change once the special directories are loaded.

directory

the logical id of special directory

Returns

the path to the specified special directory, or None if the logical id was not found. The returned string is owned by GLib and should not be modified or freed.