glib::object
Struct Interface

pub struct Interface<T: IsInterface>(/* private fields */);

Interface struct of type T for some type.

Implementations

impl<T: IsInterface> Interface<T>

pub fn type_(&self) -> Type

Get the type id for this interface.

This is equivalent to T::static_type().

pub fn instance_type(&self) -> Type

Get the type id for the instance type of this interface.

This is not equivalent to T::static_type() but is the type id of the type this specific interface belongs to.

pub fn from_class<U: IsClass>(klass: &Class<U>) -> Option<InterfaceRef<'_, T>>

Gets the interface struct for Self of klass.

This will return None if klass is not implementing Self.

pub fn from_type(type_: Type) -> Option<InterfaceRef<'static, T>>

Gets the default interface struct for Self.

This will return None if type_ is not an interface.

pub fn default() -> InterfaceRef<'static, T>

Gets the default interface struct for Self.

pub fn parent(&self) -> Option<InterfaceRef<'_, T>>

Gets the parent interface struct, if any.

This returns the parent interface if a parent type of the instance type also implements the interface.

impl<T: IsA<Object> + IsInterface> Interface<T>

pub fn has_property(&self, property_name: &str) -> bool

Check if the interface has a property property_name of the given type_.

pub fn has_property_with_type(&self, property_name: &str, type_: Type) -> bool

Check if the interface has a property property_name of the given type_ or a subtype of it.

pub fn property_type(&self, property_name: &str) -> Option<Type>

Get the type of the property property_name of this interface.

This returns None if the property does not exist.

pub fn find_property(&self, property_name: &str) -> Option<ParamSpec>

Get the ParamSpec of the property property_name of this interface.

pub fn list_properties(&self) -> PtrSlice<ParamSpec>

Return all ParamSpec of the properties of this interface.