Struct glib::object::Object

pub struct Object(_);

The base class in the object hierarchy.

Implementations

impl Object

pub fn new<T: IsA<Object> + IsClass>(
properties: &[(&str, &dyn ToValue)]
) -> Result<T, BoolError>

Uses found in examples/

examples/gio_task/file_size/mod.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71

pub mod ffi;
mod imp;

use gtk::glib::subclass::prelude::*;
use gtk::prelude::*;
use gtk::{gio, glib};

glib::wrapper! {
    pub struct FileSize(ObjectSubclass<imp::FileSize>);
}

impl FileSize {
    pub fn new() -> Self {
        glib::Object::new(&[]).expect("Failed to create FileSize")
    }

    pub fn retrieved_size(&self) -> Option<i64> {
        let simple_object = imp::FileSize::from_instance(self);
        *simple_object.size.borrow()
    }

    #[doc(alias = "get_file_size_async")]
    pub fn file_size_async<Q: FnOnce(i64, &FileSize) + 'static>(
        &self,
        cancellable: Option<&gio::Cancellable>,
        callback: Q,
    ) {
        let closure = move |result: &gio::AsyncResult, source_object: Option<&glib::Object>| {
            let value = result
                .downcast_ref::<gio::Task>()
                .unwrap()
                .propagate_value()
                .unwrap()
                .get::<i64>()
                .unwrap();
            let source_object = source_object.unwrap().downcast_ref::<FileSize>().unwrap();
            callback(value, &source_object);
        };

        let task = gio::Task::new(
            Some(&self.upcast_ref::<glib::Object>()),
            cancellable,
            closure,
        );

        glib::MainContext::default().spawn_local(async move {
            let size = gio::File::for_path("Cargo.toml")
                .query_info_async_future("*", gio::FileQueryInfoFlags::NONE, glib::PRIORITY_DEFAULT)
                .await
                .unwrap()
                .size();

            let source_object = task
                .upcast_ref::<gio::AsyncResult>()
                .source_object()
                .unwrap();

            let source_object =
                imp::FileSize::from_instance(&source_object.downcast_ref::<FileSize>().unwrap());

            source_object.size.replace(Some(size));
            task.return_value(&size.to_value());
        });
    }
}

impl Default for FileSize {
    fn default() -> Self {
        Self::new()
    }
}

examples/list_box_model/model/mod.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41

//! Defines our custom model

mod imp;

use crate::row_data::RowData;
use glib::subclass::prelude::*;
use gtk::{gio, glib, prelude::*};

// Public part of the Model type.
glib::wrapper! {
    pub struct Model(ObjectSubclass<imp::Model>) @implements gio::ListModel;
}

// Constructor for new instances. This simply calls glib::Object::new()
impl Model {
    #[allow(clippy::new_without_default)]
    pub fn new() -> Model {
        glib::Object::new(&[]).expect("Failed to create Model")
    }

    pub fn append(&self, obj: &RowData) {
        let self_ = imp::Model::from_instance(self);
        let index = {
            // Borrow the data only once and ensure the borrow guard is dropped
            // before we emit the items_changed signal because the view
            // could call get_item / get_n_item from the signal handler to update its state
            let mut data = self_.0.borrow_mut();
            data.push(obj.clone());
            data.len() - 1
        };
        // Emits a signal that 1 item was added, 0 removed at the position index
        self.items_changed(index as u32, 0, 1);
    }

    pub fn remove(&self, index: u32) {
        let self_ = imp::Model::from_instance(self);
        self_.0.borrow_mut().remove(index as usize);
        // Emits a signal that 1 item was removed, 0 added at the position index
        self.items_changed(index, 1, 0);
    }
}

examples/composite_template/example_application_window/mod.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26

mod imp;

use glib::subclass::prelude::*;
use gtk::prelude::*;
use gtk::{gio, glib};

glib::wrapper! {
    pub struct ExampleApplicationWindow(ObjectSubclass<imp::ExampleApplicationWindow>)
        @extends gtk::Widget, gtk::Window, gtk::ApplicationWindow, @implements gio::ActionMap, gio::ActionGroup;
}

impl ExampleApplicationWindow {
    pub fn new<P: glib::IsA<gtk::Application>>(app: &P) -> Self {
        glib::Object::new(&[("application", app)])
            .expect("Failed to create ExampleApplicationWindow")
    }

    fn init_label(&self) {
        // To access fields such as template children, you must get
        // the private struct.
        let self_ = imp::ExampleApplicationWindow::from_instance(self);
        self_
            .subtitle
            .set_text("This is an example window made using composite templates");
    }
}

examples/basic_subclass/simple_application/mod.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20

mod imp;

use gio::ApplicationFlags;
use gtk::{gio, glib};

glib::wrapper! {
    pub struct SimpleApplication(ObjectSubclass<imp::SimpleApplication>)
        @extends gio::Application, gtk::Application;
}

impl SimpleApplication {
    #[allow(clippy::new_without_default)]
    pub fn new() -> Self {
        glib::Object::new(&[
            ("application-id", &"org.gtk-rs.SimpleApplication"),
            ("flags", &ApplicationFlags::empty()),
        ])
        .expect("Failed to create SimpleApp")
    }
}

examples/list_box_model/row_data/mod.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23

//! Our GObject subclass for carrying a name and count for the ListBox model
//!
//! Both name and count are stored in a RefCell to allow for interior mutability
//! and are exposed via normal GObject properties. This allows us to use property
//! bindings below to bind the values with what widgets display in the UI

mod imp;

use gtk::glib;

// Public part of the RowData type. This behaves like a normal gtk-rs-style GObject
// binding
glib::wrapper! {
    pub struct RowData(ObjectSubclass<imp::RowData>);
}

// Constructor for new instances. This simply calls glib::Object::new() with
// initial values for our two properties and then returns the new instance
impl RowData {
    pub fn new(name: &str, count: u32) -> RowData {
        glib::Object::new(&[("name", &name), ("count", &count)]).expect("Failed to create row data")
    }
}

examples/basic_subclass/simple_window/mod.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16

mod imp;

use gtk::glib;

use crate::simple_application::SimpleApplication;

glib::wrapper! {
    pub struct SimpleWindow(ObjectSubclass<imp::SimpleWindow>)
        @extends gtk::Widget, gtk::Container, gtk::Bin, gtk::Window, gtk::ApplicationWindow;
}

impl SimpleWindow {
    pub fn new(app: &SimpleApplication) -> Self {
        glib::Object::new(&[("application", app)]).expect("Failed to create SimpleWindow")
    }
}

pub fn with_type(
type_: Type,
properties: &[(&str, &dyn ToValue)]
) -> Result<Object, BoolError>

pub fn with_values(
type_: Type,
properties: &[(&str, Value)]
) -> Result<Object, BoolError>