glib::object

Struct WeakRef

pub struct WeakRef<T: ObjectType>(/* private fields */);

A weak reference to an object.

Implementations

impl<T: ObjectType> WeakRef<T>

pub fn new() -> WeakRef<T>

Create a new empty weak reference.

upgrade will always return None until an object is set on it.

pub fn set(&self, obj: Option<&T>)

Set this weak reference to the given object.

pub fn upgrade(&self) -> Option<T>

Try to upgrade this weak reference to a strong reference.

If the stored object was already destroyed or no object was set in this weak reference then None is returned.