Trait glib::object::ObjectClassExt

pub unsafe trait ObjectClassExt {
    // Provided methods
    fn has_property(&self, property_name: &str, type_: Option<Type>) -> bool { ... }
    fn property_type(&self, property_name: &str) -> Option<Type> { ... }
    fn find_property(&self, property_name: &str) -> Option<ParamSpec> { ... }
    fn list_properties(&self) -> PtrSlice<ParamSpec> { ... }
}

Trait for class methods on Object and subclasses of it.

Provided Methods

fn has_property(&self, property_name: &str, type_: Option<Type>) -> bool

Check if the object class has a property property_name of the given type_.

If no type is provided then only the existence of the property is checked.

fn property_type(&self, property_name: &str) -> Option<Type>

Get the type of the property property_name of this object class.

This returns None if the property does not exist.

fn find_property(&self, property_name: &str) -> Option<ParamSpec>

Get the ParamSpec of the property property_name of this object class.

fn list_properties(&self) -> PtrSlice<ParamSpec>

Return all ParamSpec of the properties of this object class.