glib::object

Type Alias ObjectClass

pub type ObjectClass = Class<Object>;

Aliased Type

struct ObjectClass(/* private fields */);

Implementations

impl ObjectClass

pub fn has_property(&self, property_name: &str, type_: Option<Type>) -> bool

Check if the object class has a property property_name of the given type_.

If no type is provided then only the existence of the property is checked.

pub fn property_type(&self, property_name: &str) -> Option<Type>

Get the type of the property property_name of this object class.

This returns None if the property does not exist.

pub fn find_property(&self, property_name: &str) -> Option<ParamSpec>

Get the ParamSpec of the property property_name of this object class.

pub fn list_properties(&self) -> PtrSlice<ParamSpec>

Return all ParamSpec of the properties of this object class.