Module glib::property

Macros

impl_atomic 🔒

Traits

Property

A type that can be used as a property. It covers every type which have an associated ParamSpec (HasParamSpec) and some useful types wrapping HasParamSpec. The definition is recursive, so you can nest many Propertys together. The final ParamSpec will be the one of the innermost type

PropertyGet

A container type implementing this trait can be read by the default getter generated by the Props macro.

PropertySet

A container type implementing this trait can be written by the default setter generated by the Props macro.

PropertySetNested

A container type implementing this trait can be written by the default setter generated by the Props macro. It takes a FnOnce(&mut Self::Value) so that the caller may access nested fields of a struct by doing ${Self::Value}.member