glib::property
Trait PropertySet

pub trait PropertySet {
    type SetValue;

    // Required method
    fn set(&self, v: Self::SetValue);
}

A container type implementing this trait can be written by the default setter generated by the Properties macro.

Required Associated Types

type SetValue

Required Methods

fn set(&self, v: Self::SetValue)