glib::property
Trait PropertySetNested

pub trait PropertySetNested {
    type SetNestedValue;

    // Required method
    fn set_nested<F: FnOnce(&mut Self::SetNestedValue)>(&self, f: F);
}

A container type implementing this trait can be written by the default setter generated by the Properties macro. It takes a FnOnce(&mut Self::Value) so that the caller may access nested fields of a struct by doing ${Self::Value}.member

Required Associated Types

type SetNestedValue

Required Methods

fn set_nested<F: FnOnce(&mut Self::SetNestedValue)>(&self, f: F)

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.