Module glib::send_unique

Structs

BorrowError

Ref

SendUniqueCell
Allows sending reference counted objects that don't implement Send to other threads as long as only a single reference to the object exists.

Traits

SendUnique
Like Send but only if we have the unique reference to the object