Struct glib::send_unique::SendUniqueCell

pub struct SendUniqueCell<T: SendUnique> { /* fields omitted */ }

Allows sending reference counted objects that don't implement Send to other threads as long as only a single reference to the object exists.

Methods

impl<T: SendUnique> SendUniqueCell<T>

pub fn new(obj: T) -> Result<Self, T>

Create a new SendUniqueCell out of obj

Fails if obj is not unique at this time

pub fn borrow(&self) -> Ref<T>

Borrow the contained object or panic if borrowing is not possible at this time

pub fn try_borrow(&self) -> Result<Ref<T>, BorrowError>

Try borrowing the contained object

Borrowing is possible as long as only a single reference to the object exists, or it is borrowed from the same thread currently

pub fn into_inner(self) -> T

Extract the contained object or panic if it is not possible at this time

pub fn try_into_inner(self) -> Result<T, Self>

Try extracing the contained object

Borrowing is possible as long as only a single reference to the object exists, or it is borrowed from the same thread currently