glib::shared

Trait SharedMemoryManager

pub trait SharedMemoryManager {
    type Target;

    // Required methods
    unsafe fn ref_(ptr: *mut Self::Target);
    unsafe fn unref(ptr: *mut Self::Target);
}

Required Associated Types

type Target

Required Methods

unsafe fn ref_(ptr: *mut Self::Target)

Safety

Callers are responsible for ensuring that a matching call to unref is made at an appropriate time.

unsafe fn unref(ptr: *mut Self::Target)

Safety

Callers are responsible for ensuring that a matching call to ref was made before this is called, and that the pointer is not used after the unref call.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.