glib

Module source

Structs

Pid

Process identifier

Priority

The priority of sources

SourceId

The id of a source that is returned by idle_add and timeout_add.

Functions

child_watch_add

Adds a closure to be called by the main loop the returned Source is attached to when a child process exits.

child_watch_add_local

Adds a closure to be called by the main loop the returned Source is attached to when a child process exits.

child_watch_source_new

Adds a closure to be called by the main loop the returned Source is attached to when a child process exits.

destroy_closure 🔒 ⚠

destroy_closure_child_watch 🔒 ⚠

destroy_closure_child_watch_local 🔒 ⚠

destroy_closure_local 🔒 ⚠

fnmut_callback_wrapper 🔒

Transform a generic FnOnce into a closure that can be used as callback in various glib methods

fnmut_callback_wrapper_local 🔒

Transform a generic FnOnce into a closure that can be used as callback in various glib methods

idle_add

Adds a closure to be called by the default main loop when it’s idle.

idle_add_full

Adds a closure to be called by the default main loop when it’s idle.

idle_add_local

Adds a closure to be called by the default main loop when it’s idle.

idle_add_local_full

Adds a closure to be called by the default main loop when it’s idle.

idle_add_local_once

Adds a closure to be called by the default main loop when it’s idle.

idle_add_once

Adds a closure to be called by the default main loop when it’s idle.

idle_source_new

Adds a closure to be called by the main loop the return Source is attached to when it’s idle.

into_raw 🔒

into_raw_child_watch 🔒

into_raw_child_watch_local 🔒

into_raw_local 🔒

timeout_add

Adds a closure to be called by the default main loop at regular intervals with millisecond granularity.

timeout_add_full

Adds a closure to be called by the default main loop at regular intervals with millisecond granularity.

timeout_add_local

Adds a closure to be called by the default main loop at regular intervals with millisecond granularity.

timeout_add_local_full

Adds a closure to be called by the default main loop at regular intervals with millisecond granularity.

timeout_add_local_once

Adds a closure to be called by the default main loop at regular intervals with millisecond granularity.

timeout_add_once

Adds a closure to be called by the default main loop at regular intervals with millisecond granularity.

timeout_add_seconds

Adds a closure to be called by the default main loop at regular intervals with second granularity.

timeout_add_seconds_local

Adds a closure to be called by the default main loop at regular intervals with second granularity.

timeout_add_seconds_local_once

Adds a closure to be called by the default main loop at regular intervals with second granularity.

timeout_add_seconds_once

Adds a closure to be called by the default main loop at regular intervals with second granularity.

timeout_source_new

Adds a closure to be called by the main loop the returned Source is attached to at regular intervals with millisecond granularity.

timeout_source_new_seconds

Adds a closure to be called by the main loop the returned Source is attached to at regular intervals with second granularity.

trampoline 🔒 ⚠

trampoline_child_watch 🔒 ⚠

trampoline_child_watch_local 🔒 ⚠

trampoline_local 🔒 ⚠