Function glib::source::child_watch_source_new

pub fn child_watch_source_new<F>(
    pid: Pid, 
    name: Option<&str>, 
    priority: Priority, 
    func: F
) -> Source where
    F: FnMut(Pid, i32) + Send + 'static, 

Adds a closure to be called by the main loop the returned Source is attached to when a child process exits.

func will be called when pid exits