Function glib::source::idle_add

pub fn idle_add<F>(func: F) -> SourceId where
    F: FnMut() -> Continue + Send + 'static, 

Adds a closure to be called by the default main loop when it’s idle.

func will be called repeatedly until it returns Continue(false).

The default main loop almost always is the main loop of the main thread. Thus the closure is called on the main thread. Adds a function to be called whenever there are no higher priority events pending to the default main loop. The function is given the default idle priority, G_PRIORITY_DEFAULT_IDLE. If the function returns false it is automatically removed from the list of event sources and will not be called again.

See [memory management of sources][mainloop-memory-management] for details on how to handle the return value and memory management of data.

This internally creates a main loop source using g_idle_source_new() and attaches it to the global MainContext using g_source_attach(), so the callback will be invoked in whichever thread is running that main context. You can do these steps manually if you need greater control or to use a custom main context.

function

function to call

Returns

the ID (greater than 0) of the event source.