glib::source

Function idle_add_local_full

pub fn idle_add_local_full<F>(priority: Priority, func: F) -> SourceId
where
    F: FnMut() -> ControlFlow + 'static,

Adds a closure to be called by the default main loop when it’s idle.

func will be called repeatedly with priority until it returns ControlFlow::Break.

The default main loop almost always is the main loop of the main thread. Thus, the closure is called on the main thread.

Different to idle_add(), this does not require func to be Send but can only be called from the thread that owns the main context.

This function panics if called from a different thread than the one that owns the default main context.