Function glib::source::source_remove

pub fn source_remove(source_id: SourceId)

Removes the source with the given id source_id from the default main context.

It is a programmer error to attempt to remove a non-existent source. Note: source id are reused.

For historical reasons, the native function always returns true, so we ignore it here.