Function glib::source::timeout_add_local

pub fn timeout_add_local<F>(interval: Duration, func: F) -> SourceId where
    F: FnMut() -> Continue + 'static, 

Adds a closure to be called by the default main loop at regular intervals with millisecond granularity.

func will be called repeatedly every interval milliseconds until it returns Continue(false). Precise timing is not guaranteed, the timeout may be delayed by other events. Prefer timeout_add_seconds when millisecond precision is not necessary.

The default main loop almost always is the main loop of the main thread. Thus the closure is called on the main thread.

Different to timeout_add(), this does not require func to be Send but can only be called from the thread that owns the main context.

This function panics if called from a different thread than the one that owns the main context.

Uses found in examples/

examples/list_store/main.rs
≺ ≻↕

  1
  2
  3
  4
  5
  6
  7
  8
  9
 10
 11
 12
 13
 14
 15
 16
 17
 18
 19
 20
 21
 22
 23
 24
 25
 26
 27
 28
 29
 30
 31
 32
 33
 34
 35
 36
 37
 38
 39
 40
 41
 42
 43
 44
 45
 46
 47
 48
 49
 50
 51
 52
 53
 54
 55
 56
 57
 58
 59
 60
 61
 62
 63
 64
 65
 66
 67
 68
 69
 70
 71
 72
 73
 74
 75
 76
 77
 78
 79
 80
 81
 82
 83
 84
 85
 86
 87
 88
 89
 90
 91
 92
 93
 94
 95
 96
 97
 98
 99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214
215
216
217
218
219
220
221
222
223
224
225
226
227
228
229
230
231
232
233
234
235
236
237
238
239
240
241
242
243
244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280
281
282
283
284
285
286
287
288
289
290
291
292
293
294
295
296
297
298
299
300
301
302
303
304
305
306
307
308
309
310
311
312
313
314
315
316
317
318

use gtk::glib;
use gtk::prelude::*;

use std::rc::Rc;
use std::time::Duration;

fn main() {
    let application = gtk::Application::new(
        Some("com.github.gtk-rs.examples.list-store"),
        Default::default(),
    );

    application.connect_startup(build_ui);

    application.run();
}

#[derive(Debug)]
#[repr(i32)]
enum Columns {
    Fixed = 0,
    Number,
    Severity,
    Description,
    Pulse,
    Icon,
    Active,
    Sensitive,
}

fn build_ui(application: &gtk::Application) {
    let window = gtk::ApplicationWindow::new(application);

    window.set_title("List Store");
    window.set_border_width(10);
    window.set_position(gtk::WindowPosition::Center);
    window.set_default_size(280, 250);

    let vbox = gtk::Box::new(gtk::Orientation::Vertical, 8);
    window.add(&vbox);

    let label = gtk::Label::new(Some(
        "This is the bug list (note: not based on real data, it would be \
         nice to have a nice ODBC interface to bugzilla or so, though).",
    ));
    vbox.add(&label);

    let sw = gtk::ScrolledWindow::new(None::<&gtk::Adjustment>, None::<&gtk::Adjustment>);
    sw.set_shadow_type(gtk::ShadowType::EtchedIn);
    sw.set_policy(gtk::PolicyType::Never, gtk::PolicyType::Automatic);
    vbox.add(&sw);

    let model = Rc::new(create_model());
    let treeview = gtk::TreeView::with_model(&*model);
    treeview.set_vexpand(true);
    treeview.set_search_column(Columns::Description as i32);

    sw.add(&treeview);

    add_columns(&model, &treeview);

    window.show_all();

    glib::timeout_add_local(
        Duration::from_millis(80),
        glib::clone!(@weak model => @default-return glib::Continue(false), move || {
            spinner_timeout(&model);
            glib::Continue(false)
        }),
    );
}

struct Data {
    fixed: bool,
    number: u32,
    severity: String,
    description: String,
}

fn create_model() -> gtk::ListStore {
    let col_types: [glib::Type; 8] = [
        glib::Type::BOOL,
        glib::Type::U32,
        glib::Type::STRING,
        glib::Type::STRING,
        glib::Type::U32,
        glib::Type::STRING,
        glib::Type::BOOL,
        glib::Type::BOOL,
    ];

    let data: [Data; 14] = [
        Data {
            fixed: false,
            number: 60482,
            severity: "Normal".to_string(),
            description: "scrollable notebooks and hidden tabs".to_string(),
        },
        Data {
            fixed: false,
            number: 60620,
            severity: "Critical".to_string(),
            description: "gdk_surface_clear_area (gdksurface-win32.c) is not thread-safe"
                .to_string(),
        },
        Data {
            fixed: false,
            number: 50214,
            severity: "Major".to_string(),
            description: "Xft support does not clean up correctly".to_string(),
        },
        Data {
            fixed: true,
            number: 52877,
            severity: "Major".to_string(),
            description: "GtkFileSelection needs a refresh method. ".to_string(),
        },
        Data {
            fixed: false,
            number: 56070,
            severity: "Normal".to_string(),
            description: "Can't click button after setting in sensitive".to_string(),
        },
        Data {
            fixed: true,
            number: 56355,
            severity: "Normal".to_string(),
            description: "GtkLabel - Not all changes propagate correctly".to_string(),
        },
        Data {
            fixed: false,
            number: 50055,
            severity: "Normal".to_string(),
            description: "Rework width/height computations for TreeView".to_string(),
        },
        Data {
            fixed: false,
            number: 58278,
            severity: "Normal".to_string(),
            description: "gtk_dialog_set_response_sensitive () doesn't work".to_string(),
        },
        Data {
            fixed: false,
            number: 55767,
            severity: "Normal".to_string(),
            description: "Getters for all setters".to_string(),
        },
        Data {
            fixed: false,
            number: 56925,
            severity: "Normal".to_string(),
            description: "Gtkcalender size".to_string(),
        },
        Data {
            fixed: false,
            number: 56221,
            severity: "Normal".to_string(),
            description: "Selectable label needs right-click copy menu".to_string(),
        },
        Data {
            fixed: true,
            number: 50939,
            severity: "Normal".to_string(),
            description: "Add shift clicking to GtkTextView".to_string(),
        },
        Data {
            fixed: false,
            number: 6112,
            severity: "Normal".to_string(),
            description: "netscape-like collapsable toolbars".to_string(),
        },
        Data {
            fixed: false,
            number: 1,
            severity: "Normal".to_string(),
            description: "First bug :=)".to_string(),
        },
    ];

    let store = gtk::ListStore::new(&col_types);

    for (d_idx, d) in data.iter().enumerate() {
        let icon_name = if d_idx == 1 || d_idx == 3 {
            "battery-caution-charging-symbolic"
        } else {
            ""
        };

        let sensitive = d_idx != 3;

        let values: [(u32, &dyn ToValue); 8] = [
            (0, &d.fixed),
            (1, &d.number),
            (2, &d.severity),
            (3, &d.description),
            (4, &0u32),
            (5, &icon_name),
            (6, &false),
            (7, &sensitive),
        ];
        store.set(&store.append(), &values);
    }

    store
}

fn fixed_toggled<W: IsA<gtk::CellRendererToggle>>(
    model: &gtk::ListStore,
    _w: &W,
    path: gtk::TreePath,
) {
    let iter = model.iter(&path).unwrap();
    let mut fixed = model
        .value(&iter, Columns::Fixed as i32)
        .get::<bool>()
        .unwrap_or_else(|err| {
            panic!(
                "ListStore value for {:?} at path {}: {}",
                Columns::Fixed,
                path,
                err
            )
        });
    fixed = !fixed;
    model.set_value(&iter, Columns::Fixed as u32, &fixed.to_value());
}

fn add_columns(model: &Rc<gtk::ListStore>, treeview: &gtk::TreeView) {
    // Column for fixed toggles
    {
        let renderer = gtk::CellRendererToggle::new();
        let model_clone = model.clone();
        renderer.connect_toggled(move |w, path| fixed_toggled(&model_clone, w, path));
        let column = gtk::TreeViewColumn::new();
        column.pack_start(&renderer, true);
        column.set_title("Fixed?");
        column.add_attribute(&renderer, "active", Columns::Fixed as i32);
        column.set_sizing(gtk::TreeViewColumnSizing::Fixed);
        column.set_fixed_width(50);
        treeview.append_column(&column);
    }

    // Column for bug numbers
    {
        let renderer = gtk::CellRendererText::new();
        let column = gtk::TreeViewColumn::new();
        column.pack_start(&renderer, true);
        column.set_title("Bug number");
        column.add_attribute(&renderer, "text", Columns::Number as i32);
        column.set_sort_column_id(Columns::Number as i32);
        treeview.append_column(&column);
    }

    // Column for severities
    {
        let renderer = gtk::CellRendererText::new();
        let column = gtk::TreeViewColumn::new();
        column.pack_start(&renderer, true);
        column.set_title("Severity");
        column.add_attribute(&renderer, "text", Columns::Severity as i32);
        column.set_sort_column_id(Columns::Severity as i32);
        treeview.append_column(&column);
    }

    // Column for description
    {
        let renderer = gtk::CellRendererText::new();
        let column = gtk::TreeViewColumn::new();
        column.pack_start(&renderer, true);
        column.set_title("Description");
        column.add_attribute(&renderer, "text", Columns::Description as i32);
        column.set_sort_column_id(Columns::Description as i32);
        treeview.append_column(&column);
    }

    // Column for spinner
    {
        let renderer = gtk::CellRendererSpinner::new();
        let column = gtk::TreeViewColumn::new();
        column.pack_start(&renderer, true);
        column.set_title("Spinning");
        column.add_attribute(&renderer, "pulse", Columns::Pulse as i32);
        column.add_attribute(&renderer, "active", Columns::Active as i32);
        treeview.append_column(&column);
    }

    // Column for symbolic icon
    {
        let renderer = gtk::CellRendererPixbuf::new();
        let column = gtk::TreeViewColumn::new();
        column.pack_start(&renderer, true);
        column.set_title("Symbolic icon");
        column.add_attribute(&renderer, "icon-name", Columns::Icon as i32);
        column.add_attribute(&renderer, "sensitive", Columns::Sensitive as i32);
        column.set_sort_column_id(Columns::Icon as i32);
        treeview.append_column(&column);
    }
}

fn spinner_timeout(model: &gtk::ListStore) -> Continue {
    let iter = model.iter_first().unwrap();
    let pulse = model
        .value(&iter, Columns::Pulse as i32)
        .get::<u32>()
        .unwrap_or_else(|err| {
            panic!(
                "ListStore value for {:?} at first entry: {}",
                Columns::Pulse,
                err
            )
        })
        .wrapping_add(1);

    model.set_value(&iter, Columns::Pulse as i32 as u32, &pulse.to_value());
    model.set_value(&iter, Columns::Active as i32 as u32, &true.to_value());

    Continue(true)
}

examples/progress_tracker/main.rs
≺ ≻↕

  1
  2
  3
  4
  5
  6
  7
  8
  9
 10
 11
 12
 13
 14
 15
 16
 17
 18
 19
 20
 21
 22
 23
 24
 25
 26
 27
 28
 29
 30
 31
 32
 33
 34
 35
 36
 37
 38
 39
 40
 41
 42
 43
 44
 45
 46
 47
 48
 49
 50
 51
 52
 53
 54
 55
 56
 57
 58
 59
 60
 61
 62
 63
 64
 65
 66
 67
 68
 69
 70
 71
 72
 73
 74
 75
 76
 77
 78
 79
 80
 81
 82
 83
 84
 85
 86
 87
 88
 89
 90
 91
 92
 93
 94
 95
 96
 97
 98
 99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214
215
216

use gtk::prelude::*;
use gtk::{gio, glib};

use std::cell::{Cell, RefCell};
use std::rc::Rc;
use std::thread;
use std::time::Duration;

pub fn main() {
    glib::set_program_name(Some("Progress Tracker"));

    let application = gtk::Application::new(
        Some("com.github.progress-tracker"),
        gio::ApplicationFlags::empty(),
    );

    application.connect_startup(|app| {
        let application = Application::new(app);

        let application_container = RefCell::new(Some(application));
        app.connect_shutdown(move |_| {
            let application = application_container
                .borrow_mut()
                .take()
                .expect("Shutdown called multiple times");
            // Here we could do whatever we need to do for shutdown now
            drop(application);
        });
    });

    application.connect_activate(|_| {});
    application.run();
}

pub struct Application {
    pub widgets: Rc<Widgets>,
}

impl Application {
    pub fn new(app: &gtk::Application) -> Self {
        let app = Application {
            widgets: Rc::new(Widgets::new(app)),
        };

        app.connect_progress();

        app
    }

    fn connect_progress(&self) {
        let active = Rc::new(Cell::new(false));
        self.widgets.main_view.button.connect_clicked(
            glib::clone!(@weak self.widgets as widgets => move |_| {
                if active.get() {
                    return;
                }

                active.set(true);

                let (tx, rx) = glib::MainContext::channel(glib::PRIORITY_DEFAULT);
                thread::spawn(move || {
                    for v in 1..=10 {
                        let _ = tx.send(Some(v));
                        thread::sleep(Duration::from_millis(500));
                    }
                    let _ = tx.send(None);
                });

                rx.attach(None, glib::clone!(@weak active, @weak widgets => @default-return glib::Continue(false), move |value| match value {
                    Some(value) => {
                        widgets
                            .main_view
                            .progress
                            .set_fraction(f64::from(value) / 10.0);

                        if value == 10 {
                            widgets
                                .view_stack
                                .set_visible_child(&widgets.complete_view.container);

                            glib::timeout_add_local(Duration::from_millis(1500), glib::clone!(@weak widgets => @default-return glib::Continue(false), move || {
                                widgets.main_view.progress.set_fraction(0.0);
                                widgets
                                    .view_stack
                                    .set_visible_child(&widgets.main_view.container);
                                glib::Continue(false)
                            }));
                        }

                        glib::Continue(true)
                    }
                    None => {
                        active.set(false);
                        glib::Continue(false)
                    }
                }));
            }),
        );
    }
}

pub struct Widgets {
    pub window: gtk::ApplicationWindow,
    pub header: Header,
    pub view_stack: gtk::Stack,
    pub main_view: MainView,
    pub complete_view: CompleteView,
}

impl Widgets {
    pub fn new(application: &gtk::Application) -> Self {
        let complete_view = CompleteView::new();
        let main_view = MainView::new();

        let view_stack = gtk::Stack::new();
        view_stack.set_border_width(6);
        view_stack.set_vexpand(true);
        view_stack.set_hexpand(true);
        view_stack.add(&main_view.container);
        view_stack.add(&complete_view.container);

        let header = Header::new();

        let window = gtk::ApplicationWindow::new(application);
        window.set_icon_name(Some("package-x-generic"));
        window.set_window_position(gtk::WindowPosition::Center);
        window.set_titlebar(Some(&header.container));
        window.add(&view_stack);
        window.show_all();
        window.set_default_size(500, 250);
        window.connect_delete_event(move |window, _| {
            window.close();
            Inhibit(false)
        });

        Self {
            window,
            header,
            view_stack,
            main_view,
            complete_view,
        }
    }
}

pub struct Header {
    container: gtk::HeaderBar,
}

impl Header {
    #[allow(clippy::new_without_default)]
    pub fn new() -> Self {
        let container = gtk::HeaderBar::new();
        container.set_title(Some("Progress Tracker"));
        container.set_show_close_button(true);

        Self { container }
    }
}

pub struct CompleteView {
    pub container: gtk::Grid,
}

impl CompleteView {
    #[allow(clippy::new_without_default)]
    pub fn new() -> Self {
        let label = gtk::Label::new(None);
        label.set_markup("Task complete");
        label.set_halign(gtk::Align::Center);
        label.set_valign(gtk::Align::Center);
        label.set_vexpand(true);
        label.set_hexpand(true);

        let container = gtk::Grid::new();
        container.set_vexpand(true);
        container.set_hexpand(true);
        container.add(&label);

        Self { container }
    }
}

pub struct MainView {
    pub container: gtk::Grid,
    pub progress: gtk::ProgressBar,
    pub button: gtk::Button,
}

impl MainView {
    #[allow(clippy::new_without_default)]
    pub fn new() -> Self {
        let progress = gtk::ProgressBar::new();
        progress.set_text(Some("Progress Bar"));
        progress.set_show_text(true);
        progress.set_hexpand(true);

        let button = gtk::Button::new();
        button.set_label("start");
        button.set_halign(gtk::Align::Center);

        let container = gtk::Grid::new();
        container.attach(&progress, 0, 0, 1, 1);
        container.attach(&button, 0, 1, 1, 1);
        container.set_row_spacing(12);
        container.set_border_width(6);
        container.set_vexpand(true);
        container.set_hexpand(true);

        Self {
            container,
            progress,
            button,
        }
    }
}