Function glib::source::timeout_add_once

pub fn timeout_add_once<F>(interval: Duration, func: F) -> SourceId
where
    F: FnOnce() + Send + 'static,

Adds a closure to be called by the default main loop at regular intervals with millisecond granularity.

func will be called repeatedly every interval milliseconds until it returns ControlFlow::Break. Precise timing is not guaranteed, the timeout may be delayed by other events. Prefer timeout_add_seconds when millisecond precision is not necessary.

The default main loop almost always is the main loop of the main thread. Thus, the closure is called on the main thread.

In comparison to timeout_add(), this only requires func to be FnOnce, and will automatically return ControlFlow::Break.