Function glib::source::timeout_add_seconds_local

pub fn timeout_add_seconds_local<F>(interval: u32, func: F) -> SourceId where
    F: FnMut() -> Continue + 'static, 

Adds a closure to be called by the default main loop at regular intervals with second granularity.

func will be called repeatedly every interval seconds until it returns Continue(false). Precise timing is not guaranteed, the timeout may be delayed by other events.

The default main loop almost always is the main loop of the main thread. Thus the closure is called on the main thread.

Different to timeout_add_seconds(), this does not require func to be Send but can only be called from the thread that owns the main context.

This function panics if called from a different thread than the one that owns the main context.

Uses found in examples/

examples/clock/main.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44

use chrono::Local;
use gtk::glib;
use gtk::prelude::*;

fn current_time() -> String {
    return format!("{}", Local::now().format("%Y-%m-%d %H:%M:%S"));
}

fn build_ui(application: &gtk::Application) {
    let window = gtk::ApplicationWindow::new(application);

    window.set_title("First GTK+ Clock");
    window.set_border_width(10);
    window.set_position(gtk::WindowPosition::Center);
    window.set_default_size(260, 40);

    let time = current_time();
    let label = gtk::Label::new(None);
    label.set_text(&time);

    window.add(&label);

    window.show_all();

    // we are using a closure to capture the label (else we could also use a normal function)
    let tick = move || {
        let time = current_time();
        label.set_text(&time);
        // we could return glib::Continue(false) to stop our clock after this tick
        glib::Continue(true)
    };

    // executes the closure once every second
    glib::timeout_add_seconds_local(1, tick);
}

fn main() {
    let application =
        gtk::Application::new(Some("com.github.gtk-rs.examples.clock"), Default::default());

    application.connect_activate(build_ui);

    application.run();
}