Function glib::source::timeout_add_seconds_local_once

pub fn timeout_add_seconds_local_once<F>(interval: u32, func: F) -> SourceIdwhere
    F: FnOnce() + 'static,

Adds a closure to be called by the default main loop at regular intervals with second granularity.

func will be called repeatedly every interval seconds until it returns Continue(false). Precise timing is not guaranteed, the timeout may be delayed by other events.

The default main loop almost always is the main loop of the main thread. Thus, the closure is called on the main thread.

Different to timeout_add_seconds(), this does not require func to be Send but can only be called from the thread that owns the main context.

This function panics if called from a different thread than the one that owns the main context.

In comparison to timeout_add_seconds_local(), this only requires func to be FnOnce, and will automatically return Continue(false).