Function glib::source::timeout_source_new_seconds

pub fn timeout_source_new_seconds<F>(
    interval: u32,
    name: Option<&str>,
    priority: Priority,
    func: F,
) -> Source
where
    F: FnMut() -> ControlFlow + Send + 'static,

Adds a closure to be called by the main loop the returned Source is attached to at regular intervals with second granularity.

func will be called repeatedly every interval seconds until it returns ControlFlow::Break. Precise timing is not guaranteed, the timeout may be delayed by other events.