Function glib::source::unix_fd_add

pub fn unix_fd_add<F>(fd: RawFd, condition: IOCondition, func: F) -> SourceId
where
    F: FnMut(RawFd, IOCondition) -> ControlFlow + Send + 'static,

Available on Unix only.

Adds a closure to be called by the main loop the returned Source is attached to whenever a UNIX file descriptor reaches the given IO condition.

func will be called repeatedly while the file descriptor matches the given IO condition until it returns ControlFlow::Break.

The default main loop almost always is the main loop of the main thread. Thus, the closure is called on the main thread.