Struct glib::ByteArray

pub struct ByteArray(_);

Contains the public fields of a GByteArray.

Implementations

impl ByteArray

pub fn new() -> Self

Creates a new ByteArray with a reference count of 1.

Returns

the new ByteArray

pub fn with_capacity(size: usize) -> Self

pub fn into_gbytes(self) -> Bytes

pub fn append<T: ?Sized + AsRef<[u8]>>(&self, data: &T) -> &Self

Adds the given bytes to the end of the ByteArray. The array will grow in size automatically if necessary.

array

a ByteArray

data

the byte data to be added

len

the number of bytes to add

Returns

the ByteArray

pub fn prepend<T: ?Sized + AsRef<[u8]>>(&self, data: &T) -> &Self

Adds the given data to the start of the ByteArray. The array will grow in size automatically if necessary.

array

a ByteArray

data

the byte data to be added

len

the number of bytes to add

Returns

the ByteArray

pub fn remove_index(&self, index: usize)

Removes the byte at the given index from a ByteArray. The following bytes are moved down one place.

array

a ByteArray

index_

the index of the byte to remove

Returns

the ByteArray

pub fn remove_index_fast(&self, index: usize)

Removes the byte at the given index from a ByteArray. The last element in the array is used to fill in the space, so this function does not preserve the order of the ByteArray. But it is faster than remove_index().

array

a ByteArray

index_

the index of the byte to remove

Returns

the ByteArray

pub fn remove_range(&self, index: usize, length: usize)

Removes the given number of bytes starting at the given index from a ByteArray. The following elements are moved to close the gap.

array

a ByteArray

index_

the index of the first byte to remove

length

the number of bytes to remove

Returns

the ByteArray

pub unsafe fn set_size(&self, size: usize)

Sets the size of the ByteArray, expanding it if necessary.

array

a ByteArray

length

the new size of the ByteArray

Returns

the ByteArray

pub fn sort<F: FnMut(&u8, &u8) -> Ordering>(&self, compare_func: F)

Sorts a byte array, using compare_func which should be a qsort()-style comparison function (returns less than zero for first arg is less than second arg, zero for equal, greater than zero if first arg is greater than second arg).

If two array elements compare equal, their order in the sorted array is undefined. If you want equal elements to keep their order (i.e. you want a stable sort) you can write a comparison function that, if two elements would otherwise compare equal, compares them by their addresses.

array

a ByteArray

compare_func

comparison function