Struct glib::ConstructSyncOnceCell

pub struct ConstructSyncOnceCell<T>(_);

Wrapper around a container type to be used for custom glib properties which get set during the construct phase. This is especially useful when used alongside the Properties macro.

Implementations

impl<T> ConstructSyncOnceCell<T>

pub fn new(value: T) -> Self

pub fn new_empty() -> Self