Struct glib::EnumClass

pub struct EnumClass(_);

Representation of an enum for dynamically, at runtime, querying the values of the enum and using them.

Methods

impl EnumClass

pub fn new(type_: Type) -> Option<Self>

Create a new EnumClass from a Type.

Returns None if type_ is not representing an enum.

pub fn type_(&self) -> Type

Type of the enum.

pub fn get_value(&self, value: i32) -> Option<EnumValue>

Gets EnumValue by integer value, if existing.

Returns None if the enum does not contain any value with value.

pub fn get_value_by_name(&self, name: &str) -> Option<EnumValue>

Gets EnumValue by string name name, if existing.

Returns None if the enum does not contain any value with name name.

pub fn get_value_by_nick(&self, nick: &str) -> Option<EnumValue>

Gets EnumValue by string nick nick, if existing.

Returns None if the enum does not contain any value with nick nick.

pub fn get_values(&self) -> Vec<EnumValue>

Gets all EnumValue of this EnumClass.

pub fn to_value(&self, value: i32) -> Option<Value>

Converts integer value to a Value, if part of the enum.

pub fn to_value_by_name(&self, name: &str) -> Option<Value>

Converts string name name to a Value, if part of the enum.

pub fn to_value_by_nick(&self, nick: &str) -> Option<Value>

Converts string nick nick to a Value, if part of the enum.