Struct glib::GStr

pub struct GStr(/* private fields */);

Representation of a borrowed GString.

This type is very similar to std::ffi::CStr, but with one added constraint: the string must also be valid UTF-8.

Implementations

impl GStr

pub fn from_utf8_with_nul(bytes: &[u8]) -> Result<&Self, GStrError>

Creates a GLib string wrapper from a byte slice.

This function will cast the provided bytes to a GStr wrapper after ensuring that the byte slice is valid UTF-8 and is nul-terminated.

pub fn from_utf8_with_nul_checked(bytes: &[u8]) -> Result<&Self, GStrError>

Creates a GLib string wrapper from a byte slice, checking for interior nul-bytes.

This function will cast the provided bytes to a GStr wrapper after ensuring that the byte slice is valid UTF-8, is nul-terminated, and does not contain any interior nul-bytes.

pub const unsafe fn from_utf8_with_nul_unchecked(bytes: &[u8]) -> &Self

Unsafely creates a GLib string wrapper from a byte slice.

This function will cast the provided bytes to a GStr wrapper without performing any sanity checks.

Safety

The provided slice must be valid UTF-8 and nul-terminated. It is undefined behavior to pass a slice that does not uphold those conditions.

pub fn from_utf8_until_nul(bytes: &[u8]) -> Result<&Self, GStrError>

Creates a GLib string wrapper from a byte slice, truncating it at the first nul-byte.

This function will cast the provided bytes to a GStr wrapper after ensuring that the byte slice is valid UTF-8 and contains at least one nul-byte.

pub fn from_str_with_nul(s: &str) -> Result<&Self, GStrError>

Creates a GLib string wrapper from a string slice.

The string slice must be terminated with a nul-byte.

This function will cast the provided bytes to a GStr wrapper after ensuring that the string slice is nul-terminated.

pub fn from_str_with_nul_checked(s: &str) -> Result<&Self, GStrError>

Creates a GLib string wrapper from a string slice, checking for interior nul-bytes.

The string slice must be terminated with a nul-byte.

This function will cast the provided bytes to a GStr wrapper after ensuring that the string slice is nul-terminated and does not contain any interior nul-bytes.

pub const unsafe fn from_str_with_nul_unchecked(s: &str) -> &Self

Unsafely creates a GLib string wrapper from a string slice. The string slice must be terminated with a nul-byte.

This function will cast the provided string slice to a GStr without performing any sanity checks.

Safety

The provided string slice must be nul-terminated. It is undefined behavior to pass a slice that does not uphold those conditions.

pub fn from_str_until_nul(s: &str) -> Result<&Self, GStrError>

Creates a GLib string wrapper from a string slice, truncating it at the first nul-byte.

The string slice must contain at least one nul-byte.

This function will cast the provided bytes to a GStr wrapper after ensuring that the string slice contains at least one nul-byte.

pub unsafe fn from_ptr<'a>(ptr: *const c_char) -> &'a Self

Wraps a raw C string with a safe GLib string wrapper. The provided C string must be valid UTF-8 and nul-terminated. All constraints from CStr::from_ptr also apply here.

Safety

See CStr::from_ptr.

pub unsafe fn from_ptr_checked<'a>(ptr: *const c_char) -> Option<&'a Self>

Wraps a raw C string with a safe GLib string wrapper. The provided C string must be nul-terminated. All constraints from std::ffi::CStr::from_ptr also apply here.

If the string is valid UTF-8 then it is directly returned, otherwise None is returned.

pub unsafe fn from_ptr_lossy<'a>(ptr: *const c_char) -> Cow<'a, Self>

Wraps a raw C string with a safe GLib string wrapper. The provided C string must be nul-terminated. All constraints from std::ffi::CStr::from_ptr also apply here.

If the string is valid UTF-8 then it is directly returned otherwise a copy is created with every invalid character replaced by the Unicode replacement character (U+FFFD).

pub const fn as_bytes_with_nul(&self) -> &[u8]

Converts this GLib string to a byte slice containing the trailing 0 byte.

This function is the equivalent of GStr::as_bytes except that it will retain the trailing nul terminator instead of chopping it off.

pub const fn as_bytes(&self) -> &[u8]

Converts this GLib string to a byte slice.

The returned slice will not contain the trailing nul terminator that this GLib string has.

pub const fn as_ptr(&self) -> *const c_char

Returns the inner pointer to this GLib string.

The returned pointer will be valid for as long as self is, and points to a contiguous region of memory terminated with a 0 byte to represent the end of the string.

WARNING

The returned pointer is read-only; writing to it (including passing it to C code that writes to it) causes undefined behavior. It is your responsibility to make sure that the underlying memory is not freed too early.

pub const fn as_str(&self) -> &str

Converts this GLib string to a string slice.

pub fn to_cstr(&self) -> Result<&CStr, GStrInteriorNulError>

Converts this GLib string to a C string slice, checking for interior nul-bytes.

Returns Err if the string contains any interior nul-bytes.

pub fn to_cstr_until_nul(&self) -> &CStr

Converts this GLib string to a C string slice, truncating it at the first nul-byte.

pub const unsafe fn to_cstr_unchecked(&self) -> &CStr

Converts this GLib string to a C string slice, without checking for interior nul-bytes.

Safety

self must not contain any interior nul-bytes besides the final terminating nul-byte. It is undefined behavior to call this on a string that contains interior nul-bytes.

pub fn collate(&self, other: impl IntoGStr) -> Ordering

pub const NONE: Option<&'static GStr> = None