Struct glib::GString

pub struct GString(/* private fields */);

A type representing an owned, C-compatible, nul-terminated UTF-8 string.

GString is to &GStr as String is to &str: the former in each pair are owned strings; the latter are borrowed references.

This type is similar to std::ffi::CString, but with some special behavior. When debug assertions are enabled, From<String> will panic if there are interior nul-bytes. In production builds, no checks will be made for interior nul-bytes, and strings that contain interior nul-bytes will simply end at first nul-byte when converting to a C string.

The constructors beginning with from_utf8 and from_string can also be used to further control how interior nul-bytes are handled.

Implementations

impl GString

pub fn new() -> Self

Creates a new empty GString.

Does not allocate.

pub fn format(args: Arguments<'_>) -> Self

Formats an Arguments into a GString.

This function is the same as std::fmt::format, except it returns a GString. The Arguments instance can be created with the format_args! macro.

Please note that using gformat! might be preferable.

pub fn from_utf8(bytes: Vec<u8>) -> Result<Self, FromUtf8Error>

Creates a GLib string by consuming a byte vector.

Takes ownership of bytes. Returns Err if it contains invalid UTF-8.

A trailing nul-byte will be appended by this function.

pub fn from_utf8_checked( bytes: Vec<u8>, ) -> Result<Self, GStringFromError<Vec<u8>>>

Creates a GLib string by consuming a byte vector, checking for interior nul-bytes.

Takes ownership of bytes, as long as it is valid UTF-8 and does not contain any interior nul-bytes. Otherwise, Err is returned.

A trailing nul-byte will be appended by this function.

pub unsafe fn from_utf8_unchecked(v: Vec<u8>) -> Self

Unsafely creates a GLib string by consuming a byte vector, without checking for UTF-8 or interior nul-bytes.

A trailing nul-byte will be appended by this function.

Safety

The byte vector must not contain invalid UTF-8 characters. It is undefined behavior to pass a vector that contains invalid UTF-8.

pub fn from_utf8_with_nul( bytes: Vec<u8>, ) -> Result<Self, GStringFromError<Vec<u8>>>

Creates a GLib string by consuming a nul-terminated byte vector, without checking for interior nul-bytes.

Takes ownership of bytes. Returns Err if it contains invalid UTF-8 or does not have a trailing nul-byte.

pub fn from_utf8_with_nul_checked( bytes: Vec<u8>, ) -> Result<Self, GStringFromError<Vec<u8>>>

Creates a GLib string by consuming a nul-terminated byte vector.

Takes ownership of bytes. Returns Err if it contains invalid UTF-8, does not have a trailing nul-byte, or contains interior nul-bytes.

pub unsafe fn from_utf8_with_nul_unchecked(v: Vec<u8>) -> Self

Creates a GLib string by consuming a byte vector, without checking for UTF-8, a trailing nul-byte, or interior nul-bytes.

Safety

The byte vector must not contain invalid UTF-8 characters, and must have a trailing nul-byte. It is undefined behavior to pass a vector that does not uphold those conditions.

pub fn from_utf8_until_nul( bytes: Vec<u8>, ) -> Result<Self, GStringFromError<Vec<u8>>>

Creates a GLib string by consuming a nul-terminated byte vector, truncating it at the first nul-byte.

Takes ownership of bytes. Returns Err if it contains invalid UTF-8 or does not contain at least one nul-byte.

pub fn from_string_checked( s: String, ) -> Result<Self, GStringInteriorNulError<String>>

Creates a GLib string by consuming a string, checking for interior nul-bytes.

Takes ownership of s, as long as it does not contain any interior nul-bytes. Otherwise, Err is returned.

A trailing nul-byte will be appended by this function.

pub fn from_string_unchecked(s: String) -> Self

Creates a GLib string by consuming a string, without checking for interior nul-bytes.

A trailing nul-byte will be appended by this function.

pub unsafe fn from_ptr_lossy<'a>(ptr: *const c_char) -> Cow<'a, GStr>

Wraps a raw C string with a safe GLib string wrapper. The provided C string must be nul-terminated. All constraints from std::ffi::CStr::from_ptr also apply here.

If the string is valid UTF-8 then it is directly returned otherwise a copy is created with every invalid character replaced by the Unicode replacement character (U+FFFD).

pub unsafe fn from_ptr_and_len_unchecked(ptr: *const c_char, len: usize) -> Self

Wraps a raw C string with a safe GLib string wrapper. The provided C string must be nul-terminated. All constraints from std::ffi::CStr::from_ptr also apply here.

len is the length without the nul-terminator, i.e. if len == 0 is passed then *ptr must be the nul-terminator.

pub fn as_str(&self) -> &str

Return the GString as string slice.

pub fn as_gstr(&self) -> &GStr

Extracts the GStr containing the entire string.

pub fn as_ptr(&self) -> *const c_char

Return the underlying pointer of the GString.

pub fn into_bytes(self) -> Vec<u8>

Consumes the GString and returns the underlying byte buffer.

The returned buffer is not guaranteed to contain a trailing nul-byte.

pub fn into_bytes_with_nul(self) -> Vec<u8>

Consumes the GString and returns the underlying byte buffer, with trailing nul-byte.