Struct glib::JoinError

pub struct JoinError(/* private fields */);

Task failure from awaiting a JoinHandle.

Implementations

impl JoinError

pub fn is_cancelled(&self) -> bool

Returns true if the handle was cancelled.

pub fn is_panic(&self) -> bool

Returns true if the task terminated with a panic.

pub fn into_panic(self) -> Box<dyn Any + Send + 'static>

Converts the error into a panic result.

Panics

Panics if the error is not a panic error. Use is_panic to check first if the error represents a panic.

pub fn try_into_panic(self) -> Result<Box<dyn Any + Send + 'static>, Self>

Attempts to convert the error into a panic result.

Returns Err(self) if the error is not a panic result.