Struct glib::KeyFile

pub struct KeyFile(_);

The GKeyFile struct contains only private data and should not be accessed directly.

Implementations

impl KeyFile

pub fn new() -> KeyFile

Creates a new empty crate::KeyFile object. Use Self::load_from_file(), Self::load_from_data(), Self::load_from_dirs() or Self::load_from_data_dirs() to read an existing key file.

Returns

an empty crate::KeyFile.

pub fn comment(
&self,
group_name: Option<&str>,
key: Option<&str>
) -> Result<GString, Error>

pub fn double(&self, group_name: &str, key: &str) -> Result<f64, Error>

pub fn double_list(
&self,
group_name: &str,
key: &str
) -> Result<Vec<f64>, Error>

pub fn groups(&self) -> (Vec<GString>, usize)

pub fn int64(&self, group_name: &str, key: &str) -> Result<i64, Error>

pub fn integer(&self, group_name: &str, key: &str) -> Result<i32, Error>

pub fn integer_list(
&self,
group_name: &str,
key: &str
) -> Result<Vec<i32>, Error>

pub fn keys(&self, group_name: &str) -> Result<(Vec<GString>, usize), Error>

pub fn locale_for_key(
&self,
group_name: &str,
key: &str,
locale: Option<&str>
) -> Option<GString>

This is supported on crate feature v2_56 only.

pub fn start_group(&self) -> Option<GString>

pub fn uint64(&self, group_name: &str, key: &str) -> Result<u64, Error>

pub fn value(&self, group_name: &str, key: &str) -> Result<GString, Error>

pub fn has_group(&self, group_name: &str) -> bool

Looks whether the key file has the group group_name.

group_name

a group name

Returns

true if group_name is a part of self, false otherwise.

pub fn load_from_bytes(
&self,
bytes: &Bytes,
flags: KeyFileFlags
) -> Result<(), Error>

This is supported on crate feature v2_50 only.

Loads a key file from the data in bytes into an empty crate::KeyFile structure. If the object cannot be created then error is set to a crate::KeyFileError.

bytes

a crate::Bytes

flags

flags from crate::KeyFileFlags

Returns

true if a key file could be loaded, false otherwise

pub fn load_from_data(
&self,
data: &str,
flags: KeyFileFlags
) -> Result<(), Error>

Loads a key file from memory into an empty crate::KeyFile structure. If the object cannot be created then error is set to a crate::KeyFileError.

data

key file loaded in memory

length

the length of data in bytes (or (gsize)-1 if data is nul-terminated)

flags

flags from crate::KeyFileFlags

Returns

true if a key file could be loaded, false otherwise

pub fn load_from_file<P: AsRef<Path>>(
&self,
file: P,
flags: KeyFileFlags
) -> Result<(), Error>

Loads a key file into an empty crate::KeyFile structure.

If the OS returns an error when opening or reading the file, a G_FILE_ERROR is returned. If there is a problem parsing the file, a G_KEY_FILE_ERROR is returned.

This function will never return a crate::KeyFileError::NotFound error. If the file is not found, crate::FileError::Noent is returned.

file

the path of a filename to load, in the GLib filename encoding

flags

flags from crate::KeyFileFlags

Returns

true if a key file could be loaded, false otherwise

pub fn remove_comment(
&self,
group_name: Option<&str>,
key: Option<&str>
) -> Result<(), Error>

Removes a comment above key from group_name. If key is None then comment will be removed above group_name. If both key and group_name are None, then comment will be removed above the first group in the file.

group_name

a group name, or None

key

a key

Returns

true if the comment was removed, false otherwise

pub fn remove_group(&self, group_name: &str) -> Result<(), Error>

Removes the specified group, group_name, from the key file.

group_name

a group name

Returns

true if the group was removed, false otherwise

pub fn remove_key(&self, group_name: &str, key: &str) -> Result<(), Error>

Removes key in group_name from the key file.

group_name

a group name

key

a key name to remove

Returns

true if the key was removed, false otherwise

pub fn set_boolean(&self, group_name: &str, key: &str, value: bool)

Associates a new boolean value with key under group_name. If key cannot be found then it is created.

group_name

a group name

key

a key

value

true or false

pub fn set_comment(
&self,
group_name: Option<&str>,
key: Option<&str>,
comment: &str
) -> Result<(), Error>

Places a comment above key from group_name.

If key is None then comment will be written above group_name. If both key and group_name are None, then comment will be written above the first group in the file.

Note that this function prepends a ‘#’ comment marker to each line of comment.

group_name

a group name, or None

key

a key

comment

a comment

Returns

true if the comment was written, false otherwise

pub fn set_double(&self, group_name: &str, key: &str, value: f64)

Associates a new double value with key under group_name. If key cannot be found then it is created.

group_name

a group name

key

a key

value

a double value

pub fn set_int64(&self, group_name: &str, key: &str, value: i64)

Associates a new integer value with key under group_name. If key cannot be found then it is created.

group_name

a group name

key

a key

value

an integer value

pub fn set_integer(&self, group_name: &str, key: &str, value: i32)

Associates a new integer value with key under group_name. If key cannot be found then it is created.

group_name

a group name

key

a key

value

an integer value

pub fn set_list_separator(&self, separator: Char)

Sets the character which is used to separate values in lists. Typically ‘;’ or ‘,’ are used as separators. The default list separator is ‘;’.

separator

the separator

pub fn set_locale_string(
&self,
group_name: &str,
key: &str,
locale: &str,
string: &str
)

Associates a string value for key and locale under group_name. If the translation for key cannot be found then it is created.

group_name

a group name

key

a key

locale

a locale identifier

string

a string

pub fn set_string(&self, group_name: &str, key: &str, string: &str)

Associates a new string value with key under group_name. If key cannot be found then it is created. If group_name cannot be found then it is created. Unlike Self::set_value(), this function handles characters that need escaping, such as newlines.

group_name

a group name

key

a key

string

a string

pub fn set_uint64(&self, group_name: &str, key: &str, value: u64)

Associates a new integer value with key under group_name. If key cannot be found then it is created.

group_name

a group name

key

a key

value

an integer value

pub fn set_value(&self, group_name: &str, key: &str, value: &str)

Associates a new value with key under group_name.

If key cannot be found then it is created. If group_name cannot be found then it is created. To set an UTF-8 string which may contain characters that need escaping (such as newlines or spaces), use Self::set_string().

group_name

a group name

key

a key

value

a string

impl KeyFile

pub fn save_to_file<T: AsRef<Path>>(&self, filename: T) -> Result<(), Error>

Writes the contents of self to filename using g_file_set_contents(). If you need stricter guarantees about durability of the written file than are provided by g_file_set_contents(), use g_file_set_contents_full() with the return value of Self::to_data().

This function can fail for any of the reasons that g_file_set_contents() may fail.

filename

the name of the file to write to

Returns

true if successful, else false with error set

pub fn load_from_data_dirs<T: AsRef<Path>>(
&self,
file: T,
flags: KeyFileFlags
) -> Result<PathBuf, Error>

This function looks for a key file named file in the paths returned from g_get_user_data_dir() and g_get_system_data_dirs(), loads the file into self and returns the file’s full path in full_path. If the file could not be loaded then an error is set to either a crate::FileError or crate::KeyFileError.

file

a relative path to a filename to open and parse

full_path

return location for a string containing the full path of the file, or None

flags

flags from crate::KeyFileFlags

Returns

true if a key file could be loaded, false otherwise

pub fn load_from_dirs<T: AsRef<Path>, U: AsRef<Path>>(
&self,
file: T,
search_dirs: &[U],
flags: KeyFileFlags
) -> Result<PathBuf, Error>

This function looks for a key file named file in the paths specified in search_dirs, loads the file into self and returns the file’s full path in full_path.

If the file could not be found in any of the search_dirs, crate::KeyFileError::NotFound is returned. If the file is found but the OS returns an error when opening or reading the file, a G_FILE_ERROR is returned. If there is a problem parsing the file, a G_KEY_FILE_ERROR is returned.

file

a relative path to a filename to open and parse

search_dirs

None-terminated array of directories to search

full_path

return location for a string containing the full path of the file, or None

flags

flags from crate::KeyFileFlags

Returns

true if a key file could be loaded, false otherwise

pub fn to_data(&self) -> GString

This function outputs self as a string.

Note that this function never reports an error, so it is safe to pass None as error.

length

return location for the length of the returned string, or None

Returns

a newly allocated string holding the contents of the crate::KeyFile

pub fn boolean(&self, group_name: &str, key: &str) -> Result<bool, Error>

pub fn has_key(&self, group_name: &str, key: &str) -> Result<bool, Error>

Looks whether the key file has the key key in the group group_name.

Note that this function does not follow the rules for crate::Error strictly; the return value both carries meaning and signals an error. To use this function, you must pass a crate::Error pointer in error, and check whether it is not None to see if an error occurred.

Language bindings should use [Self::get_value()] to test whether or not a key exists.

group_name

a group name

key

a key name

Returns

true if key is a part of group_name, false otherwise

pub fn boolean_list(
&self,
group_name: &str,
key: &str
) -> Result<Vec<bool>, Error>

pub fn string(&self, group_name: &str, key: &str) -> Result<GString, Error>

pub fn string_list(
&self,
group_name: &str,
key: &str
) -> Result<Vec<GString>, Error>

pub fn locale_string(
&self,
group_name: &str,
key: &str,
locale: Option<&str>
) -> Result<GString, Error>

pub fn locale_string_list(
&self,
group_name: &str,
key: &str,
locale: Option<&str>
) -> Result<Vec<GString>, Error>