Struct glib::MainLoop

pub struct MainLoop(_);

Implementations

impl MainLoop

pub fn new(context: Option<&MainContext>, is_running: bool) -> MainLoop

Uses found in examples/

examples/gio_futures_await/main.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73

use glib::clone;
use gtk::prelude::*;
use gtk::{gio, glib};

use futures::prelude::*;

use std::str;

fn main() {
    let c = glib::MainContext::default();
    let l = glib::MainLoop::new(Some(&c), false);

    c.push_thread_default();

    let file = gio::File::for_path("Cargo.toml");

    let future = clone!(@strong l => async move {
        match read_file(file).await {
            Ok(()) => (),
            Err(err) => eprintln!("Got error: {}", err),
        }
        l.quit();
    });

    c.spawn_local(future);

    l.run();

    c.pop_thread_default();
}

/// Throughout our chained futures, we convert all errors to strings
/// via map_err() return them directly.
async fn read_file(file: gio::File) -> Result<(), String> {
    // Try to open the file.
    let strm = file
        .read_async_future(glib::PRIORITY_DEFAULT)
        .map_err(|err| format!("Failed to open file: {}", err))
        .await?;

    // If opening the file succeeds, we asynchronously loop and
    // read the file in up to 64 byte chunks and re-use the same
    // vec for each read.
    let mut buf = vec![0; 64];
    let mut idx = 0;

    loop {
        let (b, len) = strm
            .read_async_future(buf, glib::PRIORITY_DEFAULT)
            .map_err(|(_buf, err)| format!("Failed to read from stream: {}", err))
            .await?;

        // Once 0 is returned, we know that we're done with reading, otherwise
        // loop again and read another chunk.
        if len == 0 {
            break;
        }

        buf = b;

        println!("line {}: {:?}", idx, str::from_utf8(&buf[0..len]).unwrap());

        idx += 1;
    }

    // Asynchronously close the stream in the end.
    let _ = strm
        .close_async_future(glib::PRIORITY_DEFAULT)
        .map_err(|err| format!("Failed to close stream: {}", err))
        .await?;

    Ok(())
}

examples/gio_futures/main.rs
≺ ≻↕

  1
  2
  3
  4
  5
  6
  7
  8
  9
 10
 11
 12
 13
 14
 15
 16
 17
 18
 19
 20
 21
 22
 23
 24
 25
 26
 27
 28
 29
 30
 31
 32
 33
 34
 35
 36
 37
 38
 39
 40
 41
 42
 43
 44
 45
 46
 47
 48
 49
 50
 51
 52
 53
 54
 55
 56
 57
 58
 59
 60
 61
 62
 63
 64
 65
 66
 67
 68
 69
 70
 71
 72
 73
 74
 75
 76
 77
 78
 79
 80
 81
 82
 83
 84
 85
 86
 87
 88
 89
 90
 91
 92
 93
 94
 95
 96
 97
 98
 99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126

use gtk::prelude::*;
use gtk::{gio, glib};

use std::str;

use futures::prelude::*;

fn main() {
    let c = glib::MainContext::default();
    let l = glib::MainLoop::new(Some(&c), false);

    c.push_thread_default();

    let file = gio::File::for_path("Cargo.toml");

    let l_clone = l.clone();
    c.spawn_local(
        read_and_print_file(&file)
            // Once all is done we quit the main loop and in case of an
            // error first print that error.
            .map(move |res| {
                if let Err(err) = res {
                    eprintln!("Got error: {}", err);
                }

                l_clone.quit();
            }),
    );

    l.run();

    c.pop_thread_default();
}

/// Throughout our chained futures, we convert all errors to strings
/// via map_err() and print them at the very end.
///
/// Open the file for reading, and if that succeeds read the whole file from
/// the resulting input stream.
fn read_and_print_file(
    file: &gio::File,
) -> impl Future<Output = Result<(), String>> + std::marker::Unpin {
    file.read_async_future(glib::PRIORITY_DEFAULT)
        .map_err(|err| format!("Failed to open file: {}", err))
        .and_then(read_and_print_chunks)
}

/// Read the input stream in chunks of 64 bytes, always into the same buffer
/// without re-allocating it all the time. Continue until the end of the file
/// or an error happens.
fn read_and_print_chunks(
    strm: gio::FileInputStream,
) -> impl Future<Output = Result<(), String>> + std::marker::Unpin {
    let buf = vec![0; 64];
    let idx = 0;

    // We use unfold() here, which takes some initialization data and a
    // closure that is returning an item and the next state, or None to
    // finish the stream
    futures::stream::unfold(Some((buf, idx)), move |buf_and_idx| {
        // If None was returned from the last iteration then the last iteration
        // was closing the input stream or an error happened, and now we only
        // have to finish the stream created by unfold().
        //
        // Otherwise we got the buffer to read to and the index of the next line
        // from the previous iteration.
        let (buf, idx) = match buf_and_idx {
            None => {
                return futures::future::Either::Left(futures::future::ready(None));
            }
            Some(buf_and_idx) => buf_and_idx,
        };

        // Read and print the next chunk
        futures::future::Either::Right(read_and_print_next_chunk(&strm, buf, idx).map(move |res| {
            match res {
                // And error happened, return the error from this stream and then finish on the
                // next iteration.
                Err(err) => Some((Err(err), None)),
                // The input stream was closed, return Ok(()) from this stream and then finish on
                // the next iteration.
                Ok(None) => Some((Ok(()), None)),
                // A chunk was successfully read and printed, return Ok(()) from this stream and
                // then continue with the next iteration.
                Ok(Some(buf)) => Some((Ok(()), Some((buf, idx + 1)))),
            }
        }))
    })
    // Convert the stream into a simple future that collects all items and
    // returns Ok(()), or short-circuits on the very first error and returns it
    .try_for_each(|_| futures::future::ok(()))
}

/// Read the next chunk into the buffer and print it out, or return an error. If
/// the input stream is finished, close the stream.
///
/// After reading successfully we return the buffer again so it can be used in the
/// next iteration.
fn read_and_print_next_chunk(
    strm: &gio::FileInputStream,
    buf: Vec<u8>,
    idx: usize,
) -> impl Future<Output = Result<Option<Vec<u8>>, String>> + std::marker::Unpin {
    let strm_clone = strm.clone();
    strm.read_async_future(buf, glib::PRIORITY_DEFAULT)
        .map_err(|(_buf, err)| format!("Failed to read from stream: {}", err))
        .and_then(move |(buf, len)| {
            println!("line {}: {:?}", idx, str::from_utf8(&buf[0..len]).unwrap());

            // 0 is only returned when the input stream is finished, in which case
            // we drop the buffer and close the stream asynchronously.
            //
            // Otherwise we simply return the buffer again so it can be read into
            // in the next iteration.
            if len == 0 {
                futures::future::Either::Left(
                    strm_clone
                        .close_async_future(glib::PRIORITY_DEFAULT)
                        .map_err(|err| format!("Failed to close stream: {}", err))
                        .map_ok(|_| None),
                )
            } else {
                futures::future::Either::Right(futures::future::ok(Some(buf)))
            }
        })
}

examples/gio_task/main.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99

mod file_size;

use futures_channel::oneshot;
use glib::clone;
use gtk::glib::translate::*;
use gtk::{gio, glib};

use file_size::FileSize;

fn main() {
    let main_context = glib::MainContext::default();
    let main_loop = glib::MainLoop::new(Some(&main_context), false);
    let (send_safe, recv_safe) = oneshot::channel();
    let (send_unsafe, recv_unsafe) = oneshot::channel();

    main_context.push_thread_default();
    main_context.invoke_local(move || {
        run_unsafe(send_unsafe);
    });
    main_context.invoke_local(move || {
        run_safe(send_safe);
    });

    main_context.spawn_local(clone!(@strong main_loop => async move {
        recv_safe.await.unwrap();
        recv_unsafe.await.unwrap();
        main_loop.quit();
    }));
    main_loop.run();
    main_context.pop_thread_default();
}

// This function mimicks what the C code using the exported async/finish API would do. It first defines a
// callback of type GAsyncResult that internally calls get_file_size_finish to retrieve the Task
// result, and simply prints it out. Then it invokes the my_file_size_get_file_size_async method, passing the
// callback to it as parameter.
fn run_unsafe(send: oneshot::Sender<()>) {
    let simple_object = FileSize::new();
    let cancellable = gio::Cancellable::new();

    // The callback to be passed to my_file_size_get_file_size_async
    unsafe extern "C" fn c_callback(
        source_object: *mut glib::object::GObject,
        result: *mut gio::ffi::GAsyncResult,
        user_data: glib::ffi::gpointer,
    ) {
        let mut error = std::ptr::null_mut();
        let ret = file_size::ffi::my_file_size_get_file_size_finish(
            source_object as *mut file_size::ffi::FileSize,
            result,
            &mut error,
        );
        if !error.is_null() {
            eprintln!("Task returned error!");
            return;
        }

        println!("Unsafe callback - Returned value from task: {}", ret);
        println!(
            "Unsafe callback - FileSize::size: {}",
            file_size::ffi::my_file_size_get_retrieved_size(
                source_object as *mut file_size::ffi::FileSize
            )
        );

        Box::from_raw(user_data as *mut oneshot::Sender<()>)
            .send(())
            .unwrap();
    }

    // The actual call to my_file_size_get_file_size_async
    unsafe {
        file_size::ffi::my_file_size_get_file_size_async(
            simple_object.to_glib_none().0,
            cancellable.to_glib_none().0,
            Some(c_callback),
            Box::into_raw(Box::new(send)) as glib::ffi::gpointer,
        );
    }
}

// This function is the "safe" counterpart of "run_unsafe", using only safe rust
// bindings to accomplish the same task
fn run_safe(send: oneshot::Sender<()>) {
    let simple_object = FileSize::new();
    let cancellable = gio::Cancellable::new();

    let closure = move |value: i64, source_object: &FileSize| {
        println!("Safe callback - Returned value from task: {}", value);
        println!(
            "Safe callback - FileSize::size: {}",
            source_object.retrieved_size().unwrap()
        );

        send.send(()).unwrap();
    };

    simple_object.file_size_async(Some(&cancellable), closure);
}

pub fn context(&self) -> MainContext

pub fn is_running(&self) -> bool

pub fn quit(&self)

Uses found in examples/

examples/gio_futures_await/main.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73

use glib::clone;
use gtk::prelude::*;
use gtk::{gio, glib};

use futures::prelude::*;

use std::str;

fn main() {
    let c = glib::MainContext::default();
    let l = glib::MainLoop::new(Some(&c), false);

    c.push_thread_default();

    let file = gio::File::for_path("Cargo.toml");

    let future = clone!(@strong l => async move {
        match read_file(file).await {
            Ok(()) => (),
            Err(err) => eprintln!("Got error: {}", err),
        }
        l.quit();
    });

    c.spawn_local(future);

    l.run();

    c.pop_thread_default();
}

/// Throughout our chained futures, we convert all errors to strings
/// via map_err() return them directly.
async fn read_file(file: gio::File) -> Result<(), String> {
    // Try to open the file.
    let strm = file
        .read_async_future(glib::PRIORITY_DEFAULT)
        .map_err(|err| format!("Failed to open file: {}", err))
        .await?;

    // If opening the file succeeds, we asynchronously loop and
    // read the file in up to 64 byte chunks and re-use the same
    // vec for each read.
    let mut buf = vec![0; 64];
    let mut idx = 0;

    loop {
        let (b, len) = strm
            .read_async_future(buf, glib::PRIORITY_DEFAULT)
            .map_err(|(_buf, err)| format!("Failed to read from stream: {}", err))
            .await?;

        // Once 0 is returned, we know that we're done with reading, otherwise
        // loop again and read another chunk.
        if len == 0 {
            break;
        }

        buf = b;

        println!("line {}: {:?}", idx, str::from_utf8(&buf[0..len]).unwrap());

        idx += 1;
    }

    // Asynchronously close the stream in the end.
    let _ = strm
        .close_async_future(glib::PRIORITY_DEFAULT)
        .map_err(|err| format!("Failed to close stream: {}", err))
        .await?;

    Ok(())
}

examples/gio_futures/main.rs
≺ ≻↕

  1
  2
  3
  4
  5
  6
  7
  8
  9
 10
 11
 12
 13
 14
 15
 16
 17
 18
 19
 20
 21
 22
 23
 24
 25
 26
 27
 28
 29
 30
 31
 32
 33
 34
 35
 36
 37
 38
 39
 40
 41
 42
 43
 44
 45
 46
 47
 48
 49
 50
 51
 52
 53
 54
 55
 56
 57
 58
 59
 60
 61
 62
 63
 64
 65
 66
 67
 68
 69
 70
 71
 72
 73
 74
 75
 76
 77
 78
 79
 80
 81
 82
 83
 84
 85
 86
 87
 88
 89
 90
 91
 92
 93
 94
 95
 96
 97
 98
 99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126

use gtk::prelude::*;
use gtk::{gio, glib};

use std::str;

use futures::prelude::*;

fn main() {
    let c = glib::MainContext::default();
    let l = glib::MainLoop::new(Some(&c), false);

    c.push_thread_default();

    let file = gio::File::for_path("Cargo.toml");

    let l_clone = l.clone();
    c.spawn_local(
        read_and_print_file(&file)
            // Once all is done we quit the main loop and in case of an
            // error first print that error.
            .map(move |res| {
                if let Err(err) = res {
                    eprintln!("Got error: {}", err);
                }

                l_clone.quit();
            }),
    );

    l.run();

    c.pop_thread_default();
}

/// Throughout our chained futures, we convert all errors to strings
/// via map_err() and print them at the very end.
///
/// Open the file for reading, and if that succeeds read the whole file from
/// the resulting input stream.
fn read_and_print_file(
    file: &gio::File,
) -> impl Future<Output = Result<(), String>> + std::marker::Unpin {
    file.read_async_future(glib::PRIORITY_DEFAULT)
        .map_err(|err| format!("Failed to open file: {}", err))
        .and_then(read_and_print_chunks)
}

/// Read the input stream in chunks of 64 bytes, always into the same buffer
/// without re-allocating it all the time. Continue until the end of the file
/// or an error happens.
fn read_and_print_chunks(
    strm: gio::FileInputStream,
) -> impl Future<Output = Result<(), String>> + std::marker::Unpin {
    let buf = vec![0; 64];
    let idx = 0;

    // We use unfold() here, which takes some initialization data and a
    // closure that is returning an item and the next state, or None to
    // finish the stream
    futures::stream::unfold(Some((buf, idx)), move |buf_and_idx| {
        // If None was returned from the last iteration then the last iteration
        // was closing the input stream or an error happened, and now we only
        // have to finish the stream created by unfold().
        //
        // Otherwise we got the buffer to read to and the index of the next line
        // from the previous iteration.
        let (buf, idx) = match buf_and_idx {
            None => {
                return futures::future::Either::Left(futures::future::ready(None));
            }
            Some(buf_and_idx) => buf_and_idx,
        };

        // Read and print the next chunk
        futures::future::Either::Right(read_and_print_next_chunk(&strm, buf, idx).map(move |res| {
            match res {
                // And error happened, return the error from this stream and then finish on the
                // next iteration.
                Err(err) => Some((Err(err), None)),
                // The input stream was closed, return Ok(()) from this stream and then finish on
                // the next iteration.
                Ok(None) => Some((Ok(()), None)),
                // A chunk was successfully read and printed, return Ok(()) from this stream and
                // then continue with the next iteration.
                Ok(Some(buf)) => Some((Ok(()), Some((buf, idx + 1)))),
            }
        }))
    })
    // Convert the stream into a simple future that collects all items and
    // returns Ok(()), or short-circuits on the very first error and returns it
    .try_for_each(|_| futures::future::ok(()))
}

/// Read the next chunk into the buffer and print it out, or return an error. If
/// the input stream is finished, close the stream.
///
/// After reading successfully we return the buffer again so it can be used in the
/// next iteration.
fn read_and_print_next_chunk(
    strm: &gio::FileInputStream,
    buf: Vec<u8>,
    idx: usize,
) -> impl Future<Output = Result<Option<Vec<u8>>, String>> + std::marker::Unpin {
    let strm_clone = strm.clone();
    strm.read_async_future(buf, glib::PRIORITY_DEFAULT)
        .map_err(|(_buf, err)| format!("Failed to read from stream: {}", err))
        .and_then(move |(buf, len)| {
            println!("line {}: {:?}", idx, str::from_utf8(&buf[0..len]).unwrap());

            // 0 is only returned when the input stream is finished, in which case
            // we drop the buffer and close the stream asynchronously.
            //
            // Otherwise we simply return the buffer again so it can be read into
            // in the next iteration.
            if len == 0 {
                futures::future::Either::Left(
                    strm_clone
                        .close_async_future(glib::PRIORITY_DEFAULT)
                        .map_err(|err| format!("Failed to close stream: {}", err))
                        .map_ok(|_| None),
                )
            } else {
                futures::future::Either::Right(futures::future::ok(Some(buf)))
            }
        })
}

examples/gio_task/main.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99

mod file_size;

use futures_channel::oneshot;
use glib::clone;
use gtk::glib::translate::*;
use gtk::{gio, glib};

use file_size::FileSize;

fn main() {
    let main_context = glib::MainContext::default();
    let main_loop = glib::MainLoop::new(Some(&main_context), false);
    let (send_safe, recv_safe) = oneshot::channel();
    let (send_unsafe, recv_unsafe) = oneshot::channel();

    main_context.push_thread_default();
    main_context.invoke_local(move || {
        run_unsafe(send_unsafe);
    });
    main_context.invoke_local(move || {
        run_safe(send_safe);
    });

    main_context.spawn_local(clone!(@strong main_loop => async move {
        recv_safe.await.unwrap();
        recv_unsafe.await.unwrap();
        main_loop.quit();
    }));
    main_loop.run();
    main_context.pop_thread_default();
}

// This function mimicks what the C code using the exported async/finish API would do. It first defines a
// callback of type GAsyncResult that internally calls get_file_size_finish to retrieve the Task
// result, and simply prints it out. Then it invokes the my_file_size_get_file_size_async method, passing the
// callback to it as parameter.
fn run_unsafe(send: oneshot::Sender<()>) {
    let simple_object = FileSize::new();
    let cancellable = gio::Cancellable::new();

    // The callback to be passed to my_file_size_get_file_size_async
    unsafe extern "C" fn c_callback(
        source_object: *mut glib::object::GObject,
        result: *mut gio::ffi::GAsyncResult,
        user_data: glib::ffi::gpointer,
    ) {
        let mut error = std::ptr::null_mut();
        let ret = file_size::ffi::my_file_size_get_file_size_finish(
            source_object as *mut file_size::ffi::FileSize,
            result,
            &mut error,
        );
        if !error.is_null() {
            eprintln!("Task returned error!");
            return;
        }

        println!("Unsafe callback - Returned value from task: {}", ret);
        println!(
            "Unsafe callback - FileSize::size: {}",
            file_size::ffi::my_file_size_get_retrieved_size(
                source_object as *mut file_size::ffi::FileSize
            )
        );

        Box::from_raw(user_data as *mut oneshot::Sender<()>)
            .send(())
            .unwrap();
    }

    // The actual call to my_file_size_get_file_size_async
    unsafe {
        file_size::ffi::my_file_size_get_file_size_async(
            simple_object.to_glib_none().0,
            cancellable.to_glib_none().0,
            Some(c_callback),
            Box::into_raw(Box::new(send)) as glib::ffi::gpointer,
        );
    }
}

// This function is the "safe" counterpart of "run_unsafe", using only safe rust
// bindings to accomplish the same task
fn run_safe(send: oneshot::Sender<()>) {
    let simple_object = FileSize::new();
    let cancellable = gio::Cancellable::new();

    let closure = move |value: i64, source_object: &FileSize| {
        println!("Safe callback - Returned value from task: {}", value);
        println!(
            "Safe callback - FileSize::size: {}",
            source_object.retrieved_size().unwrap()
        );

        send.send(()).unwrap();
    };

    simple_object.file_size_async(Some(&cancellable), closure);
}

pub fn run(&self)

Uses found in examples/

examples/gio_futures_await/main.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73

use glib::clone;
use gtk::prelude::*;
use gtk::{gio, glib};

use futures::prelude::*;

use std::str;

fn main() {
    let c = glib::MainContext::default();
    let l = glib::MainLoop::new(Some(&c), false);

    c.push_thread_default();

    let file = gio::File::for_path("Cargo.toml");

    let future = clone!(@strong l => async move {
        match read_file(file).await {
            Ok(()) => (),
            Err(err) => eprintln!("Got error: {}", err),
        }
        l.quit();
    });

    c.spawn_local(future);

    l.run();

    c.pop_thread_default();
}

/// Throughout our chained futures, we convert all errors to strings
/// via map_err() return them directly.
async fn read_file(file: gio::File) -> Result<(), String> {
    // Try to open the file.
    let strm = file
        .read_async_future(glib::PRIORITY_DEFAULT)
        .map_err(|err| format!("Failed to open file: {}", err))
        .await?;

    // If opening the file succeeds, we asynchronously loop and
    // read the file in up to 64 byte chunks and re-use the same
    // vec for each read.
    let mut buf = vec![0; 64];
    let mut idx = 0;

    loop {
        let (b, len) = strm
            .read_async_future(buf, glib::PRIORITY_DEFAULT)
            .map_err(|(_buf, err)| format!("Failed to read from stream: {}", err))
            .await?;

        // Once 0 is returned, we know that we're done with reading, otherwise
        // loop again and read another chunk.
        if len == 0 {
            break;
        }

        buf = b;

        println!("line {}: {:?}", idx, str::from_utf8(&buf[0..len]).unwrap());

        idx += 1;
    }

    // Asynchronously close the stream in the end.
    let _ = strm
        .close_async_future(glib::PRIORITY_DEFAULT)
        .map_err(|err| format!("Failed to close stream: {}", err))
        .await?;

    Ok(())
}

examples/gio_futures/main.rs
≺ ≻↕

  1
  2
  3
  4
  5
  6
  7
  8
  9
 10
 11
 12
 13
 14
 15
 16
 17
 18
 19
 20
 21
 22
 23
 24
 25
 26
 27
 28
 29
 30
 31
 32
 33
 34
 35
 36
 37
 38
 39
 40
 41
 42
 43
 44
 45
 46
 47
 48
 49
 50
 51
 52
 53
 54
 55
 56
 57
 58
 59
 60
 61
 62
 63
 64
 65
 66
 67
 68
 69
 70
 71
 72
 73
 74
 75
 76
 77
 78
 79
 80
 81
 82
 83
 84
 85
 86
 87
 88
 89
 90
 91
 92
 93
 94
 95
 96
 97
 98
 99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126

use gtk::prelude::*;
use gtk::{gio, glib};

use std::str;

use futures::prelude::*;

fn main() {
    let c = glib::MainContext::default();
    let l = glib::MainLoop::new(Some(&c), false);

    c.push_thread_default();

    let file = gio::File::for_path("Cargo.toml");

    let l_clone = l.clone();
    c.spawn_local(
        read_and_print_file(&file)
            // Once all is done we quit the main loop and in case of an
            // error first print that error.
            .map(move |res| {
                if let Err(err) = res {
                    eprintln!("Got error: {}", err);
                }

                l_clone.quit();
            }),
    );

    l.run();

    c.pop_thread_default();
}

/// Throughout our chained futures, we convert all errors to strings
/// via map_err() and print them at the very end.
///
/// Open the file for reading, and if that succeeds read the whole file from
/// the resulting input stream.
fn read_and_print_file(
    file: &gio::File,
) -> impl Future<Output = Result<(), String>> + std::marker::Unpin {
    file.read_async_future(glib::PRIORITY_DEFAULT)
        .map_err(|err| format!("Failed to open file: {}", err))
        .and_then(read_and_print_chunks)
}

/// Read the input stream in chunks of 64 bytes, always into the same buffer
/// without re-allocating it all the time. Continue until the end of the file
/// or an error happens.
fn read_and_print_chunks(
    strm: gio::FileInputStream,
) -> impl Future<Output = Result<(), String>> + std::marker::Unpin {
    let buf = vec![0; 64];
    let idx = 0;

    // We use unfold() here, which takes some initialization data and a
    // closure that is returning an item and the next state, or None to
    // finish the stream
    futures::stream::unfold(Some((buf, idx)), move |buf_and_idx| {
        // If None was returned from the last iteration then the last iteration
        // was closing the input stream or an error happened, and now we only
        // have to finish the stream created by unfold().
        //
        // Otherwise we got the buffer to read to and the index of the next line
        // from the previous iteration.
        let (buf, idx) = match buf_and_idx {
            None => {
                return futures::future::Either::Left(futures::future::ready(None));
            }
            Some(buf_and_idx) => buf_and_idx,
        };

        // Read and print the next chunk
        futures::future::Either::Right(read_and_print_next_chunk(&strm, buf, idx).map(move |res| {
            match res {
                // And error happened, return the error from this stream and then finish on the
                // next iteration.
                Err(err) => Some((Err(err), None)),
                // The input stream was closed, return Ok(()) from this stream and then finish on
                // the next iteration.
                Ok(None) => Some((Ok(()), None)),
                // A chunk was successfully read and printed, return Ok(()) from this stream and
                // then continue with the next iteration.
                Ok(Some(buf)) => Some((Ok(()), Some((buf, idx + 1)))),
            }
        }))
    })
    // Convert the stream into a simple future that collects all items and
    // returns Ok(()), or short-circuits on the very first error and returns it
    .try_for_each(|_| futures::future::ok(()))
}

/// Read the next chunk into the buffer and print it out, or return an error. If
/// the input stream is finished, close the stream.
///
/// After reading successfully we return the buffer again so it can be used in the
/// next iteration.
fn read_and_print_next_chunk(
    strm: &gio::FileInputStream,
    buf: Vec<u8>,
    idx: usize,
) -> impl Future<Output = Result<Option<Vec<u8>>, String>> + std::marker::Unpin {
    let strm_clone = strm.clone();
    strm.read_async_future(buf, glib::PRIORITY_DEFAULT)
        .map_err(|(_buf, err)| format!("Failed to read from stream: {}", err))
        .and_then(move |(buf, len)| {
            println!("line {}: {:?}", idx, str::from_utf8(&buf[0..len]).unwrap());

            // 0 is only returned when the input stream is finished, in which case
            // we drop the buffer and close the stream asynchronously.
            //
            // Otherwise we simply return the buffer again so it can be read into
            // in the next iteration.
            if len == 0 {
                futures::future::Either::Left(
                    strm_clone
                        .close_async_future(glib::PRIORITY_DEFAULT)
                        .map_err(|err| format!("Failed to close stream: {}", err))
                        .map_ok(|_| None),
                )
            } else {
                futures::future::Either::Right(futures::future::ok(Some(buf)))
            }
        })
}

examples/gio_task/main.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99

mod file_size;

use futures_channel::oneshot;
use glib::clone;
use gtk::glib::translate::*;
use gtk::{gio, glib};

use file_size::FileSize;

fn main() {
    let main_context = glib::MainContext::default();
    let main_loop = glib::MainLoop::new(Some(&main_context), false);
    let (send_safe, recv_safe) = oneshot::channel();
    let (send_unsafe, recv_unsafe) = oneshot::channel();

    main_context.push_thread_default();
    main_context.invoke_local(move || {
        run_unsafe(send_unsafe);
    });
    main_context.invoke_local(move || {
        run_safe(send_safe);
    });

    main_context.spawn_local(clone!(@strong main_loop => async move {
        recv_safe.await.unwrap();
        recv_unsafe.await.unwrap();
        main_loop.quit();
    }));
    main_loop.run();
    main_context.pop_thread_default();
}

// This function mimicks what the C code using the exported async/finish API would do. It first defines a
// callback of type GAsyncResult that internally calls get_file_size_finish to retrieve the Task
// result, and simply prints it out. Then it invokes the my_file_size_get_file_size_async method, passing the
// callback to it as parameter.
fn run_unsafe(send: oneshot::Sender<()>) {
    let simple_object = FileSize::new();
    let cancellable = gio::Cancellable::new();

    // The callback to be passed to my_file_size_get_file_size_async
    unsafe extern "C" fn c_callback(
        source_object: *mut glib::object::GObject,
        result: *mut gio::ffi::GAsyncResult,
        user_data: glib::ffi::gpointer,
    ) {
        let mut error = std::ptr::null_mut();
        let ret = file_size::ffi::my_file_size_get_file_size_finish(
            source_object as *mut file_size::ffi::FileSize,
            result,
            &mut error,
        );
        if !error.is_null() {
            eprintln!("Task returned error!");
            return;
        }

        println!("Unsafe callback - Returned value from task: {}", ret);
        println!(
            "Unsafe callback - FileSize::size: {}",
            file_size::ffi::my_file_size_get_retrieved_size(
                source_object as *mut file_size::ffi::FileSize
            )
        );

        Box::from_raw(user_data as *mut oneshot::Sender<()>)
            .send(())
            .unwrap();
    }

    // The actual call to my_file_size_get_file_size_async
    unsafe {
        file_size::ffi::my_file_size_get_file_size_async(
            simple_object.to_glib_none().0,
            cancellable.to_glib_none().0,
            Some(c_callback),
            Box::into_raw(Box::new(send)) as glib::ffi::gpointer,
        );
    }
}

// This function is the "safe" counterpart of "run_unsafe", using only safe rust
// bindings to accomplish the same task
fn run_safe(send: oneshot::Sender<()>) {
    let simple_object = FileSize::new();
    let cancellable = gio::Cancellable::new();

    let closure = move |value: i64, source_object: &FileSize| {
        println!("Safe callback - Returned value from task: {}", value);
        println!(
            "Safe callback - FileSize::size: {}",
            source_object.retrieved_size().unwrap()
        );

        send.send(()).unwrap();
    };

    simple_object.file_size_async(Some(&cancellable), closure);
}