Struct glib::MarkupParseContext

pub struct MarkupParseContext { /* private fields */ }

A parse context is used to parse a stream of bytes that you expect to contain marked-up text.

See g_markup_parse_context_new(), #GMarkupParser, and so on for more details.

GLib type: Shared boxed type with reference counted clone semantics.

Implementations

impl MarkupParseContext

pub fn as_ptr(&self) -> *mut GMarkupParseContext

Return the inner pointer to the underlying C value.

pub unsafe fn from_glib_ptr_borrow(ptr: &*mut GMarkupParseContext) -> &Self

Borrows the underlying C value.

impl MarkupParseContext

pub fn end_parse(&self) -> Result<(), Error>

Signals to the #GMarkupParseContext that all data has been fed into the parse context with g_markup_parse_context_parse().

This function reports an error if the document isn’t complete, for example if elements are still open.

Returns

true on success, false if an error was set

pub fn element(&self) -> GString

Retrieves the name of the currently open element.

If called from the start_element or end_element handlers this will give the element_name as passed to those functions. For the parent elements, see g_markup_parse_context_get_element_stack().

Returns

the name of the currently open element, or None

pub fn element_stack(&self) -> Vec<GString>

Retrieves the element stack from the internal state of the parser.

The returned #GSList is a list of strings where the first item is the currently open tag (as would be returned by g_markup_parse_context_get_element()) and the next item is its immediate parent.

This function is intended to be used in the start_element and end_element handlers where g_markup_parse_context_get_element() would merely return the name of the element that is being processed.

Returns

the element stack, which must not be modified

pub fn position(&self) -> (i32, i32)

Retrieves the current line number and the number of the character on that line. Intended for use in error messages; there are no strict semantics for what constitutes the “current” line number other than “the best number we could come up with for error messages.”

Returns

line_number

return location for a line number, or None

char_number

return location for a char-on-line number, or None

pub fn parse(&self, text: &str) -> Result<(), Error>

Feed some data to the #GMarkupParseContext.

The data need not be valid UTF-8; an error will be signaled if it’s invalid. The data need not be an entire document; you can feed a document into the parser incrementally, via multiple calls to this function. Typically, as you receive data from a network connection or file, you feed each received chunk of data into this function, aborting the process if an error occurs. Once an error is reported, no further data may be fed to the #GMarkupParseContext; all errors are fatal.

text

chunk of text to parse

text_len

length of @text in bytes

Returns

false if an error occurred, true on success