Struct glib::MatchInfo

pub struct MatchInfo { /* private fields */ }

A GMatchInfo is an opaque struct used to return information about matches.

Implementations

impl MatchInfo

pub fn as_ptr(&self) -> *mut GMatchInfo

Return the inner pointer to the underlying C value.

pub unsafe fn from_glib_ptr_borrow<'a>( ptr: *const *const GMatchInfo ) -> &'a Self

Borrows the underlying C value.

impl MatchInfo

pub fn fetch(&self, match_num: i32) -> Option<GString>

Retrieves the text matching the match_num’th capturing parentheses. 0 is the full text of the match, 1 is the first paren set, 2 the second, and so on.

If match_num is a valid sub pattern but it didn’t match anything (e.g. sub pattern 1, matching “b” against “(a)?b”) then an empty string is returned.

If the match was obtained using the DFA algorithm, that is using Regex::match_all() or Regex::match_all_full(), the retrieved string is not that of a set of parentheses but that of a matched substring. Substrings are matched in reverse order of length, so 0 is the longest match.

The string is fetched from the string passed to the match function, so you cannot call this function after freeing the string.

match_num

number of the sub expression

Returns

The matched substring, or None if an error occurred. You have to free the string yourself

pub fn fetch_all(&self) -> Vec<GString>

Bundles up pointers to each of the matching substrings from a match and stores them in an array of gchar pointers. The first element in the returned array is the match number 0, i.e. the entire matched text.

If a sub pattern didn’t match anything (e.g. sub pattern 1, matching “b” against “(a)?b”) then an empty string is inserted.

If the last match was obtained using the DFA algorithm, that is using Regex::match_all() or Regex::match_all_full(), the retrieved strings are not that matched by sets of parentheses but that of the matched substring. Substrings are matched in reverse order of length, so the first one is the longest match.

The strings are fetched from the string passed to the match function, so you cannot call this function after freeing the string.

Returns

a None-terminated array of gchar * pointers. It must be freed using g_strfreev(). If the previous match failed None is returned

pub fn fetch_pos(&self, match_num: i32) -> Option<(i32, i32)>

Retrieves the position in bytes of the match_num’th capturing parentheses. 0 is the full text of the match, 1 is the first paren set, 2 the second, and so on.

If match_num is a valid sub pattern but it didn’t match anything (e.g. sub pattern 1, matching “b” against “(a)?b”) then start_pos and end_pos are set to -1 and true is returned.

If the match was obtained using the DFA algorithm, that is using Regex::match_all() or Regex::match_all_full(), the retrieved position is not that of a set of parentheses but that of a matched substring. Substrings are matched in reverse order of length, so 0 is the longest match.

match_num

number of the sub expression

Returns

true if the position was fetched, false otherwise. If the position cannot be fetched, start_pos and end_pos are left unchanged

start_pos

pointer to location where to store the start position, or None

end_pos

pointer to location where to store the end position, or None

pub fn match_count(&self) -> i32

Retrieves the number of matched substrings (including substring 0, that is the whole matched text), so 1 is returned if the pattern has no substrings in it and 0 is returned if the match failed.

If the last match was obtained using the DFA algorithm, that is using Regex::match_all() or Regex::match_all_full(), the retrieved count is not that of the number of capturing parentheses but that of the number of matched substrings.

Returns

Number of matched substrings, or -1 if an error occurred

pub fn regex(&self) -> Regex

Returns Regex object used in self. It belongs to Glib and must not be freed. Use g_regex_ref() if you need to keep it after you free self object.

Returns

Regex object used in self

pub fn string(&self) -> GString

Returns the string searched with self. This is the string passed to Regex::match_() or Regex::replace() so you may not free it before calling this function.

Returns

the string searched with self

pub fn is_partial_match(&self) -> bool

Usually if the string passed to g_regex_match*() matches as far as it goes, but is too short to match the entire pattern, false is returned. There are circumstances where it might be helpful to distinguish this case from other cases in which there is no match.

Consider, for example, an application where a human is required to type in data for a field with specific formatting requirements. An example might be a date in the form ddmmmyy, defined by the pattern “^\d?\d(jan|feb|mar|apr|may|jun|jul|aug|sep|oct|nov|dec)\d\d$”. If the application sees the user’s keystrokes one by one, and can check that what has been typed so far is potentially valid, it is able to raise an error as soon as a mistake is made.

GRegex supports the concept of partial matching by means of the RegexMatchFlags::PARTIAL_SOFT and RegexMatchFlags::PARTIAL_HARD flags. When they are used, the return code for Regex::match_() or Regex::match_full() is, as usual, true for a complete match, false otherwise. But, when these functions return false, you can check if the match was partial calling is_partial_match().

The difference between RegexMatchFlags::PARTIAL_SOFT and RegexMatchFlags::PARTIAL_HARD is that when a partial match is encountered with RegexMatchFlags::PARTIAL_SOFT, matching continues to search for a possible complete match, while with RegexMatchFlags::PARTIAL_HARD matching stops at the partial match. When both RegexMatchFlags::PARTIAL_SOFT and RegexMatchFlags::PARTIAL_HARD are set, the latter takes precedence.

There were formerly some restrictions on the pattern for partial matching. The restrictions no longer apply.

See pcrepartial(3) for more information on partial matching.

Returns

true if the match was partial, false otherwise

pub fn matches(&self) -> bool

Returns whether the previous match operation succeeded.

Returns

true if the previous match operation succeeded, false otherwise

pub fn next(&self) -> Result<(), Error>

Scans for the next match using the same parameters of the previous call to Regex::match_full() or Regex::match_() that returned self.

The match is done on the string passed to the match function, so you cannot free it before calling this function.

Returns

true is the string matched, false otherwise

impl MatchInfo

pub fn expand_references( &self, string_to_expand: impl IntoGStr ) -> Result<Option<GString>, Error>

Returns a new string containing the text in string_to_expand with references and escape sequences expanded. References refer to the last match done with string against regex and have the same syntax used by Regex::replace().

The string_to_expand must be UTF-8 encoded even if RegexCompileFlags::RAW was passed to Regex::new().

The backreferences are extracted from the string passed to the match function, so you cannot call this function after freeing the string.

self may be None in which case string_to_expand must not contain references. For instance “foo\n” does not refer to an actual pattern and ‘\n’ merely will be replaced with \n character, while to expand “\0” (whole match) one needs the result of a match. Use Regex::check_replacement() to find out whether string_to_expand contains references.

string_to_expand

the string to expand

Returns

the expanded string, or None if an error occurred

pub fn fetch_named(&self, name: impl IntoGStr) -> Option<GString>

Retrieves the text matching the capturing parentheses named name.

If name is a valid sub pattern name but it didn’t match anything (e.g. sub pattern “X”, matching “b” against “(?P<X>a)?b”) then an empty string is returned.

The string is fetched from the string passed to the match function, so you cannot call this function after freeing the string.

name

name of the subexpression

Returns

The matched substring, or None if an error occurred. You have to free the string yourself

pub fn fetch_named_pos(&self, name: impl IntoGStr) -> Option<(i32, i32)>

Retrieves the position in bytes of the capturing parentheses named name.

If name is a valid sub pattern name but it didn’t match anything (e.g. sub pattern “X”, matching “b” against “(?P<X>a)?b”) then start_pos and end_pos are set to -1 and true is returned.

name

name of the subexpression

Returns

true if the position was fetched, false otherwise. If the position cannot be fetched, start_pos and end_pos are left unchanged.

start_pos

pointer to location where to store the start position, or None

end_pos

pointer to location where to store the end position, or None