Struct glib::Permission

pub struct Permission {
    // some fields omitted
}

Methods

impl Permission

fn get_allowed(&self) -> bool

Gets the value of the 'allowed' property. This property is true if the caller currently has permission to perform the action that permission represents the permission to perform.

fn get_can_acquire(&self) -> bool

Gets the value of the 'can-acquire' property. This property is true if it is generally possible to acquire the permission by calling g_permission_acquire().

fn get_can_release(&self) -> bool

Gets the value of the 'can-release' property. This property is true if it is generally possible to release the permission by calling g_permission_release().

fn impl_update(&self, allowed: bool, can_acquire: bool, can_release: bool)

This function is called by the GPermission implementation to update the properties of the permission. You should never call this function except from a GPermission implementation.

GObject notify signals are generated, as appropriate.