glib

Struct Regex

pub struct Regex { /* private fields */ }

A GRegex is a compiled form of a regular expression.

After instantiating a GRegex, you can use its methods to find matches in a string, replace matches within a string, or split the string at matches.

GRegex implements regular expression pattern matching using syntax and semantics (such as character classes, quantifiers, and capture groups) similar to Perl regular expression. See the PCRE documentation for details.

A typical scenario for regex pattern matching is to check if a string matches a pattern. The following statements implement this scenario.

⚠️ The following code is in { .c } ⚠️

const char *regex_pattern = ".*GLib.*";
const char *string_to_search = "You will love the GLib implementation of regex";
g_autoptr(GMatchInfo) match_info = NULL;
g_autoptr(GRegex) regex = NULL;

regex = g_regex_new (regex_pattern, G_REGEX_DEFAULT, G_REGEX_MATCH_DEFAULT, NULL);
g_assert (regex != NULL);

if (g_regex_match (regex, string_to_search, G_REGEX_MATCH_DEFAULT, &match_info))
  {
    int start_pos, end_pos;
    g_match_info_fetch_pos (match_info, 0, &start_pos, &end_pos);
    g_print ("Match successful! Overall pattern matches bytes %d to %d\n", start_pos, end_pos);
  }
else
  {
    g_print ("No match!\n");
  }

The constructor for GRegex includes two sets of bitmapped flags:

RegexCompileFlags—These flags control how GLib compiles the regex. There are options for case sensitivity, multiline, ignoring whitespace, etc.

RegexMatchFlags—These flags control GRegex’s matching behavior, such as anchoring and customizing definitions for newline characters.

Some regex patterns include backslash assertions, such as \d (digit) or \D (non-digit). The regex pattern must escape those backslashes. For example, the pattern "\\d\\D" matches a digit followed by a non-digit.

GLib’s implementation of pattern matching includes a start_position argument for some of the match, replace, and split methods. Specifying a start position provides flexibility when you want to ignore the first n characters of a string, but want to incorporate backslash assertions at character n - 1. For example, a database field contains inconsistent spelling for a job title: healthcare provider and health-care provider. The database manager wants to make the spelling consistent by adding a hyphen when it is missing. The following regex pattern tests for the string care preceded by a non-word boundary character (instead of a hyphen) and followed by a space.

⚠️ The following code is in { .c } ⚠️

const char *regex_pattern = "\\Bcare\\s";

An efficient way to match with this pattern is to start examining at start_position 6 in the string healthcare or health-care.

⚠️ The following code is in { .c } ⚠️

const char *regex_pattern = "\\Bcare\\s";
const char *string_to_search = "healthcare provider";
g_autoptr(GMatchInfo) match_info = NULL;
g_autoptr(GRegex) regex = NULL;

regex = g_regex_new (
  regex_pattern,
  G_REGEX_DEFAULT,
  G_REGEX_MATCH_DEFAULT,
  NULL);
g_assert (regex != NULL);

g_regex_match_full (
  regex,
  string_to_search,
  -1,
  6, // position of 'c' in the test string.
  G_REGEX_MATCH_DEFAULT,
  &match_info,
  NULL);

The method match_full() (and other methods implementing start_pos) allow for lookback before the start position to determine if the previous character satisfies an assertion.

Unless you set the [flags@GLib.RegexCompileFlags.RAW] as one of the GRegexCompileFlags, all the strings passed to GRegex methods must be encoded in UTF-8. The lengths and the positions inside the strings are in bytes and not in characters, so, for instance, \xc3\xa0 (i.e., à) is two bytes long but it is treated as a single character. If you set G_REGEX_RAW, the strings can be non-valid UTF-8 strings and a byte is treated as a character, so \xc3\xa0 is two bytes and two characters long.

Regarding line endings, \n matches a \n character, and \r matches a \r character. More generally, \R matches all typical line endings: CR + LF (\r\n), LF (linefeed, U+000A, \n), VT (vertical tab, U+000B, \v), FF (formfeed, U+000C, \f), CR (carriage return, U+000D, \r), NEL (next line, U+0085), LS (line separator, U+2028), and PS (paragraph separator, U+2029).

The behaviour of the dot, circumflex, and dollar metacharacters are affected by newline characters. By default, GRegex matches any newline character matched by \R. You can limit the matched newline characters by specifying the [flags@GLib.RegexMatchFlags.NEWLINE_CR], [flags@GLib.RegexMatchFlags.NEWLINE_LF], and [flags@GLib.RegexMatchFlags.NEWLINE_CRLF] compile options, and with [flags@GLib.RegexMatchFlags.NEWLINE_ANY], [flags@GLib.RegexMatchFlags.NEWLINE_CR], [flags@GLib.RegexMatchFlags.NEWLINE_LF] and [flags@GLib.RegexMatchFlags.NEWLINE_CRLF] match options. These settings are also relevant when compiling a pattern if [flags@GLib.RegexCompileFlags.EXTENDED] is set and an unescaped # outside a character class is encountered. This indicates a comment that lasts until after the next newline.

Because GRegex does not modify its internal state between creation and destruction, you can create and modify the same GRegex instance from different threads. In contrast, MatchInfo is not thread safe.

The regular expression low-level functionalities are obtained through the excellent PCRE library written by Philip Hazel.

GLib type: Shared boxed type with reference counted clone semantics.

Implementations

impl Regex

pub fn as_ptr(&self) -> *mut GRegex

Return the inner pointer to the underlying C value.

pub unsafe fn from_glib_ptr_borrow(ptr: &*mut GRegex) -> &Self

Borrows the underlying C value.

impl Regex

pub fn new( pattern: &str, compile_options: RegexCompileFlags, match_options: RegexMatchFlags, ) -> Result<Option<Regex>, Error>

Compiles the regular expression to an internal form, and does the initial setup of the #GRegex structure.

pattern

the regular expression

compile_options

compile options for the regular expression, or 0

match_options

match options for the regular expression, or 0

Returns

a #GRegex structure or None if an error occurred. Call g_regex_unref() when you are done with it

pub fn capture_count(&self) -> i32

Returns the number of capturing subpatterns in the pattern.

Returns

the number of capturing subpatterns

pub fn compile_flags(&self) -> RegexCompileFlags

Returns the compile options that @self was created with.

Depending on the version of PCRE that is used, this may or may not include flags set by option expressions such as (?i) found at the top-level within the compiled pattern.

Returns

flags from #GRegexCompileFlags

pub fn has_cr_or_lf(&self) -> bool

Checks whether the pattern contains explicit CR or LF references.

Returns

true if the pattern contains explicit CR or LF references

pub fn match_flags(&self) -> RegexMatchFlags

Returns the match options that @self was created with.

Returns

flags from #GRegexMatchFlags

pub fn max_backref(&self) -> i32

Returns the number of the highest back reference in the pattern, or 0 if the pattern does not contain back references.

Returns

the number of the highest back reference

pub fn max_lookbehind(&self) -> i32

Gets the number of characters in the longest lookbehind assertion in the pattern. This information is useful when doing multi-segment matching using the partial matching facilities.

Returns

the number of characters in the longest lookbehind assertion.

pub fn pattern(&self) -> GString

Gets the pattern string associated with @self, i.e. a copy of the string passed to g_regex_new().

Returns

the pattern of @self

impl Regex

pub fn string_number(&self, name: impl IntoGStr) -> i32

Retrieves the number of the subexpression named @name.

name

name of the subexpression

Returns

The number of the subexpression or -1 if @name does not exists

pub fn escape_nul(string: impl IntoGStr) -> GString

Escapes the nul characters in @string to “\x00”. It can be used to compile a regex with embedded nul characters.

For completeness, @length can be -1 for a nul-terminated string. In this case the output string will be of course equal to @string.

string

the string to escape

length

the length of @string

Returns

a newly-allocated escaped string

pub fn escape_string(string: impl IntoGStr) -> GString

Escapes the special characters used for regular expressions in @string, for instance “a.b*c” becomes “a.b*c”. This function is useful to dynamically generate regular expressions.

@string can contain nul characters that are replaced with “\0”, in this case remember to specify the correct length of @string in @length.

string

the string to escape

length

the length of @string, in bytes, or -1 if @string is nul-terminated

Returns

a newly-allocated escaped string

pub fn check_replacement(replacement: impl IntoGStr) -> Result<bool, Error>

Checks whether @replacement is a valid replacement string (see g_regex_replace()), i.e. that all escape sequences in it are valid.

If @has_references is not None then @replacement is checked for pattern references. For instance, replacement text ‘foo\n’ does not contain references and may be evaluated without information about actual match, but ‘\0\1’ (whole match followed by first subpattern) requires valid #GMatchInfo object.

replacement

the replacement string

Returns

whether @replacement is a valid replacement string

has_references

location to store information about references in @replacement or None

pub fn match_simple( pattern: impl IntoGStr, string: impl IntoGStr, compile_options: RegexCompileFlags, match_options: RegexMatchFlags, ) -> bool

Scans for a match in @string for @pattern.

This function is equivalent to g_regex_match() but it does not require to compile the pattern with g_regex_new(), avoiding some lines of code when you need just to do a match without extracting substrings, capture counts, and so on.

If this function is to be called on the same @pattern more than once, it’s more efficient to compile the pattern once with g_regex_new() and then use g_regex_match().

pattern

the regular expression

string

the string to scan for matches

compile_options

compile options for the regular expression, or 0

match_options

match options, or 0

Returns

true if the string matched, false otherwise

pub fn replace( &self, string: impl IntoGStr, start_position: i32, replacement: impl IntoGStr, match_options: RegexMatchFlags, ) -> Result<GString, Error>

Replaces all occurrences of the pattern in @self with the replacement text. Backreferences of the form \number or \g<number> in the replacement text are interpolated by the number-th captured subexpression of the match, \g<name> refers to the captured subexpression with the given name. \0 refers to the complete match, but \0 followed by a number is the octal representation of a character. To include a literal \ in the replacement, write \\\\.

There are also escapes that changes the case of the following text:

\l: Convert to lower case the next character

\u: Convert to upper case the next character

\L: Convert to lower case till \E

\U: Convert to upper case till \E

\E: End case modification

If you do not need to use backreferences use g_regex_replace_literal().

The @replacement string must be UTF-8 encoded even if RegexCompileFlags::RAW was passed to g_regex_new(). If you want to use not UTF-8 encoded strings you can use g_regex_replace_literal().

Setting @start_position differs from just passing over a shortened string and setting RegexMatchFlags::NOTBOL in the case of a pattern that begins with any kind of lookbehind assertion, such as “\b”.

string

the string to perform matches against

start_position

starting index of the string to match, in bytes

replacement

text to replace each match with

match_options

options for the match

Returns

a newly allocated string containing the replacements

pub fn match_all<'input>( &self, string: &'input GStr, match_options: RegexMatchFlags, ) -> Result<MatchInfo<'input>, Error>

Using the standard algorithm for regular expression matching only the longest match in the string is retrieved. This function uses a different algorithm so it can retrieve all the possible matches. For more documentation see g_regex_match_all_full().

A #GMatchInfo structure, used to get information on the match, is stored in @match_info if not None. Note that if @match_info is not None then it is created even if the function returns false, i.e. you must free it regardless if regular expression actually matched.

@string is not copied and is used in #GMatchInfo internally. If you use any #GMatchInfo method (except g_match_info_free()) after freeing or modifying @string then the behaviour is undefined.

string

the string to scan for matches

match_options

match options

Returns

true is the string matched, false otherwise

match_info

pointer to location where to store the #GMatchInfo, or None if you do not need it

pub fn match_all_full<'input>( &self, string: &'input GStr, start_position: i32, match_options: RegexMatchFlags, ) -> Result<MatchInfo<'input>, Error>

Using the standard algorithm for regular expression matching only the longest match in the @string is retrieved, it is not possible to obtain all the available matches. For instance matching "<a> <b> <c>" against the pattern "<.*>" you get "<a> <b> <c>".

This function uses a different algorithm (called DFA, i.e. deterministic finite automaton), so it can retrieve all the possible matches, all starting at the same point in the string. For instance matching "<a> <b> <c>" against the pattern "<.*>" you would obtain three matches: "<a> <b> <c>", "<a> <b>" and "<a>".

The number of matched strings is retrieved using g_match_info_get_match_count(). To obtain the matched strings and their position you can use, respectively, g_match_info_fetch() and g_match_info_fetch_pos(). Note that the strings are returned in reverse order of length; that is, the longest matching string is given first.

Note that the DFA algorithm is slower than the standard one and it is not able to capture substrings, so backreferences do not work.

Setting @start_position differs from just passing over a shortened string and setting RegexMatchFlags::NOTBOL in the case of a pattern that begins with any kind of lookbehind assertion, such as “\b”.

Unless RegexCompileFlags::RAW is specified in the options, @string must be valid UTF-8.

A #GMatchInfo structure, used to get information on the match, is stored in @match_info if not None. Note that if @match_info is not None then it is created even if the function returns false, i.e. you must free it regardless if regular expression actually matched.

@string is not copied and is used in #GMatchInfo internally. If you use any #GMatchInfo method (except g_match_info_free()) after freeing or modifying @string then the behaviour is undefined.

string

the string to scan for matches

start_position

starting index of the string to match, in bytes

match_options

match options

Returns

true is the string matched, false otherwise

match_info

pointer to location where to store the #GMatchInfo, or None if you do not need it

pub fn match_<'input>( &self, string: &'input GStr, match_options: RegexMatchFlags, ) -> Result<MatchInfo<'input>, Error>

Scans for a match in @string for the pattern in @self. The @match_options are combined with the match options specified when the @self structure was created, letting you have more flexibility in reusing #GRegex structures.

Unless RegexCompileFlags::RAW is specified in the options, @string must be valid UTF-8.

A #GMatchInfo structure, used to get information on the match, is stored in @match_info if not None. Note that if @match_info is not None then it is created even if the function returns false, i.e. you must free it regardless if regular expression actually matched.

To retrieve all the non-overlapping matches of the pattern in string you can use g_match_info_next().

⚠️ The following code is in C ⚠️

static void
print_uppercase_words (const gchar *string)
{
  // Print all uppercase-only words.
  GRegex *regex;
  GMatchInfo *match_info;

  regex = g_regex_new ("[A-Z]+", G_REGEX_DEFAULT, G_REGEX_MATCH_DEFAULT, NULL);
  g_regex_match (regex, string, 0, &match_info);
  while (g_match_info_matches (match_info))
    {
      gchar *word = g_match_info_fetch (match_info, 0);
      g_print ("Found: %s\n", word);
      g_free (word);
      g_match_info_next (match_info, NULL);
    }
  g_match_info_free (match_info);
  g_regex_unref (regex);
}

@string is not copied and is used in #GMatchInfo internally. If you use any #GMatchInfo method (except g_match_info_free()) after freeing or modifying @string then the behaviour is undefined.

string

the string to scan for matches

match_options

match options

Returns

true is the string matched, false otherwise

match_info

pointer to location where to store the #GMatchInfo, or None if you do not need it

pub fn match_full<'input>( &self, string: &'input GStr, start_position: i32, match_options: RegexMatchFlags, ) -> Result<MatchInfo<'input>, Error>

Scans for a match in @string for the pattern in @self. The @match_options are combined with the match options specified when the @self structure was created, letting you have more flexibility in reusing #GRegex structures.

Setting @start_position differs from just passing over a shortened string and setting RegexMatchFlags::NOTBOL in the case of a pattern that begins with any kind of lookbehind assertion, such as “\b”.

Unless RegexCompileFlags::RAW is specified in the options, @string must be valid UTF-8.

A #GMatchInfo structure, used to get information on the match, is stored in @match_info if not None. Note that if @match_info is not None then it is created even if the function returns false, i.e. you must free it regardless if regular expression actually matched.

@string is not copied and is used in #GMatchInfo internally. If you use any #GMatchInfo method (except g_match_info_free()) after freeing or modifying @string then the behaviour is undefined.

To retrieve all the non-overlapping matches of the pattern in string you can use g_match_info_next().

⚠️ The following code is in C ⚠️

static void
print_uppercase_words (const gchar *string)
{
  // Print all uppercase-only words.
  GRegex *regex;
  GMatchInfo *match_info;
  GError *error = NULL;

  regex = g_regex_new ("[A-Z]+", G_REGEX_DEFAULT, G_REGEX_MATCH_DEFAULT, NULL);
  g_regex_match_full (regex, string, -1, 0, 0, &match_info, &error);
  while (g_match_info_matches (match_info))
    {
      gchar *word = g_match_info_fetch (match_info, 0);
      g_print ("Found: %s\n", word);
      g_free (word);
      g_match_info_next (match_info, &error);
    }
  g_match_info_free (match_info);
  g_regex_unref (regex);
  if (error != NULL)
    {
      g_printerr ("Error while matching: %s\n", error->message);
      g_error_free (error);
    }
}

string

the string to scan for matches

start_position

starting index of the string to match, in bytes

match_options

match options

Returns

true is the string matched, false otherwise

match_info

pointer to location where to store the #GMatchInfo, or None if you do not need it

pub fn replace_literal( &self, string: impl IntoGStr, start_position: i32, replacement: impl IntoGStr, match_options: RegexMatchFlags, ) -> Result<GString, Error>

Replaces all occurrences of the pattern in @self with the replacement text. @replacement is replaced literally, to include backreferences use g_regex_replace().

Setting @start_position differs from just passing over a shortened string and setting RegexMatchFlags::NOTBOL in the case of a pattern that begins with any kind of lookbehind assertion, such as “\b”.

string

the string to perform matches against

start_position

starting index of the string to match, in bytes

replacement

text to replace each match with

match_options

options for the match

Returns

a newly allocated string containing the replacements

pub fn split( &self, string: impl IntoGStr, match_options: RegexMatchFlags, ) -> PtrSlice<GStringPtr>

Breaks the string on the pattern, and returns an array of the tokens. If the pattern contains capturing parentheses, then the text for each of the substrings will also be returned. If the pattern does not match anywhere in the string, then the whole string is returned as the first token.

As a special case, the result of splitting the empty string “” is an empty vector, not a vector containing a single string. The reason for this special case is that being able to represent an empty vector is typically more useful than consistent handling of empty elements. If you do need to represent empty elements, you’ll need to check for the empty string before calling this function.

A pattern that can match empty strings splits @string into separate characters wherever it matches the empty string between characters. For example splitting “ab c” using as a separator “\s*”, you will get “a”, “b” and “c”.

string

the string to split with the pattern

match_options

match time option flags

Returns

a None-terminated gchar ** array. Free it using g_strfreev()

pub fn split_full( &self, string: impl IntoGStr, start_position: i32, match_options: RegexMatchFlags, max_tokens: i32, ) -> Result<PtrSlice<GStringPtr>, Error>

Breaks the string on the pattern, and returns an array of the tokens. If the pattern contains capturing parentheses, then the text for each of the substrings will also be returned. If the pattern does not match anywhere in the string, then the whole string is returned as the first token.

As a special case, the result of splitting the empty string “” is an empty vector, not a vector containing a single string. The reason for this special case is that being able to represent an empty vector is typically more useful than consistent handling of empty elements. If you do need to represent empty elements, you’ll need to check for the empty string before calling this function.

A pattern that can match empty strings splits @string into separate characters wherever it matches the empty string between characters. For example splitting “ab c” using as a separator “\s*”, you will get “a”, “b” and “c”.

Setting @start_position differs from just passing over a shortened string and setting RegexMatchFlags::NOTBOL in the case of a pattern that begins with any kind of lookbehind assertion, such as “\b”.

string

the string to split with the pattern

start_position

starting index of the string to match, in bytes

match_options

match time option flags

max_tokens

the maximum number of tokens to split @string into. If this is less than 1, the string is split completely

Returns

a None-terminated gchar ** array. Free it using g_strfreev()

pub fn split_simple( pattern: impl IntoGStr, string: impl IntoGStr, compile_options: RegexCompileFlags, match_options: RegexMatchFlags, ) -> PtrSlice<GStringPtr>

Breaks the string on the pattern, and returns an array of the tokens. If the pattern contains capturing parentheses, then the text for each of the substrings will also be returned. If the pattern does not match anywhere in the string, then the whole string is returned as the first token.

This function is equivalent to g_regex_split() but it does not require to compile the pattern with g_regex_new(), avoiding some lines of code when you need just to do a split without extracting substrings, capture counts, and so on.

If this function is to be called on the same @pattern more than once, it’s more efficient to compile the pattern once with g_regex_new() and then use g_regex_split().

As a special case, the result of splitting the empty string “” is an empty vector, not a vector containing a single string. The reason for this special case is that being able to represent an empty vector is typically more useful than consistent handling of empty elements. If you do need to represent empty elements, you’ll need to check for the empty string before calling this function.

A pattern that can match empty strings splits @string into separate characters wherever it matches the empty string between characters. For example splitting “ab c” using as a separator “\s*”, you will get “a”, “b” and “c”.

pattern

the regular expression

string

the string to scan for matches

compile_options

compile options for the regular expression, or 0

match_options

match options, or 0

Returns

a None-terminated array of strings. Free it using g_strfreev()