glib

Struct ThreadPool

pub struct ThreadPool(/* private fields */);

Implementations

impl ThreadPool

pub fn shared(max_threads: Option<u32>) -> Result<Self, Error>

pub fn exclusive(max_threads: u32) -> Result<Self, Error>

pub fn push<T: Send + 'static, F: FnOnce() -> T + Send + 'static>( &self, func: F, ) -> Result<ThreadHandle<T>, Error>

pub fn push_future<T: Send + 'static, F: FnOnce() -> T + Send + 'static>( &self, func: F, ) -> Result<impl Future<Output = Result<T>> + Send + Sync + 'static + use<T, F>, Error>

Available on crate feature futures only.

pub fn set_max_threads(&self, max_threads: Option<u32>) -> Result<(), Error>

pub fn max_threads(&self) -> Option<u32>

pub fn num_threads(&self) -> u32

pub fn unprocessed(&self) -> u32

pub fn set_max_unused_threads(max_threads: Option<u32>)

Sets the maximal number of unused threads to @max_threads. If @max_threads is -1, no limit is imposed on the number of unused threads.

The default value is 8 since GLib 2.84. Previously the default value was 2.

max_threads

maximal number of unused threads Sets the maximal number of unused threads to @max_threads. If @max_threads is -1, no limit is imposed on the number of unused threads.

The default value is 8 since GLib 2.84. Previously the default value was 2.

max_threads

maximal number of unused threads

pub fn max_unused_threads() -> Option<u32>

pub fn num_unused_threads() -> u32

pub fn stop_unused_threads()

Stops all currently unused threads. This does not change the maximal number of unused threads. This function can be used to regularly stop all unused threads e.g. from g_timeout_add(). Stops all currently unused threads. This does not change the maximal number of unused threads. This function can be used to regularly stop all unused threads e.g. from g_timeout_add().

pub fn set_max_idle_time(max_idle_time: u32)

This function will set the maximum @interval that a thread waiting in the pool for new tasks can be idle for before being stopped. This function is similar to calling g_thread_pool_stop_unused_threads() on a regular timeout, except this is done on a per thread basis.

By setting @interval to 0, idle threads will not be stopped.

The default value is 15000 (15 seconds).

interval

the maximum @interval (in milliseconds) a thread can be idle This function will set the maximum @interval that a thread waiting in the pool for new tasks can be idle for before being stopped. This function is similar to calling g_thread_pool_stop_unused_threads() on a regular timeout, except this is done on a per thread basis.

By setting @interval to 0, idle threads will not be stopped.

The default value is 15000 (15 seconds).

interval

the maximum @interval (in milliseconds) a thread can be idle

pub fn max_idle_time() -> u32