Struct glib::TimeZone

pub struct TimeZone(_);

crate::TimeZone is an opaque structure whose members cannot be accessed directly.

Implementations

impl TimeZone

pub fn new(identifier: Option<&str>) -> TimeZone

A version of g_time_zone_new_identifier() which returns the UTC time zone if identifier could not be parsed or loaded.

If you need to check whether identifier was loaded successfully, use g_time_zone_new_identifier().

Deprecated since 2.68

Use g_time_zone_new_identifier() instead, as it provides error reporting. Change your code to handle a potentially None return value.

identifier

a timezone identifier

Returns

the requested timezone

pub fn new_local() -> TimeZone

Creates a crate::TimeZone corresponding to local time. The local time zone may change between invocations to this function; for example, if the system administrator changes it.

This is equivalent to calling Self::new() with the value of the TZ environment variable (including the possibility of None).

You should release the return value by calling [Self::unref()] when you are done with it.

Returns

the local timezone

pub fn new_offset(seconds: i32) -> TimeZone

This is supported on crate feature v2_58 only.

Creates a crate::TimeZone corresponding to the given constant offset from UTC, in seconds.

This is equivalent to calling Self::new() with a string in the form [+|-]hh[:mm[:ss]].

seconds

offset to UTC, in seconds

Returns

a timezone at the given offset from UTC

pub fn new_utc() -> TimeZone

Creates a crate::TimeZone corresponding to UTC.

This is equivalent to calling Self::new() with a value like “Z”, “UTC”, “+00”, etc.

You should release the return value by calling [Self::unref()] when you are done with it.

Returns

the universal timezone

pub fn find_interval(&self, type_: TimeType, time_: i64) -> i32

Finds an interval within self that corresponds to the given time_. The meaning of time_ depends on type_.

If type_ is crate::TimeType::Universal then this function will always succeed (since universal time is monotonic and continuous).

Otherwise time_ is treated as local time. The distinction between crate::TimeType::Standard and crate::TimeType::Daylight is ignored except in the case that the given time_ is ambiguous. In Toronto, for example, 01:30 on November 7th 2010 occurred twice (once inside of daylight savings time and the next, an hour later, outside of daylight savings time). In this case, the different value of type_ would result in a different interval being returned.

It is still possible for this function to fail. In Toronto, for example, 02:00 on March 14th 2010 does not exist (due to the leap forward to begin daylight savings time). -1 is returned in that case.

type_

the crate::TimeType of time_

time_

a number of seconds since January 1, 1970

Returns

the interval containing time_, or -1 in case of failure

pub fn abbreviation(&self, interval: i32) -> GString

pub fn identifier(&self) -> GString

This is supported on crate feature v2_58 only.

pub fn offset(&self, interval: i32) -> i32

pub fn is_dst(&self, interval: i32) -> bool

Determines if daylight savings time is in effect during a particular interval of time in the time zone self.

interval

an interval within the timezone

Returns

true if daylight savings time is in effect