glib::subclass
Module object

Module that contains all types needed for creating a direct subclass of GObject or implementing virtual methods of it.

Traits

DerivedObjectProperties

Trait containing only the property related functions of ObjectImpl. Implemented by the Properties macro. When implementing ObjectImpl you may want to delegate the function calls to this trait.

ObjectClassSubclassExt

Extension trait for glib::Object’s class struct.

ObjectImpl

Trait for implementors of glib::Object subclasses.

ObjectImplExt