Trait glib::subclass::object::DerivedObjectProperties

pub trait DerivedObjectProperties: ObjectSubclass {
    // Provided methods
    fn derived_properties() -> &'static [ParamSpec] { ... }
    fn derived_set_property(
        &self,
        _id: usize,
        _value: &Value,
        _pspec: &ParamSpec,
    ) { ... }
    fn derived_property(&self, _id: usize, _pspec: &ParamSpec) -> Value { ... }
}

Trait containing only the property related functions of ObjectImpl. Implemented by the Properties macro. When implementing ObjectImpl you may want to delegate the function calls to this trait.

Provided Methods

fn derived_properties() -> &'static [ParamSpec]

Properties installed for this type.

fn derived_set_property(&self, _id: usize, _value: &Value, _pspec: &ParamSpec)

Similar to ObjectImpl but auto-generated by the Properties macro to allow handling more complex use-cases.

fn derived_property(&self, _id: usize, _pspec: &ParamSpec) -> Value

Similar to ObjectImpl but auto-generated by the Properties macro to allow handling more complex use-cases.

Object Safety

This trait is not object safe.