glib::subclass::type_module

Trait TypeModuleImpl

pub trait TypeModuleImpl: ObjectImpl + ObjectSubclass<Type: IsA<Object> + IsA<TypeModule>> {
    // Required methods
    fn load(&self) -> bool;
    fn unload(&self);
}

Required Methods

fn load(&self) -> bool

Loads the module, registers one or more object subclasses using register_dynamic_type and registers one or more object interfaces using register_dynamic_interface (see TypeModule).

fn unload(&self)

Unloads the module (see TypeModuleExt::unuse).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.