Module glib::subclass::types

Module that contains the basic infrastructure for subclassing GObject.

Macros

interface_list_trait 🔒

interface_list_trait_impl 🔒

Structs

InitializingObject

An object that is currently being initialized.

InitializingType

A newly registered glib::Type that is currently still being initialized.

PrivateStruct 🔒

Struct used for the instance private data of the GObject.

TypeData

Type-specific data that is filled in during type creation.

Constants

INIT_TYPE_DATA

Traits

ClassStruct

Trait implemented by structs that implement a GObject C class struct.

FromObject

Helper trait for macros to access a subclass or its wrapper.

InstanceStruct

Trait implemented by structs that implement a GObject C instance struct.

InstanceStructExt

Trait implemented by any type implementing InstanceStruct to return the implementation, private Rust struct.

InterfaceList

Trait for a type list of interfaces.

IsImplementable

Trait for implementable interfaces.

IsSubclassable

Trait for subclassable class structs.

IsSubclassableExt

ObjectSubclass

The central trait for subclassing a GObject type.

ObjectSubclassExt

Extension methods for all ObjectSubclass impls.

ObjectSubclassIsExt

Trait implemented by any type implementing ObjectSubclassIs to return the implementation, private Rust struct.

ObjectSubclassType

Type methods required for an ObjectSubclass implementation.

Functions

class_init 🔒 ⚠

finalize 🔒 ⚠

instance_init 🔒 ⚠

interface_init 🔒 ⚠

offset_ptr_by_bytes 🔒

Offset ptr by offset bytes and cast the result to *const U.

offset_ptr_by_bytes_mut 🔒

Offset ptr by offset bytes and cast the result to *mut U.

register_dynamic_type

Registers a glib::Type ID for T as a dynamic type.

register_type

Register a glib::Type ID for T.

signal_chain_from_overridden 🔒 ⚠

signal_override_class_handler 🔒 ⚠