Trait glib::subclass::types::ObjectSubclassExt

pub trait ObjectSubclassExt: ObjectSubclass {
    fn get_instance(&self) -> Self::Type;

    fn from_instance(obj: &Self::Type) -> &Self;

    fn get_instance_data<U: Any + Send + Sync + 'static>(
        &self, 
        type_: Type
    ) -> Option<&U>;
}

Extension methods for all ObjectSubclass impls.

Required methods

fn get_instance(&self) -> Self::Type

Returns the corresponding object instance.

fn from_instance(obj: &Self::Type) -> &Self

Returns the implementation from an instance.

fn get_instance_data<U: Any + Send + Sync + 'static>(
&self,
type_: Type
) -> Option<&U>

Returns a pointer to the instance implementation specific data.

This is used for the subclassing infrastructure to store additional instance data.

Loading content...