Module glib::thread_guard

Structs

ThreadGuard

Thread guard that only gives access to the contained value on the thread it was created on.

Functions

thread_id

Returns a unique ID for the current thread.