glib::translate

Trait TransparentType

pub unsafe trait TransparentType: Clone + Sized {
    type GlibType;
}

Trait for types that have the same memory representation as their FFI type.

Values of types implementing this trait can be transmuted directly to the FFI type, references to pointers to the FFI type.

Required Associated Types

type GlibType

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.