Enum glib::types::Type

pub enum Type {
    Invalid,
    Unit,
    I8,
    U8,
    Bool,
    I32,
    U32,
    ILong,
    ULong,
    I64,
    U64,
    F32,
    F64,
    String,
    Pointer,
    Variant,
    BaseInterface,
    BaseEnum,
    BaseFlags,
    BaseBoxed,
    BaseParamSpec,
    BaseObject,
    Other(usize),
}

A GLib or GLib-based library type

Variants

Invalid

An invalid Type used as error return value in some functions

Unit

The fundamental type corresponding to the unit type ()

I8

The fundamental type corresponding to i8

U8

The fundamental type corresponding to u8

Bool

The fundamental type corresponding to bool

I32

The fundamental type corresponding to i32

U32

The fundamental type corresponding to u32

ILong

The fundamental type corresponding to C long

ULong

The fundamental type corresponding to C unsigned long

I64

The fundamental type corresponding to i64

U64

The fundamental type corresponding to u64

F32

The fundamental type corresponding to f32

F64

The fundamental type corresponding to f64

String

The fundamental type corresponding to String

Pointer

The fundamental type corresponding to a pointer

Variant

The fundamental type of GVariant

BaseInterface

The fundamental type from which all interfaces are derived

BaseEnum

The fundamental type from which all enumeration types are derived

BaseFlags

The fundamental type from which all flags types are derived

BaseBoxed

The fundamental type from which all boxed types are derived

BaseParamSpec

The fundamental type from which all GParamSpec types are derived

BaseObject

The fundamental type from which all objects are derived

Other(usize)

A non-fundamental type identified by value of type usize

Methods

impl Type

pub fn name(&self) -> String

pub fn qname(&self) -> Quark

pub fn is_a(&self, other: &Type) -> bool

pub fn parent(&self) -> Option<Self>

pub fn children(&self) -> Vec<Self>

pub fn interfaces(&self) -> Vec<Self>

pub fn interface_prerequisites(&self) -> Vec<Self>

pub fn from_name<'a, P: Into<&'a str>>(name: P) -> Option<Self>