Struct glib::types::Type

pub struct Type(_);

A GLib or GLib-based library type

Implementations

impl Type

pub const INVALID: Self

An invalid Type used as error return value in some functions

pub const UNIT: Self

The fundamental type corresponding to the unit type ()

pub const I8: Self

The fundamental type corresponding to i8

pub const U8: Self

The fundamental type corresponding to u8

pub const BOOL: Self

The fundamental type corresponding to bool

pub const I32: Self

The fundamental type corresponding to i32

pub const U32: Self

The fundamental type corresponding to u32

pub const I_LONG: Self

The fundamental type corresponding to C long

pub const U_LONG: Self

The fundamental type corresponding to C unsigned long

pub const I64: Self

The fundamental type corresponding to i64

pub const U64: Self

The fundamental type corresponding to u64

pub const F32: Self

The fundamental type corresponding to f32

pub const F64: Self

The fundamental type corresponding to f64

pub const STRING: Self

The fundamental type corresponding to String

pub const POINTER: Self

The fundamental type corresponding to a pointer

pub const VARIANT: Self

The fundamental type of GVariant

pub const INTERFACE: Self

The fundamental type from which all interfaces are derived

pub const ENUM: Self

The fundamental type from which all enumeration types are derived

pub const FLAGS: Self

The fundamental type from which all flags types are derived

pub const BOXED: Self

The fundamental type from which all boxed types are derived

pub const PARAM_SPEC: Self

The fundamental type from which all GParamSpec types are derived

pub const OBJECT: Self

The fundamental type from which all objects are derived

pub fn name<'a>(self) -> &'a str

pub fn qname(self) -> Quark

pub fn is_a(self, other: Self) -> bool

pub fn parent(self) -> Option<Self>

pub fn children(self) -> Vec<Self>

pub fn interfaces(self) -> Vec<Self>

pub fn interface_prerequisites(self) -> Vec<Self>

pub fn from_name<'a, P: Into<&'a str>>(name: P) -> Option<Self>

pub fn is_valid(self) -> bool

Checks that the type is not INVALID

Uses found in examples/

examples/composite_template/example_application_window/imp.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51

use gtk::glib;
use gtk::subclass::prelude::*;
use gtk::{prelude::*, CompositeTemplate};

/// The private struct, which can hold widgets and other data.
#[derive(Debug, Default, CompositeTemplate)]
#[template(file = "composite_template.ui")]
pub struct ExampleApplicationWindow {
    // The #[template_child] attribute tells the CompositeTemplate macro
    // that a field is meant to be a child within the template.
    #[template_child]
    pub headerbar: TemplateChild<gtk::HeaderBar>,
    #[template_child]
    pub label: TemplateChild<gtk::Label>,
    // You can specify the optional `id` argument if the id is not the same
    // as the identifier
    #[template_child(id = "subtitle_label")]
    pub subtitle: TemplateChild<gtk::Label>,
}

#[glib::object_subclass]
impl ObjectSubclass for ExampleApplicationWindow {
    const NAME: &'static str = "ExampleApplicationWindow";
    type Type = super::ExampleApplicationWindow;
    type ParentType = gtk::ApplicationWindow;

    // Within class_init() you must set the template.
    // The CompositeTemplate derive macro provides a convenience function
    // bind_template() to set the template and bind all children at once.
    fn class_init(klass: &mut Self::Class) {
        Self::bind_template(klass);
    }

    // You must call `Widget`'s `init_template()` within `instance_init()`.
    fn instance_init(obj: &glib::subclass::InitializingObject<Self>) {
        obj.init_template();
    }
}

impl ObjectImpl for ExampleApplicationWindow {
    fn constructed(&self, obj: &Self::Type) {
        obj.init_label();
        self.parent_constructed(obj);
    }
}

impl WidgetImpl for ExampleApplicationWindow {}
impl ContainerImpl for ExampleApplicationWindow {}
impl BinImpl for ExampleApplicationWindow {}
impl WindowImpl for ExampleApplicationWindow {}
impl ApplicationWindowImpl for ExampleApplicationWindow {}

examples/basic_subclass/simple_window/imp.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77

use glib::clone;
use gtk::glib;
use gtk::prelude::*;
use gtk::subclass::prelude::*;
use once_cell::unsync::OnceCell;
use std::cell::Cell;

#[derive(Debug)]
struct WindowWidgets {
    headerbar: gtk::HeaderBar,
    increment: gtk::Button,
    label: gtk::Label,
}

#[derive(Debug, Default)]
pub struct SimpleWindow {
    widgets: OnceCell<WindowWidgets>,
    counter: Cell<u64>,
}

#[glib::object_subclass]
impl ObjectSubclass for SimpleWindow {
    const NAME: &'static str = "SimpleWindow";
    type Type = super::SimpleWindow;
    type ParentType = gtk::ApplicationWindow;
}

impl ObjectImpl for SimpleWindow {
    // Here we are overriding the glib::Objcet::contructed
    // method. Its what gets called when we create our Object
    // and where we can initialize things.
    fn constructed(&self, obj: &Self::Type) {
        self.parent_constructed(obj);

        let headerbar = gtk::HeaderBar::new();
        let increment = gtk::Button::with_label("Increment!");
        let label = gtk::Label::new(Some("Press the Increment Button!"));

        headerbar.set_title(Some("Hello World!"));
        headerbar.set_show_close_button(true);
        headerbar.pack_start(&increment);

        // Connect our method `on_increment_clicked` to be called
        // when the increment button is clicked.
        increment.connect_clicked(clone!(@weak obj => move |_| {
            let priv_ = SimpleWindow::from_instance(&obj);
            priv_.on_increment_clicked();
        }));

        obj.add(&label);
        obj.set_titlebar(Some(&headerbar));
        obj.set_default_size(640, 480);

        self.widgets
            .set(WindowWidgets {
                headerbar,
                increment,
                label,
            })
            .expect("Failed to initialize window state");
    }
}

impl SimpleWindow {
    fn on_increment_clicked(&self) {
        self.counter.set(self.counter.get() + 1);
        let w = self.widgets.get().unwrap();
        w.label
            .set_text(&format!("Counter is {}", self.counter.get()));
    }
}

impl WidgetImpl for SimpleWindow {}
impl ContainerImpl for SimpleWindow {}
impl BinImpl for SimpleWindow {}
impl WindowImpl for SimpleWindow {}
impl ApplicationWindowImpl for SimpleWindow {}

examples/list_box_model/row_data/imp.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85

use glib::subclass::prelude::*;
use gtk::{glib, prelude::*};
use std::cell::RefCell;

// The actual data structure that stores our values. This is not accessible
// directly from the outside.
#[derive(Default)]
pub struct RowData {
    name: RefCell<Option<String>>,
    count: RefCell<u32>,
}

// Basic declaration of our type for the GObject type system
#[glib::object_subclass]
impl ObjectSubclass for RowData {
    const NAME: &'static str = "RowData";
    type Type = super::RowData;
    type ParentType = glib::Object;
}

// The ObjectImpl trait provides the setters/getters for GObject properties.
// Here we need to provide the values that are internally stored back to the
// caller, or store whatever new value the caller is providing.
//
// This maps between the GObject properties and our internal storage of the
// corresponding values of the properties.
impl ObjectImpl for RowData {
    fn properties() -> &'static [glib::ParamSpec] {
        use once_cell::sync::Lazy;
        static PROPERTIES: Lazy<Vec<glib::ParamSpec>> = Lazy::new(|| {
            vec![
                glib::ParamSpec::new_string(
                    "name",
                    "Name",
                    "Name",
                    None, // Default value
                    glib::ParamFlags::READWRITE,
                ),
                glib::ParamSpec::new_uint(
                    "count",
                    "Count",
                    "Count",
                    0,
                    100,
                    0, // Allowed range and default value
                    glib::ParamFlags::READWRITE,
                ),
            ]
        });

        PROPERTIES.as_ref()
    }

    fn set_property(
        &self,
        _obj: &Self::Type,
        _id: usize,
        value: &glib::Value,
        pspec: &glib::ParamSpec,
    ) {
        match pspec.name() {
            "name" => {
                let name = value
                    .get()
                    .expect("type conformity checked by `Object::set_property`");
                self.name.replace(name);
            }
            "count" => {
                let count = value
                    .get()
                    .expect("type conformity checked by `Object::set_property`");
                self.count.replace(count);
            }
            _ => unimplemented!(),
        }
    }

    fn property(&self, _obj: &Self::Type, _id: usize, pspec: &glib::ParamSpec) -> glib::Value {
        match pspec.name() {
            "name" => self.name.borrow().to_value(),
            "count" => self.count.borrow().to_value(),
            _ => unimplemented!(),
        }
    }
}

examples/list_box_model/model/imp.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37

//! Defines the implementation of our model

use gio::subclass::prelude::*;
use gtk::{gio, glib, prelude::*};

use std::cell::RefCell;

use crate::row_data::RowData;

#[derive(Debug, Default)]
pub struct Model(pub RefCell<Vec<RowData>>);

/// Basic declaration of our type for the GObject type system
#[glib::object_subclass]
impl ObjectSubclass for Model {
    const NAME: &'static str = "Model";
    type Type = super::Model;
    type ParentType = glib::Object;
    type Interfaces = (gio::ListModel,);
}

impl ObjectImpl for Model {}

impl ListModelImpl for Model {
    fn item_type(&self, _list_model: &Self::Type) -> glib::Type {
        RowData::static_type()
    }
    fn n_items(&self, _list_model: &Self::Type) -> u32 {
        self.0.borrow().len() as u32
    }
    fn item(&self, _list_model: &Self::Type, position: u32) -> Option<glib::Object> {
        self.0
            .borrow()
            .get(position as usize)
            .map(|o| o.clone().upcast::<glib::Object>())
    }
}

examples/basic_subclass/simple_application/imp.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61

use gtk::glib;
use gtk::prelude::*;
use gtk::subclass::prelude::*;
use once_cell::unsync::OnceCell;

use crate::simple_window::SimpleWindow;

#[derive(Debug, Default)]
pub struct SimpleApplication {
    window: OnceCell<SimpleWindow>,
}

#[glib::object_subclass]
impl ObjectSubclass for SimpleApplication {
    const NAME: &'static str = "SimpleApplication";
    type Type = super::SimpleApplication;
    type ParentType = gtk::Application;
}

impl ObjectImpl for SimpleApplication {}

/// When our application starts, the `startup` signal will be fired.
/// This gives us a chance to perform initialisation tasks that are not directly
/// related to showing a new window. After this, depending on how
/// the application is started, either `activate` or `open` will be called next.
impl ApplicationImpl for SimpleApplication {
    /// `gio::Application::activate` is what gets called when the
    /// application is launched by the desktop environment and
    /// aksed to present itself.
    fn activate(&self, app: &Self::Type) {
        let app = app.downcast_ref::<super::SimpleApplication>().unwrap();
        let priv_ = SimpleApplication::from_instance(app);
        let window = priv_
            .window
            .get()
            .expect("Should always be initiliazed in gio_application_startup");
        window.show_all();
        window.present();
    }

    /// `gio::Application` is bit special. It does not get initialized
    /// when `new` is called and the object created, but rather
    /// once the `startup` signal is emitted and the `gio::Application::startup`
    /// is called.
    ///
    /// Due to this, we create and initialize the `SimpleWindow` widget
    /// here. Widgets can't be created before `startup` has been called.
    fn startup(&self, app: &Self::Type) {
        self.parent_startup(app);

        let app = app.downcast_ref::<super::SimpleApplication>().unwrap();
        let priv_ = SimpleApplication::from_instance(app);
        let window = SimpleWindow::new(&app);
        priv_
            .window
            .set(window)
            .expect("Failed to initialize application window");
    }
}

impl GtkApplicationImpl for SimpleApplication {}

examples/gio_task/file_size/imp.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18

use glib::subclass::prelude::*;
use gtk::glib;

// FileSize is a simple object that will just contain the read file size.
// Initially the optional size field will be initialized to None.
#[derive(Default)]
pub struct FileSize {
    pub size: std::cell::RefCell<Option<i64>>,
}

#[glib::object_subclass]
impl ObjectSubclass for FileSize {
    const NAME: &'static str = "FileSize";
    type ParentType = glib::Object;
    type Type = super::FileSize;
}

impl ObjectImpl for FileSize {}