Trait glib::unichar::Unichar

pub trait Unichar: Sealed + Copy + Into<u32> + UnsafeFrom<u32> {

    // Provided methods
    fn unicode_type(self) -> UnicodeType { ... }
    fn break_type(self) -> UnicodeBreakType { ... }
    fn script(self) -> UnicodeScript { ... }
    fn combining_class(self) -> u8 { ... }
    fn is_mark(self) -> bool { ... }
    fn is_graphical(self) -> bool { ... }
    fn is_punctuation(self) -> bool { ... }
    fn is_titlecase(self) -> bool { ... }
    fn is_defined(self) -> bool { ... }
    fn is_wide(self) -> bool { ... }
    fn is_wide_cjk(self) -> bool { ... }
    fn is_zero_width(self) -> bool { ... }
    fn to_titlecase(self) -> Self { ... }
    fn mirror_char(self) -> Option<Self> { ... }
    fn fully_decompose(self, decomposition_kind: DecompositionKind) -> Vec<Self> { ... }
    fn decompose(self) -> CharacterDecomposition { ... }
    fn compose(a: char, b: char) -> Option<Self> { ... }

}

This trait provides access to Unicode character classification and manipulations functions provided by GLib that do not exist in the standard library

Provided Methods

fn unicode_type(self) -> UnicodeType

fn break_type(self) -> UnicodeBreakType

fn script(self) -> UnicodeScript

fn combining_class(self) -> u8

fn is_mark(self) -> bool

fn is_graphical(self) -> bool

fn is_punctuation(self) -> bool

fn is_titlecase(self) -> bool

fn is_defined(self) -> bool

fn is_wide(self) -> bool

fn is_wide_cjk(self) -> bool

fn is_zero_width(self) -> bool

fn to_titlecase(self) -> Self

fn mirror_char(self) -> Option<Self>

fn fully_decompose(self, decomposition_kind: DecompositionKind) -> Vec<Self>

fn decompose(self) -> CharacterDecomposition

fn compose(a: char, b: char) -> Option<Self>