Struct glib::value::AnyValue

pub struct AnyValue { /* fields omitted */ }

A container type that allows storing any 'static type that implements Any and Clone to be stored in a Value.

See the module documentation for more details.

Examples

use glib::prelude::*; // or `use gtk::prelude::*;`
use glib::{AnyValue, Value};

// Store a Rust string inside a Value
let v = AnyValue::new(String::from("123")).to_value();

// Retrieve the Rust String from the the Value again
let any_v = v.get::<&AnyValue>()
    .expect("Value did not actually contain an AnyValue");
assert_eq!(any_v.downcast_ref::<String>(), Some(&String::from("123")));

Methods

impl AnyValue

pub fn new<T: Any + Clone + 'static>(val: T) -> Self

Create a new AnyValue from val

pub fn downcast<T: Any + Clone + 'static>(self) -> Result<T, Self>

Attempt the value to its concrete type.