Trait glib::value::FromValueOptional

pub trait FromValueOptional<'a>: StaticType + Sized {
    unsafe fn from_value_optional(value: &'a Value) -> Option<Self>;
}

Extracts a value.

Types that don’t support a None value always return Some.

Required methods

unsafe fn from_value_optional(value: &'a Value) -> Option<Self>

Safety

The caller is responsible for ensuring the given Value is of a suitable type for this conversion.

Loading content...