Trait glib::value::SetValue

pub trait SetValue: StaticType {
    unsafe fn set_value(value: &mut Value, new_value: &Self);
}

Sets a value.

Required methods

unsafe fn set_value(value: &mut Value, new_value: &Self)

Safety

The caller is responsible for ensuring the given Value is of a suitable type for this conversion.

Loading content...