Trait glib::value::SetValueOptional

pub trait SetValueOptional: SetValue {
    unsafe fn set_value_optional(value: &mut Value, new_value: Option<&Self>);
}

Sets a value.

Only implemented for types that support a None value.

Required methods

unsafe fn set_value_optional(value: &mut Value, new_value: Option<&Self>)

Safety

The caller is responsible for ensuring the given Value is of a suitable type for this conversion.

Loading content...