Crate glib

Rust GLib and GObject bindings

Rust bindings and wrappers for GLib, part of gtk-rs-core.

GLib 2.56 is the lowest supported version for the underlying library.

This library contains bindings to GLib and GObject types and APIs as well as common building blocks used in both handmade and machine generated bindings to GTK and other GLib-based libraries.

It is the foundation for higher level libraries with uniform Rusty (safe and strongly typed) APIs. It avoids exposing GLib-specific data types where possible and is not meant to provide comprehensive GLib bindings, which would often amount to duplicating the Rust Standard Library or other utility crates.

Minimum supported Rust version

Currently, the minimum supported Rust version is 1.64.0.

Dynamic typing

Most types in the GLib family have Type identifiers. Their corresponding Rust types implement the StaticType trait.

A dynamically typed Value can carry values of any StaticType. Variants can carry values of StaticVariantType.

Errors

Errors are represented by Error, which can carry values from various error domains such as FileError.

Objects

Each class and interface has a corresponding smart pointer struct representing an instance of that type (e.g. Object for GObject or gtk::Widget for GtkWidget). They are reference counted and feature interior mutability similarly to Rust’s Rc<RefCell<T>> idiom. Consequently, cloning objects is cheap and their methods never require mutable borrows. Two smart pointers are equal if they point to the same object.

The root of the object hierarchy is Object. Inheritance and subtyping is denoted with the IsA marker trait. The Cast trait enables upcasting and downcasting.

Interfaces and non-leaf classes also have corresponding traits (e.g. ObjectExt or gtk::WidgetExt), which are blanketly implemented for all their subtypes.

You can create new subclasses of Object or other object types. Look at the module’s documentation for further details and a code example.

Under the hood

GLib-based libraries largely operate on pointers to various boxed or reference counted structures so the bindings have to implement corresponding smart pointers (wrappers), which encapsulate resource management and safety checks. Such wrappers are defined via the wrapper macro, which uses abstractions defined in the wrapper, boxed, shared and object modules.

The translate module defines and partly implements conversions between high level Rust types (including the aforementioned wrappers) and their FFI counterparts.

Documentation

Rust API - Stable

Rust API - Development

C API

GTK Installation instructions

Using

We recommend using crates from crates.io, as demonstrated here.

If you want to track the bleeding edge, use the git dependency instead:

[dependencies]
glib = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "glib" }

Avoid mixing versioned and git crates like this:

# This will not compile
[dependencies]
glib = "0.13"
glib = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "glib" }

License

glib is available under the MIT License, please refer to it.

Re-exports

pub use ffi;

pub use gobject_ffi;

pub use self::closure::Closure;

pub use self::closure::RustClosure;

pub use self::error::BoolError;

pub use self::error::Error;

pub use self::object::BorrowedObject;

pub use self::object::Cast;

pub use self::object::CastNone;

pub use self::object::Class;

pub use self::object::InitiallyUnowned;

pub use self::object::Interface;

pub use self::object::IsA;

pub use self::object::Object;

pub use self::object::ObjectExt;

pub use self::object::ObjectType;

pub use self::object::SendWeakRef;

pub use self::object::WeakRef;

pub use self::signal::signal_handler_block;

pub use self::signal::signal_handler_disconnect;

pub use self::signal::signal_handler_unblock;

pub use self::signal::signal_stop_emission_by_name;

pub use self::signal::SignalHandlerId;

pub use self::types::ILong;

pub use self::types::Pointer;

pub use self::types::StaticType;

pub use self::types::StaticTypeExt;

pub use self::types::Type;

pub use self::types::ULong;

pub use self::value::BoxedValue;

pub use self::value::SendValue;

pub use self::value::ToSendValue;

pub use self::value::ToValue;

pub use self::value::Value;

pub use self::variant::FixedSizeVariantArray;

pub use self::variant::FixedSizeVariantType;

pub use self::variant::FromVariant;

pub use self::variant::StaticVariantType;

pub use self::variant::ToVariant;

pub use self::variant::Variant;

pub use self::FileError;

pub use collections::ptr_slice::IntoPtrSlice;

pub use collections::strv::IntoStrV;

pub use collections::List;

pub use collections::PtrSlice;

pub use collections::SList;

pub use collections::Slice;

pub use collections::StrV;

pub use self::char::*;

pub use self::source::*;

Modules

boxed

IMPL Boxed wrapper implementation.

boxed_inline

IMPL BoxedInline wrapper implementation.

char

clone

closure

collections

error

Error binding and helper trait.

object

IMPL Object wrapper implementation and Object binding.

prelude

Traits and essential types intended for blanket imports.

shared

IMPL Shared (reference counted) wrapper implementation.

signal

IMPL Low level signal support.

source

subclass

Module containing infrastructure for subclassing GObjects and registering boxed types.

thread_guard

translate

Translation between GLib/GLib-based FFI types and their Rust counterparts.

types

Runtime type information.

value

Value binding and helper traits.

variant

Variant binding and helper traits.

wrapper

IMPL The wrapper! macro and miscellaneous wrapper traits.

Macros

bool_error

Generic error used for functions that fail without any further information

clone

Macro for passing variables as strong or weak references into a closure.

closure

Macro for creating a Closure object. This is a wrapper around Closure::new that automatically type checks its arguments at run-time.

closure_local

The same as closure! but uses Closure::new_local as a constructor. This is useful for closures which can’t be sent across threads. See the documentation of closure! for details.

debuglog_macros

A macro which behaves exactly as log::debug! except that it sets the current log target to the contents of a G_LOG_DOMAIN constant (and fails to build if not defined).

errorlog_macros

A macro which behaves exactly as log::error! except that it sets the current log target to the contents of a G_LOG_DOMAIN constant (and fails to build if not defined).

function_name

This macro returns the name of the enclosing function. As the internal implementation is based on the std::any::type_name, this macro derives all the limitations of this function.

g_critical

Macro used to log using GLib logging system. It uses g_log.

g_debug

Macro used to log using GLib logging system. It uses g_log.

g_error

Macro used to log using GLib logging system. It uses g_log.

g_info

Macro used to log using GLib logging system. It uses g_log.

g_log

Macro used to log using GLib logging system. It uses g_log.

g_message

Macro used to log using GLib logging system. It uses g_log.

g_print

Macro used to print messages. It uses g_print.

g_printerr

Macro used to print error messages. It uses g_printerr.

g_warning

Macro used to log using GLib logging system. It uses g_log.

gformat

Creates a GString using interpolation of runtime expressions.

glib_boxed_inline_wrapper

Wrapper implementations for BoxedInline types. See wrapper!.

glib_boxed_wrapper

Wrapper implementations for Boxed types. See wrapper!.

glib_object_wrapper

ObjectType implementations for Object types. See wrapper!.

glib_shared_wrapper

Wrapper implementations for shared types. See wrapper!.

gstr

Converts a static string literal into a static nul-terminated string.

infolog_macros

A macro which behaves exactly as log::info! except that it sets the current log target to the contents of a G_LOG_DOMAIN constant (and fails to build if not defined).

log_structured

Macro used to log using GLib structured logging system.

result_from_gboolean

tracelog_macros

A macro which behaves exactly as log::trace! except that it sets the current log target to the contents of a G_LOG_DOMAIN constant (and fails to build if not defined).

warnlog_macros

A macro which behaves exactly as log::warn! except that it sets the current log target to the contents of a G_LOG_DOMAIN constant (and fails to build if not defined).

wrapper

Defines a wrapper type and implements the appropriate traits.

Structs

Binding

Binding is the representation of a binding between a property on a Object instance (or source) and another property on another Object instance (or target).

BindingFlags

Flags to be passed to ObjectExt::bind_property() or [ObjectExt::bind_property_full()][crate::prelude::ObjectExt::bind_property_full()].

BindingGroupv2_72

The BindingGroup can be used to bind multiple properties from an object collectively.

BindingGroupBuilderv2_72

Builder for binding group bindings.

BoxedAnyObject

This is a subclass of glib::object::Object capable of storing any Rust type. It let’s you insert a Rust type anywhere a glib::object::Object is needed. The inserted value can then be borrowed as a Rust type, by using the various provided methods.

ByteArray

Bytes

A shared immutable byte slice (the equivalent of Rc<[u8]>).

Checksum

CollationKey

A CollationKey allows ordering strings using the linguistically correct rules for the current locale.

Date

DateTime

EnumClass

Representation of an enum for dynamically, at runtime, querying the values of the enum and using them.

EnumValue

Representation of a single enum value of an EnumClass.

ExitCode

FileSetContentsFlagsv2_66

FilenameCollationKey

A FilenameCollationKey allows ordering file names using the linguistically correct rules for the current locale. Compared to CollationKey, filename collation keys take into consideration dots and other characters commonly found in file names.

FlagsBuilder

Builder for conveniently setting/unsetting flags and returning a Value.

FlagsClass

Representation of a flags for dynamically, at runtime, querying the values of the enum and using them

FlagsValue

Representation of a single flags value of a FlagsClass.

FormatSizeFlags

FutureWithTimeoutError

The error returned when a future times out.

GStr

Representation of a borrowed GString.

GStrInteriorNulError

Error type indicating that a buffer had unexpected nul-bytes.

GString

A type representing an owned, C-compatible, nul-terminated UTF-8 string.

GStringBuilder

A mutable text buffer that grows automatically.

GStringInteriorNulError

Error type indicating that a buffer had unexpected nul-bytes.

GStringNoTrailingNulError

Error type indicating that a buffer did not have a trailing nul-byte.

GStringPtr

NULL-terminated UTF-8 string as stored in [StrV].

GStringUtf8Error

Error type indicating that a buffer had invalid UTF-8.

GlibLogger

An implementation of a log compatible logger which logs over glib logging facilities.

IConv

IOCondition

JoinError

Task failure from awaiting a JoinHandle.

JoinHandle

A handle to a task running on a MainContext.

KeyFile

KeyFileFlags

LogField

Structure representing a single field in a structured log entry.

LogHandlerId

LogLevelFlags

LogLevels

MainContext

MainContextAcquireGuard

MainContextFlagsv2_72

MainLoop

MarkupParseContext

OptionFlags

ParamFlags

ParamSpec

ParamSpecBoolean

ParamSpecBooleanBuilder

ParamSpecBoxed

ParamSpecBoxedBuilder

ParamSpecChar

ParamSpecCharBuilder

ParamSpecDouble

ParamSpecDoubleBuilder

ParamSpecEnum

ParamSpecEnumBuilder

ParamSpecFlags

ParamSpecFlagsBuilder

ParamSpecFloat

ParamSpecFloatBuilder

ParamSpecGType

ParamSpecGTypeBuilder

ParamSpecInt

ParamSpecInt64

ParamSpecInt64Builder

ParamSpecIntBuilder

ParamSpecLong

ParamSpecLongBuilder

ParamSpecObject

ParamSpecObjectBuilder

ParamSpecOverride

ParamSpecOverrideBuilder

ParamSpecParam

ParamSpecParamBuilder

ParamSpecPointer

ParamSpecPointerBuilder

ParamSpecString

ParamSpecStringBuilder

ParamSpecUChar

ParamSpecUCharBuilder

ParamSpecUInt

ParamSpecUInt64

ParamSpecUInt64Builder

ParamSpecUIntBuilder

ParamSpecULong

ParamSpecULongBuilder

ParamSpecUnichar

ParamSpecUnicharBuilder

ParamSpecValueArray

ParamSpecValueArrayBuilder

ParamSpecVariant

ParamSpecVariantBuilder

Quark

Receiver

A Receiver that can be attached to a main context to receive items from its corresponding Sender or SyncSender.

Sender

A Sender that can be used to send items to the corresponding main context receiver.

SignalFlags

SignalGroupv2_74

Source

SourceFuture

Represents a Future around a glib::Source. The future will be resolved once the source has provided a value

SourceStream

Represents a Stream around a glib::Source. The stream will be provide all values that are provided by the source

SpawnFlags

SpawnWithinJoinHandle

Variant of JoinHandle that is returned from MainContext::spawn_from_within.

SyncSender

A SyncSender that can be used to send items to the corresponding main context receiver.

ThreadHandle

A handle to a thread running on a ThreadPool.

ThreadPool

TimeSpan

A value representing an interval of time, in microseconds.

TimeZone

Uriv2_66

UriFlagsv2_66

Flags that describe a URI.

UriHideFlagsv2_66

Flags describing what parts of the URI to hide in Uri::to_string_partial(). Note that PASSWORD and AUTH_PARAMS will only work if the Uri was parsed with the corresponding flags.

UriParamsFlagsv2_66

Flags modifying the way parameters are handled by g_uri_parse_params() and GUriParamsIter.

ValueArray

VariantDict

VariantDict is a mutable key/value store where the keys are always strings and the values are Variants.

VariantIter

Iterator over items in a variant.

VariantStrIter

Iterator over items in a variant of type as.

VariantTy

Describes Variant types.

VariantTyIterator

An iterator over the individual components of a tuple VariantTy.

VariantType

Describes Variant types.

Enums

ChecksumType

ConvertError

CvtError

A wrapper for ConvertError that can hold an offset into the input string.

DateMonth

DateWeekday

FileError

GStrError

Error type holding all possible failures when creating a GStr reference.

GStringFromError

Error type holding all possible failures when creating a GString.

GlibLoggerDomain

Enumeration of the possible domain handling behaviours for a GlibLogger.

GlibLoggerFormat

Enumeration of the possible formatting behaviours for a GlibLogger.

IConvError

A wrapper for std::io::Error that can hold an offset into an input string.

KeyFileError

LogLevel

LogWriterOutput

MarkupError

NormalizeMode

OptionArg

SeekType

TimeType

UnicodeScript

UriErrorv2_66

Error codes returned by Uri methods.

UserDirectory

These are logical ids for special directories which are defined depending on the platform used. You should use user_special_dir() to retrieve the full path associated to the logical id.

VariantClass

The range of possible top-level types of Variant instances.

Constants

CLONE_MACRO_LOG_DOMAIN

This is the log domain used by the clone! macro. If you want to use a custom logger (it prints to stdout by default), you can set your own logger using the corresponding log functions.

NONE_STR

Statics

CSET_A_2_Z

CSET_DIGITS

CSET_a_2_z

KEY_FILE_DESKTOP_GROUP

KEY_FILE_DESKTOP_KEY_ACTIONS

KEY_FILE_DESKTOP_KEY_CATEGORIES

KEY_FILE_DESKTOP_KEY_COMMENT

KEY_FILE_DESKTOP_KEY_DBUS_ACTIVATABLE

KEY_FILE_DESKTOP_KEY_EXEC

KEY_FILE_DESKTOP_KEY_GENERIC_NAME

KEY_FILE_DESKTOP_KEY_HIDDEN

KEY_FILE_DESKTOP_KEY_ICON

KEY_FILE_DESKTOP_KEY_MIME_TYPE

KEY_FILE_DESKTOP_KEY_NAME

KEY_FILE_DESKTOP_KEY_NOT_SHOW_IN

KEY_FILE_DESKTOP_KEY_NO_DISPLAY

KEY_FILE_DESKTOP_KEY_ONLY_SHOW_IN

KEY_FILE_DESKTOP_KEY_PATH

KEY_FILE_DESKTOP_KEY_STARTUP_NOTIFY

KEY_FILE_DESKTOP_KEY_STARTUP_WM_CLASS

KEY_FILE_DESKTOP_KEY_TERMINAL

KEY_FILE_DESKTOP_KEY_TRY_EXEC

KEY_FILE_DESKTOP_KEY_TYPE

KEY_FILE_DESKTOP_KEY_URL

KEY_FILE_DESKTOP_KEY_VERSION

KEY_FILE_DESKTOP_TYPE_APPLICATION

KEY_FILE_DESKTOP_TYPE_DIRECTORY

KEY_FILE_DESKTOP_TYPE_LINK

OPTION_REMAINING

STR_DELIMITERS

TEST_OPTION_ISOLATE_DIRSv2_60

URI_RESERVED_CHARS_GENERIC_DELIMITERS

URI_RESERVED_CHARS_SUBCOMPONENT_DELIMITERS

g_param_spec_types

Traits

HasParamSpec

IntoGStr

A trait to accept both &str or &GStr as an argument.

IntoOptionalGStr

A trait to accept both Option<&str> or Option<&GStr> as an argument.

ParamSpecBuilderExt

A trait implemented by the various ParamSpec builder types.

ParamSpecType

Property

A type that can be used as a property. It covers every type which have an associated ParamSpec (HasParamSpec) and some useful types wrapping HasParamSpec. The definition is recursive, so you can nest many Propertys together. The final ParamSpec will be the one of the innermost type

PropertyGet

A container type implementing this trait can be read by the default getter generated by the Props macro.

PropertySet

A container type implementing this trait can be written by the default setter generated by the Props macro.

PropertySetNested

A container type implementing this trait can be written by the default setter generated by the Props macro. It takes a FnOnce(&mut Self::Value) so that the caller may access nested fields of a struct by doing ${Self::Value}.member

Functions

access

application_name

base64_decode

base64_encode

casefold

charset

Obtain the character set for the current locale.

check_version

child_watch_future

Create a Future that will resolve once the child process with the given pid exits

child_watch_future_with_priority

Create a Future that will resolve once the child process with the given pid exits

codeset

compute_checksum_for_bytes

compute_checksum_for_data

compute_hmac_for_bytes

compute_hmac_for_data

console_charsetv2_62

convert

convert_with_fallback

current_dir

dcgettext

dgettext

dngettext

dpgettext

This function is a variant of dgettext() which supports a disambiguating message context. GNU gettext uses the ‘\004’ character to separate the message context and message id in msgctxtid. If 0 is passed as msgidoffset, this function will fall back to trying to use the deprecated convention of using “|” as a separation character.

dpgettext2

This function is a variant of dgettext() which supports a disambiguating message context. GNU gettext uses the ‘\004’ character to separate the message context and message id in msgctxtid.

environ

environ_getenv

file_get_contents

Reads an entire file into allocated memory, with good error checking.

file_open_tmp

Opens a file for writing in the preferred directory for temporary files (as returned by tmp_dir()).

file_set_contents

Writes all of contents to a file named filename. This is a convenience wrapper around calling file_set_contents_full() with flags set to G_FILE_SET_CONTENTS_CONSISTENT | G_FILE_SET_CONTENTS_ONLY_EXISTING and mode set to 0666.

file_set_contents_fullv2_66

Writes all of contents to a file named filename, with good error checking. If a file called filename already exists it will be overwritten.

filename_charsets

filename_display_basename

Returns the display basename for the particular filename, guaranteed to be valid UTF-8. The display name might not be identical to the filename, for instance there might be problems converting it to UTF-8, and some files can be translated in the display.

filename_display_name

Converts a filename into a valid UTF-8 string. The conversion is not necessarily reversible, so you should keep the original around and use the return value of this function only for display purposes. Unlike filename_to_utf8(), the result is guaranteed to be non-None even if the filename actually isn’t in the GLib file name encoding.

filename_from_uri

Converts an escaped ASCII-encoded URI to a local filename in the encoding used for filenames.

filename_from_utf8

Converts a string from UTF-8 to the encoding GLib uses for filenames. Note that on Windows GLib uses UTF-8 for filenames; on other platforms, this function indirectly depends on the [current locale][setlocale].

filename_to_uri

Converts an absolute filename to an escaped ASCII-encoded URI, with the path component following Section 3.3. of RFC 2396.

filename_to_utf8

Converts a string which is in the encoding used by GLib for filenames into a UTF-8 string. Note that on Windows GLib uses UTF-8 for filenames; on other platforms, this function indirectly depends on the [current locale][setlocale].

find_program_in_path

format_size

format_size_full

future_with_timeout

Add a timeout to a Future.

future_with_timeout_with_priority

Add a timeout to a Future.

getenv

Returns the value of an environment variable.

home_dir

host_name

Return a name for the machine.

hostname_is_ascii_encoded

Tests if hostname contains segments with an ASCII-compatible encoding of an Internationalized Domain Name. If this returns true, you should decode the hostname with hostname_to_unicode() before displaying it to the user.

hostname_is_ip_address

Tests if hostname is the string form of an IPv4 or IPv6 address. (Eg, “192.168.0.1”.)

hostname_is_non_ascii

Tests if hostname contains Unicode characters. If this returns true, you need to encode the hostname with hostname_to_ascii() before using it in non-IDN-aware contexts.

hostname_to_ascii

Converts hostname to its canonical ASCII form; an ASCII-only string containing no uppercase letters and not ending with a trailing dot.

hostname_to_unicode

Converts hostname to its canonical presentation form; a UTF-8 string in Unicode normalization form C, containing no uppercase letters, no forbidden characters, and no ASCII-encoded segments, and not ending with a trailing dot.

interval_stream

Create a Stream that will provide a value every given number of milliseconds.

interval_stream_seconds

Create a Stream that will provide a value every given number of seconds.

interval_stream_seconds_with_priority

Create a Stream that will provide a value every given number of seconds.

interval_stream_with_priority

Create a Stream that will provide a value every given number of milliseconds.

is_canonical_pspec_name

language_names

Computes a list of applicable locale names, which can be used to e.g. construct locale-dependent filenames or search paths. The returned list is sorted from most desirable to least desirable and always contains the default locale “C”.

language_names_with_categoryv2_58

Computes a list of applicable locale names with a locale category name, which can be used to construct the fallback locale-dependent filenames or search paths. The returned list is sorted from most desirable to least desirable and always contains the default locale “C”.

listenv

Gets the names of all variables set in the environment.

locale_from_utf8

Converts a string from UTF-8 to the encoding used for strings by the C runtime (usually the same as that used by the operating system) in the [current locale][setlocale]. On Windows this means the system codepage.

locale_to_utf8

Converts a string which is in the encoding used for strings by the C runtime (usually the same as that used by the operating system) in the [current locale][setlocale] into a UTF-8 string.

locale_variants

Returns a list of derived variants of locale, which can be used to e.g. construct locale-dependent filenames or search paths. The returned list is sorted from most desirable to least desirable. This function handles territory, charset and extra locale modifiers. See setlocale(3) for information about locales and their format.

log_default_handler

log_remove_handler

log_set_always_fatal

log_set_default_handler

To set back the default print handler, use the log_unset_default_handler function.

log_set_fatal_mask

log_set_handler

log_set_writer_func

log_structured_array

log_unset_default_handler

To set the default print handler, use the log_set_default_handler function.

log_variant

log_writer_default

log_writer_default_set_use_stderr⚠v2_68

Sets whether GLib log functions output to stderr or stdout.

log_writer_default_would_dropv2_68

log_writer_format_fields

log_writer_is_journald

log_writer_journald

log_writer_standard_streams

log_writer_supports_color

main_current_source

Returns the currently firing source for this thread.

main_depth

Returns the depth of the stack of calls to MainContext::dispatch() on any MainContext in the current thread. That is, when called from the toplevel, it gives 0. When called from within a callback from MainContext::iteration() (or MainLoop::run(), etc.) it returns 1. When called from within a callback to a recursive call to MainContext::iteration(), it returns 2. And so forth.

markup_escape_text

Escapes text so that the markup parser will parse it verbatim. Less than, greater than, ampersand, etc. are replaced with the corresponding entities. This function would typically be used when writing out a file to be parsed with the markup parser.

mkdir_with_parents

Create a directory if it doesn’t already exist. Create intermediate parent directories as needed, too.

mkdtemp

Creates a temporary directory. See the mkdtemp() documentation on most UNIX-like systems.

mkdtemp_full

Creates a temporary directory. See the mkdtemp() documentation on most UNIX-like systems.

mkstemp

Opens a temporary file. See the mkstemp() documentation on most UNIX-like systems.

mkstemp_full

Opens a temporary file. See the mkstemp() documentation on most UNIX-like systems.

monotonic_time

Queries the system monotonic time.

normalize

num_processors

Determine the approximate number of threads that the system will schedule simultaneously for this process. This is intended to be used as a parameter to [ThreadPool::new()][crate::ThreadPool::new()] for CPU bound tasks and similar cases.

on_error_query

Prompts the user with [E]xit, [H]alt, show [S]tack trace or [P]roceed. This function is intended to be used for debugging use only. The following example shows how it can be used together with the g_log() functions.

on_error_stack_trace

Invokes gdb, which attaches to the current process and shows a stack trace. Called by on_error_query() when the “[S]tack trace” option is selected. You can get the current process’s program name with g_get_prgname(), assuming that you have called gtk_init() or gdk_init().

os_infov2_64

Get information about the operating system.

prgname

program_name

Same as get_prgname().

random_double

Returns a random gdouble equally distributed over the range [0..1).

random_double_range

Returns a random gdouble equally distributed over the range [begin..end).

random_int

Return a random guint32 equally distributed over the range [0..2^32-1].

random_int_range

Returns a random gint32 equally distributed over the range [begin..end-1].

random_set_seed

Sets the seed for the global random number generator, which is used by the g_random_* functions, to seed.

real_name

Gets the real name of the user. This usually comes from the user’s entry in the passwd file. The encoding of the returned string is system-defined. (On Windows, it is, however, always UTF-8.) If the real user name cannot be determined, the string “Unknown” is returned.

real_time

Queries the system wall-clock time.

reload_user_special_dirs_cache

Resets the cache used for user_special_dir(), so that the latest on-disk version is used. Call this only if you just changed the data on disk yourself.

rust_log_handler

Provides a glib log handler which routes all logging messages to the log crate.

set_application_name

Sets a human-readable name for the application. This name should be localized if possible, and is intended for display to the user. Contrast with g_set_prgname(), which sets a non-localized name. g_set_prgname() will be called automatically by gtk_init(), but set_application_name() will not.

set_prgname

set_print_handler

To set back the default print handler, use the unset_print_handler function.

set_printerr_handler

To set back the default print handler, use the unset_printerr_handler function.

set_program_name

Same as set_prgname().

setenv

Sets an environment variable. On UNIX, both the variable’s name and value can be arbitrary byte strings, except that the variable’s name cannot contain ‘=’. On Windows, they should be in UTF-8.

shell_parse_argv

shell_quote

shell_unquote

spaced_primes_closest

spawn_async

spawn_async_with_fdsv2_58 and non-Windows

spawn_async_with_pipesNon-Windows

spawn_check_exit_status

spawn_check_wait_statusv2_70

spawn_command_line_asyncUnix

system_config_dirs

Returns an ordered list of base directories in which to access system-wide configuration information.

system_data_dirs

Returns an ordered list of base directories in which to access system-wide application data.

timeout_future

Create a Future that will resolve after the given number of milliseconds.

timeout_future_seconds

Create a Future that will resolve after the given number of seconds.

timeout_future_seconds_with_priority

Create a Future that will resolve after the given number of seconds.

timeout_future_with_priority

Create a Future that will resolve after the given number of milliseconds.

tmp_dir

Gets the directory to use for temporary files.

unicode_script_from_iso15924

unicode_script_to_iso15924

unix_open_pipe

unix_signal_futureUnix

Create a Future that will resolve once the given UNIX signal is raised

unix_signal_future_with_priorityUnix

Create a Future that will resolve once the given UNIX signal is raised

unix_signal_streamUnix

Create a Stream that will provide a value whenever the given UNIX signal is raised

unix_signal_stream_with_priorityUnix

Create a Stream that will provide a value whenever the given UNIX signal is raised

unlink

unset_print_handler

To set the default print handler, use the set_print_handler function.

unset_printerr_handler

To set the default print handler, use the set_printerr_handler function.

unsetenv

uri_escape_string

uri_parse_scheme

uri_unescape_segment

uri_unescape_string

user_cache_dir

Returns a base directory in which to store non-essential, cached data specific to particular user.

user_config_dir

Returns a base directory in which to store user-specific application configuration information such as user preferences and settings.

user_data_dir

Returns a base directory in which to access application data such as icons that is customized for a particular user.

user_name

Gets the user name of the current user. The encoding of the returned string is system-defined. On UNIX, it might be the preferred file name encoding, or something else, and there is no guarantee that it is even consistent on a machine. On Windows, it is always UTF-8.

user_runtime_dir

Returns a directory that is unique to the current user on the local system.

user_special_dir

Returns the full path of a special directory using its logical id.

user_state_dirv2_72

Returns a base directory in which to store state files specific to particular user.

usleep

utf8_to_utf32

utf32_to_utf8

uuid_string_is_valid

uuid_string_random

Type Definitions

DateDay

DateYear

Time

Attribute Macros

flags

Attribute macro for defining flags using the bitflags crate. This macro will also define a GFlags::type_ function and the glib::Value traits.

object_interface

Macro for boilerplate of ObjectInterface implementations.

object_subclass

Macro for boilerplate of ObjectSubclass implementations.

Derive Macros

Boxed

Derive macro for defining a BoxedType::type_ function and the glib::Value traits. Optionally, the type can be marked as nullable to get an implemention of glib::value::ToValueOptional.

Downgrade

Macro for deriving implementations of glib::clone::Downgrade and glib::clone::Upgrade traits and a weak type.

Enum

Derive macro for register a rust enum in the glib type system and derive the the glib::Value traits.

ErrorDomain

Derive macro for defining a GLib error domain and its associated ErrorDomain trait.

Properties

This macro enables you to derive object properties in a quick way.

SharedBoxed

Derive macro for defining a SharedType::get_type function and the glib::Value traits. Optionally, the type can be marked as nullable to get an implemention of glib::value::ToValueOptional.

ValueDelegate

Example

Variant

Derive macro for serializing/deserializing custom structs/enums as glib::Variants.