Crate glib

Rust GLib and GObject bindings

Rust bindings and wrappers for GLib, part of gtk-rs-core.

GLib 2.56 is the lowest supported version for the underlying library.

This library contains bindings to GLib and GObject types and APIs as well as common building blocks used in both handmade and machine generated bindings to GTK and other GLib-based libraries.

It is the foundation for higher level libraries with uniform Rusty (safe and strongly typed) APIs. It avoids exposing GLib-specific data types where possible and is not meant to provide comprehensive GLib bindings, which would often amount to duplicating the Rust Standard Library or other utility crates.

Minimum supported Rust version

Currently, the minimum supported Rust version is 1.63.0.

Dynamic typing

Most types in the GLib family have Type identifiers. Their corresponding Rust types implement the StaticType trait.

A dynamically typed Value can carry values of any StaticType. Variants can carry values of StaticVariantType.

Errors

Errors are represented by Error, which can carry values from various error domains such as FileError.

Objects

Each class and interface has a corresponding smart pointer struct representing an instance of that type (e.g. Object for GObject or gtk::Widget for GtkWidget). They are reference counted and feature interior mutability similarly to Rust’s Rc<RefCell<T>> idiom. Consequently, cloning objects is cheap and their methods never require mutable borrows. Two smart pointers are equal if they point to the same object.

The root of the object hierarchy is Object. Inheritance and subtyping is denoted with the IsA marker trait. The Cast trait enables upcasting and downcasting.

Interfaces and non-leaf classes also have corresponding traits (e.g. ObjectExt or gtk::WidgetExt), which are blanketly implemented for all their subtypes.

You can create new subclasses of Object or other object types. Look at the module’s documentation for further details and a code example.

Under the hood

GLib-based libraries largely operate on pointers to various boxed or reference counted structures so the bindings have to implement corresponding smart pointers (wrappers), which encapsulate resource management and safety checks. Such wrappers are defined via the wrapper macro, which uses abstractions defined in the wrapper, boxed, shared and object modules.

The translate module defines and partly implements conversions between high level Rust types (including the aforementioned wrappers) and their FFI counterparts.

Documentation

Rust API - Stable

Rust API - Development

C API

GTK Installation instructions

Using

We recommend using crates from crates.io, as demonstrated here.

If you want to track the bleeding edge, use the git dependency instead:

[dependencies]
glib = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "glib" }

Avoid mixing versioned and git crates like this:

# This will not compile
[dependencies]
glib = "0.13"
glib = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "glib" }

License

glib is available under the MIT License, please refer to it.

Re-exports

pub use ffi;

pub use gobject_ffi;

pub use self::closure::Closure;

pub use self::closure::RustClosure;

pub use self::error::BoolError;

pub use self::error::Error;

pub use self::object::BorrowedObject;

pub use self::object::Cast;

pub use self::object::CastNone;

pub use self::object::Class;

pub use self::object::InitiallyUnowned;

pub use self::object::Interface;

pub use self::object::IsA;

pub use self::object::Object;

pub use self::object::ObjectExt;

pub use self::object::ObjectType;

pub use self::object::SendWeakRef;

pub use self::object::WeakRef;

pub use self::signal::signal_handler_block;

pub use self::signal::signal_handler_disconnect;

pub use self::signal::signal_handler_unblock;

pub use self::signal::signal_stop_emission_by_name;

pub use self::signal::SignalHandlerId;

pub use self::types::ILong;

pub use self::types::Pointer;

pub use self::types::StaticType;

pub use self::types::StaticTypeExt;

pub use self::types::Type;

pub use self::types::ULong;

pub use self::value::BoxedValue;

pub use self::value::SendValue;

pub use self::value::ToSendValue;

pub use self::value::ToValue;

pub use self::value::Value;

pub use self::variant::FixedSizeVariantArray;

pub use self::variant::FixedSizeVariantType;

pub use self::variant::FromVariant;

pub use self::variant::StaticVariantType;

pub use self::variant::ToVariant;

pub use self::variant::Variant;

pub use collections::List;

pub use collections::PtrSlice;

pub use collections::SList;

pub use collections::Slice;

pub use self::char::*;

pub use self::source::*;

Modules

boxed

IMPL Boxed wrapper implementation.

boxed_inline

IMPL BoxedInline wrapper implementation.

char

clone

closure

collections

error

Error binding and helper trait.

object

IMPL Object wrapper implementation and Object binding.

prelude

Traits and essential types intended for blanket imports.

shared

IMPL Shared (reference counted) wrapper implementation.

signal

IMPL Low level signal support.

source

subclass

Module containing infrastructure for subclassing GObjects and registering boxed types.

thread_guard

translate

Translation between GLib/GLib-based FFI types and their Rust counterparts.

types

Runtime type information.

value

Value binding and helper traits.

variant

Variant binding and helper traits.

wrapper

IMPL The wrapper! macro and miscellaneous wrapper traits.

Macros

bool_error

Generic error used for functions that fail without any further information

clone

Macro for passing variables as strong or weak references into a closure.

closure

Macro for creating a Closure object. This is a wrapper around Closure::new that automatically type checks its arguments at run-time.

closure_local

The same as closure! but uses Closure::new_local as a constructor. This is useful for closures which can’t be sent across threads. See the documentation of closure! for details.

debuglog_macros

A macro which behaves exactly as log::debug! except that it sets the current log target to the contents of a G_LOG_DOMAIN constant (and fails to build if not defined).

errorlog_macros

A macro which behaves exactly as log::error! except that it sets the current log target to the contents of a G_LOG_DOMAIN constant (and fails to build if not defined).

g_critical

Macro used to log using GLib logging system. It uses g_log.

g_debug

Macro used to log using GLib logging system. It uses g_log.

g_error

Macro used to log using GLib logging system. It uses g_log.

g_info

Macro used to log using GLib logging system. It uses g_log.

g_log

Macro used to log using GLib logging system. It uses g_log.

g_message

Macro used to log using GLib logging system. It uses g_log.

g_print

Macro used to print messages. It uses g_print.

g_printerr

Macro used to print error messages. It uses g_printerr.

g_warning

Macro used to log using GLib logging system. It uses g_log.

glib_boxed_inline_wrapper

Wrapper implementations for BoxedInline types. See wrapper!.

glib_boxed_wrapper

Wrapper implementations for Boxed types. See wrapper!.

glib_object_wrapper

ObjectType implementations for Object types. See wrapper!.

glib_shared_wrapper

Wrapper implementations for shared types. See wrapper!.

gstr

Converts a static string literal into a static nul-terminated string.

infolog_macros

A macro which behaves exactly as log::info! except that it sets the current log target to the contents of a G_LOG_DOMAIN constant (and fails to build if not defined).

log_structured

Macro used to log using GLib structured logging system.

result_from_gboolean

tracelog_macros

A macro which behaves exactly as log::trace! except that it sets the current log target to the contents of a G_LOG_DOMAIN constant (and fails to build if not defined).

warnlog_macros

A macro which behaves exactly as log::warn! except that it sets the current log target to the contents of a G_LOG_DOMAIN constant (and fails to build if not defined).

wrapper

Defines a wrapper type and implements the appropriate traits.

Structs

Binding

Binding is the representation of a binding between a property on a Object instance (or source) and another property on another Object instance (or target).

BindingFlags

Flags to be passed to ObjectExt::bind_property() or [ObjectExt::bind_property_full()][crate::prelude::ObjectExt::bind_property_full()].

BindingGroupv2_72

The BindingGroup can be used to bind multiple properties from an object collectively.

BindingGroupBuilderv2_72

Builder for binding group bindings.

BoxedAnyObject

This is a subclass of glib::object::Object capable of storing any Rust type. It let’s you insert a Rust type anywhere a glib::object::Object is needed. The inserted value can then be borrowed as a Rust type, by using the various provided methods.

ByteArray

Bytes

A shared immutable byte slice (the equivalent of Rc<[u8]>).

Checksum

CollationKey

A CollationKey allows ordering strings using the linguistically correct rules for the current locale.

Date

DateTime

EnumClass

Representation of an enum for dynamically, at runtime, querying the values of the enum and using them.

EnumValue

Representation of a single enum value of an EnumClass.

FileSetContentsFlagsv2_66

FilenameCollationKey

A FilenameCollationKey allows ordering file names using the linguistically correct rules for the current locale. Compared to CollationKey, filename collation keys take into consideration dots and other characters commonly found in file names.

FlagsBuilder

Builder for conveniently setting/unsetting flags and returning a Value.

FlagsClass

Representation of a flags for dynamically, at runtime, querying the values of the enum and using them

FlagsValue

Representation of a single flags value of a FlagsClass.

FormatSizeFlags

FutureWithTimeoutError

The error returned when a future times out.

GStr

Representation of a borrowed GString.

GString

A type representing an owned, C-compatible, nul-terminated UTF-8 string.

GStringBuilder

A mutable text buffer that grows automatically.

GlibLogger

An implementation of a log compatible logger which logs over glib logging facilities.

IConv

IOCondition

KeyFile

KeyFileFlags

LogField

Structure representing a single field in a structured log entry.

LogHandlerId

LogLevelFlags

LogLevels

MainContext

MainContextAcquireGuard

MainContextFlagsv2_72

MainLoop

MarkupParseContext

OptionFlags

ParamFlags

Through the ParamFlags flag values, certain aspects of parameters can be configured.

ParamSpec

ParamSpec is an object structure that encapsulates the metadata required to specify parameters, such as e.g. Object properties.

ParamSpecBoolean

ParamSpecBooleanBuilder

ParamSpecBoxed

ParamSpecBoxedBuilder

ParamSpecChar

ParamSpecCharBuilder

ParamSpecDouble

ParamSpecDoubleBuilder

ParamSpecEnum

ParamSpecEnumBuilder

ParamSpecFlags

ParamSpecFlagsBuilder

ParamSpecFloat

ParamSpecFloatBuilder

ParamSpecGType

ParamSpecGTypeBuilder

ParamSpecInt

ParamSpecInt64

ParamSpecInt64Builder

ParamSpecIntBuilder

ParamSpecLong

ParamSpecLongBuilder

ParamSpecObject

ParamSpecObjectBuilder

ParamSpecOverride

ParamSpecOverrideBuilder

ParamSpecParam

ParamSpecParamBuilder

ParamSpecPointer

ParamSpecPointerBuilder

ParamSpecString

ParamSpecStringBuilder

ParamSpecUChar

ParamSpecUCharBuilder

ParamSpecUInt

ParamSpecUInt64

ParamSpecUInt64Builder

ParamSpecUIntBuilder

ParamSpecULong

ParamSpecULongBuilder

ParamSpecUnichar

ParamSpecUnicharBuilder

ParamSpecValueArray

ParamSpecValueArrayBuilder

ParamSpecVariant

ParamSpecVariantBuilder

Quark

Receiver

A Receiver that can be attached to a main context to receive items from its corresponding Sender or SyncSender.

Sender

A Sender that can be used to send items to the corresponding main context receiver.

SignalFlags

The signal flags are used to specify a signal’s behaviour.

SignalGroupv2_74

SignalGroup manages to simplify the process of connecting many signals to a Object as a group. As such there is no API to disconnect a signal from the group.

Source

SourceFuture

Represents a Future around a glib::Source. The future will be resolved once the source has provided a value

SourceStream

Represents a Stream around a glib::Source. The stream will be provide all values that are provided by the source

SpawnFlags

SyncSender

A SyncSender that can be used to send items to the corresponding main context receiver.

ThreadPool

TimeSpan

A value representing an interval of time, in microseconds.

TimeZone

Uriv2_66

UriFlagsv2_66

Flags that describe a URI.

UriHideFlagsv2_66

Flags describing what parts of the URI to hide in Uri::to_string_partial(). Note that PASSWORD and AUTH_PARAMS will only work if the Uri was parsed with the corresponding flags.

UriParamsFlagsv2_66

Flags modifying the way parameters are handled by g_uri_parse_params() and GUriParamsIter.

ValueArray

VariantDict

VariantDict is a mutable key/value store where the keys are always strings and the values are Variants.

VariantIter

Iterator over items in a variant.

VariantStrIter

Iterator over items in a variant of type as.

VariantTy

Describes Variant types.

VariantTyIterator

An iterator over the individual components of a tuple VariantTy.

VariantType

Describes Variant types.

Enums

ChecksumType

ConvertError

CvtError

A wrapper for ConvertError that can hold an offset into the input string.

DateMonth

DateWeekday

FileError

GlibLoggerDomain

Enumeration of the possible domain handling behaviours for a GlibLogger.

GlibLoggerFormat

Enumeration of the possible formatting behaviours for a GlibLogger.

IConvError

A wrapper for std::io::Error that can hold an offset into an input string.

KeyFileError

LogLevel

LogWriterOutput

MarkupError

OptionArg

SeekType

TimeType

UnicodeScript

UriErrorv2_66

Error codes returned by Uri methods.

UserDirectory

These are logical ids for special directories which are defined depending on the platform used. You should use user_special_dir() to retrieve the full path associated to the logical id.

VariantClass

The range of possible top-level types of Variant instances.

Constants

CLONE_MACRO_LOG_DOMAIN

This is the log domain used by the clone! macro. If you want to use a custom logger (it prints to stdout by default), you can set your own logger using the corresponding log functions.

Statics

CSET_A_2_Z

CSET_DIGITS

CSET_a_2_z

KEY_FILE_DESKTOP_GROUP

KEY_FILE_DESKTOP_KEY_ACTIONS

KEY_FILE_DESKTOP_KEY_CATEGORIES

KEY_FILE_DESKTOP_KEY_COMMENT

KEY_FILE_DESKTOP_KEY_DBUS_ACTIVATABLE

KEY_FILE_DESKTOP_KEY_EXEC

KEY_FILE_DESKTOP_KEY_GENERIC_NAME

KEY_FILE_DESKTOP_KEY_HIDDEN

KEY_FILE_DESKTOP_KEY_ICON

KEY_FILE_DESKTOP_KEY_MIME_TYPE

KEY_FILE_DESKTOP_KEY_NAME

KEY_FILE_DESKTOP_KEY_NOT_SHOW_IN

KEY_FILE_DESKTOP_KEY_NO_DISPLAY

KEY_FILE_DESKTOP_KEY_ONLY_SHOW_IN

KEY_FILE_DESKTOP_KEY_PATH

KEY_FILE_DESKTOP_KEY_STARTUP_NOTIFY

KEY_FILE_DESKTOP_KEY_STARTUP_WM_CLASS

KEY_FILE_DESKTOP_KEY_TERMINAL

KEY_FILE_DESKTOP_KEY_TRY_EXEC

KEY_FILE_DESKTOP_KEY_TYPE

KEY_FILE_DESKTOP_KEY_URL

KEY_FILE_DESKTOP_KEY_VERSION

KEY_FILE_DESKTOP_TYPE_APPLICATION

KEY_FILE_DESKTOP_TYPE_DIRECTORY

KEY_FILE_DESKTOP_TYPE_LINK

OPTION_REMAINING

STR_DELIMITERS

TEST_OPTION_ISOLATE_DIRSv2_60

URI_RESERVED_CHARS_GENERIC_DELIMITERS

URI_RESERVED_CHARS_SUBCOMPONENT_DELIMITERS

g_param_spec_types

Traits

ParamSpecBuilderExt

A trait implemented by the various ParamSpec builder types.

ParamSpecType

Functions

access

application_name

base64_decode

base64_encode

charset

Obtain the character set for the current locale.

check_version

child_watch_future

Create a Future that will resolve once the child process with the given pid exits

child_watch_future_with_priority

Create a Future that will resolve once the child process with the given pid exits

codeset

compute_checksum_for_bytes

compute_checksum_for_data

compute_hmac_for_bytes

compute_hmac_for_data

console_charsetv2_62

convert

convert_with_fallback

current_dir

dcgettext

dgettext

dngettext

dpgettext

dpgettext2

environ

environ_getenv

file_get_contents

file_open_tmp

file_set_contents

file_set_contents_fullv2_66

filename_charsets

filename_display_basename

filename_display_name

filename_from_uri

filename_from_utf8

filename_to_uri

filename_to_utf8

find_program_in_path

format_size

format_size_full

future_with_timeout

Add a timeout to a Future.

future_with_timeout_with_priority

Add a timeout to a Future.

getenv

home_dir

host_name

hostname_is_ascii_encoded

hostname_is_ip_address

hostname_is_non_ascii

hostname_to_ascii

hostname_to_unicode

interval_stream

Create a Stream that will provide a value every given number of milliseconds.

interval_stream_seconds

Create a Stream that will provide a value every given number of seconds.

interval_stream_seconds_with_priority

Create a Stream that will provide a value every given number of seconds.

interval_stream_with_priority

Create a Stream that will provide a value every given number of milliseconds.

is_canonical_pspec_name

language_names

language_names_with_categoryv2_58

listenv

locale_from_utf8

locale_to_utf8

locale_variants

log_default_handler

log_remove_handler

log_set_always_fatal

log_set_default_handler

To set back the default print handler, use the log_unset_default_handler function.

log_set_fatal_mask

log_set_handler

log_set_writer_func

log_structured_array

log_unset_default_handler

To set the default print handler, use the log_set_default_handler function.

log_variant

log_writer_default

log_writer_default_set_use_stderr⚠v2_68

Sets whether GLib log functions output to stderr or stdout.

log_writer_default_would_dropv2_68

log_writer_format_fields

log_writer_is_journald

log_writer_journald

log_writer_standard_streams

log_writer_supports_color

main_current_source

main_depth

markup_escape_text

mkdir_with_parents

mkdtemp

mkdtemp_full

mkstemp

mkstemp_full

monotonic_time

num_processors

on_error_query

on_error_stack_trace

os_infov2_64

prgname

program_name

Same as get_prgname().

random_double

random_double_range

random_int

random_int_range

Returns a random gint32 equally distributed over the range [begin..end-1].

random_set_seed

Sets the seed for the global random number generator, which is used by the g_random_* functions, to seed.

real_name

real_time

reload_user_special_dirs_cache

Resets the cache used for user_special_dir(), so that the latest on-disk version is used. Call this only if you just changed the data on disk yourself.

rust_log_handler

Provides a glib log handler which routes all logging messages to the log crate.

set_application_name

Sets a human-readable name for the application. This name should be localized if possible, and is intended for display to the user. Contrast with g_set_prgname(), which sets a non-localized name. g_set_prgname() will be called automatically by gtk_init(), but set_application_name() will not.

set_prgname

set_print_handler

To set back the default print handler, use the unset_print_handler function.

set_printerr_handler

To set back the default print handler, use the unset_printerr_handler function.

set_program_name

Same as set_prgname().

setenv

Sets an environment variable. On UNIX, both the variable’s name and value can be arbitrary byte strings, except that the variable’s name cannot contain ‘=’. On Windows, they should be in UTF-8.

shell_parse_argv

Parses a command line into an argument vector, in much the same way the shell would, but without many of the expansions the shell would perform (variable expansion, globs, operators, filename expansion, etc. are not supported).

shell_quote

Quotes a string so that the shell (/bin/sh) will interpret the quoted string to mean unquoted_string.

shell_unquote

Unquotes a string as the shell (/bin/sh) would.

spaced_primes_closest

Gets the smallest prime number from a built-in array of primes which is larger than num. This is used within GLib to calculate the optimum size of a GHashTable.

spawn_async

Executes a child program asynchronously.

spawn_async_with_fdsv2_58 and non-Windows

spawn_async_with_pipesNon-Windows

spawn_check_exit_status

spawn_check_wait_statusv2_70

spawn_command_line_asyncUnix

system_config_dirs

system_data_dirs

timeout_future

Create a Future that will resolve after the given number of milliseconds.

timeout_future_seconds

Create a Future that will resolve after the given number of seconds.

timeout_future_seconds_with_priority

Create a Future that will resolve after the given number of seconds.

timeout_future_with_priority

Create a Future that will resolve after the given number of milliseconds.

tmp_dir

unicode_script_from_iso15924

unicode_script_to_iso15924

unix_open_pipe

unix_signal_futureUnix

Create a Future that will resolve once the given UNIX signal is raised

unix_signal_future_with_priorityUnix

Create a Future that will resolve once the given UNIX signal is raised

unix_signal_streamUnix

Create a Stream that will provide a value whenever the given UNIX signal is raised

unix_signal_stream_with_priorityUnix

Create a Stream that will provide a value whenever the given UNIX signal is raised

unlink

unset_print_handler

To set the default print handler, use the set_print_handler function.

unset_printerr_handler

To set the default print handler, use the set_printerr_handler function.

unsetenv

uri_escape_string

uri_parse_scheme

uri_unescape_segment

uri_unescape_string

user_cache_dir

user_config_dir

user_data_dir

user_name

user_runtime_dir

user_special_dir

user_state_dirv2_72

usleep

uuid_string_is_valid

uuid_string_random

Type Definitions

DateDay

DateYear

Time

Attribute Macros

flags

Attribute macro for defining flags using the bitflags crate. This macro will also define a GFlags::type_ function and the glib::Value traits.

object_interface

Macro for boilerplate of ObjectInterface implementations.

object_subclass

Macro for boilerplate of ObjectSubclass implementations.

Derive Macros

Boxed

Derive macro for defining a BoxedType::type_ function and the glib::Value traits.

Downgrade

Macro for deriving implementations of glib::clone::Downgrade and glib::clone::Upgrade traits and a weak type.

Enum

Derive macro for register a rust enum in the glib type system and derive the the glib::Value traits.

ErrorDomain

Derive macro for defining a GLib error domain and its associated ErrorDomain trait.

SharedBoxed

Derive macro for defining a SharedType::get_type function and the glib::Value traits.

Variant

Derive macro for serializing/deserializing custom structs/enums as glib::Variants.