Crate glib

Bindings and wrappers for GLib

Reexports

pub use self::glib_container::GlibContainer;

pub use self::source::{Continue, idle_add, timeout_add, timeout_add_seconds};

pub use self::traits::FFIGObject;

pub use types::Type;

pub use self::date::{TimeVal, Time, Date, Year, Month, Weekday, Day};

Modules

date
Date and Time Functions — calendrical calculations and miscellaneous time stuff

ffi

glib_container

object
Types that facilitate representing GObject descendants.

signal

source
Manages available sources of events for the main loop

traits

translate
Translation between GLib/GLib-based FFI types and their Rust counterparts.

types

Structs

Elem

Error
The Error structure contains information about an error that has occurred.

List

ParamSpec

Permission

RevElem

SElem

SList

TypeClass

TypeInstance

TypeInterface

Value
Possible improvement : store a function pointer inside the struct and make the struct templated

Enums

ParamFlags

Traits

ValuePublic

Functions

to_bool

to_gboolean

Type Definitions

AppInfo