glib::clone

Trait Upgrade

pub trait Upgrade
where
    Self: Sized,
{
    type Strong;

    // Required method
    fn upgrade(&self) -> Option<Self::Strong>;
}

Trait for generalizing upgrading a weak reference to a strong reference.

Required Associated Types

type Strong

Strong reference type.

Required Methods

fn upgrade(&self) -> Option<Self::Strong>

Try upgrading a weak reference to a strong reference.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.