Trait glib::clone::Upgrade

pub trait Upgrade where
    Self: Sized, {
    type Strong;
    fn upgrade(&self) -> Option<Self::Strong>;
}

Trait for generalizing upgrading a weak reference to a strong reference.

Associated Types

type Strong

Strong reference type.

Required methods

fn upgrade(&self) -> Option<Self::Strong>

Try upgrading a weak reference to a strong reference.