Enum glib::UnicodeScript

#[non_exhaustive]
pub enum UnicodeScript {

    InvalidCode,
    Common,
    Inherited,
    Arabic,
    Armenian,
    Bengali,
    Bopomofo,
    Cherokee,
    Coptic,
    Cyrillic,
    Deseret,
    Devanagari,
    Ethiopic,
    Georgian,
    Gothic,
    Greek,
    Gujarati,
    Gurmukhi,
    Han,
    Hangul,
    Hebrew,
    Hiragana,
    Kannada,
    Katakana,
    Khmer,
    Lao,
    Latin,
    Malayalam,
    Mongolian,
    Myanmar,
    Ogham,
    OldItalic,
    Oriya,
    Runic,
    Sinhala,
    Syriac,
    Tamil,
    Telugu,
    Thaana,
    Thai,
    Tibetan,
    CanadianAboriginal,
    Yi,
    Tagalog,
    Hanunoo,
    Buhid,
    Tagbanwa,
    Braille,
    Cypriot,
    Limbu,
    Osmanya,
    Shavian,
    LinearB,
    TaiLe,
    Ugaritic,
    NewTaiLue,
    Buginese,
    Glagolitic,
    Tifinagh,
    SylotiNagri,
    OldPersian,
    Kharoshthi,
    Unknown,
    Balinese,
    Cuneiform,
    Phoenician,
    PhagsPa,
    Nko,
    KayahLi,
    Lepcha,
    Rejang,
    Sundanese,
    Saurashtra,
    Cham,
    OlChiki,
    Vai,
    Carian,
    Lycian,
    Lydian,
    Avestan,
    Bamum,
    EgyptianHieroglyphs,
    ImperialAramaic,
    InscriptionalPahlavi,
    InscriptionalParthian,
    Javanese,
    Kaithi,
    Lisu,
    MeeteiMayek,
    OldSouthArabian,
    OldTurkic,
    Samaritan,
    TaiTham,
    TaiViet,
    Batak,
    Brahmi,
    Mandaic,
    Chakma,
    MeroiticCursive,
    MeroiticHieroglyphs,
    Miao,
    Sharada,
    SoraSompeng,
    Takri,
    BassaVah,
    CaucasianAlbanian,
    Duployan,
    Elbasan,
    Grantha,
    Khojki,
    Khudawadi,
    LinearA,
    Mahajani,
    Manichaean,
    MendeKikakui,
    Modi,
    Mro,
    Nabataean,
    OldNorthArabian,
    OldPermic,
    PahawhHmong,
    Palmyrene,
    PauCinHau,
    PsalterPahlavi,
    Siddham,
    Tirhuta,
    WarangCiti,
    Ahom,
    AnatolianHieroglyphs,
    Hatran,
    Multani,
    OldHungarian,
    Signwriting,
    Adlam,
    Bhaiksuki,
    Marchen,
    Newa,
    Osage,
    Tangut,
    MasaramGondi,
    Nushu,
    Soyombo,
    ZanabazarSquare,
    Dogra,
    GunjalaGondi,
    HanifiRohingya,
    Makasar,
    Medefaidrin,
    OldSogdian,
    Sogdian,
    Elymaic,
    Nandinagari,
    NyiakengPuachueHmong,
    Wancho,
    Chorasmian,
    DivesAkuru,
    KhitanSmallScript,
    Yezidi,
    CyproMinoan,
    OldUyghur,
    Tangsa,
    Toto,
    Vithkuqi,
    Math,

}

The UnicodeScript enumeration identifies different writing systems. The values correspond to the names as defined in the Unicode standard. The enumeration has been added in GLib 2.14, and is interchangeable with PangoScript.

Note that new types may be added in the future. Applications should be ready to handle unknown values. See Unicode Standard Annex 24: Script names.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

InvalidCode

a value never returned from g_unichar_get_script()

Common

a character used by multiple different scripts

Inherited

a mark glyph that takes its script from the base glyph to which it is attached

Arabic

Arabic

Armenian

Armenian

Bengali

Bengali

Bopomofo

Bopomofo

Cherokee

Cherokee

Coptic

Coptic

Cyrillic

Cyrillic

Deseret

Deseret

Devanagari

Devanagari

Ethiopic

Ethiopic

Georgian

Georgian

Gothic

Gothic

Greek

Greek

Gujarati

Gujarati

Gurmukhi

Gurmukhi

Han

Han

Hangul

Hangul

Hebrew

Hebrew

Hiragana

Hiragana

Kannada

Kannada

Katakana

Katakana

Khmer

Khmer

Lao

Lao

Latin

Latin

Malayalam

Malayalam

Mongolian

Mongolian

Myanmar

Myanmar

Ogham

Ogham

OldItalic

Old Italic

Oriya

Oriya

Runic

Runic

Sinhala

Sinhala

Syriac

Syriac

Tamil

Tamil

Telugu

Telugu

Thaana

Thaana

Thai

Thai

Tibetan

Tibetan

CanadianAboriginal

Canadian Aboriginal

Yi

Yi

Tagalog

Tagalog

Hanunoo

Hanunoo

Buhid

Buhid

Tagbanwa

Tagbanwa

Braille

Braille

Cypriot

Cypriot

Limbu

Limbu

Osmanya

Osmanya

Shavian

Shavian

LinearB

Linear B

TaiLe

Tai Le

Ugaritic

Ugaritic

NewTaiLue

New Tai Lue

Buginese

Buginese

Glagolitic

Glagolitic

Tifinagh

Tifinagh

SylotiNagri

Syloti Nagri

OldPersian

Old Persian

Kharoshthi

Kharoshthi

Unknown

an unassigned code point

Balinese

Balinese

Cuneiform

Cuneiform

Phoenician

Phoenician

PhagsPa

Phags-pa

Nko

N’Ko

KayahLi

Kayah Li. Since 2.16.3

Lepcha

Lepcha. Since 2.16.3

Rejang

Rejang. Since 2.16.3

Sundanese

Sundanese. Since 2.16.3

Saurashtra

Saurashtra. Since 2.16.3

Cham

Cham. Since 2.16.3

OlChiki

Ol Chiki. Since 2.16.3

Vai

Vai. Since 2.16.3

Carian

Carian. Since 2.16.3

Lycian

Lycian. Since 2.16.3

Lydian

Lydian. Since 2.16.3

Avestan

Avestan. Since 2.26

Bamum

Bamum. Since 2.26

EgyptianHieroglyphs

Egyptian Hieroglpyhs. Since 2.26

ImperialAramaic

Imperial Aramaic. Since 2.26

InscriptionalPahlavi

Inscriptional Pahlavi. Since 2.26

InscriptionalParthian

Inscriptional Parthian. Since 2.26

Javanese

Javanese. Since 2.26

Kaithi

Kaithi. Since 2.26

Lisu

Lisu. Since 2.26

MeeteiMayek

Meetei Mayek. Since 2.26

OldSouthArabian

Old South Arabian. Since 2.26

OldTurkic

Old Turkic. Since 2.28

Samaritan

Samaritan. Since 2.26

TaiTham

Tai Tham. Since 2.26

TaiViet

Tai Viet. Since 2.26

Batak

Batak. Since 2.28

Brahmi

Brahmi. Since 2.28

Mandaic

Mandaic. Since 2.28

Chakma

Chakma. Since: 2.32

MeroiticCursive

Meroitic Cursive. Since: 2.32

MeroiticHieroglyphs

Meroitic Hieroglyphs. Since: 2.32

Miao

Miao. Since: 2.32

Sharada

Sharada. Since: 2.32

SoraSompeng

Sora Sompeng. Since: 2.32

Takri

Takri. Since: 2.32

BassaVah

Bassa. Since: 2.42

CaucasianAlbanian

Caucasian Albanian. Since: 2.42

Duployan

Duployan. Since: 2.42

Elbasan

Elbasan. Since: 2.42

Grantha

Grantha. Since: 2.42

Khojki

Kjohki. Since: 2.42

Khudawadi

Khudawadi, Sindhi. Since: 2.42

LinearA

Linear A. Since: 2.42

Mahajani

Mahajani. Since: 2.42

Manichaean

Manichaean. Since: 2.42

MendeKikakui

Mende Kikakui. Since: 2.42

Modi

Modi. Since: 2.42

Mro

Mro. Since: 2.42

Nabataean

Nabataean. Since: 2.42

OldNorthArabian

Old North Arabian. Since: 2.42

OldPermic

Old Permic. Since: 2.42

PahawhHmong

Pahawh Hmong. Since: 2.42

Palmyrene

Palmyrene. Since: 2.42

PauCinHau

Pau Cin Hau. Since: 2.42

PsalterPahlavi

Psalter Pahlavi. Since: 2.42

Siddham

Siddham. Since: 2.42

Tirhuta

Tirhuta. Since: 2.42

WarangCiti

Warang Citi. Since: 2.42

Ahom

Ahom. Since: 2.48

AnatolianHieroglyphs

Anatolian Hieroglyphs. Since: 2.48

Hatran

Hatran. Since: 2.48

Multani

Multani. Since: 2.48

OldHungarian

Old Hungarian. Since: 2.48

Signwriting

Signwriting. Since: 2.48

Adlam

Adlam. Since: 2.50

Bhaiksuki

Bhaiksuki. Since: 2.50

Marchen

Marchen. Since: 2.50

Newa

Newa. Since: 2.50

Osage

Osage. Since: 2.50

Tangut

Tangut. Since: 2.50

MasaramGondi

Masaram Gondi. Since: 2.54

Nushu

Nushu. Since: 2.54

Soyombo

Soyombo. Since: 2.54

ZanabazarSquare

Zanabazar Square. Since: 2.54

Dogra

Dogra. Since: 2.58

GunjalaGondi

Gunjala Gondi. Since: 2.58

HanifiRohingya

Hanifi Rohingya. Since: 2.58

Makasar

Makasar. Since: 2.58

Medefaidrin

Medefaidrin. Since: 2.58

OldSogdian

Old Sogdian. Since: 2.58

Sogdian

Sogdian. Since: 2.58

Elymaic

Elym. Since: 2.62

Nandinagari

Nand. Since: 2.62

NyiakengPuachueHmong

Rohg. Since: 2.62

Wancho

Wcho. Since: 2.62

Chorasmian

Chorasmian. Since: 2.66

DivesAkuru

Dives Akuru. Since: 2.66

KhitanSmallScript

Khitan small script. Since: 2.66

Yezidi

Yezidi. Since: 2.66

CyproMinoan

Available on crate feature v2_72 only.

Cypro-Minoan. Since: 2.72

OldUyghur

Available on crate feature v2_72 only.

Old Uyghur. Since: 2.72

Tangsa

Available on crate feature v2_72 only.

Tangsa. Since: 2.72

Toto

Available on crate feature v2_72 only.

Toto. Since: 2.72

Vithkuqi

Available on crate feature v2_72 only.

Vithkuqi. Since: 2.72

Math

Available on crate feature v2_72 only.

Mathematical notation. Since: 2.72