Module glib::enums

Structs

EnumClass

Representation of an enum for dynamically, at runtime, querying the values of the enum and using them.

EnumTypeChecker

EnumValue

Representation of a single enum value of an EnumClass. A structure which contains a single enum value, its name, and its nickname.

EnumerationValues

Representation of enumeration values wrapped by EnumerationValuesStorage. Easier to use because don’t have a size parameter to be specify. Should be used only to register an enum or flags as a dynamic type. see TypePluginRegisterImpl::register_dynamic_enum(), TypePluginRegisterImpl::register_dynamic_flags() and TypePluginImpl::complete_type_info(). Field is intentionally private to ensure other modules will not access the enum (or flags) values by this way. Use EnumClass::values() or EnumClass::value() to get the enum values. Use FlagsClass::values() or FlagsClass::value() to get the flags values.

EnumerationValuesStorage

Storage of enumeration values terminated by a zero value. Should be used only as a storage location for EnumValue or FlagsValue when registering an enum or flags as a dynamic type. see TypePluginRegisterImpl::register_dynamic_enum(), TypePluginRegisterImpl::register_dynamic_flags() and TypePluginImpl::complete_type_info(). Inner is intentionally private to ensure other modules will not access the enum (or flags) values by this way. Use EnumClass::values() or EnumClass::value() to get the enum values. Use FlagsClass::values() or FlagsClass::value() to get the flags values.

FlagsBuilder

Builder for conveniently setting/unsetting flags and returning a Value.

FlagsClass

Representation of a flags for dynamically, at runtime, querying the values of the enum and using them

FlagsTypeChecker

FlagsValue

Representation of a single flags value of a FlagsClass. A structure which contains a single flags value, its name, and its nickname.

InvalidEnumError

An error returned from the get function on a Value for enum types.

InvalidFlagsError

An error returned from the get function on a Value for flags types.

ParseFlagsError

Enums

UserDirectory

These are logical ids for special directories which are defined depending on the platform used. You should use g_get_user_special_dir() to retrieve the full path associated to the logical id.

Traits

EnumerationValue

helper trait to define the zero value and the associated GLib type.

Type Aliases

EnumValues

Representation of enum values wrapped by EnumValuesStorage

EnumValuesStorage

Storage of enum values.

FlagsValues

Representation of flags values wrapped by FlagsValuesStorage

FlagsValuesStorage

Storage of flags values.