Struct glib::error::Error

pub struct Error { /* private fields */ }

A generic error capable of representing various error domains (types). The GError structure contains information about an error that has occurred.

GLib type: Boxed type with copy-on-clone semantics.

Implementations

impl Error

pub fn as_ptr(&self) -> *mut GError

Return the inner pointer to the underlying C value.

pub unsafe fn from_glib_ptr_borrow(ptr: &*mut GError) -> &Self

Borrows the underlying C value.

pub unsafe fn from_glib_ptr_borrow_mut(ptr: &mut *mut GError) -> &mut Self

Borrows the underlying C value mutably.

impl Error

pub fn new<T: ErrorDomain>(error: T, message: &str) -> Error

Creates an error with supplied error enum variant and message. Creates a new #GError with the given @domain and @code, and a message formatted with @format.

domain

error domain

code

error code

format

printf()-style format for error message

Returns

a new #GError

pub fn is<T: ErrorDomain>(&self) -> bool

Checks if the error domain matches T.

pub fn domain(&self) -> Quark

Returns the error domain quark

pub fn matches<T: ErrorDomain>(&self, err: T) -> bool

Checks if the error matches the specified domain and error code. Returns true if @self matches @domain and @code, false otherwise. In particular, when @self is None, false will be returned.

If @domain contains a FAILED (or otherwise generic) error code, you should generally not check for it explicitly, but should instead treat any not-explicitly-recognized error code as being equivalent to the FAILED code. This way, if the domain is extended in the future to provide a more specific error code for a certain case, your code will still work.

domain

an error domain

code

an error code

Returns

whether @self has @domain and @code

pub fn kind<T: ErrorDomain>(&self) -> Option<T>

Tries to convert to a specific error enum.

Returns Some if the error belongs to the enumโs error domain and None otherwise.

Examples

if let Some(file_error) = error.kind::<FileError>() {
    match file_error {
        FileError::Exist => ...
        FileError::Isdir => ...
        ...
    }
}

pub fn message(&self) -> &str

Returns the error message

Most of the time you can simply print the error since it implements the Display trait, but you can use this method if you need to have the message as a &str.