glib::error

Trait ErrorDomain

pub trait ErrorDomain: Copy {
    // Required methods
    fn domain() -> Quark;
    fn code(self) -> i32;
    fn from(code: i32) -> Option<Self>
       where Self: Sized;
}

GLib error domain.

This trait is implemented by error enums that represent error domains (types).

Required Methods

fn domain() -> Quark

Returns the quark identifying the error domain.

As returned from g_some_error_quark.

fn code(self) -> i32

Gets the integer representation of the variant.

fn from(code: i32) -> Option<Self>
where Self: Sized,

Tries to convert an integer code to an enum variant.

By convention, the Failed variant, if present, is a catch-all, i.e. any unrecognized codes map to it.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.