glib

Function compute_hmac_for_bytes

pub fn compute_hmac_for_bytes(
    digest_type: ChecksumType,
    key: &Bytes,
    data: &Bytes,
) -> GString

Computes the HMAC for a binary @data. This is a convenience wrapper for g_hmac_new(), g_hmac_get_string() and g_hmac_unref().

The hexadecimal string returned will be in lower case.

digest_type

a #GChecksumType to use for the HMAC

key

the key to use in the HMAC

data

binary blob to compute the HMAC of

Returns

the HMAC of the binary data as a string in hexadecimal. The returned string should be freed with g_free() when done using it. Computes the HMAC for a binary @data. This is a convenience wrapper for g_hmac_new(), g_hmac_get_string() and g_hmac_unref().

The hexadecimal string returned will be in lower case.

digest_type

a #GChecksumType to use for the HMAC

key

the key to use in the HMAC

data

binary blob to compute the HMAC of

Returns

the HMAC of the binary data as a string in hexadecimal. The returned string should be freed with g_free() when done using it.