Function glib::dpgettext

pub fn dpgettext(
    domain: Option<&str>,
    msgctxtid: &str,
    msgidoffset: usize,
) -> GString

This function is a variant of g_dgettext() which supports a disambiguating message context. GNU gettext uses the ‘\004’ character to separate the message context and message id in @msgctxtid. If 0 is passed as @msgidoffset, this function will fall back to trying to use the deprecated convention of using “|” as a separation character.

This uses g_dgettext() internally. See that functions for differences with dgettext() proper.

Applications should normally not use this function directly, but use the C_() macro for translations with context.

domain

the translation domain to use, or None to use the domain set with textdomain()

msgctxtid

a combined message context and message id, separated by a \004 character

msgidoffset

the offset of the message id in @msgctxid

Returns

The translated string