glib
Function file_set_contents_full

pub fn file_set_contents_full(
    filename: impl AsRef<Path>,
    contents: &[u8],
    flags: FileSetContentsFlags,
    mode: i32,
) -> Result<(), Error>

Available on crate feature v2_66 only.

Writes all of @contents to a file named @filename, with good error checking. If a file called @filename already exists it will be overwritten.

@flags control the properties of the write operation: whether it’s atomic, and what the tradeoff is between returning quickly or being resilient to system crashes.

As this function performs file I/O, it is recommended to not call it anywhere where blocking would cause problems, such as in the main loop of a graphical application. In particular, if @flags has any value other than FileSetContentsFlags::NONE then this function may call fsync().

If FileSetContentsFlags::CONSISTENT is set in @flags, the operation is atomic in the sense that it is first written to a temporary file which is then renamed to the final name.

Notes:

On UNIX, if @filename already exists hard links to @filename will break. Also since the file is recreated, existing permissions, access control lists, metadata etc. may be lost. If @filename is a symbolic link, the link itself will be replaced, not the linked file.

On UNIX, if @filename already exists and is non-empty, and if the system supports it (via a journalling filesystem or equivalent), and if FileSetContentsFlags::CONSISTENT is set in @flags, the fsync() call (or equivalent) will be used to ensure atomic replacement: @filename will contain either its old contents or @contents, even in the face of system power loss, the disk being unsafely removed, etc.

On UNIX, if @filename does not already exist or is empty, there is a possibility that system power loss etc. after calling this function will leave @filename empty or full of NUL bytes, depending on the underlying filesystem, unless FileSetContentsFlags::DURABLE and FileSetContentsFlags::CONSISTENT are set in @flags.

On Windows renaming a file will not remove an existing file with the new name, so on Windows there is a race condition between the existing file being removed and the temporary file being renamed.

On Windows there is no way to remove a file that is open to some process, or mapped into memory. Thus, this function will fail if @filename already exists and is open.

If the call was successful, it returns true. If the call was not successful, it returns false and sets @error. The error domain is G_FILE_ERROR. Possible error codes are those in the #GFileError enumeration.

Note that the name for the temporary file is constructed by appending up to 7 characters to @filename.

If the file didn’t exist before and is created, it will be given the permissions from @mode. Otherwise, the permissions of the existing file may be chanComplete version of [bind_property()][Self::bind_property()].

Creates a binding between source_property on self and target_property on target, allowing you to set the transformation functions to be used by the binding.

If flags contains BindingFlags::BIDIRECTIONAL then the binding will be mutual: if target_property on target changes then the source_property on self will be updated as well. The transform_from function is only used in case of bidirectional bindings, otherwise it will be ignored

The binding will automatically be removed when either the self or the target instances are finalized. This will release the reference that is being held on the Binding instance; if you want to hold on to the Binding instance, you will need to hold a reference to it.

To remove the binding, call Binding::unbind().

A Object can have multiple bindings.

The same user_data parameter will be used for both transform_to and transform_from transformation functions; the notify function will be called once, when the binding is removed. If you need different data for each transformation function, please use [bind_property_with_closures()][Self::bind_property_with_closures()] instead.

source_property

the property on self to bind

target

the target Object

target_property

the property on target to bind

flags

flags to pass to Binding

transform_to

the transformation function from the self to the target, or None to use the default

transform_from

the transformation function from the target to the self, or None to use the default

notify

a function to call when disposing the binding, to free resources used by the transformation functions, or None if not required

Returns

the Binding instance representing the binding between the two Object instances. The binding is released whenever the Binding reference count reaches zero. Writes all of @contents to a file named @filename, with good error checking. If a file called @filename already exists it will be overwritten.

@flags control the properties of the write operation: whether it’s atomic, and what the tradeoff is between returning quickly or being resilient to system crashes.

As this function performs file I/O, it is recommended to not call it anywhere where blocking would cause problems, such as in the main loop of a graphical application. In particular, if @flags has any value other than FileSetContentsFlags::NONE then this function may call fsync().

If FileSetContentsFlags::CONSISTENT is set in @flags, the operation is atomic in the sense that it is first written to a temporary file which is then renamed to the final name.

Notes:

On UNIX, if @filename already exists hard links to @filename will break. Also since the file is recreated, existing permissions, access control lists, metadata etc. may be lost. If @filename is a symbolic link, the link itself will be replaced, not the linked file.

On UNIX, if @filename already exists and is non-empty, and if the system supports it (via a journalling filesystem or equivalent), and if FileSetContentsFlags::CONSISTENT is set in @flags, the fsync() call (or equivalent) will be used to ensure atomic replacement: @filename will contain either its old contents or @contents, even in the face of system power loss, the disk being unsafely removed, etc.

On UNIX, if @filename does not already exist or is empty, there is a possibility that system power loss etc. after calling this function will leave @filename empty or full of NUL bytes, depending on the underlying filesystem, unless FileSetContentsFlags::DURABLE and FileSetContentsFlags::CONSISTENT are set in @flags.

On Windows renaming a file will not remove an existing file with the new name, so on Windows there is a race condition between the existing file being removed and the temporary file being renamed.

On Windows there is no way to remove a file that is open to some process, or mapped into memory. Thus, this function will fail if @filename already exists and is open.

If the call was successful, it returns true. If the call was not successful, it returns false and sets @error. The error domain is G_FILE_ERROR. Possible error codes are those in the #GFileError enumeration.

Note that the name for the temporary file is constructed by appending up to 7 characters to @filename.

If the file didn’t exist before and is created, it will be given the permissions from @mode. Otherwise, the permissions of the existing file may be chanComplete version of [bind_property()][Self::bind_property()].

Creates a binding between source_property on self and target_property on target, allowing you to set the transformation functions to be used by the binding.

If flags contains BindingFlags::BIDIRECTIONAL then the binding will be mutual: if target_property on target changes then the source_property on self will be updated as well. The transform_from function is only used in case of bidirectional bindings, otherwise it will be ignored

The binding will automatically be removed when either the self or the target instances are finalized. This will release the reference that is being held on the Binding instance; if you want to hold on to the Binding instance, you will need to hold a reference to it.

To remove the binding, call Binding::unbind().

A Object can have multiple bindings.

The same user_data parameter will be used for both transform_to and transform_from transformation functions; the notify function will be called once, when the binding is removed. If you need different data for each transformation function, please use [bind_property_with_closures()][Self::bind_property_with_closures()] instead.

source_property

the property on self to bind

target

the target Object

target_property

the property on target to bind

flags

flags to pass to Binding

transform_to

the transformation function from the self to the target, or None to use the default

transform_from

the transformation function from the target to the self, or None to use the default

notify

a function to call when disposing the binding, to free resources used by the transformation functions, or None if not required

Returns

the Binding instance representing the binding between the two Object instances. The binding is released whenever the Binding reference count reaches zero.