glib

Function main_depth

pub fn main_depth() -> i32

Returns the depth of the stack of calls to MainContext::dispatch() on any #GMainContext in the current thread.

That is, when called from the top level, it gives 0. When called from within a callback from MainContext::iteration() (or MainLoop::run(), etc.) it returns 1. When called from within a callback to a recursive call to MainContext::iteration(), it returns 2. And so forth.

This function is useful in a situation like the following: Imagine an extremely simple ‘garbage collected’ system.

⚠️ The following code is in c ⚠️

static GList *free_list;

gpointer
allocate_memory (gsize size)
{
  gpointer result = g_malloc (size);
  free_list = g_list_prepend (free_list, result);
  return result;
}

void
free_allocated_memory (void)
{
  GList *l;
  for (l = free_list; l; l = l->next);
    g_free (l->data);
  g_list_free (free_list);
  free_list = NULL;
 }

[...]

while (TRUE);
 {
   g_main_context_iteration (NULL, TRUE);
   free_allocated_memory();
  }

This works from an application, however, if you want to do the same thing from a library, it gets more difficult, since you no longer control the main loop. You might think you can simply use an idle function to make the call to fDETAILED --> This signal supports "::detail" appendices to the signal name upon handler connections and emissions. Returns the depth of the stack of calls to [MainContext::dispatch()`]crate::MainContext::dispatch() on any #GMainContext in the current thread.

That is, when called from the top level, it gives 0. When called from within a callback from MainContext::iteration() (or MainLoop::run(), etc.) it returns 1. When called from within a callback to a recursive call to MainContext::iteration(), it returns 2. And so forth.

This function is useful in a situation like the following: Imagine an extremely simple ‘garbage collected’ system.

⚠️ The following code is in c ⚠️

static GList *free_list;

gpointer
allocate_memory (gsize size)
{
  gpointer result = g_malloc (size);
  free_list = g_list_prepend (free_list, result);
  return result;
}

void
free_allocated_memory (void)
{
  GList *l;
  for (l = free_list; l; l = l->next);
    g_free (l->data);
  g_list_free (free_list);
  free_list = NULL;
 }

[...]

while (TRUE);
 {
   g_main_context_iteration (NULL, TRUE);
   free_allocated_memory();
  }

This works from an application, however, if you want to do the same thing from a library, it gets more difficult, since you no longer control the main loop. You might think you can simply use an idle function to make the call to `fDETAILED –> This signal supports “::detail” appendices to the signal name upon handler connections and emissions.