Function glib::on_error_stack_trace

pub fn on_error_stack_trace(prg_name: &str)

Invokes gdb, which attaches to the current process and shows a stack trace. Called by on_error_query() when the “[S]tack trace” option is selected. You can get the current process’s program name with g_get_prgname(), assuming that you have called gtk_init() or gdk_init().

This function may cause different actions on non-UNIX platforms.

When running on Windows, this function is not called by on_error_query(). If called directly, it will raise an exception, which will crash the program. If the G_DEBUGGER environment variable is set, a debugger will be invoked to attach and handle that exception (see Running GLib Applications).

prg_name

the program name, needed by gdb for the “[S]tack trace” option