glib

Function rust_log_handler

pub fn rust_log_handler(domain: Option<&str>, level: LogLevel, message: &str)

Available on crate feature log only.

Provides a glib log handler which routes all logging messages to the log crate.

In order to use this function, glib must be built with the log feature enabled.

Use this function if you want to use the log crate as the main logging output in your application, and want to route all logging happening in glib to the log crate. If you want the opposite, use GlibLogger.

NOTE: This should never be used when GlibLogger is registered as a logger, otherwise a stack overflow will occur.

glib::log_set_default_handler(glib::rust_log_handler);