glib

Function setenv

pub unsafe fn setenv(
    variable: impl AsRef<OsStr>,
    value: impl AsRef<OsStr>,
    overwrite: bool,
) -> Result<(), BoolError>

Sets an environment variable. On UNIX, both the variable’s name and value can be arbitrary byte strings, except that the variable’s name cannot contain ‘=’. On Windows, they should be in UTF-8.

Note that on some systems, when variables are overwritteAn interface that handles the lifecycle of dynamically loaded types.

The GObject type system supports dynamic loading of types. It goes as follows:

The type is initially introduced (usually upon loading the module the first time, or by your main application that knows what modules introduces what types), like this: ⚠️ The following code is in c ⚠️

   new_type_id = g_type_register_dynamic (parent_type_id,
                                          "TypeName",
                                          new_type_plugin,
                                          type_flags);

where new_type_plugin is an implementation of the GTypePlugin interface.

The type’s implementation is referenced, e.g. through [funcObject.ref] or through [funcObject] (this is being called by [ctorObject.new]) or through one of the above done on a type derived from new_type_id.

This causes the type system to load the type’s implementation by calling [methodObject.use] and [methodObject.complete_type_info] on new_type_plugin.

At some point the type’s implementation isn’t required anymore, e.g. after [methodObject.unref] or [funcObject] (called when the reference count of an instance drops to zero).

This causes the type system to throw away the information retrieved from [methodObject.complete_type_info] and then it calls [methodObject.unuse] on new_type_plugin.

Things may repeat from the second step.

So basically, you need to implement a GTypePlugin type that carries a use_count, once use_count goes from zero to one, you need to load the implementation to successfully handle the upcoming [methodObject.complete_type_info] call. Later, maybe after succeeding use/unuse calls, once use_count drops to zero, you can unload the implementation again. The type system makes sure to call [methodObject.use] and [methodObject.complete_type_info] again when the type is needed again.

[classObject] is an implementation of GTypePlugin that already implements most of this except for the actual module loading and unloading. It even handles multiple registered types per module.

Implements

TypePluginExt Sets an environment variable. On UNIX, both the variable’s name and value can be arbitrary byte strings, except that the variable’s name cannot contain ‘=’. On Windows, they should be in UTF-8.

Note that on some systems, when variables are overwritteAn interface that handles the lifecycle of dynamically loaded types.

The GObject type system supports dynamic loading of types. It goes as follows:

The type is initially introduced (usually upon loading the module the first time, or by your main application that knows what modules introduces what types), like this: ⚠️ The following code is in c ⚠️

   new_type_id = g_type_register_dynamic (parent_type_id,
                                          "TypeName",
                                          new_type_plugin,
                                          type_flags);

where new_type_plugin is an implementation of the GTypePlugin interface.

The type’s implementation is referenced, e.g. through [funcObject.ref] or through [funcObject] (this is being called by [ctorObject.new]) or through one of the above done on a type derived from new_type_id.

This causes the type system to load the type’s implementation by calling [methodObject.use] and [methodObject.complete_type_info] on new_type_plugin.

At some point the type’s implementation isn’t required anymore, e.g. after [methodObject.unref] or [funcObject] (called when the reference count of an instance drops to zero).

This causes the type system to throw away the information retrieved from [methodObject.complete_type_info] and then it calls [methodObject.unuse] on new_type_plugin.

Things may repeat from the second step.

So basically, you need to implement a GTypePlugin type that carries a use_count, once use_count goes from zero to one, you need to load the implementation to successfully handle the upcoming [methodObject.complete_type_info] call. Later, maybe after succeeding use/unuse calls, once use_count drops to zero, you can unload the implementation again. The type system makes sure to call [methodObject.use] and [methodObject.complete_type_info] again when the type is needed again.

[classObject] is an implementation of GTypePlugin that already implements most of this except for the actual module loading and unloading. It even handles multiple registered types per module.

Implements

TypePluginExt