glib

Function shell_quote

pub fn shell_quote(unquoted_string: impl AsRef<OsStr>) -> OsString

Quotes a string so that the shell (/bin/sh) will interpret the quoted string to mean @unquoted_string.

If you pass a filename to the shell, for example, you should first quote it with this function.

The return value must be freed with g_free().

The quoting style used is undefined (single or double quotes may be used).

unquoted_string

a literal string

Returns

quoted string