glib

Function spawn_command_line_async

pub fn spawn_command_line_async(
    command_line: impl AsRef<OsStr>,
) -> Result<(), Error>

Available on Unix only.

A simple version of g_spawn_async() that parses a command line with g_shell_parse_argv() and passes it to g_spawn_async().

Runs a command line in the background. Unlike g_spawn_async(), the SpawnFlags::SEARCH_PATH flag is enabled, other flags are not. Note that SpawnFlags::SEARCH_PATH can have security implications, so consider using g_spawn_async() directly if appropriate. Possible errors are those from g_shell_parse_argv() and g_spawn_async().

The same concerns on Windows apply as for g_spawn_command_line_sync().

command_line

a command line

Returns

true on success, false if error is set