Function glib::spawn_command_line_async

pub fn spawn_command_line_async<P: AsRef<OsStr>>(
    command_line: P
) -> Result<(), Error>

This is supported on Unix only.

A simple version of spawn_async() that parses a command line with shell_parse_argv() and passes it to spawn_async(). Runs a command line in the background. Unlike spawn_async(), the SpawnFlags::SEARCH_PATH flag is enabled, other flags are not. Note that SpawnFlags::SEARCH_PATH can have security implications, so consider using spawn_async() directly if appropriate. Possible errors are those from shell_parse_argv() and spawn_async().

The same concerns on Windows apply as for g_spawn_command_line_sync().

command_line

a command line

Returns

true on success, false if error is set