Module glib::functions

Functions

access
A wrapper for the POSIX access() function. This function is used to test a pathname for one or several of read, write or execute permissions, or just existence.

application_name
Gets a human-readable name for the application, as set by g_set_application_name(). This name should be localized if possible, and is intended for display to the user. Contrast with g_get_prgname(), which gets a non-localized name. If g_set_application_name() has not been called, returns the result of g_get_prgname() (which may be None if g_set_prgname() has also not been called).

assert_warning

assertion_message

assertion_message_cmpstr

base64_decode
Decode a sequence of Base-64 encoded text into binary data. Note that the returned binary data is not necessarily zero-terminated, so it should not be used as a character string.

base64_encode
Encode a sequence of binary data into its Base-64 stringified representation.

bit_nth_lsf
Find the position of the first bit set in mask, searching from (but not including) nth_bit upwards. Bits are numbered from 0 (least significant) to sizeof(gulong) * 8 - 1 (31 or 63, usually). To start searching from the 0th bit, set nth_bit to -1.

bit_nth_msf
Find the position of the first bit set in mask, searching from (but not including) nth_bit downwards. Bits are numbered from 0 (least significant) to sizeof(gulong) * 8 - 1 (31 or 63, usually). To start searching from the last bit, set nth_bit to -1 or GLIB_SIZEOF_LONG * 8.

bit_storage
Gets the number of bits used to hold number, e.g. if number is 4, 3 bits are needed.

build_filenamev
Behaves exactly like g_build_filename(), but takes the path elements as a string array, instead of varargs. This function is mainly meant for language bindings.

build_pathv
Behaves exactly like g_build_path(), but takes the path elements as a string array, instead of varargs. This function is mainly meant for language bindings.

canonicalize_filename	v2_58
Gets the canonical file name from filename. All triple slashes are turned into single slashes, and all .. and .s resolved against relative_to.

chdir
A wrapper for the POSIX chdir() function. The function changes the current directory of the process to path.

check_version
Checks that the GLib library in use is compatible with the given version. Generally you would pass in the constants GLIB_MAJOR_VERSION, GLIB_MINOR_VERSION, GLIB_MICRO_VERSION as the three arguments to this function; that produces a check that the library in use is compatible with the version of GLib the application or module was compiled against.

clear_error
If err or *err is None, does nothing. Otherwise, calls [crate::Error::free()] on *err and sets *err to None.

codeset
Gets the character set for the current locale.

compute_checksum_for_bytes
Computes the checksum for a binary data. This is a convenience wrapper for crate::Checksum::new(), [crate::Checksum::get_string()] and [crate::Checksum::free()].

compute_checksum_for_data
Computes the checksum for a binary data of length. This is a convenience wrapper for crate::Checksum::new(), [crate::Checksum::get_string()] and [crate::Checksum::free()].

compute_checksum_for_string
Computes the checksum of a string.

compute_hmac_for_bytes	v2_50
Computes the HMAC for a binary data. This is a convenience wrapper for [crate::Hmac::new()], [crate::Hmac::get_string()] and [crate::Hmac::unref()].

compute_hmac_for_data
Computes the HMAC for a binary data of length. This is a convenience wrapper for [crate::Hmac::new()], [crate::Hmac::get_string()] and [crate::Hmac::unref()].

compute_hmac_for_string
Computes the HMAC for a string.

console_charset	v2_62
Obtains the character set used by the console attached to the process, which is suitable for printing output to the terminal.

dcgettext
This is a variant of g_dgettext() that allows specifying a locale category instead of always using LC_MESSAGES. See g_dgettext() for more information about how this functions differs from calling dcgettext() directly.

dgettext
This function is a wrapper of dgettext() which does not translate the message if the default domain as set with textdomain() has no translations for the current locale.

dngettext
This function is a wrapper of dngettext() which does not translate the message if the default domain as set with textdomain() has no translations for the current locale.

dpgettext
This function is a variant of g_dgettext() which supports a disambiguating message context. GNU gettext uses the ‘\004’ character to separate the message context and message id in msgctxtid. If 0 is passed as msgidoffset, this function will fall back to trying to use the deprecated convention of using “|” as a separation character.

dpgettext2
This function is a variant of g_dgettext() which supports a disambiguating message context. GNU gettext uses the ‘\004’ character to separate the message context and message id in msgctxtid.

environ
Gets the list of environment variables for the current process.

file_get_contents
Reads an entire file into allocated memory, with good error checking.

file_open_tmp
Opens a file for writing in the preferred directory for temporary files (as returned by g_get_tmp_dir()).

file_read_link
Reads the contents of the symbolic link filename like the POSIX readlink() function. The returned string is in the encoding used for filenames. Use g_filename_to_utf8() to convert it to UTF-8.

file_set_contents
Writes all of contents to a file named filename. This is a convenience wrapper around calling g_file_set_contents_full() with flags set to G_FILE_SET_CONTENTS_CONSISTENT | G_FILE_SET_CONTENTS_ONLY_EXISTING and mode set to 0666.

file_set_contents_full	v2_66
Writes all of contents to a file named filename, with good error checking. If a file called filename already exists it will be overwritten.

file_test
Returns true if any of the tests in the bitfield test are true. For example, (G_FILE_TEST_EXISTS | G_FILE_TEST_IS_DIR) will return true if the file exists; the check whether it’s a directory doesn’t matter since the existence test is true. With the current set of available tests, there’s no point passing in more than one test at a time.

filename_display_basename
Returns the display basename for the particular filename, guaranteed to be valid UTF-8. The display name might not be identical to the filename, for instance there might be problems converting it to UTF-8, and some files can be translated in the display.

filename_display_name
Converts a filename into a valid UTF-8 string. The conversion is not necessarily reversible, so you should keep the original around and use the return value of this function only for display purposes. Unlike g_filename_to_utf8(), the result is guaranteed to be non-None even if the filename actually isn’t in the GLib file name encoding.

format_size
Formats a size (for example the size of a file) into a human readable string. Sizes are rounded to the nearest size prefix (kB, MB, GB) and are displayed rounded to the nearest tenth. E.g. the file size 3292528 bytes will be converted into the string “3.2 MB”. The returned string is UTF-8, and may use a non-breaking space to separate the number and units, to ensure they aren’t separated when line wrapped.

format_size_full
Formats a size.

host_name
Return a name for the machine.

hostname_is_ascii_encoded
Tests if hostname contains segments with an ASCII-compatible encoding of an Internationalized Domain Name. If this returns true, you should decode the hostname with g_hostname_to_unicode() before displaying it to the user.

hostname_is_ip_address
Tests if hostname is the string form of an IPv4 or IPv6 address. (Eg, “192.168.0.1”.)

hostname_is_non_ascii
Tests if hostname contains Unicode characters. If this returns true, you need to encode the hostname with g_hostname_to_ascii() before using it in non-IDN-aware contexts.

hostname_to_ascii
Converts hostname to its canonical ASCII form; an ASCII-only string containing no uppercase letters and not ending with a trailing dot.

hostname_to_unicode
Converts hostname to its canonical presentation form; a UTF-8 string in Unicode normalization form C, containing no uppercase letters, no forbidden characters, and no ASCII-encoded segments, and not ending with a trailing dot.

language_names
Computes a list of applicable locale names, which can be used to e.g. construct locale-dependent filenames or search paths. The returned list is sorted from most desirable to least desirable and always contains the default locale “C”.

language_names_with_category	v2_58
Computes a list of applicable locale names with a locale category name, which can be used to construct the fallback locale-dependent filenames or search paths. The returned list is sorted from most desirable to least desirable and always contains the default locale “C”.

listenv
Gets the names of all variables set in the environment.

locale_variants
Returns a list of derived variants of locale, which can be used to e.g. construct locale-dependent filenames or search paths. The returned list is sorted from most desirable to least desirable. This function handles territory, charset and extra locale modifiers. See setlocale(3) for information about locales and their format.

main_current_source
Returns the currently firing source for this thread.

main_depth
Returns the depth of the stack of calls to crate::MainContext::dispatch() on any crate::MainContext in the current thread. That is, when called from the toplevel, it gives 0. When called from within a callback from crate::MainContext::iteration() (or crate::MainLoop::run(), etc.) it returns 1. When called from within a callback to a recursive call to crate::MainContext::iteration(), it returns 2. And so forth.

markup_escape_text
Escapes text so that the markup parser will parse it verbatim. Less than, greater than, ampersand, etc. are replaced with the corresponding entities. This function would typically be used when writing out a file to be parsed with the markup parser.

mkdir_with_parents
Create a directory if it doesn’t already exist. Create intermediate parent directories as needed, too.

mkdtemp
Creates a temporary directory. See the mkdtemp() documentation on most UNIX-like systems.

mkdtemp_full
Creates a temporary directory. See the mkdtemp() documentation on most UNIX-like systems.

mkstemp_full
Opens a temporary file. See the mkstemp() documentation on most UNIX-like systems.

monotonic_time
Queries the system monotonic time.

num_processors
Determine the approximate number of threads that the system will schedule simultaneously for this process. This is intended to be used as a parameter to [crate::ThreadPool::new()] for CPU bound tasks and similar cases.

on_error_query
Prompts the user with [E]xit, [H]alt, show [S]tack trace or [P]roceed. This function is intended to be used for debugging use only. The following example shows how it can be used together with the g_log() functions.

on_error_stack_trace
Invokes gdb, which attaches to the current process and shows a stack trace. Called by g_on_error_query() when the “[S]tack trace” option is selected. You can get the current process’s program name with g_get_prgname(), assuming that you have called gtk_init() or gdk_init().

os_info	v2_64
Get information about the operating system.

path_get_basename
Gets the last component of the filename.

path_get_dirname
Gets the directory components of a file name. For example, the directory component of /usr/bin/test is /usr/bin. The directory component of / is /.

path_is_absolute
Returns true if the given file_name is an absolute file name. Note that this is a somewhat vague concept on Windows.

path_skip_root
Returns a pointer into file_name after the root component, i.e. after the “/” in UNIX or “C:" under Windows. If file_name is not an absolute path it returns None.

pattern_match_simple
Matches a string against a pattern given as a string. If this function is to be called in a loop, it’s more efficient to compile the pattern once with [crate::PatternSpec::new()] and call g_pattern_match_string() repeatedly.

random_double
Returns a random gdouble equally distributed over the range [0..1).

random_double_range
Returns a random gdouble equally distributed over the range [begin..end).

random_int
Return a random guint32 equally distributed over the range [0..2^32-1].

random_int_range
Returns a random gint32 equally distributed over the range [begin..end-1].

random_set_seed
Sets the seed for the global random number generator, which is used by the g_random_* functions, to seed.

real_time
Queries the system wall-clock time.

reload_user_special_dirs_cache
Resets the cache used for g_get_user_special_dir(), so that the latest on-disk version is used. Call this only if you just changed the data on disk yourself.

return_if_fail_warning
Internal function used to print messages from the public g_return_if_fail() and g_return_val_if_fail() macros.

rmdir
A wrapper for the POSIX rmdir() function. The rmdir() function deletes a directory from the filesystem.

set_application_name
Sets a human-readable name for the application. This name should be localized if possible, and is intended for display to the user. Contrast with g_set_prgname(), which sets a non-localized name. g_set_prgname() will be called automatically by gtk_init(), but g_set_application_name() will not.

shell_parse_argv
Parses a command line into an argument vector, in much the same way the shell would, but without many of the expansions the shell would perform (variable expansion, globs, operators, filename expansion, etc. are not supported). The results are defined to be the same as those you would get from a UNIX98 /bin/sh, as long as the input contains none of the unsupported shell expansions. If the input does contain such expansions, they are passed through literally. Possible errors are those from the G_SHELL_ERROR domain. Free the returned vector with g_strfreev().

shell_quote
Quotes a string so that the shell (/bin/sh) will interpret the quoted string to mean unquoted_string. If you pass a filename to the shell, for example, you should first quote it with this function. The return value must be freed with g_free(). The quoting style used is undefined (single or double quotes may be used).

shell_unquote
Unquotes a string as the shell (/bin/sh) would. Only handles quotes; if a string contains file globs, arithmetic operators, variables, backticks, redirections, or other special-to-the-shell features, the result will be different from the result a real shell would produce (the variables, backticks, etc. will be passed through literally instead of being expanded). This function is guaranteed to succeed if applied to the result of g_shell_quote(). If it fails, it returns None and sets the error. The quoted_string need not actually contain quoted or escaped text; g_shell_unquote() simply goes through the string and unquotes/unescapes anything that the shell would. Both single and double quotes are handled, as are escapes including escaped newlines. The return value must be freed with g_free(). Possible errors are in the G_SHELL_ERROR domain.

spaced_primes_closest
Gets the smallest prime number from a built-in array of primes which is larger than num. This is used within GLib to calculate the optimum size of a [crate::HashTable].

spawn_async
See g_spawn_async_with_pipes() for a full description; this function simply calls the g_spawn_async_with_pipes() without any pipes.

spawn_check_exit_status
Set error if exit_status indicates the child exited abnormally (e.g. with a nonzero exit code, or via a fatal signal).

spawn_command_line_async	Unix
A simple version of g_spawn_async() that parses a command line with g_shell_parse_argv() and passes it to g_spawn_async(). Runs a command line in the background. Unlike g_spawn_async(), the [crate::SpawnFlags::SearchPath] flag is enabled, other flags are not. Note that [crate::SpawnFlags::SearchPath] can have security implications, so consider using g_spawn_async() directly if appropriate. Possible errors are those from g_shell_parse_argv() and g_spawn_async().

stpcpy
Copies a nul-terminated string into the dest buffer, include the trailing nul, and return a pointer to the trailing nul byte. This is useful for concatenating multiple strings together without having to repeatedly scan for the end.

system_config_dirs
Returns an ordered list of base directories in which to access system-wide configuration information.

system_data_dirs
Returns an ordered list of base directories in which to access system-wide application data.

unlink
A wrapper for the POSIX unlink() function. The unlink() function deletes a name from the filesystem. If this was the last link to the file and no processes have it opened, the diskspace occupied by the file is freed.

user_cache_dir
Returns a base directory in which to store non-essential, cached data specific to particular user.

user_config_dir
Returns a base directory in which to store user-specific application configuration information such as user preferences and settings.

user_data_dir
Returns a base directory in which to access application data such as icons that is customized for a particular user.

user_runtime_dir
Returns a directory that is unique to the current user on the local system.

user_special_dir
Returns the full path of a special directory using its logical id.

usleep
Pauses the current thread for the given number of microseconds.

uuid_string_is_valid	v2_52
Parses the string str and verify if it is a UUID.

uuid_string_random	v2_52
Generates a random UUID (RFC 4122 version 4) as a string. It has the same randomness guarantees as [crate::Rand], so must not be used for cryptographic purposes such as key generation, nonces, salts or one-time pads.

warn_message
Internal function used to print messages from the public g_warn_if_reached() and g_warn_if_fail() macros.