Module glib::functions

Functions

access

application_name

base64_decode

base64_encode

check_version

codeset

compute_checksum_for_bytes

compute_checksum_for_data

compute_hmac_for_bytes

compute_hmac_for_data

console_charsetv2_62

current_dir

dcgettext

dgettext

dngettext

dpgettext

dpgettext2

environ

file_set_contents

file_set_contents_fullv2_66

filename_display_basename

filename_display_name

filename_from_uri

filename_to_uri

find_program_in_path

format_size

format_size_full

getenv

home_dir

host_name

hostname_is_ascii_encoded

hostname_is_ip_address

hostname_is_non_ascii

hostname_to_ascii

hostname_to_unicode

language_names

language_names_with_categoryv2_58

listenv

locale_variants

main_current_source

main_depth

markup_escape_text

Escapes text so that the markup parser will parse it verbatim. Less than, greater than, ampersand, etc. are replaced with the corresponding entities. This function would typically be used when writing out a file to be parsed with the markup parser.

mkdir_with_parents

Create a directory if it doesn’t already exist. Create intermediate parent directories as needed, too.

monotonic_time

num_processors

on_error_query

Prompts the user with [E]xit, [H]alt, show [S]tack trace or [P]roceed. This function is intended to be used for debugging use only. The following example shows how it can be used together with the g_log() functions.

on_error_stack_trace

Invokes gdb, which attaches to the current process and shows a stack trace. Called by on_error_query() when the “[S]tack trace” option is selected. You can get the current process’s program name with g_get_prgname(), assuming that you have called gtk_init() or gdk_init().

os_infov2_64

random_double

Returns a random gdouble equally distributed over the range [0..1).

random_double_range

Returns a random gdouble equally distributed over the range [begin..end).

random_int

Return a random guint32 equally distributed over the range [0..2^32-1].

random_int_range

Returns a random gint32 equally distributed over the range [begin..end-1].

random_set_seed

Sets the seed for the global random number generator, which is used by the g_random_* functions, to seed.

real_name

real_time

reload_user_special_dirs_cache

Resets the cache used for user_special_dir(), so that the latest on-disk version is used. Call this only if you just changed the data on disk yourself.

set_application_name

Sets a human-readable name for the application. This name should be localized if possible, and is intended for display to the user. Contrast with g_set_prgname(), which sets a non-localized name. g_set_prgname() will be called automatically by gtk_init(), but set_application_name() will not.

setenv

Sets an environment variable. On UNIX, both the variable’s name and value can be arbitrary byte strings, except that the variable’s name cannot contain ‘=’. On Windows, they should be in UTF-8.

shell_parse_argv

shell_quote

shell_unquote

spaced_primes_closest

spawn_async

spawn_check_exit_statusDeprecated

spawn_check_wait_statusv2_70

spawn_command_line_asyncUnix

system_config_dirs

system_data_dirs

tmp_dir

unicode_script_from_iso15924

unicode_script_to_iso15924

unlink

unsetenv

user_cache_dir

user_config_dir

user_data_dir

user_name

user_runtime_dir

user_special_dir

user_state_dirv2_72

usleep

uuid_string_is_valid

uuid_string_random