Module glib::gstring

Structs

GStr

Representation of a borrowed GString.

GStrInteriorNulError

Error type indicating that a buffer had unexpected nul-bytes.

GString

A type representing an owned, C-compatible, nul-terminated UTF-8 string.

GStringInteriorNulError

Error type indicating that a buffer had unexpected nul-bytes.

GStringNoTrailingNulError

Error type indicating that a buffer did not have a trailing nul-byte.

GStringPtr

NULL-terminated UTF-8 string as stored in StrV.

GStringUtf8Error

Error type indicating that a buffer had invalid UTF-8.

Enums

GStrError

Error type holding all possible failures when creating a GStr reference.

GStringFromError

Error type holding all possible failures when creating a GString.

Inner 🔒

Constants

INLINE_LEN 🔒

MAX_STACK_ALLOCATION 🔒

NONE_STR

Traits

IntoGStr

A trait to accept both &str or &GStr as an argument.

IntoOptionalGStr

A trait to accept both Option<&str> or Option<&GStr> as an argument.