Module glib::log

Structs

LogField

Structure representing a single field in a structured log entry.

LogHandlerId

LogLevels

Enums

LogLevel

Statics

WRITER_FUNC 🔒

Functions

default_handler 🔒

log_default_handler

log_remove_handler

log_set_always_fatal

log_set_default_handler

To set back the default print handler, use the log_unset_default_handler function.

log_set_fatal_mask

log_set_handler

log_set_writer_func

log_structured_array

log_unset_default_handler

To set the default print handler, use the log_set_default_handler function.

log_variant

log_writer_default

log_writer_format_fields

log_writer_is_journald

log_writer_journald

log_writer_standard_streams

log_writer_supports_color

print_handler 🔒

printerr_handler 🔒

set_print_handler

To set back the default print handler, use the unset_print_handler function.

set_printerr_handler

To set back the default print handler, use the unset_printerr_handler function.

to_log_flags 🔒

unset_print_handler

To set the default print handler, use the set_print_handler function.

unset_printerr_handler

To set the default print handler, use the set_printerr_handler function.

Type Aliases

LogCallback 🔒

PrintCallback 🔒

WriterCallback 🔒