Module glib::object

IMPL Object wrapper implementation and Object binding.

Structs

BindingBuilder

ClassRef

InitiallyUnowned

InitiallyUnownedClass

Object
The base class in the object hierarchy.

ObjectClass

ObjectRef

SendWeakRef
A weak reference to the object it was created for that can be sent to different threads even for object types that don't implement Send.

WeakRef

Traits

CanDowncast
Marker trait for the statically known possibility of downcasting from Self to T.

Cast
Upcasting and downcasting support.

IsA
Declares the "is a" relationship.

IsClassFor
Trait for mapping a class struct type to its corresponding instance type.

ObjectExt

ObjectType
Implemented by types representing glib::Object and subclasses of it.

UnsafeFrom
Unsafe variant of the From trait.