Module glib::object

IMPL Object wrapper implementation and Object binding.

Structs

BindingBuilder

Builder for object property bindings.

Class

Class struct of type T.

ClassRef

Reference to the class struct of type T.

InitiallyUnowned

Interface

Interface struct of type T for some type.

InterfaceRef

Reference to a class struct of type T.

Object

The base class in the object hierarchy.

ObjectBuilder

ObjectRef

ObjectValueTypeChecker

Value type checker for object types.

PropertyNotificationFreezeGuard

SendWeakRef

A weak reference to the object it was created for that can be sent to different threads even for object types that don’t implement Send.

WeakRef

A weak reference to an object.

Traits

CanDowncast

Marker trait for the statically known possibility of downcasting from Self to T.

Cast

Upcasting and downcasting support.

IsA

Declares the “is a” relationship.

IsClass

Trait implemented by class types.

IsInterface

Trait implemented by interface types.

ObjectExt

ObjectSubclassIs

Automatically implemented by ObjectSubclass variants of wrapper!

ObjectType

Implemented by types representing glib::Object and subclasses of it.

ParentClassIs

Type Definitions

ObjectClass