Module glib::object

IMPL Object wrapper implementation and Object binding.

Structs

BindingBuilder

Class

ClassRef

InitiallyUnowned

Object
The base class in the object hierarchy.

ObjectRef

SendWeakRef
A weak reference to the object it was created for that can be sent to different threads even for object types that don't implement Send.

WeakRef

Traits

CanDowncast
Marker trait for the statically known possibility of downcasting from Self to T.

Cast
Upcasting and downcasting support.

IsA
Declares the "is a" relationship.

ObjectExt

ObjectSubclassIs
Automatically implemented by ObjectSubclass variants of wrapper!

ObjectType
Implemented by types representing glib::Object and subclasses of it.

ParentClassIs

UnsafeFrom
Unsafe variant of the From trait.

Type Definitions

ObjectClass