Struct glib::object::Class

pub struct Class<T: IsClass>(/* private fields */);

Class struct of type T.

Implementations

impl<T: IsClass> Class<T>

pub fn type_(&self) -> Type

Get the type id for this class.

This is not equivalent to T::static_type() but is the type of the subclass of T where this class belongs to.

pub fn upcast_ref<U: IsClass>(&self) -> &Class<U>
where T: IsA<U>,

Casts this class to a reference to a parent type’s class.

pub fn upcast_ref_mut<U: IsClass>(&mut self) -> &mut Class<U>
where T: IsA<U>,

Casts this class to a mutable reference to a parent type’s class.

pub fn downcast_ref<U: IsClass + IsA<T>>(&self) -> Option<&Class<U>>

Casts this class to a reference to a child type’s class or fails if this class is not implementing the child class.

pub fn downcast_ref_mut<U: IsClass + IsA<T>>(&mut self) -> Option<&mut Class<U>>

Casts this class to a mutable reference to a child type’s class or fails if this class is not implementing the child class.

pub fn from_type(type_: Type) -> Option<ClassRef<'static, T>>

Gets the class struct for Self of type_.

This will return None if type_ is not a subclass of Self.

pub fn parent(&self) -> Option<ClassRef<'_, T>>

Gets the parent class struct, if any.