Struct glib::object::Class

#[repr(transparent)]pub struct Class<T: ObjectType>(_);

Implementations

impl Class<Object>

pub fn has_property<'a, N: Into<&'a str>>(
&self,
property_name: N,
type_: Option<Type>
) -> bool

pub fn get_property_type<'a, N: Into<&'a str>>(
&self,
property_name: N
) -> Option<Type>

pub fn find_property<'a, N: Into<&'a str>>(
&self,
property_name: N
) -> Option<ParamSpec>

pub fn list_properties(&self) -> Vec<ParamSpec>

impl<T: ObjectType> Class<T>

pub fn get_type(&self) -> Type

Get the type id for this class.

pub fn upcast_ref<U: ObjectType>(&self) -> &Class<U> where
T: IsA<U>,

Casts this class to a reference to a parent type's class.

pub fn upcast_ref_mut<U: ObjectType>(&mut self) -> &mut Class<U> where
T: IsA<U>,

Casts this class to a mutable reference to a parent type's class.

pub fn downcast_ref<U: ObjectType>(&self) -> Option<&Class<U>> where
U: IsA<T>,

Casts this class to a reference to a child type's class or fails if this class is not implementing the child class.

pub fn downcast_ref_mut<U: ObjectType>(&mut self) -> Option<&mut Class<U>> where
U: IsA<T>,

Casts this class to a mutable reference to a child type's class or fails if this class is not implementing the child class.

pub fn from_type(type_: Type) -> Option<ClassRef<T>>

Gets the class struct corresponding to type_.

This will return None if type_ is not a subclass of Self.