glib::object

Struct SendWeakRef

pub struct SendWeakRef<T: ObjectType>(/* private fields */);

A weak reference to the object it was created for that can be sent to different threads even for object types that donโt implement Send.

Trying to upgrade the weak reference from another thread than the one where it was created on will panic but dropping or cloning can be done safely from any thread.

Implementations

impl<T: ObjectType> SendWeakRef<T>

pub fn into_weak_ref(self) -> WeakRef<T>