Trait glib::object::ObjectExt

pub trait ObjectExt: ObjectType {
    fn is<T: StaticType>(&self) -> bool;

    fn get_type(&self) -> Type;

    fn get_object_class(&self) -> &ObjectClass;

    fn set_property<'a, N: Into<&'a str>, V: ToValue>(
        &self, 
        property_name: N, 
        value: &V
    ) -> Result<(), BoolError>;

    fn set_property_generic<'a, N: Into<&'a str>>(
        &self, 
        property_name: N, 
        value: &Value
    ) -> Result<(), BoolError>;

    fn set_properties(
        &self, 
        property_values: &[(&str, &dyn ToValue)]
    ) -> Result<(), BoolError>;

    fn set_properties_generic(
        &self, 
        property_values: &[(&str, Value)]
    ) -> Result<(), BoolError>;

    fn get_property<'a, N: Into<&'a str>>(
        &self, 
        property_name: N
    ) -> Result<Value, BoolError>;

    fn has_property<'a, N: Into<&'a str>>(
        &self, 
        property_name: N, 
        type_: Option<Type>
    ) -> bool;

    fn get_property_type<'a, N: Into<&'a str>>(
        &self, 
        property_name: N
    ) -> Option<Type>;

    fn find_property<'a, N: Into<&'a str>>(
        &self, 
        property_name: N
    ) -> Option<ParamSpec>;

    fn list_properties(&self) -> Vec<ParamSpec>;

    unsafe fn set_qdata<QD: 'static>(&self, key: Quark, value: QD);

    unsafe fn get_qdata<QD: 'static>(&self, key: Quark) -> Option<&QD>;

    unsafe fn steal_qdata<QD: 'static>(&self, key: Quark) -> Option<QD>;

    unsafe fn set_data<QD: 'static>(&self, key: &str, value: QD);

    unsafe fn get_data<QD: 'static>(&self, key: &str) -> Option<&QD>;

    unsafe fn steal_data<QD: 'static>(&self, key: &str) -> Option<QD>;

    fn block_signal(&self, handler_id: &SignalHandlerId);

    fn unblock_signal(&self, handler_id: &SignalHandlerId);

    fn stop_signal_emission(&self, signal_name: &str);

    fn connect<'a, N, F>(
        &self, 
        signal_name: N, 
        after: bool, 
        callback: F
    ) -> Result<SignalHandlerId, BoolError>
    where
        N: Into<&'a str>,
        F: Fn(&[Value]) -> Option<Value> + Send + Sync + 'static;

    fn connect_local<'a, N, F>(
        &self, 
        signal_name: N, 
        after: bool, 
        callback: F
    ) -> Result<SignalHandlerId, BoolError>
    where
        N: Into<&'a str>,
        F: Fn(&[Value]) -> Option<Value> + 'static;

    unsafe fn connect_unsafe<'a, N, F>(
        &self, 
        signal_name: N, 
        after: bool, 
        callback: F
    ) -> Result<SignalHandlerId, BoolError>
    where
        N: Into<&'a str>,
        F: Fn(&[Value]) -> Option<Value>;

    fn emit<'a, N: Into<&'a str>>(
        &self, 
        signal_name: N, 
        args: &[&dyn ToValue]
    ) -> Result<Option<Value>, BoolError>;

    fn emit_generic<'a, N: Into<&'a str>>(
        &self, 
        signal_name: N, 
        args: &[Value]
    ) -> Result<Option<Value>, BoolError>;

    fn disconnect(&self, handler_id: SignalHandlerId);

    fn connect_notify<F: Fn(&Self, &ParamSpec) + Send + Sync + 'static>(
        &self, 
        name: Option<&str>, 
        f: F
    ) -> SignalHandlerId;

    fn connect_notify_local<F: Fn(&Self, &ParamSpec) + 'static>(
        &self, 
        name: Option<&str>, 
        f: F
    ) -> SignalHandlerId;

    unsafe fn connect_notify_unsafe<F: Fn(&Self, &ParamSpec)>(
        &self, 
        name: Option<&str>, 
        f: F
    ) -> SignalHandlerId;

    fn notify<'a, N: Into<&'a str>>(&self, property_name: N);

    fn notify_by_pspec(&self, pspec: &ParamSpec);

    fn downgrade(&self) -> WeakRef<Self>;

    fn bind_property<'a, O: ObjectType, N: Into<&'a str>, M: Into<&'a str>>(
        &'a self, 
        source_property: N, 
        target: &'a O, 
        target_property: M
    ) -> BindingBuilder<'a>;

    fn ref_count(&self) -> u32;
}

Required methods

fn is<T: StaticType>(&self) -> bool

Returns true if the object is an instance of (can be cast to) T.

fn get_type(&self) -> Type

fn get_object_class(&self) -> &ObjectClass

fn set_property<'a, N: Into<&'a str>, V: ToValue>(
&self,
property_name: N,
value: &V
) -> Result<(), BoolError>

fn set_property_generic<'a, N: Into<&'a str>>(
&self,
property_name: N,
value: &Value
) -> Result<(), BoolError>

fn set_properties(
&self,
property_values: &[(&str, &dyn ToValue)]
) -> Result<(), BoolError>

fn set_properties_generic(
&self,
property_values: &[(&str, Value)]
) -> Result<(), BoolError>

fn get_property<'a, N: Into<&'a str>>(
&self,
property_name: N
) -> Result<Value, BoolError>

fn has_property<'a, N: Into<&'a str>>(
&self,
property_name: N,
type_: Option<Type>
) -> bool

fn get_property_type<'a, N: Into<&'a str>>(
&self,
property_name: N
) -> Option<Type>

fn find_property<'a, N: Into<&'a str>>(
&self,
property_name: N
) -> Option<ParamSpec>

fn list_properties(&self) -> Vec<ParamSpec>

unsafe fn set_qdata<QD: 'static>(&self, key: Quark, value: QD)

Safety

This function doesn't store type information

unsafe fn get_qdata<QD: 'static>(&self, key: Quark) -> Option<&QD>

Safety

The caller is responsible for ensuring the returned value is of a suitable type

unsafe fn steal_qdata<QD: 'static>(&self, key: Quark) -> Option<QD>

Safety

The caller is responsible for ensuring the returned value is of a suitable type

unsafe fn set_data<QD: 'static>(&self, key: &str, value: QD)

Safety

This function doesn't store type information

unsafe fn get_data<QD: 'static>(&self, key: &str) -> Option<&QD>

Safety

The caller is responsible for ensuring the returned value is of a suitable type

unsafe fn steal_data<QD: 'static>(&self, key: &str) -> Option<QD>

Safety

The caller is responsible for ensuring the returned value is of a suitable type

fn block_signal(&self, handler_id: &SignalHandlerId)

fn unblock_signal(&self, handler_id: &SignalHandlerId)

fn stop_signal_emission(&self, signal_name: &str)

fn connect<'a, N, F>(
&self,
signal_name: N,
after: bool,
callback: F
) -> Result<SignalHandlerId, BoolError> where
N: Into<&'a str>,
F: Fn(&[Value]) -> Option<Value> + Send + Sync + 'static,

fn connect_local<'a, N, F>(
&self,
signal_name: N,
after: bool,
callback: F
) -> Result<SignalHandlerId, BoolError> where
N: Into<&'a str>,
F: Fn(&[Value]) -> Option<Value> + 'static,

unsafe fn connect_unsafe<'a, N, F>(
&self,
signal_name: N,
after: bool,
callback: F
) -> Result<SignalHandlerId, BoolError> where
N: Into<&'a str>,
F: Fn(&[Value]) -> Option<Value>,

fn emit<'a, N: Into<&'a str>>(
&self,
signal_name: N,
args: &[&dyn ToValue]
) -> Result<Option<Value>, BoolError>

fn emit_generic<'a, N: Into<&'a str>>(
&self,
signal_name: N,
args: &[Value]
) -> Result<Option<Value>, BoolError>

fn disconnect(&self, handler_id: SignalHandlerId)

fn connect_notify<F: Fn(&Self, &ParamSpec) + Send + Sync + 'static>(
&self,
name: Option<&str>,
f: F
) -> SignalHandlerId

fn connect_notify_local<F: Fn(&Self, &ParamSpec) + 'static>(
&self,
name: Option<&str>,
f: F
) -> SignalHandlerId

unsafe fn connect_notify_unsafe<F: Fn(&Self, &ParamSpec)>(
&self,
name: Option<&str>,
f: F
) -> SignalHandlerId

fn notify<'a, N: Into<&'a str>>(&self, property_name: N)

fn notify_by_pspec(&self, pspec: &ParamSpec)

fn downgrade(&self) -> WeakRef<Self>

fn bind_property<'a, O: ObjectType, N: Into<&'a str>, M: Into<&'a str>>(
&'a self,
source_property: N,
target: &'a O,
target_property: M
) -> BindingBuilder<'a>

fn ref_count(&self) -> u32

Loading content...