Trait glib::prelude::TypePluginExt

pub trait TypePluginExt: IsA<TypePlugin> + Sealed + 'static {
    // Provided methods
    fn complete_interface_info(
        &self,
        instance_type: Type,
        interface_type: Type,
    ) -> InterfaceInfo { ... }
    fn complete_type_info(&self, g_type: Type) -> (TypeInfo, TypeValueTable) { ... }
    fn unuse(&self) { ... }
    fn use_(&self) { ... }
}

Trait containing all TypePlugin methods.

Implementors

TypeModule, TypePlugin

Provided Methods

fn complete_interface_info( &self, instance_type: Type, interface_type: Type, ) -> InterfaceInfo

Calls the complete_interface_info function from the GTypePluginClass of self. There should be no need to use this function outside of the GObject type system itself.

instance_type

the GType of an instantiatable type to which the interface is added

interface_type

the GType of the interface whose info is completed

info

the InterfaceInfo to fill in

fn complete_type_info(&self, g_type: Type) -> (TypeInfo, TypeValueTable)

Calls the complete_type_info function from the GTypePluginClass of self. There should be no need to use this function outside of the GObject type system itself.

g_type

the GType whose info is completed

info

the TypeInfo struct to fill in

value_table

the TypeValueTable to fill in

fn unuse(&self)

Calls the unuse_plugin function from the GTypePluginClass of self. There should be no need to use this function outside of the GObject type system itself.

fn use_(&self)

Object Safety

This trait is not object safe.