Trait glib::property::Property

pub trait Property {
    type Value: HasParamSpec;
}

A type that can be used as a property. It covers every type which have an associated ParamSpec (HasParamSpec) and some useful types wrapping HasParamSpec. The definition is recursive, so you can nest many Propertys together. The final ParamSpec will be the one of the innermost type

Required Associated Types

type Value: HasParamSpec