glib::property

Trait PropertyGet

pub trait PropertyGet {
    type Value;

    // Required method
    fn get<R, F: Fn(&Self::Value) -> R>(&self, f: F) -> R;
}

A container type implementing this trait can be read by the default getter generated by the Properties macro.

Required Associated Types

type Value

Required Methods

fn get<R, F: Fn(&Self::Value) -> R>(&self, f: F) -> R

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.