Module glib::signal

IMPL Low level signal support.

Structs

Inhibit

Whether to propagate the signal to the default handler.

SignalHandlerId

The id of a signal that is returned by connect.

Functions

connect_raw⚠

signal_handler_block

signal_handler_disconnect

signal_handler_unblock

signal_has_handler_pending

signal_stop_emission_by_name