Module glib::source

Manages available sources of events for the main loop

Structs

Continue
Return type of idle and timeout functions.

Functions

idle_add
Adds a function to be called whenever there are no higher priority events pending to the default main loop.

timeout_add
Sets a function to be called at regular intervals, with the default priority, PRIORITY_DEFAULT.

timeout_add_seconds
Sets a function to be called at regular intervals with the default priority, PRIORITY_DEFAULT.