glib::source
Function timeout_add_seconds

pub fn timeout_add_seconds<F>(interval: u32, func: F) -> SourceId
where
    F: FnMut() -> ControlFlow + Send + 'static,

Adds a closure to be called by the default main loop at regular intervals with second granularity.

func will be called repeatedly every interval seconds until it returns ControlFlow::Break. Precise timing is not guaranteed, the timeout may be delayed by other events.

The default main loop almost always is the main loop of the main thread. Thus, the closure is called on the main thread.