Module glib::source_futures

Structs

SourceFuture
Represents a Future around a glib::Source. The future will be resolved once the source has provided a value

SourceStream
Represents a Stream around a glib::Source. The stream will be provide all values that are provided by the source

Functions

child_watch_future
Create a Future that will resolve once the child process with the given pid exits

child_watch_future_with_priority
Create a Future that will resolve once the child process with the given pid exits

interval_stream
Create a Stream that will provide a value every given number of milliseconds.

interval_stream_seconds
Create a Stream that will provide a value every given number of seconds.

interval_stream_seconds_with_priority
Create a Stream that will provide a value every given number of seconds.

interval_stream_with_priority
Create a Stream that will provide a value every given number of milliseconds.

timeout_future
Create a Future that will resolve after the given number of milliseconds.

timeout_future_seconds
Create a Future that will resolve after the given number of seconds.

timeout_future_seconds_with_priority
Create a Future that will resolve after the given number of seconds.

timeout_future_with_priority
Create a Future that will resolve after the given number of milliseconds.

unix_signal_future
Create a Future that will resolve once the given UNIX signal is raised

unix_signal_future_with_priority
Create a Future that will resolve once the given UNIX signal is raised

unix_signal_stream
Create a Stream that will provide a value whenever the given UNIX signal is raised

unix_signal_stream_with_priority
Create a Stream that will provide a value whenever the given UNIX signal is raised