Struct glib::DateTime

pub struct DateTime(_);

GDateTime is an opaque structure whose members cannot be accessed directly.

Implementations

impl DateTime

pub fn new(
tz: &TimeZone,
year: i32,
month: i32,
day: i32,
hour: i32,
minute: i32,
seconds: f64
) -> Result<DateTime, BoolError>

Creates a new crate::DateTime corresponding to the given date and time in the time zone tz.

The year must be between 1 and 9999, month between 1 and 12 and day between 1 and 28, 29, 30 or 31 depending on the month and the year.

hour must be between 0 and 23 and minute must be between 0 and 59.

seconds must be at least 0.0 and must be strictly less than 60.0. It will be rounded down to the nearest microsecond.

If the given time is not representable in the given time zone (for example, 02:30 on March 14th 2010 in Toronto, due to daylight savings time) then the time will be rounded up to the nearest existing time (in this case, 03:00). If this matters to you then you should verify the return value for containing the same as the numbers you gave.

In the case that the given time is ambiguous in the given time zone (for example, 01:30 on November 7th 2010 in Toronto, due to daylight savings time) then the time falling within standard (ie: non-daylight) time is taken.

It not considered a programmer error for the values to this function to be out of range, but in the case that they are, the function will return None.

You should release the return value by calling [Self::unref()] when you are done with it.

tz

a crate::TimeZone

year

the year component of the date

month

the month component of the date

day

the day component of the date

hour

the hour component of the date

minute

the minute component of the date

seconds

the number of seconds past the minute

Returns

a new crate::DateTime, or None

pub fn from_iso8601(
text: &str,
default_tz: Option<&TimeZone>
) -> Result<DateTime, BoolError>

This is supported on crate feature v2_56 only.

pub fn from_unix_local(t: i64) -> Result<DateTime, BoolError>

pub fn from_unix_utc(t: i64) -> Result<DateTime, BoolError>

pub fn new_local(
year: i32,
month: i32,
day: i32,
hour: i32,
minute: i32,
seconds: f64
) -> Result<DateTime, BoolError>

Creates a new crate::DateTime corresponding to the given date and time in the local time zone.

This call is equivalent to calling Self::new() with the time zone returned by crate::TimeZone::new_local().

year

the year component of the date

month

the month component of the date

day

the day component of the date

hour

the hour component of the date

minute

the minute component of the date

seconds

the number of seconds past the minute

Returns

a crate::DateTime, or None

pub fn new_now(tz: &TimeZone) -> Result<DateTime, BoolError>

Creates a crate::DateTime corresponding to this exact instant in the given time zone tz. The time is as accurate as the system allows, to a maximum accuracy of 1 microsecond.

This function will always succeed unless GLib is still being used after the year 9999.

You should release the return value by calling [Self::unref()] when you are done with it.

tz

a crate::TimeZone

Returns

a new crate::DateTime, or None

pub fn new_now_local() -> Result<DateTime, BoolError>

Creates a crate::DateTime corresponding to this exact instant in the local time zone.

This is equivalent to calling Self::new_now() with the time zone returned by crate::TimeZone::new_local().

Returns

a new crate::DateTime, or None

pub fn new_now_utc() -> Result<DateTime, BoolError>

Creates a crate::DateTime corresponding to this exact instant in UTC.

This is equivalent to calling Self::new_now() with the time zone returned by crate::TimeZone::new_utc().

Returns

a new crate::DateTime, or None

pub fn new_utc(
year: i32,
month: i32,
day: i32,
hour: i32,
minute: i32,
seconds: f64
) -> Result<DateTime, BoolError>

Creates a new crate::DateTime corresponding to the given date and time in UTC.

This call is equivalent to calling Self::new() with the time zone returned by crate::TimeZone::new_utc().

year

the year component of the date

month

the month component of the date

day

the day component of the date

hour

the hour component of the date

minute

the minute component of the date

seconds

the number of seconds past the minute

Returns

a crate::DateTime, or None

pub fn add(&self, timespan: TimeSpan) -> Result<DateTime, BoolError>

Creates a copy of self and adds the specified timespan to the copy.

timespan

a crate::TimeSpan

Returns

the newly created crate::DateTime which should be freed with [Self::unref()], or None

pub fn add_days(&self, days: i32) -> Result<DateTime, BoolError>

Creates a copy of self and adds the specified number of days to the copy. Add negative values to subtract days.

days

the number of days

Returns

the newly created crate::DateTime which should be freed with [Self::unref()], or None

pub fn add_full(
&self,
years: i32,
months: i32,
days: i32,
hours: i32,
minutes: i32,
seconds: f64
) -> Result<DateTime, BoolError>

Creates a new crate::DateTime adding the specified values to the current date and time in self. Add negative values to subtract.

years

the number of years to add

months

the number of months to add

days

the number of days to add

hours

the number of hours to add

minutes

the number of minutes to add

seconds

the number of seconds to add

Returns

the newly created crate::DateTime which should be freed with [Self::unref()], or None

pub fn add_hours(&self, hours: i32) -> Result<DateTime, BoolError>

Creates a copy of self and adds the specified number of hours. Add negative values to subtract hours.

hours

the number of hours to add

Returns

the newly created crate::DateTime which should be freed with [Self::unref()], or None

pub fn add_minutes(&self, minutes: i32) -> Result<DateTime, BoolError>

Creates a copy of self adding the specified number of minutes. Add negative values to subtract minutes.

minutes

the number of minutes to add

Returns

the newly created crate::DateTime which should be freed with [Self::unref()], or None

pub fn add_months(&self, months: i32) -> Result<DateTime, BoolError>

Creates a copy of self and adds the specified number of months to the copy. Add negative values to subtract months.

The day of the month of the resulting crate::DateTime is clamped to the number of days in the updated calendar month. For example, if adding 1 month to 31st January 2018, the result would be 28th February 2018. In 2020 (a leap year), the result would be 29th February.

months

the number of months

Returns

the newly created crate::DateTime which should be freed with [Self::unref()], or None

pub fn add_seconds(&self, seconds: f64) -> Result<DateTime, BoolError>

Creates a copy of self and adds the specified number of seconds. Add negative values to subtract seconds.

seconds

the number of seconds to add

Returns

the newly created crate::DateTime which should be freed with [Self::unref()], or None

pub fn add_weeks(&self, weeks: i32) -> Result<DateTime, BoolError>

Creates a copy of self and adds the specified number of weeks to the copy. Add negative values to subtract weeks.

weeks

the number of weeks

Returns

the newly created crate::DateTime which should be freed with [Self::unref()], or None

pub fn add_years(&self, years: i32) -> Result<DateTime, BoolError>

Creates a copy of self and adds the specified number of years to the copy. Add negative values to subtract years.

As with Self::add_months(), if the resulting date would be 29th February on a non-leap year, the day will be clamped to 28th February.

years

the number of years

Returns

the newly created crate::DateTime which should be freed with [Self::unref()], or None

pub fn difference(&self, begin: &DateTime) -> TimeSpan

Calculates the difference in time between self and begin. The crate::TimeSpan that is returned is effectively self - begin (ie: positive if the first parameter is larger).

begin

a crate::DateTime

Returns

the difference between the two crate::DateTime, as a time span expressed in microseconds.

pub fn format(&self, format: &str) -> Result<GString, BoolError>

Creates a newly allocated string representing the requested format.

The format strings understood by this function are a subset of the strftime() format language as specified by C99. The %D, %U and %W conversions are not supported, nor is the ‘E’ modifier. The GNU extensions %k, %l, %s and %P are supported, however, as are the ‘0’, ‘_’ and ‘-’ modifiers. The Python extension %f is also supported.

In contrast to strftime(), this function always produces a UTF-8 string, regardless of the current locale. Note that the rendering of many formats is locale-dependent and may not match the strftime() output exactly.

The following format specifiers are supported:

%a: the abbreviated weekday name according to the current locale

%A: the full weekday name according to the current locale

%b: the abbreviated month name according to the current locale

%B: the full month name according to the current locale

%c: the preferred date and time representation for the current locale

%C: the century number (year/100) as a 2-digit integer (00-99)

%d: the day of the month as a decimal number (range 01 to 31)

%e: the day of the month as a decimal number (range 1 to 31)

%F: equivalent to %Y-%m-%d (the ISO 8601 date format)

%g: the last two digits of the ISO 8601 week-based year as a decimal number (00-99). This works well with %V and %u.

%G: the ISO 8601 week-based year as a decimal number. This works well with %V and %u.

%h: equivalent to %b

%H: the hour as a decimal number using a 24-hour clock (range 00 to 23)

%I: the hour as a decimal number using a 12-hour clock (range 01 to 12)

%j: the day of the year as a decimal number (range 001 to 366)

%k: the hour (24-hour clock) as a decimal number (range 0 to 23); single digits are preceded by a blank

%l: the hour (12-hour clock) as a decimal number (range 1 to 12); single digits are preceded by a blank

%m: the month as a decimal number (range 01 to 12)

%M: the minute as a decimal number (range 00 to 59)

%f: the microsecond as a decimal number (range 000000 to 999999)

%p: either “AM” or “PM” according to the given time value, or the corresponding strings for the current locale. Noon is treated as “PM” and midnight as “AM”. Use of this format specifier is discouraged, as many locales have no concept of AM/PM formatting. Use %c or %X instead.

%P: like %p but lowercase: “am” or “pm” or a corresponding string for the current locale. Use of this format specifier is discouraged, as many locales have no concept of AM/PM formatting. Use %c or %X instead.

%r: the time in a.m. or p.m. notation. Use of this format specifier is discouraged, as many locales have no concept of AM/PM formatting. Use %c or %X instead.

%R: the time in 24-hour notation (%H:%M)

%s: the number of seconds since the Epoch, that is, since 1970-01-01 00:00:00 UTC

%S: the second as a decimal number (range 00 to 60)

%t: a tab character

%T: the time in 24-hour notation with seconds (%H:%M:%S)

%u: the ISO 8601 standard day of the week as a decimal, range 1 to 7, Monday being 1. This works well with %G and %V.

%V: the ISO 8601 standard week number of the current year as a decimal number, range 01 to 53, where week 1 is the first week that has at least 4 days in the new year. See [Self::get_week_of_year()]. This works well with %G and %u.

%w: the day of the week as a decimal, range 0 to 6, Sunday being 0. This is not the ISO 8601 standard format – use %u instead.

%x: the preferred date representation for the current locale without the time

%X: the preferred time representation for the current locale without the date

%y: the year as a decimal number without the century

%Y: the year as a decimal number including the century

%z: the time zone as an offset from UTC (+hhmm)

%:z: the time zone as an offset from UTC (+hh:mm). This is a gnulib strftime() extension. Since: 2.38

%::z: the time zone as an offset from UTC (+hh:mm:ss). This is a gnulib strftime() extension. Since: 2.38

%:::z: the time zone as an offset from UTC, with : to necessary precision (e.g., -04, +05:30). This is a gnulib strftime() extension. Since: 2.38

%Z: the time zone or name or abbreviation

%%: a literal % character

Some conversion specifications can be modified by preceding the conversion specifier by one or more modifier characters. The following modifiers are supported for many of the numeric conversions:

O: Use alternative numeric symbols, if the current locale supports those.

_: Pad a numeric result with spaces. This overrides the default padding for the specifier.

-: Do not pad a numeric result. This overrides the default padding for the specifier.

0: Pad a numeric result with zeros. This overrides the default padding for the specifier.

Additionally, when O is used with B, b, or h, it produces the alternative form of a month name. The alternative form should be used when the month name is used without a day number (e.g., standalone). It is required in some languages (Baltic, Slavic, Greek, and more) due to their grammatical rules. For other languages there is no difference. %OB is a GNU and BSD strftime() extension expected to be added to the future POSIX specification, %Ob and %Oh are GNU strftime() extensions. Since: 2.56

format

a valid UTF-8 string, containing the format for the crate::DateTime

Returns

a newly allocated string formatted to the requested format or None in the case that there was an error (such as a format specifier not being supported in the current locale). The string should be freed with g_free().

pub fn format_iso8601(&self) -> Result<GString, BoolError>

This is supported on crate feature v2_62 only.

Format self in ISO 8601 format, including the date, time and time zone, and return that as a UTF-8 encoded string.

Since GLib 2.66, this will output to sub-second precision if needed.

Returns

a newly allocated string formatted in ISO 8601 format or None in the case that there was an error. The string should be freed with g_free().

pub fn day_of_month(&self) -> i32

pub fn day_of_week(&self) -> i32

pub fn day_of_year(&self) -> i32

pub fn hour(&self) -> i32

pub fn microsecond(&self) -> i32

pub fn minute(&self) -> i32

pub fn month(&self) -> i32

pub fn second(&self) -> i32

pub fn seconds(&self) -> f64

pub fn timezone(&self) -> TimeZone

This is supported on crate feature v2_58 only.

pub fn timezone_abbreviation(&self) -> GString

pub fn utc_offset(&self) -> TimeSpan

pub fn week_numbering_year(&self) -> i32

pub fn week_of_year(&self) -> i32

pub fn year(&self) -> i32

pub fn ymd(&self) -> (i32, i32, i32)

pub fn is_daylight_savings(&self) -> bool

Determines if daylight savings time is in effect at the time and in the time zone of self.

Returns

true if daylight savings time is in effect

pub fn to_local(&self) -> Result<DateTime, BoolError>

Creates a new crate::DateTime corresponding to the same instant in time as self, but in the local time zone.

This call is equivalent to calling Self::to_timezone() with the time zone returned by crate::TimeZone::new_local().

Returns

the newly created crate::DateTime which should be freed with [Self::unref()], or None

pub fn to_timezone(&self, tz: &TimeZone) -> Result<DateTime, BoolError>

Create a new crate::DateTime corresponding to the same instant in time as self, but in the time zone tz.

This call can fail in the case that the time goes out of bounds. For example, converting 0001-01-01 00:00:00 UTC to a time zone west of Greenwich will fail (due to the year 0 being out of range).

tz

the new crate::TimeZone

Returns

the newly created crate::DateTime which should be freed with [Self::unref()], or None

pub fn to_unix(&self) -> i64

Gives the Unix time corresponding to self, rounding down to the nearest second.

Unix time is the number of seconds that have elapsed since 1970-01-01 00:00:00 UTC, regardless of the time zone associated with self.

Returns

the Unix time corresponding to self

pub fn to_utc(&self) -> Result<DateTime, BoolError>

Creates a new crate::DateTime corresponding to the same instant in time as self, but in UTC.

This call is equivalent to calling Self::to_timezone() with the time zone returned by crate::TimeZone::new_utc().

Returns

the newly created crate::DateTime which should be freed with [Self::unref()], or None