Struct glib::FlagsClass

pub struct FlagsClass(_);

Representation of a flags for dynamically, at runtime, querying the values of the enum and using them

Methods

impl FlagsClass

pub fn new(type_: Type) -> Option<Self>

Create a new FlagsClass from a Type

Returns None if type_ is not representing a flags type.

pub fn type_(&self) -> Type

Type of the flags.

pub fn get_value(&self, value: u32) -> Option<FlagsValue>

Gets FlagsValue by integer value, if existing.

Returns None if the flags do not contain any value with value.

pub fn get_value_by_name(&self, name: &str) -> Option<FlagsValue>

Gets FlagsValue by string name name, if existing.

Returns None if the flags do not contain any value with name name.

pub fn get_value_by_nick(&self, nick: &str) -> Option<FlagsValue>

Gets FlagsValue by string nick nick, if existing.

Returns None if the flags do not contain any value with nick nick.

pub fn get_values(&self) -> Vec<FlagsValue>

Gets all FlagsValue of this FlagsClass.

pub fn to_value(&self, value: u32) -> Option<Value>

Converts integer value to a Value, if part of the flags.

pub fn to_value_by_name(&self, name: &str) -> Option<Value>

Converts string name name to a Value, if part of the flags.

pub fn to_value_by_nick(&self, nick: &str) -> Option<Value>

Converts string nick nick to a Value, if part of the flags.

pub fn is_set(&self, value: &Value, f: u32) -> bool

Checks if the flags corresponding to integer f is set in value.

pub fn is_set_by_name(&self, value: &Value, name: &str) -> bool

Checks if the flags corresponding to string name name is set in value.

pub fn is_set_by_nick(&self, value: &Value, nick: &str) -> bool

Checks if the flags corresponding to string nick nick is set in value.

pub fn set(&self, value: Value, f: u32) -> Result<Value, Value>

Sets flags value corresponding to integer f in value, if part of that flags. If the flag is already set, it will succeed without doing any changes.

Returns Ok(value) with the flag set if successful, or Err(value) with the original value otherwise.

pub fn set_by_name(&self, value: Value, name: &str) -> Result<Value, Value>

Sets flags value corresponding to string name name in value, if part of that flags. If the flag is already set, it will succeed without doing any changes.

Returns Ok(value) with the flag set if successful, or Err(value) with the original value otherwise.

pub fn set_by_nick(&self, value: Value, nick: &str) -> Result<Value, Value>

Sets flags value corresponding to string nick nick in value, if part of that flags. If the flag is already set, it will succeed without doing any changes.

Returns Ok(value) with the flag set if successful, or Err(value) with the original value otherwise.

pub fn unset(&self, value: Value, f: u32) -> Result<Value, Value>

Unsets flags value corresponding to integer f in value, if part of that flags. If the flag is already unset, it will succeed without doing any changes.

Returns Ok(value) with the flag unset if successful, or Err(value) with the original value otherwise.

pub fn unset_by_name(&self, value: Value, name: &str) -> Result<Value, Value>

Unsets flags value corresponding to string name name in value, if part of that flags. If the flag is already unset, it will succeed without doing any changes.

Returns Ok(value) with the flag unset if successful, or Err(value) with the original value otherwise.

pub fn unset_by_nick(&self, value: Value, nick: &str) -> Result<Value, Value>

Unsets flags value corresponding to string nick nick in value, if part of that flags. If the flag is already unset, it will succeed without doing any changes.

Returns Ok(value) with the flag unset if successful, or Err(value) with the original value otherwise.

pub fn builder(&self) -> FlagsBuilder

Returns a new FlagsBuilder for conveniently setting/unsetting flags and building a Value.

pub fn builder_with_value(&self, value: Value) -> Option<FlagsBuilder>

Returns a new FlagsBuilder for conveniently setting/unsetting flags and building a Value. The Value is initialized with value.