Struct glib::MainContext

pub struct MainContext(_);

Methods

impl MainContext

pub fn new() -> MainContext

pub fn acquire(&self) -> bool

pub fn dispatch(&self)

pub fn is_owner(&self) -> bool

pub fn iteration(&self, may_block: bool) -> bool

pub fn pending(&self) -> bool

pub fn pop_thread_default(&self)

pub fn push_thread_default(&self)

pub fn release(&self)

pub fn wakeup(&self)

pub fn default() -> MainContext

pub fn get_thread_default() -> Option<MainContext>

pub fn ref_thread_default() -> MainContext

impl MainContext

pub fn prepare(&self) -> (bool, i32)

pub fn find_source_by_id(&self, source_id: &SourceId) -> Option<Source>

pub fn invoke<F>(&self, func: F) where
F: FnOnce() + Send + 'static,

pub fn invoke_with_priority<F>(&self, priority: Priority, func: F) where
F: FnOnce() + Send + 'static,

pub fn with_thread_default<R, F: Sized>(&self, func: F) -> R where
F: FnOnce() -> R,

Calls closure with context configured as the thread default one.

Thread default context is changed in panic-safe manner by calling push_thread_default before calling closure and pop_thread_default afterwards regardless of whether closure panicked or not.

impl MainContext

pub fn spawn<F: Future<Item = (), Error = Never> + Send + 'static>(&self, f: F)

Spawn a new infallible Future on the main context.

This can be called from any thread and will execute the future from the thread where main context is running, e.g. via a MainLoop.

pub fn spawn_local<F: Future<Item = (), Error = Never> + 'static>(&self, f: F)

Spawn a new infallible Future on the main context.

The given Future does not have to be Send.

This can be called only from the thread where the main context is running, e.g. from any other Future that is executed on this main context, or after calling push_thread_default or acquire on the main context.

pub fn spawn_with_priority<F: Future<Item = (), Error = Never> + Send + 'static>(
&self,
priority: Priority,
f: F
)

Spawn a new infallible Future on the main context, with a non-default priority.

This can be called from any thread and will execute the future from the thread where main context is running, e.g. via a MainLoop.

pub fn spawn_local_with_priority<F: Future<Item = (), Error = Never> + 'static>(
&self,
priority: Priority,
f: F
)

Spawn a new infallible Future on the main context, with a non-default priority.

The given Future does not have to be Send.

This can be called only from the thread where the main context is running, e.g. from any other Future that is executed on this main context, or after calling push_thread_default or acquire on the main context.

pub fn block_on<F: Future>(&self, f: F) -> Result<F::Item, F::Error>

Runs a new, infallible Future on the main context and block until it finished, returning the result of the Future.

The given Future does not have to be Send or 'static.

This must only be called if no MainLoop or anything else is running on this specific main context.