Struct glib::Uri

#[repr(transparent)]
pub struct Uri { /* private fields */ }

Available on crate feature v2_66 only.

Implementations

impl Uri

pub fn as_ptr(&self) -> *mut GUri

Return the inner pointer to the underlying C value.

impl Uri

pub fn auth_params(&self) -> Option<GString>

pub fn flags(&self) -> UriFlags

pub fn fragment(&self) -> Option<GString>

pub fn host(&self) -> Option<GString>

pub fn password(&self) -> Option<GString>

pub fn path(&self) -> GString

pub fn port(&self) -> i32

pub fn query(&self) -> Option<GString>

pub fn scheme(&self) -> GString

pub fn user(&self) -> Option<GString>

pub fn userinfo(&self) -> Option<GString>

Gets self’s userinfo, which may contain %-encoding, depending on the flags with which self was created.

Returns

self’s userinfo.

pub fn parse_relative(&self, uri_ref: &str, flags: UriFlags) -> Result<Uri, Error>

Parses uri_ref according to flags and, if it is a [relative URI][relative-absolute-uris], resolves it relative to self. If the result is not a valid absolute URI, it will be discarded, and an error returned.

uri_ref

a string representing a relative or absolute URI

flags

flags describing how to parse uri_ref

Returns

a new Uri, or NULL on error.

pub fn to_str(&self) -> GString

Returns a string representing self.

This is not guaranteed to return a string which is identical to the string that self was parsed from. However, if the source URI was syntactically correct (according to RFC 3986), and it was parsed with UriFlags::ENCODED, then to_str() is guaranteed to return a string which is at least semantically equivalent to the source URI (according to RFC 3986).

If self might contain sensitive details, such as authentication parameters, or private data in its query string, and the returned string is going to be logged, then consider using to_string_partial() to redact parts.

Returns

a string representing self, which the caller must free.

pub fn to_string_partial(&self, flags: UriHideFlags) -> GString

Returns a string representing self, subject to the options in flags. See to_str() and UriHideFlags for more details.

flags

flags describing what parts of self to hide

Returns

a string representing self, which the caller must free.

pub fn build(
flags: UriFlags,
scheme: &str,
userinfo: Option<&str>,
host: Option<&str>,
port: i32,
path: &str,
query: Option<&str>,
fragment: Option<&str>
) -> Uri

Creates a new Uri from the given components according to flags.

See also build_with_user(), which allows specifying the components of the “userinfo” separately.

flags

flags describing how to build the Uri

scheme

the URI scheme

userinfo

the userinfo component, or None

host

the host component, or None

port

the port, or -1

path

the path component

query

the query component, or None

fragment

the fragment, or None

Returns

a new Uri

pub fn build_with_user(
flags: UriFlags,
scheme: &str,
user: Option<&str>,
password: Option<&str>,
auth_params: Option<&str>,
host: Option<&str>,
port: i32,
path: &str,
query: Option<&str>,
fragment: Option<&str>
) -> Uri

Creates a new Uri from the given components according to flags (UriFlags::HAS_PASSWORD is added unconditionally). The flags must be coherent with the passed values, in particular use %-encoded values with UriFlags::ENCODED.

In contrast to build(), this allows specifying the components of the ‘userinfo’ field separately. Note that user must be non-None if either password or auth_params is non-None.

flags

flags describing how to build the Uri

scheme

the URI scheme

user

the user component of the userinfo, or None

password

the password component of the userinfo, or None

auth_params

the auth params of the userinfo, or None

host

the host component, or None

port

the port, or -1

path

the path component

query

the query component, or None

fragment

the fragment, or None

Returns

a new Uri

pub fn escape_bytes(
unescaped: &[u8],
reserved_chars_allowed: Option<&str>
) -> GString

Escapes arbitrary data for use in a URI.

Normally all characters that are not ‘unreserved’ (i.e. ASCII alphanumerical characters plus dash, dot, underscore and tilde) are escaped. But if you specify characters in reserved_chars_allowed they are not escaped. This is useful for the ‘reserved’ characters in the URI specification, since those are allowed unescaped in some portions of a URI.

Though technically incorrect, this will also allow escaping nul bytes as %``00.

unescaped

the unescaped input data.

reserved_chars_allowed

a string of reserved characters that are allowed to be used, or None.

Returns

an escaped version of unescaped. The returned string should be freed when no longer needed.

pub fn escape_string(
unescaped: &str,
reserved_chars_allowed: Option<&str>,
allow_utf8: bool
) -> GString

Escapes a string for use in a URI.

Normally all characters that are not “unreserved” (i.e. ASCII alphanumerical characters plus dash, dot, underscore and tilde) are escaped. But if you specify characters in reserved_chars_allowed they are not escaped. This is useful for the “reserved” characters in the URI specification, since those are allowed unescaped in some portions of a URI.

unescaped

the unescaped input string.

reserved_chars_allowed

a string of reserved characters that are allowed to be used, or None.

allow_utf8

true if the result can include UTF-8 characters.

Returns

an escaped version of unescaped. The returned string should be freed when no longer needed.

pub fn is_valid(uri_string: &str, flags: UriFlags) -> Result<(), Error>

Parses uri_string according to flags, to determine whether it is a valid [absolute URI][relative-absolute-uris], i.e. it does not need to be resolved relative to another URI using parse_relative().

If it’s not a valid URI, an error is returned explaining how it’s invalid.

See split(), and the definition of UriFlags, for more information on the effect of flags.

uri_string

a string containing an absolute URI

flags

flags for parsing uri_string

Returns

true if uri_string is a valid absolute URI, false on error.

pub fn join(
flags: UriFlags,
scheme: Option<&str>,
userinfo: Option<&str>,
host: Option<&str>,
port: i32,
path: &str,
query: Option<&str>,
fragment: Option<&str>
) -> GString

Joins the given components together according to flags to create an absolute URI string. path may not be None (though it may be the empty string).

When host is present, path must either be empty or begin with a slash (/) character. When host is not present, path cannot begin with two slash characters (//). See RFC 3986, section 3.

See also join_with_user(), which allows specifying the components of the ‘userinfo’ separately.

UriFlags::HAS_PASSWORD and UriFlags::HAS_AUTH_PARAMS are ignored if set in flags.

flags

flags describing how to build the URI string

scheme

the URI scheme, or None

userinfo

the userinfo component, or None

host

the host component, or None

port

the port, or -1

path

the path component

query

the query component, or None

fragment

the fragment, or None

Returns

an absolute URI string

pub fn join_with_user(
flags: UriFlags,
scheme: Option<&str>,
user: Option<&str>,
password: Option<&str>,
auth_params: Option<&str>,
host: Option<&str>,
port: i32,
path: &str,
query: Option<&str>,
fragment: Option<&str>
) -> GString

Joins the given components together according to flags to create an absolute URI string. path may not be None (though it may be the empty string).

In contrast to join(), this allows specifying the components of the ‘userinfo’ separately. It otherwise behaves the same.

UriFlags::HAS_PASSWORD and UriFlags::HAS_AUTH_PARAMS are ignored if set in flags.

flags

flags describing how to build the URI string

scheme

the URI scheme, or None

user

the user component of the userinfo, or None

password

the password component of the userinfo, or None

auth_params

the auth params of the userinfo, or None

host

the host component, or None

port

the port, or -1

path

the path component

query

the query component, or None

fragment

the fragment, or None

Returns

an absolute URI string

pub fn list_extract_uris(uri_list: &str) -> Vec<GString>

Splits an URI list conforming to the text/uri-list mime type defined in RFC 2483 into individual URIs, discarding any comments. The URIs are not validated.

uri_list

an URI list

Returns

a newly allocated None-terminated list of strings holding the individual URIs. The array should be freed with g_strfreev().

pub fn parse(uri_string: &str, flags: UriFlags) -> Result<Uri, Error>

Parses uri_string according to flags. If the result is not a valid [absolute URI][relative-absolute-uris], it will be discarded, and an error returned.

uri_string

a string representing an absolute URI

flags

flags describing how to parse uri_string

Returns

a new Uri, or NULL on error.

pub fn parse_scheme(uri: &str) -> Option<GString>

Gets the scheme portion of a URI string. RFC 3986 decodes the scheme as:

URI = scheme ":" hier-part [ "?" query ] [ "#" fragment ]

Common schemes include file, https, svn+ssh, etc.

uri

a valid URI.

Returns

The ‘scheme’ component of the URI, or None on error. The returned string should be freed when no longer needed.

pub fn peek_scheme(uri: &str) -> Option<GString>

Gets the scheme portion of a URI string. RFC 3986 decodes the scheme as:

URI = scheme ":" hier-part [ "?" query ] [ "#" fragment ]

Common schemes include file, https, svn+ssh, etc.

Unlike parse_scheme(), the returned scheme is normalized to all-lowercase and does not need to be freed.

uri

a valid URI.

Returns

The ‘scheme’ component of the URI, or None on error. The returned string is normalized to all-lowercase, and interned via g_intern_string(), so it does not need to be freed.

pub fn resolve_relative(
base_uri_string: Option<&str>,
uri_ref: &str,
flags: UriFlags
) -> Result<GString, Error>

Parses uri_ref according to flags and, if it is a [relative URI][relative-absolute-uris], resolves it relative to base_uri_string. If the result is not a valid absolute URI, it will be discarded, and an error returned.

(If base_uri_string is None, this just returns uri_ref, or None if uri_ref is invalid or not absolute.)

base_uri_string

a string representing a base URI

uri_ref

a string representing a relative or absolute URI

flags

flags describing how to parse uri_ref

Returns

the resolved URI string, or NULL on error.

pub fn split(
uri_ref: &str,
flags: UriFlags
) -> Result<(Option<GString>, Option<GString>, Option<GString>, i32, GString, Option<GString>, Option<GString>), Error>

Parses uri_ref (which can be an [absolute or relative URI][relative-absolute-uris]) according to flags, and returns the pieces. Any component that doesn’t appear in uri_ref will be returned as None (but note that all URIs always have a path component, though it may be the empty string).

If flags contains UriFlags::ENCODED, then %-encoded characters in uri_ref will remain encoded in the output strings. (If not, then all such characters will be decoded.) Note that decoding will only work if the URI components are ASCII or UTF-8, so you will need to use UriFlags::ENCODED if they are not.

Note that the UriFlags::HAS_PASSWORD and UriFlags::HAS_AUTH_PARAMS flags are ignored by split(), since it always returns only the full userinfo; use split_with_user() if you want it split up.

uri_ref

a string containing a relative or absolute URI

flags

flags for parsing uri_ref

Returns

true if uri_ref parsed successfully, false on error.

scheme

on return, contains the scheme (converted to lowercase), or None

userinfo

on return, contains the userinfo, or None

host

on return, contains the host, or None

port

on return, contains the port, or -1

path

on return, contains the path

query

on return, contains the query, or None

fragment

on return, contains the fragment, or None

pub fn split_network(
uri_string: &str,
flags: UriFlags
) -> Result<(Option<GString>, Option<GString>, i32), Error>

Parses uri_string (which must be an [absolute URI][relative-absolute-uris]) according to flags, and returns the pieces relevant to connecting to a host. See the documentation for split() for more details; this is mostly a wrapper around that function with simpler arguments. However, it will return an error if uri_string is a relative URI, or does not contain a hostname component.

uri_string

a string containing an absolute URI

flags

flags for parsing uri_string

Returns

true if uri_string parsed successfully, false on error.

scheme

on return, contains the scheme (converted to lowercase), or None

host

on return, contains the host, or None

port

on return, contains the port, or -1

pub fn split_with_user(
uri_ref: &str,
flags: UriFlags
) -> Result<(Option<GString>, Option<GString>, Option<GString>, Option<GString>, Option<GString>, i32, GString, Option<GString>, Option<GString>), Error>

Parses uri_ref (which can be an [absolute or relative URI][relative-absolute-uris]) according to flags, and returns the pieces. Any component that doesn’t appear in uri_ref will be returned as None (but note that all URIs always have a path component, though it may be the empty string).

See split(), and the definition of UriFlags, for more information on the effect of flags. Note that password will only be parsed out if flags contains UriFlags::HAS_PASSWORD, and auth_params will only be parsed out if flags contains UriFlags::HAS_AUTH_PARAMS.

uri_ref

a string containing a relative or absolute URI

flags

flags for parsing uri_ref

Returns

true if uri_ref parsed successfully, false on error.

scheme

on return, contains the scheme (converted to lowercase), or None

user

on return, contains the user, or None

password

on return, contains the password, or None

auth_params

on return, contains the auth_params, or None

host

on return, contains the host, or None

port

on return, contains the port, or -1

path

on return, contains the path

query

on return, contains the query, or None

fragment

on return, contains the fragment, or None

pub fn unescape_bytes(
escaped_string: &str,
illegal_characters: Option<&str>
) -> Result<Bytes, Error>

Unescapes a segment of an escaped string as binary data.

Note that in contrast to unescape_string(), this does allow nul bytes to appear in the output.

If any of the characters in illegal_characters appears as an escaped character in escaped_string, then that is an error and None will be returned. This is useful if you want to avoid for instance having a slash being expanded in an escaped path element, which might confuse pathname handling.

escaped_string

A URI-escaped string

length

the length (in bytes) of escaped_string to escape, or -1 if it is nul-terminated.

illegal_characters

a string of illegal characters not to be allowed, or None.

Returns

an unescaped version of escaped_string or None on error (if decoding failed, using UriError::Failed error code). The returned Bytes should be unreffed when no longer needed.

pub fn unescape_segment(
escaped_string: Option<&str>,
escaped_string_end: Option<&str>,
illegal_characters: Option<&str>
) -> Option<GString>

Unescapes a segment of an escaped string.

If any of the characters in illegal_characters or the NUL character appears as an escaped character in escaped_string, then that is an error and None will be returned. This is useful if you want to avoid for instance having a slash being expanded in an escaped path element, which might confuse pathname handling.

Note: NUL byte is not accepted in the output, in contrast to unescape_bytes().

escaped_string

A string, may be None

escaped_string_end

Pointer to end of escaped_string, may be None

illegal_characters

An optional string of illegal characters not to be allowed, may be None

Returns

an unescaped version of escaped_string, or None on error. The returned string should be freed when no longer needed. As a special case if None is given for escaped_string, this function will return None.

pub fn unescape_string(
escaped_string: &str,
illegal_characters: Option<&str>
) -> Option<GString>

Unescapes a whole escaped string.

If any of the characters in illegal_characters or the NUL character appears as an escaped character in escaped_string, then that is an error and None will be returned. This is useful if you want to avoid for instance having a slash being expanded in an escaped path element, which might confuse pathname handling.

escaped_string

an escaped string to be unescaped.

illegal_characters

a string of illegal characters not to be allowed, or None.

Returns

an unescaped version of escaped_string. The returned string should be freed when no longer needed.