Struct glib::Uri

#[repr(transparent)]
pub struct Uri { /* private fields */ }

The Uri type and related functions can be used to parse URIs into their components, and build valid URIs from individual components.

Note that Uri scope is to help manipulate URIs in various applications, following RFC 3986. In particular, it doesn’t intend to cover web browser needs, and doesn’t implement the WHATWG URL standard. No APIs are provided to help prevent homograph attacks, so Uri is not suitable for formatting URIs for display to the user for making security-sensitive decisions.

Relative and absolute URIs # {relative-absolute-uris}

As defined in RFC 3986, the hierarchical nature of URIs means that they can either be ‘relative references’ (sometimes referred to as ‘relative URIs’) or ‘URIs’ (for clarity, ‘URIs’ are referred to in this documentation as ‘absolute URIs’ — although in constrast to RFC 3986, fragment identifiers are always allowed).

Relative references have one or more components of the URI missing. In particular, they have no scheme. Any other component, such as hostname, query, etc. may be missing, apart from a path, which has to be specified (but may be empty). The path may be relative, starting with ./ rather than /.

For example, a valid relative reference is ./path?query, /?queryfragment`` or //example.com.

Absolute URIs have a scheme specified. Any other components of the URI which are missing are specified as explicitly unset in the URI, rather than being resolved relative to a base URI using parse_relative().

For example, a valid absolute URI is file:///home/bob or https://search.com?query=string.

A Uri instance is always an absolute URI. A string may be an absolute URI or a relative reference; see the documentation for individual functions as to what forms they accept.

Parsing URIs

The most minimalist APIs for parsing URIs are split() and split_with_user(). These split a URI into its component parts, and return the parts; the difference between the two is that split() treats the ‘userinfo’ component of the URI as a single element, while split_with_user() can (depending on the UriFlags you pass) treat it as containing a username, password, and authentication parameters. Alternatively, split_network() can be used when you are only interested in the components that are needed to initiate a network connection to the service (scheme, host, and port).

parse() is similar to split(), but instead of returning individual strings, it returns a Uri structure (and it requires that the URI be an absolute URI).

resolve_relative() and parse_relative() allow you to resolve a relative URI relative to a base URI. resolve_relative() takes two strings and returns a string, and parse_relative() takes a Uri and a string and returns a Uri.

All of the parsing functions take a UriFlags argument describing exactly how to parse the URI; see the documentation for that type for more details on the specific flags that you can pass. If you need to choose different flags based on the type of URI, you can use peek_scheme() on the URI string to check the scheme first, and use that to decide what flags to parse it with.

For example, you might want to use UriParamsFlags::WWW_FORM when parsing the params for a web URI, so compare the result of peek_scheme() against http and https.

Building URIs

join() and join_with_user() can be used to construct valid URI strings from a set of component strings. They are the inverse of split() and split_with_user().

Similarly, build() and build_with_user() can be used to construct a Uri from a set of component strings.

As with the parsing functions, the building functions take a UriFlags argument. In particular, it is important to keep in mind whether the URI components you are using are already %-encoded. If so, you must pass the UriFlags::ENCODED flag.

file:// URIs

Note that Windows and Unix both define special rules for parsing file:// URIs (involving non-UTF-8 character sets on Unix, and the interpretation of path separators on Windows). Uri does not implement these rules. Use filename_from_uri() and filename_to_uri() if you want to properly convert between file:// URIs and local filenames.

URI Equality

Note that there is no g_uri_equal () function, because comparing URIs usefully requires scheme-specific knowledge that Uri does not have. Uri can help with normalization if you use the various encoded UriFlags as well as UriFlags::SCHEME_NORMALIZE however it is not comprehensive. For example, data:,foo and data:;base64,Zm9v resolve to the same thing according to the data: URI specification which GLib does not handle.

Implementations

impl Uri

pub fn auth_params(&self) -> Option<GString>

Available on crate feature v2_66 only.

Gets self’s authentication parameters, which may contain %-encoding, depending on the flags with which self was created. (If self was not created with UriFlags::HAS_AUTH_PARAMS then this will be None.)

Depending on the URI scheme, g_uri_parse_params() may be useful for further parsing this information.

Returns

self’s authentication parameters.

pub fn flags(&self) -> UriFlags

Available on crate feature v2_66 only.

Gets self’s flags set upon construction.

Returns

self’s flags.

pub fn fragment(&self) -> Option<GString>

Available on crate feature v2_66 only.

Gets self’s fragment, which may contain %-encoding, depending on the flags with which self was created.

Returns

self’s fragment.

pub fn host(&self) -> Option<GString>

Available on crate feature v2_66 only.

Gets self’s host. This will never have %-encoded characters, unless it is non-UTF-8 (which can only be the case if self was created with UriFlags::NON_DNS).

If self contained an IPv6 address literal, this value will be just that address, without the brackets around it that are necessary in the string form of the URI. Note that in this case there may also be a scope ID attached to the address. Eg, fe80::1234%``em1 (or fe80::1234%``25em1 if the string is still encoded).

Returns

self’s host.

pub fn password(&self) -> Option<GString>

Available on crate feature v2_66 only.

Gets self’s password, which may contain %-encoding, depending on the flags with which self was created. (If self was not created with UriFlags::HAS_PASSWORD then this will be None.)

Returns

self’s password.

pub fn path(&self) -> GString

Available on crate feature v2_66 only.

Gets self’s path, which may contain %-encoding, depending on the flags with which self was created.

Returns

self’s path.

pub fn port(&self) -> i32

Available on crate feature v2_66 only.

Gets self’s port.

Returns

self’s port, or -1 if no port was specified.

pub fn query(&self) -> Option<GString>

Available on crate feature v2_66 only.

Gets self’s query, which may contain %-encoding, depending on the flags with which self was created.

For queries consisting of a series of name=value parameters, GUriParamsIter or g_uri_parse_params() may be useful.

Returns

self’s query.

pub fn scheme(&self) -> GString

Available on crate feature v2_66 only.

Gets self’s scheme. Note that this will always be all-lowercase, regardless of the string or strings that self was created from.

Returns

self’s scheme.

pub fn user(&self) -> Option<GString>

Available on crate feature v2_66 only.

Gets the ‘username’ component of self’s userinfo, which may contain %-encoding, depending on the flags with which self was created. If self was not created with UriFlags::HAS_PASSWORD or UriFlags::HAS_AUTH_PARAMS, this is the same as userinfo().

Returns

self’s user.

pub fn userinfo(&self) -> Option<GString>

Available on crate feature v2_66 only.

Gets self’s userinfo, which may contain %-encoding, depending on the flags with which self was created.

Returns

self’s userinfo.

pub fn parse_relative(
&self,
uri_ref: &str,
flags: UriFlags
) -> Result<Uri, Error>

Available on crate feature v2_66 only.

Parses uri_ref according to flags and, if it is a [relative URI][relative-absolute-uris], resolves it relative to self. If the result is not a valid absolute URI, it will be discarded, and an error returned.

uri_ref

a string representing a relative or absolute URI

flags

flags describing how to parse uri_ref

Returns

a new Uri, or NULL on error.

pub fn to_str(&self) -> GString

Available on crate feature v2_66 only.

Returns a string representing self.

This is not guaranteed to return a string which is identical to the string that self was parsed from. However, if the source URI was syntactically correct (according to RFC 3986), and it was parsed with UriFlags::ENCODED, then to_str() is guaranteed to return a string which is at least semantically equivalent to the source URI (according to RFC 3986).

If self might contain sensitive details, such as authentication parameters, or private data in its query string, and the returned string is going to be logged, then consider using to_string_partial() to redact parts.

Returns

a string representing self, which the caller must free.

pub fn to_string_partial(&self, flags: UriHideFlags) -> GString

Available on crate feature v2_66 only.

Returns a string representing self, subject to the options in flags. See to_str() and UriHideFlags for more details.

flags

flags describing what parts of self to hide

Returns

a string representing self, which the caller must free.

pub fn build(
flags: UriFlags,
scheme: &str,
userinfo: Option<&str>,
host: Option<&str>,
port: i32,
path: &str,
query: Option<&str>,
fragment: Option<&str>
) -> Uri

Available on crate feature v2_66 only.

Creates a new Uri from the given components according to flags.

See also build_with_user(), which allows specifying the components of the “userinfo” separately.

flags

flags describing how to build the Uri

scheme

the URI scheme

userinfo

the userinfo component, or None

host

the host component, or None

port

the port, or -1

path

the path component

query

the query component, or None

fragment

the fragment, or None

Returns

a new Uri

pub fn build_with_user(
flags: UriFlags,
scheme: &str,
user: Option<&str>,
password: Option<&str>,
auth_params: Option<&str>,
host: Option<&str>,
port: i32,
path: &str,
query: Option<&str>,
fragment: Option<&str>
) -> Uri

Available on crate feature v2_66 only.

Creates a new Uri from the given components according to flags (UriFlags::HAS_PASSWORD is added unconditionally). The flags must be coherent with the passed values, in particular use %-encoded values with UriFlags::ENCODED.

In contrast to build(), this allows specifying the components of the ‘userinfo’ field separately. Note that user must be non-None if either password or auth_params is non-None.

flags

flags describing how to build the Uri

scheme

the URI scheme

user

the user component of the userinfo, or None

password

the password component of the userinfo, or None

auth_params

the auth params of the userinfo, or None

host

the host component, or None

port

the port, or -1

path

the path component

query

the query component, or None

fragment

the fragment, or None

Returns

a new Uri

pub fn escape_bytes(
unescaped: &[u8],
reserved_chars_allowed: Option<&str>
) -> GString

Available on crate feature v2_66 only.

Escapes arbitrary data for use in a URI.

Normally all characters that are not ‘unreserved’ (i.e. ASCII alphanumerical characters plus dash, dot, underscore and tilde) are escaped. But if you specify characters in reserved_chars_allowed they are not escaped. This is useful for the ‘reserved’ characters in the URI specification, since those are allowed unescaped in some portions of a URI.

Though technically incorrect, this will also allow escaping nul bytes as %``00.

unescaped

the unescaped input data.

reserved_chars_allowed

a string of reserved characters that are allowed to be used, or None.

Returns

an escaped version of unescaped. The returned string should be freed when no longer needed.

pub fn escape_string(
unescaped: &str,
reserved_chars_allowed: Option<&str>,
allow_utf8: bool
) -> GString

Available on crate feature v2_66 only.

Escapes a string for use in a URI.

Normally all characters that are not “unreserved” (i.e. ASCII alphanumerical characters plus dash, dot, underscore and tilde) are escaped. But if you specify characters in reserved_chars_allowed they are not escaped. This is useful for the “reserved” characters in the URI specification, since those are allowed unescaped in some portions of a URI.

unescaped

the unescaped input string.

reserved_chars_allowed

a string of reserved characters that are allowed to be used, or None.

allow_utf8

true if the result can include UTF-8 characters.

Returns

an escaped version of unescaped. The returned string should be freed when no longer needed.

pub fn is_valid(uri_string: &str, flags: UriFlags) -> Result<(), Error>

Available on crate feature v2_66 only.

Parses uri_string according to flags, to determine whether it is a valid [absolute URI][relative-absolute-uris], i.e. it does not need to be resolved relative to another URI using parse_relative().

If it’s not a valid URI, an error is returned explaining how it’s invalid.

See split(), and the definition of UriFlags, for more information on the effect of flags.

uri_string

a string containing an absolute URI

flags

flags for parsing uri_string

Returns

true if uri_string is a valid absolute URI, false on error.

pub fn join(
flags: UriFlags,
scheme: Option<&str>,
userinfo: Option<&str>,
host: Option<&str>,
port: i32,
path: &str,
query: Option<&str>,
fragment: Option<&str>
) -> GString

Available on crate feature v2_66 only.

Joins the given components together according to flags to create an absolute URI string. path may not be None (though it may be the empty string).

When host is present, path must either be empty or begin with a slash (/) character. When host is not present, path cannot begin with two slash characters (//). See RFC 3986, section 3.

See also join_with_user(), which allows specifying the components of the ‘userinfo’ separately.

UriFlags::HAS_PASSWORD and UriFlags::HAS_AUTH_PARAMS are ignored if set in flags.

flags

flags describing how to build the URI string

scheme

the URI scheme, or None

userinfo

the userinfo component, or None

host

the host component, or None

port

the port, or -1

path

the path component

query

the query component, or None

fragment

the fragment, or None

Returns

an absolute URI string

pub fn join_with_user(
flags: UriFlags,
scheme: Option<&str>,
user: Option<&str>,
password: Option<&str>,
auth_params: Option<&str>,
host: Option<&str>,
port: i32,
path: &str,
query: Option<&str>,
fragment: Option<&str>
) -> GString

Available on crate feature v2_66 only.

Joins the given components together according to flags to create an absolute URI string. path may not be None (though it may be the empty string).

In contrast to join(), this allows specifying the components of the ‘userinfo’ separately. It otherwise behaves the same.

UriFlags::HAS_PASSWORD and UriFlags::HAS_AUTH_PARAMS are ignored if set in flags.

flags

flags describing how to build the URI string

scheme

the URI scheme, or None

user

the user component of the userinfo, or None

password

the password component of the userinfo, or None

auth_params

the auth params of the userinfo, or None

host

the host component, or None

port

the port, or -1

path

the path component

query

the query component, or None

fragment

the fragment, or None

Returns

an absolute URI string

pub fn list_extract_uris(uri_list: &str) -> Vec<GString>

Available on crate feature v2_66 only.

Splits an URI list conforming to the text/uri-list mime type defined in RFC 2483 into individual URIs, discarding any comments. The URIs are not validated.

uri_list

an URI list

Returns

a newly allocated None-terminated list of strings holding the individual URIs. The array should be freed with g_strfreev().

pub fn parse(uri_string: &str, flags: UriFlags) -> Result<Uri, Error>

Available on crate feature v2_66 only.

Parses uri_string according to flags. If the result is not a valid [absolute URI][relative-absolute-uris], it will be discarded, and an error returned.

uri_string

a string representing an absolute URI

flags

flags describing how to parse uri_string

Returns

a new Uri, or NULL on error.

pub fn parse_scheme(uri: &str) -> Option<GString>

Available on crate feature v2_66 only.

Gets the scheme portion of a URI string. RFC 3986 decodes the scheme as:

URI = scheme ":" hier-part [ "?" query ] [ "#" fragment ]

Common schemes include file, https, svn+ssh, etc.

uri

a valid URI.

Returns

The ‘scheme’ component of the URI, or None on error. The returned string should be freed when no longer needed.

pub fn peek_scheme(uri: &str) -> Option<GString>

Available on crate feature v2_66 only.

Gets the scheme portion of a URI string. RFC 3986 decodes the scheme as:

URI = scheme ":" hier-part [ "?" query ] [ "#" fragment ]

Common schemes include file, https, svn+ssh, etc.

Unlike parse_scheme(), the returned scheme is normalized to all-lowercase and does not need to be freed.

uri

a valid URI.

Returns

The ‘scheme’ component of the URI, or None on error. The returned string is normalized to all-lowercase, and interned via g_intern_string(), so it does not need to be freed.

pub fn resolve_relative(
base_uri_string: Option<&str>,
uri_ref: &str,
flags: UriFlags
) -> Result<GString, Error>

Available on crate feature v2_66 only.

Parses uri_ref according to flags and, if it is a [relative URI][relative-absolute-uris], resolves it relative to base_uri_string. If the result is not a valid absolute URI, it will be discarded, and an error returned.

(If base_uri_string is None, this just returns uri_ref, or None if uri_ref is invalid or not absolute.)

base_uri_string

a string representing a base URI

uri_ref

a string representing a relative or absolute URI

flags

flags describing how to parse uri_ref

Returns

the resolved URI string, or NULL on error.

pub fn split(
uri_ref: &str,
flags: UriFlags
) -> Result<(Option<GString>, Option<GString>, Option<GString>, i32, GString, Option<GString>, Option<GString>), Error>

Available on crate feature v2_66 only.

Parses uri_ref (which can be an [absolute or relative URI][relative-absolute-uris]) according to flags, and returns the pieces. Any component that doesn’t appear in uri_ref will be returned as None (but note that all URIs always have a path component, though it may be the empty string).

If flags contains UriFlags::ENCODED, then %-encoded characters in uri_ref will remain encoded in the output strings. (If not, then all such characters will be decoded.) Note that decoding will only work if the URI components are ASCII or UTF-8, so you will need to use UriFlags::ENCODED if they are not.

Note that the UriFlags::HAS_PASSWORD and UriFlags::HAS_AUTH_PARAMS flags are ignored by split(), since it always returns only the full userinfo; use split_with_user() if you want it split up.

uri_ref

a string containing a relative or absolute URI

flags

flags for parsing uri_ref

Returns

true if uri_ref parsed successfully, false on error.

scheme

on return, contains the scheme (converted to lowercase), or None

userinfo

on return, contains the userinfo, or None

host

on return, contains the host, or None

port

on return, contains the port, or -1

path

on return, contains the path

query

on return, contains the query, or None

fragment

on return, contains the fragment, or None

pub fn split_network(
uri_string: &str,
flags: UriFlags
) -> Result<(Option<GString>, Option<GString>, i32), Error>

Available on crate feature v2_66 only.

Parses uri_string (which must be an [absolute URI][relative-absolute-uris]) according to flags, and returns the pieces relevant to connecting to a host. See the documentation for split() for more details; this is mostly a wrapper around that function with simpler arguments. However, it will return an error if uri_string is a relative URI, or does not contain a hostname component.

uri_string

a string containing an absolute URI

flags

flags for parsing uri_string

Returns

true if uri_string parsed successfully, false on error.

scheme

on return, contains the scheme (converted to lowercase), or None

host

on return, contains the host, or None

port

on return, contains the port, or -1

pub fn split_with_user(
uri_ref: &str,
flags: UriFlags
) -> Result<(Option<GString>, Option<GString>, Option<GString>, Option<GString>, Option<GString>, i32, GString, Option<GString>, Option<GString>), Error>

Available on crate feature v2_66 only.

Parses uri_ref (which can be an [absolute or relative URI][relative-absolute-uris]) according to flags, and returns the pieces. Any component that doesn’t appear in uri_ref will be returned as None (but note that all URIs always have a path component, though it may be the empty string).

See split(), and the definition of UriFlags, for more information on the effect of flags. Note that password will only be parsed out if flags contains UriFlags::HAS_PASSWORD, and auth_params will only be parsed out if flags contains UriFlags::HAS_AUTH_PARAMS.

uri_ref

a string containing a relative or absolute URI

flags

flags for parsing uri_ref

Returns

true if uri_ref parsed successfully, false on error.

scheme

on return, contains the scheme (converted to lowercase), or None

user

on return, contains the user, or None

password

on return, contains the password, or None

auth_params

on return, contains the auth_params, or None

host

on return, contains the host, or None

port

on return, contains the port, or -1

path

on return, contains the path

query

on return, contains the query, or None

fragment

on return, contains the fragment, or None

pub fn unescape_bytes(
escaped_string: &str,
illegal_characters: Option<&str>
) -> Result<Bytes, Error>

Available on crate feature v2_66 only.

Unescapes a segment of an escaped string as binary data.

Note that in contrast to unescape_string(), this does allow nul bytes to appear in the output.

If any of the characters in illegal_characters appears as an escaped character in escaped_string, then that is an error and None will be returned. This is useful if you want to avoid for instance having a slash being expanded in an escaped path element, which might confuse pathname handling.

escaped_string

A URI-escaped string

length

the length (in bytes) of escaped_string to escape, or -1 if it is nul-terminated.

illegal_characters

a string of illegal characters not to be allowed, or None.

Returns

an unescaped version of escaped_string or None on error (if decoding failed, using UriError::Failed error code). The returned Bytes should be unreffed when no longer needed.

pub fn unescape_segment(
escaped_string: Option<&str>,
escaped_string_end: Option<&str>,
illegal_characters: Option<&str>
) -> Option<GString>

Available on crate feature v2_66 only.

Unescapes a segment of an escaped string.

If any of the characters in illegal_characters or the NUL character appears as an escaped character in escaped_string, then that is an error and None will be returned. This is useful if you want to avoid for instance having a slash being expanded in an escaped path element, which might confuse pathname handling.

Note: NUL byte is not accepted in the output, in contrast to unescape_bytes().

escaped_string

A string, may be None

escaped_string_end

Pointer to end of escaped_string, may be None

illegal_characters

An optional string of illegal characters not to be allowed, may be None

Returns

an unescaped version of escaped_string, or None on error. The returned string should be freed when no longer needed. As a special case if None is given for escaped_string, this function will return None.

pub fn unescape_string(
escaped_string: &str,
illegal_characters: Option<&str>
) -> Option<GString>

Available on crate feature v2_66 only.

Unescapes a whole escaped string.

If any of the characters in illegal_characters or the NUL character appears as an escaped character in escaped_string, then that is an error and None will be returned. This is useful if you want to avoid for instance having a slash being expanded in an escaped path element, which might confuse pathname handling.

escaped_string

an escaped string to be unescaped.

illegal_characters

a string of illegal characters not to be allowed, or None.

Returns

an unescaped version of escaped_string. The returned string should be freed when no longer needed.