Struct glib::VariantTy

pub struct VariantTy {
    // some fields omitted
}

Describes Variant types.

This is a borrowed counterpart of VariantType. Essentially it's a str statically guaranteed to be a valid type string.

Methods

impl VariantTy

fn new(type_string: &str) -> Result<&VariantTy, ()>

Tries to create a &VariantTy from a string slice.

Returns Ok if the string is a valid type string, Err otherwise.

unsafe fn from_str_unchecked(type_string: &str) -> &VariantTy

Converts a type string into &VariantTy without any checks.

fn to_str(&self) -> &str

Converts to a string slice.