Struct glib::VariantTy

pub struct VariantTy {
    inner: str,
}

Describes Variant types.

This is a borrowed counterpart of VariantType. Essentially it’s a str statically guaranteed to be a valid type string.

Fields
inner: str
Implementations

impl VariantTy

pub fn new(type_string: &str) -> Result<&VariantTy, ()>

Tries to create a &VariantTy from a string slice.

Returns Ok if the string is a valid type string, Err otherwise.

pub unsafe fn from_str_unchecked(type_string: &str) -> &VariantTy

Converts a type string into &VariantTy without any checks.

pub fn to_str(&self) -> &str

Converts to a string slice.