glib

Struct VariantType

pub struct VariantType { /* private fields */ }

Describes Variant types.

The Variant type system (based on the D-Bus one) describes types with “type strings”. VariantType is an owned immutable type string (you can think of it as a Box<str> statically guaranteed to be a valid type string), &VariantTy is a borrowed one (like &str).

Implementations

impl VariantType

pub fn new(type_string: &str) -> Result<VariantType, BoolError>

Tries to create a VariantType from a string slice.

Returns Ok if the string is a valid type string, Err otherwise.

pub fn new_dict_entry( key_type: &VariantTy, value_type: &VariantTy, ) -> VariantType

Creates a VariantType from a key and value type.

pub fn new_array(elem_type: &VariantTy) -> VariantType

Creates a VariantType from an array element type.

pub fn new_maybe(child_type: &VariantTy) -> VariantType

Creates a VariantType from a maybe element type.

pub fn new_tuple( items: impl IntoIterator<Item = impl AsRef<VariantTy>>, ) -> VariantType

Creates a VariantType from a maybe element type.

pub fn from_string( type_string: impl Into<GString>, ) -> Result<VariantType, BoolError>

Tries to create a VariantType from an owned string.

Returns Ok if the string is a valid type string, Err otherwise.