Trait glib::subclass::interface::ObjectInterfaceExt

pub trait ObjectInterfaceExt: ObjectInterface {
    // Provided methods
    fn from_instance<T: IsA<Object>>(obj: &T) -> &Self { ... }
    fn from_obj<T: IsA<Object>>(obj: &T) -> &Self { ... }
}

Provided Methods

fn from_instance<T: IsA<Object>>(obj: &T) -> &Self

👎Deprecated: Use from_obj() instead

Get interface from an instance.

This will panic if obj does not implement the interface.

fn from_obj<T: IsA<Object>>(obj: &T) -> &Self

Get interface from an instance.

This will panic if obj does not implement the interface.

Object Safety

This trait is not object safe.