glib::subclass::shared

Trait SharedType

pub trait SharedType:
    StaticType
    + Clone
    + Sized
    + 'static {
    type RefCountedType: RefCounted;

    const NAME: &'static str;

    // Required methods
    fn into_refcounted(self) -> Self::RefCountedType;
    fn from_refcounted(this: Self::RefCountedType) -> Self;
}

Trait for defining shared types.

Links together the type name with the type itself.

See register_shared_type for registering an implementation of this trait with the type system.

Required Associated Constants

const NAME: &'static str

Shared type name.

This must be unique in the whole process.

Required Associated Types

type RefCountedType: RefCounted

The inner refcounted type

Required Methods

fn into_refcounted(self) -> Self::RefCountedType

Converts the SharedType into its inner RefCountedType

fn from_refcounted(this: Self::RefCountedType) -> Self

Constructs a SharedType from a RefCountedType

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.