glib::subclass::signal
Struct SignalBuilder

pub struct SignalBuilder { /* private fields */ }

Builder for signals.

Implementations

impl SignalBuilder

pub fn param_types( self, param_types: impl IntoIterator<Item = impl Into<SignalType>>, ) -> Self

The signal’s parameters.

pub fn return_type<T: StaticType>(self) -> Self

The signal’s returned value type.

pub fn return_type_from(self, type_: impl Into<SignalType>) -> Self

The signal’s returned value type.

pub fn run_first(self) -> Self

Run the signal class handler in the first emission stage.

pub fn run_last(self) -> Self

Run the signal class handler in the third emission stage.

pub fn run_cleanup(self) -> Self

Run the signal class handler in the last emission stage.

pub fn no_recurse(self) -> Self

Signals being emitted for an object while currently being in emission for this very object will not be emitted recursively, but instead cause the first emission to be restarted.

pub fn detailed(self) -> Self

This signal supports “::detail” appendices to the signal name upon handler connections and emissions.

pub fn action(self) -> Self

Action signals are signals that may freely be emitted on alive objects from user code.

pub fn no_hooks(self) -> Self

No emissions hooks are supported for this signal.

pub fn must_collect(self) -> Self

Varargs signal emission will always collect the arguments, even if there are no signal handlers connected.

pub fn deprecated(self) -> Self

The signal is deprecated and will be removed in a future version.

pub fn flags(self, flags: SignalFlags) -> Self

Explicitly set all flags.

This overrides previously set flags on this builder.

pub fn class_handler<F: Fn(&[Value]) -> Option<Value> + Send + Sync + 'static>( self, func: F, ) -> Self

Class handler for this signal.

pub fn accumulator<F: Fn(&SignalInvocationHint, Value, &Value) -> ControlFlow<Value, Value> + Send + Sync + 'static>( self, func: F, ) -> Self

Accumulator for the return values of the signal.

This is called if multiple signal handlers are connected to the signal for accumulating the return values into a single value.

pub fn build(self) -> Signal

Build the signal.

This does not register the signal yet, which only happens as part of object type registration.