Struct glib::subclass::signal::SignalBuilder

pub struct SignalBuilder<'a> { /* fields omitted */ }

Builder for signals.

Implementations

impl<'a> SignalBuilder<'a>

pub fn run_first(mut self: Self) -> Self

Run the signal class handler in the first emission stage.

pub fn run_last(mut self: Self) -> Self

Run the signal class handler in the third emission stage.

pub fn run_cleanup(mut self: Self) -> Self

Run the signal class handler in the last emission stage.

pub fn no_recurse(mut self: Self) -> Self

Signals being emitted for an object while currently being in emission for this very object will not be emitted recursively, but instead cause the first emission to be restarted.

pub fn detailed(mut self: Self) -> Self

This signal supports "::detail" appendices to the signal name upon handler connections and emissions.

pub fn action(mut self: Self) -> Self

Action signals are signals that may freely be emitted on alive objects from user code.

pub fn no_hooks(mut self: Self) -> Self

No emissions hooks are supported for this signal.

pub fn must_collect(mut self: Self) -> Self

Varargs signal emission will always collect the arguments, even if there are no signal handlers connected.

pub fn deprecated(mut self: Self) -> Self

The signal is deprecated and will be removed in a future version.

pub fn flags(mut self: Self, flags: SignalFlags) -> Self

Explicitly set all flags.

This overrides previously set flags on this builder.

pub fn class_handler<F: Fn(&SignalClassHandlerToken, &[Value]) -> Option<Value> + Send + Sync + 'static>(
mut self: Self,
func: F
) -> Self

Class handler for this signal.

pub fn accumulator<F: Fn(&SignalInvocationHint, &mut Value, &Value) -> bool + Send + Sync + 'static>(
mut self: Self,
func: F
) -> Self

Accumulator for the return values of the signal.

This is called if multiple signal handlers are connected to the signal for accumulating the return values into a single value.

pub fn build(self) -> Signal

Build the signal.

This does not register the signal yet, which only happens as part of object type registration.