Trait glib::subclass::type_module::TypeModuleImpl

pub trait TypeModuleImpl: ObjectImpl + TypeModuleImplExt {
    // Required methods
    fn load(&self) -> bool;
    fn unload(&self);
}

Required Methods

fn load(&self) -> bool

Loads the module, registers one or more object subclasses using register_dynamic_type and registers one or more object interfaces using register_dynamic_interface (see TypeModule).

loads the module and registers one or more types using [TypeModuleExtManual::register_type()][crate::prelude::TypeModuleExtManual::register_type()].

fn unload(&self)

Unloads the module (see TypeModuleExt::unuse).

unloads the module

Object Safety

This trait is not object safe.