Module glib::subclass::types

Module that contains the basic infrastructure for subclassing GObject.

Structs

InitializingType
A newly registered glib::Type that is currently still being initialized.

SignalClassHandlerToken

SignalInvocationHint

TypeData
Type-specific data that is filled in during type creation.

Traits

ClassStruct
Trait implemented by structs that implement a GObject C class struct.

InstanceStruct
Trait implemented by structs that implement a GObject C instance struct.

IsImplementable
Trait for implementable interfaces.

IsSubclassable
Trait for subclassable class structs.

ObjectSubclass
The central trait for subclassing a GObject type.

Functions

register_type
Register a glib::Type ID for T.