Module glib::subclass::types

Module that contains the basic infrastructure for subclassing GObject.

Structs

InitializingObject

An object that is currently being initialized.

InitializingType

A newly registered glib::Type that is currently still being initialized.

TypeData

Type-specific data that is filled in during type creation.

Constants

INIT_TYPE_DATA

Traits

ClassStruct

Trait implemented by structs that implement a GObject C class struct.

FromObject

Helper trait for macros to access a subclass or its wrapper.

InstanceStruct

Trait implemented by structs that implement a GObject C instance struct.

InterfaceList

Trait for a type list of interfaces.

IsImplementable

Trait for implementable interfaces.

IsSubclassable

Trait for subclassable class structs.

IsSubclassableExt

ObjectSubclass

The central trait for subclassing a GObject type.

ObjectSubclassExt

Extension methods for all ObjectSubclass impls.

ObjectSubclassIsExt

Trait implemented by any type implementing ObjectSubclassIs to return the implementation, private Rust struct.

ObjectSubclassType

Type methods required for an ObjectSubclass implementation.

Functions

register_type

Register a glib::Type ID for T.