Module glib::subclass::types

Module that contains the basic infrastructure for subclassing GObject.

Structs

InitializingType

A newly registered glib::Type that is currently still being initialized.

SignalClassHandlerToken

SignalInvocationHint

TypeData

Type-specific data that is filled in during type creation.

Traits

ClassStruct

Trait implemented by structs that implement a GObject C class struct.

InstanceStruct

Trait implemented by structs that implement a GObject C instance struct.

IsImplementable

Trait for implementable interfaces.

IsSubclassable

Trait for subclassable class structs.

ObjectSubclass

The central trait for subclassing a GObject type.

Functions

add_signal⚠

add_signal_with_accumulator⚠

add_signal_with_class_handler⚠

add_signal_with_class_handler_and_accumulator⚠

class_init⚠

finalize⚠

instance_init⚠

register_type

Register a glib::Type ID for T.

signal_chain_from_overridden⚠

signal_override_class_handler⚠