Struct glib::translate::Stash

pub struct Stash<'a, P: Copy, T: ?Sized + ToGlibPtr<'a, P>>(pub P, pub <T as ToGlibPtr<'a, P>>::Storage);

Helper type that stores temporary values used for translation.

P is the foreign type pointer and the first element of the tuple.

T is the Rust type that is translated.

The second element of the tuple is the temporary storage defined by the implementation of ToGlibPtr<P> for T

Say you want to pass a *mut GdkWindowAttr to a foreign function. The Stash will own a GdkWindowAttr and a CString that GdkWindowAttr::title points into.

impl <'a> ToGlibPtr<'a, *mut glib_ffi::GdkWindowAttr> for WindowAttr {
    type Storage = (Box<glib_ffi::GdkWindowAttr>, Stash<'a, *const c_char, Option<String>>);

    fn to_glib_none(&'a self) -> Stash<*mut glib_ffi::GdkWindowAttr, WindowAttr> {
        let title = self.title.to_glib_none();

        let mut attrs = Box::new(glib_ffi::GdkWindowAttr {
            title: title.0,
            // ....
        });

        Stash(&mut *attrs, (attrs, title))
    }
}