Trait glib::translate::FromGlibPtrBorrow

pub trait FromGlibPtrBorrow<P: Ptr>: Sized {
    unsafe fn from_glib_borrow(_ptr: P) -> Borrowed<Self> { ... }
}

Translate from a pointer type by borrowing, without affecting the refcount.

The purpose of this trait is to access values inside callbacks without changing their reference status. The obtained borrow must not be accessed outside of the scope of the callback, and called procedures must not store any references to the underlying data. Safe Rust code must never obtain a mutable Rust reference.

Safety

The implementation of this trait as well as the returned type must satisfy the same constraints together. They must not take ownership of the underlying value, copy it, and should not change its rerefence count. If it does, it must properly release obtained references.

The returned value, when dropped, must leave the underlying value in the same state as before from_glib_borrow was called:

it must not be dropped,

it must be the same type of reference, e.g. still floating.

For more information, refer to module level documentation.

Provided methods

unsafe fn from_glib_borrow(_ptr: P) -> Borrowed<Self>

Safety

See trait level notes on safety

Uses found in examples/

examples/gio_task/file_size/ffi.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86

use gtk::glib::subclass::prelude::*;
use gtk::glib::translate::*;
use gtk::prelude::*;
use gtk::{gio, glib};

pub type FileSize = <super::imp::FileSize as ObjectSubclass>::Instance;

/// # Safety
///
/// This is the ffi method to asynchronously get the file size. It accepts a callback of type
/// GAsyncReadyCallback, that will be invoked when the async operation completes. Typically,
/// this callback will invoke the get_file_size_finish method (implemented below) to get the
/// Task result and perform some operation with it.
#[no_mangle]
pub unsafe extern "C" fn my_file_size_get_file_size_async(
    this: *mut FileSize,
    cancellable: *mut gio::ffi::GCancellable,
    callback: gio::ffi::GAsyncReadyCallback,
    user_data: glib::ffi::gpointer,
) {
    let cancellable = gio::Cancellable::from_glib_borrow(cancellable);
    let closure = move |result: &gio::AsyncResult, source_object: Option<&glib::Object>| {
        let result: *mut gio::ffi::GAsyncResult = result.to_glib_none().0;
        let source_object: *mut glib::object::GObject = source_object.to_glib_none().0;
        callback.unwrap()(source_object, result, user_data)
    };

    let source_object = &super::FileSize::from_glib_borrow(this);
    let task = gio::Task::new(
        Some(&source_object.upcast_ref::<glib::Object>()),
        Some(cancellable.as_ref()),
        closure,
    );

    glib::MainContext::default().spawn_local(async move {
        let size = gio::File::for_path("Cargo.toml")
            .query_info_async_future("*", gio::FileQueryInfoFlags::NONE, glib::PRIORITY_DEFAULT)
            .await
            .unwrap()
            .size();

        let source_object = task
            .upcast_ref::<gio::AsyncResult>()
            .source_object()
            .unwrap();

        let source_object = super::imp::FileSize::from_instance(
            &source_object.downcast_ref::<super::FileSize>().unwrap(),
        );

        source_object.size.replace(Some(size));
        task.return_value(&size.to_value());
    });
}

/// # Safety
///
/// This method will typically be invoked in the callback passed to my_file_size_get_file_size_async in order
/// to propagate the Task result.
#[no_mangle]
pub unsafe extern "C" fn my_file_size_get_file_size_finish(
    _this: *mut FileSize,
    result: *mut gio::ffi::GAsyncResult,
    _error: *mut *mut glib::ffi::GError,
) -> i64 {
    gio::AsyncResult::from_glib_borrow(result)
        .downcast_ref::<gio::Task>()
        .unwrap()
        .propagate_value()
        .unwrap()
        .get::<i64>()
        .unwrap()
}

/// # Safety
///
/// Simple getter
#[no_mangle]
pub unsafe extern "C" fn my_file_size_get_retrieved_size(this: *mut FileSize) -> i64 {
    let simple_object = super::FileSize::from_glib_borrow(this);
    let simple_object = super::imp::FileSize::from_instance(
        &simple_object.downcast_ref::<super::FileSize>().unwrap(),
    );
    let x = *simple_object.size.borrow();
    x.unwrap_or(-1)
}