Trait glib::translate::FromGlibPtrFull

pub trait FromGlibPtrFull<P: Ptr>: Sized {
    // Required method
    unsafe fn from_glib_full(ptr: P) -> Self;
}

Translate from a pointer type which is annotated with transfer full. This transfers the ownership of the value to the Rust side.

Because ownership can only be transferred if something is already referenced, this is unsuitable for floating references.

Safety

The implementation of this trait should not alter the reference count or make copies of the underlying value. Values obtained using this trait must be properly released on drop() by the implementing type.

For more information, refer to module level documentation.

Required Methods

unsafe fn from_glib_full(ptr: P) -> Self

Safety

See trait level notes on safety

Object Safety

This trait is not object safe.