Trait glib::translate::FromGlibPtrNone

pub trait FromGlibPtrNone<P: Ptr>: Sized {
    // Required method
    unsafe fn from_glib_none(ptr: P) -> Self;
}

Translate from a pointer type which is annotated with transfer none. The resulting value is referenced at least once, by the bindings.

This is suitable for floating references, which become strong references. It is also suitable for acquiring non-gobject values, like gchar*.

Safety

The implementation of this trait should acquire a reference to the value in a way appropriate to the type, e.g. by increasing the reference count or copying. Values obtained using this trait must be properly released on drop() by the implementing type.

For more information, refer to module level documentation.

Required Methods

unsafe fn from_glib_none(ptr: P) -> Self

Safety

See trait level notes on safety

Object Safety

This trait is not object safe.