Trait glib::translate::TransparentType

pub unsafe trait TransparentType: Clone + Sized {
    type GlibType;
}

Trait for types that have the same memory representation as their FFI type.

Values of types implementing this trait can be transmuted directly to the FFI type, references to pointers to the FFI type.

Required Associated Types

type GlibType

Object Safety

This trait is not object safe.