Trait glib::translate::UnsafeFrom

pub trait UnsafeFrom<T> {
    // Required method
    unsafe fn unsafe_from(t: T) -> Self;
}

Unsafe variant of the From trait.

Required Methods

unsafe fn unsafe_from(t: T) -> Self

Safety

It is the responsibility of the caller to ensure all invariants of the T hold before this is called, and that after conversion to assume nothing other than the invariants of the output. Implementors of this must ensure that the invariants of the output type hold.

Object Safety

This trait is not object safe.