Struct glib::types::Type

#[repr(transparent)]
pub struct Type(_);

A GLib or GLib-based library type

Implementations

impl Type

pub const INVALID: Self = Self(gobject_ffi::G_TYPE_INVALID)

An invalid Type used as error return value in some functions

pub const UNIT: Self = Self(gobject_ffi::G_TYPE_NONE)

The fundamental type corresponding to the unit type ()

pub const I8: Self = Self(gobject_ffi::G_TYPE_CHAR)

The fundamental type corresponding to i8

pub const U8: Self = Self(gobject_ffi::G_TYPE_UCHAR)

The fundamental type corresponding to u8

pub const BOOL: Self = Self(gobject_ffi::G_TYPE_BOOLEAN)

The fundamental type corresponding to bool

pub const I32: Self = Self(gobject_ffi::G_TYPE_INT)

The fundamental type corresponding to i32

pub const U32: Self = Self(gobject_ffi::G_TYPE_UINT)

The fundamental type corresponding to u32

pub const I_LONG: Self = Self(gobject_ffi::G_TYPE_LONG)

The fundamental type corresponding to C long

pub const U_LONG: Self = Self(gobject_ffi::G_TYPE_ULONG)

The fundamental type corresponding to C unsigned long

pub const I64: Self = Self(gobject_ffi::G_TYPE_INT64)

The fundamental type corresponding to i64

pub const U64: Self = Self(gobject_ffi::G_TYPE_UINT64)

The fundamental type corresponding to u64

pub const F32: Self = Self(gobject_ffi::G_TYPE_FLOAT)

The fundamental type corresponding to f32

pub const F64: Self = Self(gobject_ffi::G_TYPE_DOUBLE)

The fundamental type corresponding to f64

pub const STRING: Self = Self(gobject_ffi::G_TYPE_STRING)

The fundamental type corresponding to String

pub const POINTER: Self = Self(gobject_ffi::G_TYPE_POINTER)

The fundamental type corresponding to a pointer

pub const VARIANT: Self = Self(gobject_ffi::G_TYPE_VARIANT)

The fundamental type of GVariant

pub const INTERFACE: Self = Self(gobject_ffi::G_TYPE_INTERFACE)

The fundamental type from which all interfaces are derived

pub const ENUM: Self = Self(gobject_ffi::G_TYPE_ENUM)

The fundamental type from which all enumeration types are derived

pub const FLAGS: Self = Self(gobject_ffi::G_TYPE_FLAGS)

The fundamental type from which all flags types are derived

pub const BOXED: Self = Self(gobject_ffi::G_TYPE_BOXED)

The fundamental type from which all boxed types are derived

pub const PARAM_SPEC: Self = Self(gobject_ffi::G_TYPE_PARAM)

The fundamental type from which all GParamSpec types are derived

pub const OBJECT: Self = Self(gobject_ffi::G_TYPE_OBJECT)

The fundamental type from which all objects are derived

pub fn name<'a>(self) -> &'a str

pub fn qname(self) -> Quark

pub fn is_a(self, other: Self) -> bool

pub fn parent(self) -> Option<Self>

pub fn children(self) -> Slice<Self>

pub fn interfaces(self) -> Slice<Self>

pub fn interface_prerequisites(self) -> Slice<Self>

pub fn from_name(name: &str) -> Option<Self>

pub fn is_valid(self) -> bool

Checks that the type is not INVALID