Struct glib::types::Type

#[repr(transparent)]
pub struct Type(_);

A GLib or GLib-based library type

Implementations

impl Type

pub const INVALID: Self

An invalid Type used as error return value in some functions

pub const UNIT: Self

The fundamental type corresponding to the unit type ()

pub const I8: Self

The fundamental type corresponding to i8

pub const U8: Self

The fundamental type corresponding to u8

pub const BOOL: Self

The fundamental type corresponding to bool

pub const I32: Self

The fundamental type corresponding to i32

pub const U32: Self

The fundamental type corresponding to u32

pub const I_LONG: Self

The fundamental type corresponding to C long

pub const U_LONG: Self

The fundamental type corresponding to C unsigned long

pub const I64: Self

The fundamental type corresponding to i64

pub const U64: Self

The fundamental type corresponding to u64

pub const F32: Self

The fundamental type corresponding to f32

pub const F64: Self

The fundamental type corresponding to f64

pub const STRING: Self

The fundamental type corresponding to String

pub const POINTER: Self

The fundamental type corresponding to a pointer

pub const VARIANT: Self

The fundamental type of GVariant

pub const INTERFACE: Self

The fundamental type from which all interfaces are derived

pub const ENUM: Self

The fundamental type from which all enumeration types are derived

pub const FLAGS: Self

The fundamental type from which all flags types are derived

pub const BOXED: Self

The fundamental type from which all boxed types are derived

pub const PARAM_SPEC: Self

The fundamental type from which all GParamSpec types are derived

pub const OBJECT: Self

The fundamental type from which all objects are derived

pub fn name<'a>(self) -> &'a str

pub fn qname(self) -> Quark

pub fn is_a(self, other: Self) -> bool

pub fn parent(self) -> Option<Self>

pub fn children(self) -> Slice<Self>

pub fn interfaces(self) -> Slice<Self>

pub fn interface_prerequisites(self) -> Slice<Self>

pub fn from_name(name: &str) -> Option<Self>

pub fn is_valid(self) -> bool

Checks that the type is not INVALID