Struct glib::value::AnySendValue

pub struct AnySendValue(_);

A container type that allows storing any 'static type that implements Any, Clone and Send to be stored in a Value or SendValue.

See the module documentation for more details and the AnyValue for a code example.

Methods

impl AnySendValue

pub fn new<T: Any + Clone + Send + 'static>(val: T) -> Self

Create a new AnySendValue from val.

pub fn downcast<T: Any + Clone + Send + 'static>(self) -> Result<T, Self>

Attempt the value to its concrete type.