Struct glib::value::SendValue

#[repr(C)]

pub struct SendValue(_);

A version of Value for storing Send types, that implements Send itself.

See the module documentation for more details.

Methods

impl SendValue

pub fn downcast<'a, T: FromValueOptional<'a> + SetValue + Send>(
self
) -> Result<TypedValue<T>, Self>

Tries to downcast to a TypedValue.

Returns Ok(TypedValue<T>) if the value carries a type corresponding to T and Err(self) otherwise.

pub fn downcast_ref<'a, T: FromValueOptional<'a> + SetValue>(
&self
) -> Option<&TypedValue<T>>

Tries to downcast to a &TypedValue.

Returns Some(&TypedValue<T>) if the value carries a type corresponding to T and None otherwise.