Struct glib::value::TypedValue

pub struct TypedValue<T>(_, _);

A statically typed Value.

It dereferences to Value and can be used everywhere Value references are accepted.

See the module documentation for more details.

Methods

impl<'a, T: FromValueOptional<'a> + SetValue> TypedValue<T>

pub fn get(&'a self) -> Option<T>

Returns the value.

Types that don't support a None value always return Some. See get_some.

pub fn get_some(&'a self) -> T where
T: FromValue<'a>,

Returns the value.

This method is only available for types that don't support a None value.

pub fn set<U: ?Sized + SetValueOptional>(&mut self, value: Option<&U>) where
T: Borrow<U>,

Sets the value.

This method is only available for types that support a None value.

pub fn set_none(&mut self) where
T: SetValueOptional,

Sets the value to None.

This method is only available for types that support a None value.

pub fn set_some<U: ?Sized + SetValue>(&mut self, value: &U) where
T: Borrow<U>,

Sets the value.