Struct glib::value::Value

#[repr(C)]
pub struct Value(_, _);

A generic value capable of carrying various types.

Once created the type of the value can't be changed.

Some types (e.g. String and objects) support None values while others (e.g. numeric types) don't.

Value does not implement the Send trait, but SendValue can be used instead.

See the module documentation for more details.

Methods

impl Value

pub fn from_type(type_: Type) -> Self

Creates a new Value that is initialized with type_

pub fn downcast<'a, T: FromValueOptional<'a> + SetValue>(
self
) -> Result<TypedValue<T>, Self>

Tries to downcast to a TypedValue.

Returns Ok(TypedValue<T>) if the value carries a type corresponding to T and Err(self) otherwise.

pub fn downcast_ref<'a, T: FromValueOptional<'a> + SetValue>(
&self
) -> Option<&TypedValue<T>>

Tries to downcast to a &TypedValue.

Returns Some(&TypedValue<T>) if the value carries a type corresponding to T and None otherwise.

pub fn get<'a, T: FromValueOptional<'a>>(
&'a self
) -> Result<Option<T>, GetError>

Tries to get a possibly optional value of type T.

Returns Ok if the type is correct.

pub fn get_some<'a, T: FromValue<'a>>(&'a self) -> Result<T, GetError>

Tries to get a value of type T.

This method is only available for types that don't support a None value.

Returns Ok if the type is correct.

pub fn is<'a, T: FromValueOptional<'a> + SetValue>(&self) -> bool

Returns true if the type of the value corresponds to T or is a sub-type of T.

pub fn type_(&self) -> Type

Returns the type of the value.

pub fn type_transformable(src: Type, dst: Type) -> bool

Returns whether Values of type src can be transformed to type dst.

pub fn transform<T: StaticType + SetValue>(&self) -> Option<Value>

Tries to transform the value into a value of the target type

pub fn try_into_send_value<'a, T: Send + FromValueOptional<'a> + SetValue>(
self
) -> Result<SendValue, Self>