glib::value

Struct Value

pub struct Value { /* private fields */ }

A generic value capable of carrying various types.

Once created the type of the value canโt be changed.

Some types (e.g. String and objects) support None values while others (e.g. numeric types) donโt.

Value does not implement the Send trait, but SendValue can be used instead.

See the module documentation for more details.

GLib type: Inline allocated boxed type with stack copy semantics.

Implementations

impl Value

pub fn as_ptr(&self) -> *mut GValue

pub unsafe fn from_glib_ptr_borrow<'a>(ptr: *const GValue) -> &'a Self

Borrows the underlying C value.

pub unsafe fn from_glib_ptr_borrow_mut<'a>(ptr: *mut GValue) -> &'a mut Self

Borrows the underlying C value mutably.

impl Value

pub fn from_type(type_: Type) -> Self

Creates a new Value that is initialized with type_.

Panics

If type_ canโt be stored in a Value this function panics.

pub unsafe fn from_type_unchecked(type_: Type) -> Self

Creates a new Value that is initialized with type_.

SAFETY

This must be called with a valid type_ that can be stored in Values.

pub fn for_value_type<T: ValueType>() -> Self

Creates a new Value that is initialized for a given ValueType.

pub fn from_static_str(s: &'static GStr) -> Self

Creates a new String-typed Value from a 'static string.

pub fn from_interned_str(s: &'static GStr) -> Self

Available on crate feature v2_66 only.

Creates a new String-typed Value from a 'static string that is also assumed to be interned.

pub fn get<'a, T>( &'a self, ) -> Result<T, <<T as FromValue<'a>>::Checker as ValueTypeChecker>::Error>
where T: FromValue<'a>,

Tries to get a value of type T.

Returns Ok if the type is correct.

pub fn get_owned<T>( &self, ) -> Result<T, <<T as FromValue<'_>>::Checker as ValueTypeChecker>::Error>
where T: for<'b> FromValue<'b> + 'static,

Tries to get a value of an owned type T.

pub fn is<T: StaticType>(&self) -> bool

Returns true if the type of the value corresponds to T or is a sub-type of T.

pub fn is_type(&self, type_: Type) -> bool

Returns true if the type of the value corresponds to type_ or is a sub-type of type_.

pub fn type_(&self) -> Type

Returns the type of the value.

pub fn type_transformable(src: Type, dst: Type) -> bool

Returns whether Values of type src can be transformed to type dst. Checks whether [methodObject.transform] is able to transform values of type src_type into values of type dest_type.

Note that for the types to be transformable, they must be compatible or a transformation function must be registered using [funcObject.register_transform_func].

src_type

source type

dest_type

target type

Returns

true if the transformation is possible; false otherwise of GLib.

Prior to version 2.70, g_spawn_check_exit_status() provides the same functionality, although under a misleading name.

wait_status

A platform-specific wait status as returned from g_spawn_sync()

Returns

true if child exited successfully, false otherwise (and @error will be set) Checks whether [methodObject.transform] is able to transform values of type src_type into values of type dest_type.

Note that for the types to be transformable, they must be compatible or a transformation function must be registered using [funcObject.register_transform_func].

src_type

source type

dest_type

target type

Returns

true if the transformation is possible; false otherwise of GLib.

Prior to version 2.70, g_spawn_check_exit_status() provides the same functionality, although under a misleading name.

wait_status

A platform-specific wait status as returned from g_spawn_sync()

Returns

true if child exited successfully, false otherwise (and @error will be set)

pub fn transform<T: ValueType>(&self) -> Result<Value, BoolError>

Tries to transform the value into a value of the target type Tries to cast the contents of self into a type appropriate to store in dest_value.

If a transformation is not possible, dest_value is not modified.

For example, this could transform a G_TYPE_INT value into a G_TYPE_FLOAT value.

Performing transformations between value types might incur precision loss. Especially transformations into strings might reveal seemingly arbitrary results and the format of particular transformations to strings is not guaranteed over time.

dest_value

target value

Returns

true on success; false otherwise Tries to cast the contents of self into a type appropriate to store in dest_value.

If a transformation is not possible, dest_value is not modified.

For example, this could transform a G_TYPE_INT value into a G_TYPE_FLOAT value.

Performing transformations between value types might incur precision loss. Especially transformations into strings might reveal seemingly arbitrary results and the format of particular transformations to strings is not guaranteed over time.

dest_value

target value

Returns

true on success; false otherwise

pub fn transform_with_type(&self, type_: Type) -> Result<Value, BoolError>

Tries to transform the value into a value of the target type

pub fn into_raw(self) -> GValue

Consumes Value and returns the corresponding GValue.

pub fn try_into_send_value<T: Send + StaticType>( self, ) -> Result<SendValue, Self>

Converts a Value into a SendValue. This fails if self does not store a value of type T. It is required for T to be Send to call this function.

pub unsafe fn into_send_value(self) -> SendValue

Converts a Value into a SendValue.

Safety

The type of the value contained in self must be Send.