Struct glib::value::Value

#[repr(transparent)]
pub struct Value { /* private fields */ }

A generic value capable of carrying various types.

Once created the type of the value can’t be changed.

Some types (e.g. String and objects) support None values while others (e.g. numeric types) don’t.

Value does not implement the Send trait, but SendValue can be used instead.

See the module documentation for more details.

Implementations

impl Value

pub fn as_ptr(&self) -> *mut GValue

pub unsafe fn from_glib_ptr_borrow<'a>(ptr: *const GValue) -> &'a Self

Borrows the underlying C value.

pub unsafe fn from_glib_ptr_borrow_mut<'a>(ptr: *mut GValue) -> &'a mut Self

Borrows the underlying C value mutably.

impl Value

pub fn from_type(type_: Type) -> Self

Creates a new Value that is initialized with type_.

Panics

If type_ can’t be stored in a Value this function panics.

pub unsafe fn from_type_unchecked(type_: Type) -> Self

Creates a new Value that is initialized with type_.

SAFETY

This must be called with a valid type_ that can be stored in Values.

pub fn for_value_type<T: ValueType>() -> Self

Creates a new Value that is initialized for a given ValueType.

pub fn get<'a, T>( &'a self ) -> Result<T, <<T as FromValue<'_>>::Checker as ValueTypeChecker>::Error>where T: FromValue<'a>,

Tries to get a value of type T.

Returns Ok if the type is correct.

pub fn get_owned<T>( &self ) -> Result<T, <<T as FromValue<'_>>::Checker as ValueTypeChecker>::Error>where T: for<'b> FromValue<'b> + 'static,

Tries to get a value of an owned type T.

pub fn is<T: StaticType>(&self) -> bool

Returns true if the type of the value corresponds to T or is a sub-type of T.

pub fn is_type(&self, type_: Type) -> bool

Returns true if the type of the value corresponds to type_ or is a sub-type of type_.

pub fn type_(&self) -> Type

Returns the type of the value.

pub fn type_transformable(src: Type, dst: Type) -> bool

Returns whether Values of type src can be transformed to type dst. Check whether transform() is able to transform values of type src_type into values of type dest_type. Note that for the types to be transformable, they must be compatible or a transformation function must be registered.

src_type

Source type.

dest_type

Target type.

Returns

true if the transformation is possible, false otherwise.

pub fn transform<T: ValueType>(&self) -> Result<Value, BoolError>

Tries to transform the value into a value of the target type

pub fn transform_with_type(&self, type_: Type) -> Result<Value, BoolError>

Tries to transform the value into a value of the target type

pub fn into_raw(self) -> GValue

Consumes Value and returns the corresponding GValue.

pub fn try_into_send_value<T: Send + StaticType>( self ) -> Result<SendValue, Self>