Struct glib::value::Value

pub struct Value(_);

A generic value capable of carrying various types.

Once created the type of the value can't be changed.

Some types (e.g. String and objects) support None values while others (e.g. numeric types) don't.

See the module documentation for more details.

Methods

impl Value

pub fn downcast<'a, T: FromValueOptional<'a> + SetValue>(
self
) -> Result<TypedValue<T>, Self>

Tries to downcast to a TypedValue.

Returns Ok(TypedValue<T>) if the value carries a type corresponding to T and Err(self) otherwise.

pub fn get<'a, T: FromValueOptional<'a>>(&'a self) -> Option<T>

Tries to get a value of type T.

Returns Some if the type is correct and the value is not None.

This function doesn't distinguish between type mismatches and correctly typed None values. Use downcast or is for that.

pub fn is<'a, T: FromValueOptional<'a> + SetValue>(&self) -> bool

Returns true if the type of the value corresponds to T.

pub fn type_(&self) -> Type

Returns the type of the value.

pub fn type_transformable(src: Type, dst: Type) -> bool

Returns whether Values of type src can be transformed to type dst.