Struct glib::value::Value

#[repr(transparent)]
pub struct Value { /* private fields */ }

A generic value capable of carrying various types.

Once created the type of the value can’t be changed.

Some types (e.g. String and objects) support None values while others (e.g. numeric types) don’t.

Value does not implement the Send trait, but SendValue can be used instead.

See the module documentation for more details. An opaque structure used to hold different types of values.

The data within the structure has protected scope: it is accessible only to functions within a GTypeValueTable structure, or implementations of the g_value_*() API. That is, code portions which implement new fundamental types.

Value users cannot make any assumptions about how data is stored within the 2 element data union, and the g_type member should only be accessed through the G_VALUE_TYPE() macro.

Implementations

impl Value

pub fn from_type(type_: Type) -> Self

Creates a new Value that is initialized with type_

pub fn for_value_type<T: ValueType>() -> Self

Creates a new Value that is initialized for a given ValueType.

pub fn get<'a, T>(
&'a self
) -> Result<T, <<T as FromValue<'_>>::Checker as ValueTypeChecker>::Error> where
T: FromValue<'a>,

Tries to get a value of type T.

Returns Ok if the type is correct.

pub fn get_owned<T>(
&self
) -> Result<T, <<T as FromValue<'_>>::Checker as ValueTypeChecker>::Error> where
T: for<'b> FromValue<'b> + 'static,

Tries to get a value of an owned type T.

pub fn is<T: StaticType>(&self) -> bool

Returns true if the type of the value corresponds to T or is a sub-type of T.

pub fn type_(&self) -> Type

Returns the type of the value.

pub fn type_transformable(src: Type, dst: Type) -> bool

Returns whether Values of type src can be transformed to type dst.

pub fn transform<T: ValueType>(&self) -> Result<Value, BoolError>

Tries to transform the value into a value of the target type

pub fn into_raw(self) -> GValue

Consumes Value and returns the corresponding GValue.

pub fn try_into_send_value<T: Send + StaticType>(
self
) -> Result<SendValue, Self>