Trait glib::value::FromValue

pub unsafe trait FromValue<'a>: Sized {
    type Checker: ValueTypeChecker;
    unsafe fn from_value(value: &'a Value) -> Self;
}

Trait to retrieve the contained value from a Value.

Usually this would not be used directly but from the get function on a Value

Associated Types

type Checker: ValueTypeChecker

Value type checker.

Required methods

unsafe fn from_value(value: &'a Value) -> Self

Get the contained value from a Value.

Safety

Self::Checker::check() must be called first and must not fail.