glib::value

Trait FromValue

pub unsafe trait FromValue<'a>: Sized {
    type Checker: ValueTypeChecker;

    // Required method
    unsafe fn from_value(value: &'a Value) -> Self;
}

Trait to retrieve the contained value from a Value.

Usually this would not be used directly but from the get function on a Value

Required Associated Types

type Checker: ValueTypeChecker

Value type checker.

Required Methods

unsafe fn from_value(value: &'a Value) -> Self

Get the contained value from a Value.

Safety

Self::Checker::check() must be called first and must not fail.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.