Trait glib::value::FromValue

pub trait FromValue<'a>: FromValueOptional<'a> {
    unsafe fn from_value(value: &'a Value) -> Self;
}

Extracts a value.

Only implemented for types that don't support a None value.

Required methods

unsafe fn from_value(value: &'a Value) -> Self

Safety

The caller is responsible for ensuring the given Value is of a suitable type for this conversion.

Loading content...