glib::value

Trait ValueTypeOptional

pub trait ValueTypeOptional:
    ValueType
    + ToValueOptional
    + FromValueOptional<'static>
    + StaticType { }

A type that can be stored in Values and is optional.

These are types were storing an Option is valid. Examples are String and all object types.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.