Trait glib::value::ValueTypeOptional

pub trait ValueTypeOptional:
    ValueType
    + ToValueOptional
    + FromValueOptional<'static>
    + StaticType { }

A type that can be stored in Values and is optional.

These are types were storing an Option is valid. Examples are String and all object types.

Object Safety

This trait is not object safe.