glib::variant

Struct DictEntry

pub struct DictEntry<K, V> { /* private fields */ }

A Dictionary entry.

While GVariant format allows a dictionary entry to be an independent type, typically youโll need to use this in a dictionary, which is simply an array of dictionary entries. The following code creates a dictionary:

use glib::variant::{Variant, FromVariant, DictEntry};

let entries = [
    DictEntry::new("uuid", 1000u32),
    DictEntry::new("guid", 1001u32),
];
let dict = entries.into_iter().collect::<Variant>();
assert_eq!(dict.n_children(), 2);
assert_eq!(dict.type_().as_str(), "a{su}");

Implementations

impl<K, V> DictEntry<K, V>
where K: StaticVariantType, V: StaticVariantType,

pub fn new(key: K, value: V) -> Self

pub fn key(&self) -> &K

pub fn value(&self) -> &V