Struct glib::variant::DictEntry

pub struct DictEntry<K, V> { /* fields omitted */ }

A Dictionary entry.

While GVariant format allows a dictionary entry to be an independent type, typically you’ll need to use this in a dictionary, which is simply an array of dictionary entries. The following code creates a dictionary:

use glib::{Variant, FromVariant, ToVariant};
use glib::variant::DictEntry;

let entries = vec![
    DictEntry::new("uuid", 1000u32).to_variant(),
    DictEntry::new("guid", 1001u32).to_variant(),
];
let dict = Variant::from_array::<DictEntry<&str, u32>>(&entries);
assert_eq!(dict.n_children(), 2);
assert_eq!(dict.type_().to_str(), "a{su}");

Implementations

impl<K, V> DictEntry<K, V> where
K: StaticVariantType + ToVariant + Eq + Hash,
V: StaticVariantType + ToVariant,

pub fn new(key: K, value: V) -> Self

pub fn key(&self) -> &K

pub fn value(&self) -> &V