Struct glib::variant::Variant

#[repr(transparent)]
pub struct Variant { /* private fields */ }

A generic immutable value capable of carrying various types.

See the module documentation for more details.

Implementations

impl Variant

pub fn as_ptr(&self) -> *mut GVariant

Return the inner pointer to the underlying C value.

impl Variant

pub fn type_(&self) -> &VariantTy

Returns the type of the value.

pub fn is<T: StaticVariantType>(&self) -> bool

Returns true if the type of the value corresponds to T.

pub fn is_type(&self, type_: &VariantTy) -> bool

Returns true if the type of the value corresponds to type_.

This is equivalent to self.type_().is_subtype_of(type_).

pub fn classify(&self) -> VariantClass

Returns the classification of the variant.

pub fn get<T: FromVariant>(&self) -> Option<T>

Tries to extract a value of type T.

Returns Some if T matches the variant’s type.

pub fn try_get<T: FromVariant>(&self) -> Result<T, VariantTypeMismatchError>

Tries to extract a value of type T.

pub fn from_variant(value: &Variant) -> Self

Boxes value.

pub fn as_variant(&self) -> Option<Variant>

Unboxes self.

Returns Some if self contains a Variant.

pub fn child_value(&self, index: usize) -> Variant

Reads a child item out of a container Variant instance.

Panics

if self is not a container type.

if given index is larger than number of children.

pub fn try_child_value(&self, index: usize) -> Option<Variant>

Try to read a child item out of a container Variant instance.

It returns None if self is not a container type or if the given index is larger than number of children.

pub fn try_child_get<T: StaticVariantType + FromVariant>( &self, index: usize ) -> Result<Option<T>, VariantTypeMismatchError>

Try to read a child item out of a container Variant instance.

It returns Ok(None) if self is not a container type or if the given index is larger than number of children. An error is thrown if the type does not match.

pub fn child_get<T: StaticVariantType + FromVariant>(&self, index: usize) -> T

Read a child item out of a container Variant instance.

Panics

if self is not a container type.

if given index is larger than number of children.

if the expected variant type does not match

pub fn str(&self) -> Option<&str>

Tries to extract a &str.

Returns Some if the variant has a string type (s, o or g type strings).

pub fn fixed_array<T: FixedSizeVariantType>( &self ) -> Result<&[T], VariantTypeMismatchError>

Tries to extract a &[T] from a variant of array type with a suitable element type.

Returns an error if the type is wrong.

pub fn array_from_iter<T: StaticVariantType>( children: impl IntoIterator<Item = Variant> ) -> Self

Creates a new Variant array from children.

Panics

This function panics if not all variants are of type T.

pub fn array_from_iter_with_type( type_: &VariantTy, children: impl IntoIterator<Item = impl AsRef<Variant>> ) -> Self

Creates a new Variant array from children with the specified type.

Panics

This function panics if not all variants are of type type_.

pub fn array_from_fixed_array<T: FixedSizeVariantType>(array: &[T]) -> Self

Creates a new Variant array from a fixed array.

pub fn tuple_from_iter( children: impl IntoIterator<Item = impl AsRef<Variant>> ) -> Self

Creates a new Variant tuple from children.

pub fn from_dict_entry(key: &Variant, value: &Variant) -> Self

Creates a new dictionary entry Variant.

DictEntry should be preferred over this when the types are known statically.

pub fn from_maybe<T: StaticVariantType>(child: Option<&Variant>) -> Self

Creates a new maybe Variant.

pub fn from_some(child: &Variant) -> Self

Creates a new maybe Variant from a child.

pub fn from_none(type_: &VariantTy) -> Self

Creates a new maybe Variant with Nothing.

pub fn as_maybe(&self) -> Option<Variant>

Extract the value of a maybe Variant.

Returns the child value, or None if the value is Nothing.

Panics

Panics if compiled with debug_assertions and the variant is not maybe-typed.

pub fn print(&self, type_annotate: bool) -> GString

Pretty-print the contents of this variant in a human-readable form.

A variant can be recreated from this output via Variant::parse.

pub fn parse(type_: Option<&VariantTy>, text: &str) -> Result<Self, Error>

Parses a GVariant from the text representation produced by print().

pub fn from_bytes<T: StaticVariantType>(bytes: &Bytes) -> Self

Constructs a new serialized-mode GVariant instance.

pub unsafe fn from_bytes_trusted<T: StaticVariantType>(bytes: &Bytes) -> Self

Constructs a new serialized-mode GVariant instance.

This is the same as from_bytes, except that checks on the passed data are skipped.

You should not use this function on data from external sources.

Safety

Since the data is not validated, this is potentially dangerous if called on bytes which are not guaranteed to have come from serialising another Variant. The caller is responsible for ensuring bad data is not passed in.

pub fn from_data<T: StaticVariantType, A: AsRef<[u8]>>(data: A) -> Self

Constructs a new serialized-mode GVariant instance.

pub unsafe fn from_data_trusted<T: StaticVariantType, A: AsRef<[u8]>>( data: A ) -> Self

Constructs a new serialized-mode GVariant instance.

This is the same as from_data, except that checks on the passed data are skipped.

You should not use this function on data from external sources.

Safety

Since the data is not validated, this is potentially dangerous if called on bytes which are not guaranteed to have come from serialising another Variant. The caller is responsible for ensuring bad data is not passed in.

pub fn from_bytes_with_type(bytes: &Bytes, type_: &VariantTy) -> Self

Constructs a new serialized-mode GVariant instance with a given type.

pub unsafe fn from_bytes_with_type_trusted( bytes: &Bytes, type_: &VariantTy ) -> Self

Constructs a new serialized-mode GVariant instance with a given type.

This is the same as from_bytes, except that checks on the passed data are skipped.

You should not use this function on data from external sources.

Safety

Since the data is not validated, this is potentially dangerous if called on bytes which are not guaranteed to have come from serialising another Variant. The caller is responsible for ensuring bad data is not passed in.

pub fn from_data_with_type<A: AsRef<[u8]>>(data: A, type_: &VariantTy) -> Self

Constructs a new serialized-mode GVariant instance with a given type.

pub unsafe fn from_data_with_type_trusted<A: AsRef<[u8]>>( data: A, type_: &VariantTy ) -> Self

Constructs a new serialized-mode GVariant instance with a given type.

This is the same as from_data, except that checks on the passed data are skipped.

You should not use this function on data from external sources.

Safety

Since the data is not validated, this is potentially dangerous if called on bytes which are not guaranteed to have come from serialising another Variant. The caller is responsible for ensuring bad data is not passed in.

pub fn data_as_bytes(&self) -> Bytes

Returns the serialized form of a GVariant instance.

pub fn data(&self) -> &[u8]

Returns the serialized form of a GVariant instance.

pub fn size(&self) -> usize

Returns the size of serialized form of a GVariant instance.

pub fn store(&self, data: &mut [u8]) -> Result<usize, BoolError>

Stores the serialized form of a GVariant instance into the given slice.

The slice needs to be big enough.

pub fn normal_form(&self) -> Self

Returns a copy of the variant in normal form.

pub fn byteswap(&self) -> Self

Returns a copy of the variant in the opposite endianness.

pub fn n_children(&self) -> usize

Determines the number of children in a container GVariant instance.

pub fn iter(&self) -> VariantIter

Create an iterator over items in the variant.

Note that this heap allocates a variant for each element, which can be particularly expensive for large arrays.

pub fn array_iter_str( &self ) -> Result<VariantStrIter<'_>, VariantTypeMismatchError>

Create an iterator over borrowed strings from a GVariant of type as (array of string).

This will fail if the variant is not an array of with the expected child type.

A benefit of this API over Self::iter() is that it minimizes allocation, and provides strongly typed access.

let strs = &["foo", "bar"];
let strs_variant: glib::Variant = strs.to_variant();
for s in strs_variant.array_iter_str()? {
    println!("{}", s);
}

pub fn is_container(&self) -> bool

Return whether this Variant is a container type.

pub fn is_normal_form(&self) -> bool

Return whether this Variant is in normal form.

pub fn is_object_path(string: &str) -> bool

Return whether input string is a valid VariantClass::ObjectPath.

pub fn is_signature(string: &str) -> bool

Return whether input string is a valid VariantClass::Signature.