Struct glib::variant::Variant

pub struct Variant(_);

A generic immutable value capable of carrying various types.

See the module documentation for more details.

Implementations

impl Variant

pub fn type_(&self) -> &VariantTy

Returns the type of the value.

pub fn is<T: StaticVariantType>(&self) -> bool

Returns true if the type of the value corresponds to T.

pub fn get<T: FromVariant>(&self) -> Option<T>

Tries to extract a value of type T.

Returns Some if T matches the variant's type.

pub fn variant(value: &Variant) -> Self

Boxes value.

pub fn get_variant(&self) -> Option<Variant>

Unboxes self.

Returns Some if self contains a Variant.

pub fn get_child_value(&self, index: usize) -> Variant

Reads a child item out of a container Variant instance.

Panics

if self is not a container type.

if given index is larger than number of children.

pub fn get_str(&self) -> Option<&str>

Tries to extract a &str.

Returns Some if the variant has a string type (s, o or g type strings).

pub fn array<T: StaticVariantType>(children: &[Variant]) -> Self

Creates a new GVariant array from children.

All children must be of type T.

pub fn tuple(children: &[Variant]) -> Self

Creates a new GVariant tuple from children.

pub fn maybe<T: StaticVariantType>(child: Option<&Variant>) -> Self

Creates a new maybe Variant.

pub fn from_bytes<T: StaticVariantType>(bytes: &Bytes) -> Self

Constructs a new serialised-mode GVariant instance.

pub unsafe fn from_bytes_trusted<T: StaticVariantType>(bytes: &Bytes) -> Self

Constructs a new serialised-mode GVariant instance.

This is the same as from_bytes, except that checks on the passed data are skipped.

You should not use this function on data from external sources.

Safety

Since the data is not validated, this is potentially dangerous if called on bytes which are not guaranteed to have come from serialising another Variant. The caller is responsible for ensuring bad data is not passed in.

pub fn get_data_as_bytes(&self) -> Bytes

Returns the serialised form of a GVariant instance.

pub fn n_children(&self) -> usize

Determines the number of children in a container GVariant instance.

pub fn iter(&self) -> VariantIter

Create an iterator over items in the variant.

pub fn is_container(&self) -> bool

Variant has a container type.