Struct glib::variant::Variant

pub struct Variant(_);

A generic immutable value capable of carrying various types.

See the module documentation for more details.

Methods

impl Variant

fn type_(&self) -> &VariantTy

Returns the type of the value.

fn is<T: StaticVariantType>(&self) -> bool

Returns true if the type of the value corresponds to T.

fn get<T: FromVariant>(&self) -> Option<T>

Tries to extract a value of type T.

Returns Some if T matches the variant's type.

fn get_str(&self) -> Option<&str>

Tries to extract a &str.

Returns Some if the variant has a string type (s, o or g type strings).